refacto(frontend) : tout les type props sont mis dans le .type.ts et importer dans le  index.ts
