+++ 1-note-quizz/.obsidian/workspace.json
@@ -21,6 +21,20 @@
               "title": "note"
             },
             "group": "1d1047c2b9ef73a9"
+          },
+          {
+            "id": "5ea418ada7042def",
+            "type": "leaf",
+            "state": {
+              "type": "markdown",
+              "state": {
+                "file": "destruction et assignation classique.md",
+                "mode": "source",
+                "source": false
+              },
+              "icon": "lucide-file",
+              "title": "destruction et assignation classique"
+            }
           }
         ]
       }
@@ -184,8 +198,10 @@
       "bases:Créer une nouvelle base": false
     }
   },
-  "active": "b95a2d78b9aa4f4c",
+  "active": "38a522587e224bf1",
   "lastOpenFiles": [
+    "destruction et assignation classique.md",
+    "17-04-2026/note.md",
     "Capture d’écran 2026-04-17 à 15.35.52 1.png",
     "Capture d’écran 2026-04-17 à 15.35.52.png"
   ]
