+++ 1-note-quizz/17-04-2026/note.md
@@ -14,6 +14,9 @@
 - je dois faire attention dans question modale j'ai une structure qui oblige d'avoir 4 réponse
 - j'ai découvert [[#le custom hook]] se qui me donne l'idée de pouvoir aussi organisé mes  QuestionEditModalProps
 - dans les régle de bonne pratique on sépare tout mais dans react il recommande garder les action visible
+- le format destruction et assignation classique
+
+[[destruction et assignation classique]]
 ### La règle d'or du `.metier.ts`
 
 Le fichier métier doit idéalement contenir de la logique qui pourrait fonctionner **même si tu changeais de framework** (si tu passais de React à Vue, par exemple).
