+++ 1-note-quizz/destruction et assignation classique.md
@@ -1,0 +1,49 @@
+C'est une excellente question de syntaxe ! En développement JavaScript/TypeScript (et particulièrement en React), ces deux approches produisent exactement le même résultat technique, mais elles portent des noms très différents.
+
+---
+
+## 1. Le premier format : La Destructuration (Destructuring)
+
+JavaScript
+
+```
+const { settings, actions, status, data, drafts } = props;
+```
+
+- **Nom officiel :** L'**Affectation par décomposition** (ou _Destructuring assignment_ en anglais).
+    
+- **Le concept :** C'est une syntaxe introduite avec l'ES6 qui permet d'extraire directement des propriétés d'un objet et de les stocker dans des variables individuelles en une seule ligne.
+    
+- **Pourquoi l'utiliser ?** C'est le standard moderne. C'est plus concis, plus lisible, et cela évite de répéter le nom de l'objet parent (`props.something`) partout dans le code.
+    
+
+---
+
+## 2. Le deuxième format : L'Assignation classique (Dot Notation)
+
+JavaScript
+
+```
+const settings = props.settings;
+const actions = props.actions;
+```
+
+- **Nom :** L'**Assignation explicite** ou l'accès par **Notation par point** (_Dot notation_).
+    
+- **Le concept :** C'est la méthode "traditionnelle". On accède manuellement à chaque propriété de l'objet une par une pour l'assigner à une nouvelle variable.
+    
+- **Pourquoi l'utiliser ?** Elle est parfois utilisée par les débutants car elle est très explicite, ou dans de rares cas spécifiques où l'on veut être extrêmement prudent sur ce que l'on extrait. Cependant, c'est beaucoup plus verbeux (plus long à écrire et à lire).
+    
+
+---
+
+### Résumé des différences
+
+|**Caractéristique**|**Destructuration (Format 1)**|**Assignation classique (Format 2)**|
+|---|---|---|
+|**Modernité**|Standard ES6+ (Recommandé)|Standard historique|
+|**Rapidité**|Très rapide à écrire|Laborieux pour les gros objets|
+|**Lisibilité**|Épurée et élégante|Très chargée visuellement|
+|**Usage React**|Quasi systématique pour les _props_|Rarement utilisé aujourd'hui|
+
+> **Le petit conseil en plus :** Dans ton premier exemple, tu utilises aussi la destructuration sur le résultat du hook `useQuestionEditModal`. C'est là que cette syntaxe brille vraiment : transformer 5 lignes de code répétitives en une seule ligne propre.
