+++ src/network/projet_quizz/frontend/src/composant/atomes/Badge/Badge.types.ts
@@ -1,5 +1,7 @@
 import type { ComponentChildren } from "preact";
 
-export type DeviceAuthGateProps = {
+export type BadgeProps = {
   children: ComponentChildren;
+  class?: string;
+  tone?: "flow" | "learn" | "neutral";
 };
