+++ src/network/projet_quizz/frontend/src/composant/atomes/Badge/index.ts
@@ -1,1 +1,2 @@
+export * from "./Badge.types";
 export * from "./Badge";
