+++ src/network/projet_quizz/frontend/src/composant/atomes/Button/index.ts
@@ -1,1 +1,2 @@
+export * from "./Button.types";
 export * from "./Button";
