+++ src/network/projet_quizz/frontend/src/composant/atomes/Card/Card.types.ts
@@ -1,0 +1,7 @@
+import type { ComponentChildren, ComponentProps } from "preact";
+
+export type CardProps = ComponentProps<"div"> & {
+  children: ComponentChildren;
+  class?: string;
+  padding?: "sm" | "md" | "lg";
+};
