+++ src/network/projet_quizz/frontend/src/composant/atomes/Card/index.ts
@@ -1,1 +1,2 @@
+export * from "./Card.types";
 export * from "./Card";
