+++ src/network/projet_quizz/frontend/src/composant/atomes/ProgressBar/ProgressBar.types.ts
@@ -1,0 +1,5 @@
+export type ProgressBarProps = {
+  value: number;
+  max: number;
+  class?: string;
+};
