+++ src/network/projet_quizz/frontend/src/composant/atomes/ProgressBar/index.ts
@@ -1,1 +1,2 @@
+export * from "./ProgressBar.types";
 export * from "./ProgressBar";
