+++ src/network/projet_quizz/frontend/src/composant/molecules/AnswerOption/AnswerOption.types.ts
@@ -1,3 +1,10 @@
+export type AnswerOptionStateParams = {
+  reponseId: number;
+  pickedId: number | null;
+  revealed: boolean;
+  isCorrectAnswer: boolean;
+};
+
 export type AnswerOptionProps = {
   label: string;
   reponseId: number;
@@ -8,9 +15,3 @@ export type AnswerOptionProps = {
   onPick: () => void;
 };
 
-export type AnswerOptionStateParams = {
-  reponseId: number;
-  pickedId: number | null;
-  revealed: boolean;
-  isCorrectAnswer: boolean;
-};
