+++ src/network/projet_quizz/frontend/src/composant/molecules/AnswerOption/index.ts
@@ -1,1 +1,2 @@
+export * from "./AnswerOption.types";
 export * from "./AnswerOption";
