+++ src/network/projet_quizz/frontend/src/composant/molecules/AppHeader/index.ts
@@ -1,1 +1,2 @@
+export * from "./AppHeader.types";
 export * from "./AppHeader";
