+++ src/network/projet_quizz/frontend/src/composant/molecules/CollectionCard/CollectionCard.tsx
@@ -9,7 +9,6 @@ import {
   playOrdersRequireUserId 
 } from "../../../lib/playOrder";
 import type { PlayQtype } from "../../../lib/playOrder";
-import type { CollectionUi, QuizzModuleRow } from "../../../types/quizz";
 
 import { Card } from "../../atomes/Card";
 import { Badge } from "../../atomes/Badge";
@@ -20,19 +19,7 @@ import type { PlayModeSettings } from "../PlayModePicker/PlayModePicker.types";
 
 import { buildQuestionsRoutePath } from "./CollectionCard.metier";
 import { COLLECTION_CARD_STYLES } from "./CollectionCard.styles";
-
-
-export type CollectionCardProps = {
-  collection: CollectionUi;
-  myUserId: number;
-  allModules: QuizzModuleRow[];
-  assignBusyCollectionId: number | null;
-  deleteBusyCollectionId: number | null;
-  interactionLocked?: boolean;
-  onAssign: (collectionId: number, moduleId: number) => void | Promise<void>;
-  onUnassign: (collectionId: number, moduleId: number) => void | Promise<void>;
-  onDeleteCollection?: (collection: CollectionUi) => void;
-};
+import type { CollectionCardProps } from "./CollectionCard.types";
 
 export function CollectionCard({
   collection,
