+++ src/network/projet_quizz/frontend/src/composant/molecules/CollectionCard/CollectionCard.types.ts
@@ -1,3 +1,16 @@
+import type { CollectionUi, QuizzModuleRow } from "../../../types/quizz";
 
 export type LinkedModule = { id: number };
+
+export type CollectionCardProps = {
+  collection: CollectionUi;
+  myUserId: number;
+  allModules: QuizzModuleRow[];
+  assignBusyCollectionId: number | null;
+  deleteBusyCollectionId: number | null;
+  interactionLocked?: boolean;
+  onAssign: (collectionId: number, moduleId: number) => void | Promise<void>;
+  onUnassign: (collectionId: number, moduleId: number) => void | Promise<void>;
+  onDeleteCollection?: (collection: CollectionUi) => void;
+};
