+++ src/network/projet_quizz/frontend/src/composant/molecules/CollectionCard/index.ts
@@ -1,1 +1,2 @@
+export * from "./CollectionCard.types";
 export * from "./CollectionCard";
