+++ src/network/projet_quizz/frontend/src/composant/molecules/KpiCard/KpiCard.types.ts
@@ -1,0 +1,9 @@
+import type { ComponentChildren } from "preact";
+
+export type KpiCardProps = {
+  title: string;
+  value: string;
+  hint?: string;
+  icon?: ComponentChildren;
+  accent?: "flow" | "learn";
+};
