+++ src/network/projet_quizz/frontend/src/composant/molecules/KpiCard/index.ts
@@ -1,1 +1,2 @@
+export * from "./KpiCard.types";
 export * from "./KpiCard";
