+++ src/network/projet_quizz/frontend/src/composant/molecules/PageMain/PageMain.types.ts
@@ -1,5 +1,7 @@
 import type { ComponentChildren } from "preact";
 
-export type DeviceAuthGateProps = {
+export type PageMainProps = {
   children: ComponentChildren;
+  class?: string;
+  narrow?: boolean;
 };
