+++ src/network/projet_quizz/frontend/src/composant/molecules/PageMain/index.ts
@@ -1,1 +1,2 @@
+export * from "./PageMain.types";
 export * from "./PageMain";
