+++ src/network/projet_quizz/frontend/src/composant/molecules/PlayModePicker/PlayModePicker.tsx
@@ -1,21 +1,5 @@
-import { PLAY_MODE_SORT_OPTIONS, type PlayModeSettings } from "./PlayModePicker.types";
+import { PLAY_MODE_SORT_OPTIONS, type PlayModePickerProps } from "./PlayModePicker.types";
 import { PLAY_MODE_PICKER_STYLES } from "./PlayModePicker.styles";
-
-/**
- * Props du panneau de choix du mode de lecture (filtres, tri, options KPI, mélange).
- *
- * @property idPrefix - Préfixe pour les attributs `id` et `name` des champs, afin d’éviter les collisions lorsque plusieurs pickers coexistent.
- * @property settings - Objet d’état courant du mode de jeu (`PlayModeSettings`).
- * @property onChange - Callback appelée avec un **patch partiel** à fusionner dans l’état parent après chaque interaction utilisateur.
- * @property labelAlignClass - Classes Tailwind additionnelles pour l’alignement des libellés de section (ex. `text-center`, `sm:text-end`).
- */
-export type PlayModePickerProps = {
-  idPrefix: string;
-  settings: PlayModeSettings;
-  onChange: (newSettings: Partial<PlayModeSettings>) => void;
-  labelAlignClass?: string;
-};
-
 
 export function PlayModePicker({
   idPrefix,
