+++ src/network/projet_quizz/frontend/src/composant/molecules/PlayModePicker/index.ts
@@ -1,1 +1,2 @@
+export * from "./PlayModePicker.types";
 export * from "./PlayModePicker";
