+++ src/network/projet_quizz/frontend/src/composant/molecules/PopUpInformation/PopUpInformation.types.ts
@@ -1,1 +1,11 @@
-export type PopUpInformationTypesPlaceholder = Record<string, never>;
+export type PopUpInformationProps = {
+  open: boolean;
+  title: string;
+  message: string;
+  variant?: "info" | "danger";
+  confirmLabel?: string;
+  cancelLabel?: string;
+  busy?: boolean;
+  onConfirm: () => void;
+  onCancel: () => void;
+};
