+++ src/network/projet_quizz/frontend/src/composant/molecules/PopUpInformation/index.ts
@@ -1,1 +1,2 @@
+export * from "./PopUpInformation.types";
 export * from "./PopUpInformation";
