+++ src/network/projet_quizz/frontend/src/composant/molecules/QuestionsLlmImportOptionsPanel/index.ts
@@ -1,1 +1,2 @@
+export * from "./QuestionsLlmImportOptionsPanel.types";
 export * from "./QuestionsLlmImportOptionsPanel";
