+++ src/network/projet_quizz/frontend/src/composant/molecules/QuestionsLlmImportPromptPanel/index.ts
@@ -1,2 +1,2 @@
+export * from "./QuestionsLlmImportPromptPanel.types";
 export * from "./QuestionsLlmImportPromptPanel";
-export type { QuestionsLlmImportPromptPanelProps } from "./QuestionsLlmImportPromptPanel.types";
