+++ src/network/projet_quizz/frontend/src/composant/organismes/DeviceAuthGate/index.ts
@@ -1,1 +1,2 @@
+export * from "./DeviceAuthGate.types";
 export * from "./DeviceAuthGate";
