+++ src/network/projet_quizz/frontend/src/composant/organismes/HomeView/HomeView.types.ts
@@ -1,1 +1,1 @@
-export type HomeViewTypesPlaceholder = Record<string, never>;
+export type HomeViewProps = Record<string, never>;
