+++ src/network/projet_quizz/frontend/src/composant/organismes/HomeView/index.ts
@@ -1,1 +1,2 @@
+export * from "./HomeView.types";
 export * from "./HomeView";
