+++ src/network/projet_quizz/frontend/src/composant/organismes/QuestionEditModal/index.ts
@@ -1,1 +1,2 @@
+export * from "./QuestionEditModal.types";
 export * from "./QuestionEditModal";
