+++ src/network/projet_quizz/frontend/src/composant/organismes/QuestionsCollectionContextBar/QuestionsCollectionContextBar.types.ts
@@ -1,1 +1,9 @@
-export type QuestionsCollectionContextBarTypesPlaceholder = Record<string, never>;
+import type { CollectionUi, QuizzModuleRow } from "../../../types/quizz";
+
+export type QuestionsCollectionContextBarProps = {
+  targetCollectionNumeric: number | null;
+  collections: CollectionUi[];
+  allModules: QuizzModuleRow[];
+  importTargetModuleId: number | null;
+  setImportTargetModuleId: (id: number | null) => void;
+};
