+++ src/network/projet_quizz/frontend/src/composant/organismes/QuestionsCollectionContextBar/index.ts
@@ -1,1 +1,2 @@
+export * from "./QuestionsCollectionContextBar.types";
 export * from "./QuestionsCollectionContextBar";
