+++ src/network/projet_quizz/frontend/src/composant/organismes/QuestionsView/index.ts
@@ -1,1 +1,2 @@
+export * from "./QuestionsView.types";
 export * from "./QuestionsView";
