+++ src/network/projet_quizz/frontend/src/composant/organismes/QuizResultsView/QuizResultsView.types.ts
@@ -1,7 +1,5 @@
 import type { LastQuizResult } from "../../../lib/lastQuizResult";
 
-export type QuizResultsViewProps = Record<string, never>;
-
 export type QuizResultsSummary = {
   good: number;
   total: number;
