+++ src/network/projet_quizz/frontend/src/composant/organismes/StatsDashboard/index.ts
@@ -1,1 +1,2 @@
+export * from "./StatsDashboard.types";
 export * from "./StatsDashboard";
