+++ src/network/site_compiled/assets/index.js
@@ -1,10 +1,10 @@
-(function(){let e=document.createElement(`link`).relList;if(e&&e.supports&&e.supports(`modulepreload`))return;for(let e of document.querySelectorAll(`link[rel="modulepreload"]`))n(e);new MutationObserver(e=>{for(let t of e)if(t.type===`childList`)for(let e of t.addedNodes)e.tagName===`LINK`&&e.rel===`modulepreload`&&n(e)}).observe(document,{childList:!0,subtree:!0});function t(e){let t={};return e.integrity&&(t.integrity=e.integrity),e.referrerPolicy&&(t.referrerPolicy=e.referrerPolicy),e.crossOrigin===`use-credentials`?t.credentials=`include`:e.crossOrigin===`anonymous`?t.credentials=`omit`:t.credentials=`same-origin`,t}function n(e){if(e.ep)return;e.ep=!0;let n=t(e);fetch(e.href,n)}})();var e,t,n,r,i,a,o,s,c,l,u,d,f={},p=[],m=/acit|ex(?:s|g|n|p|$)|rph|grid|ows|mnc|ntw|ine[ch]|zoo|^ord|itera/i,h=Array.isArray;function g(e,t){for(var n in t)e[n]=t[n];return e}function _(e){e&&e.parentNode&&e.parentNode.removeChild(e)}function v(t,n,r){var i,a,o,s={};for(o in n)o==`key`?i=n[o]:o==`ref`?a=n[o]:s[o]=n[o];if(arguments.length>2&&(s.children=arguments.length>3?e.call(arguments,2):r),typeof t==`function`&&t.defaultProps!=null)for(o in t.defaultProps)s[o]===void 0&&(s[o]=t.defaultProps[o]);return y(t,s,i,a,null)}function y(e,r,i,a,o){var s={type:e,props:r,key:i,ref:a,__k:null,__:null,__b:0,__e:null,__c:null,constructor:void 0,__v:o??++n,__i:-1,__u:0};return o==null&&t.vnode!=null&&t.vnode(s),s}function b(e){return e.children}function x(e,t){this.props=e,this.context=t}function S(e,t){if(t==null)return e.__?S(e.__,e.__i+1):null;for(var n;t<e.__k.length;t++)if((n=e.__k[t])!=null&&n.__e!=null)return n.__e;return typeof e.type==`function`?S(e):null}function C(e){if(e.__P&&e.__d){var n=e.__v,r=n.__e,i=[],a=[],o=g({},n);o.__v=n.__v+1,t.vnode&&t.vnode(o),ne(e.__P,o,n,e.__n,e.__P.namespaceURI,32&n.__u?[r]:null,i,r??S(n),!!(32&n.__u),a),o.__v=n.__v,o.__.__k[o.__i]=o,re(i,o,a),n.__e=n.__=null,o.__e!=r&&w(o)}}function w(e){if((e=e.__)!=null&&e.__c!=null)return e.__e=e.__c.base=null,e.__k.some(function(t){if(t!=null&&t.__e!=null)return e.__e=e.__c.base=t.__e}),w(e)}function T(e){(!e.__d&&(e.__d=!0)&&r.push(e)&&!E.__r++||i!=t.debounceRendering)&&((i=t.debounceRendering)||a)(E)}function E(){try{for(var e,t=1;r.length;)r.length>t&&r.sort(o),e=r.shift(),t=r.length,C(e)}finally{r.length=E.__r=0}}function ee(e,t,n,r,i,a,o,s,c,l,u){var d,m,h,g,_,v,y,b=r&&r.__k||p,x=t.length;for(c=D(n,t,b,c,x),d=0;d<x;d++)(h=n.__k[d])!=null&&(m=h.__i!=-1&&b[h.__i]||f,h.__i=d,v=ne(e,h,m,i,a,o,s,c,l,u),g=h.__e,h.ref&&m.ref!=h.ref&&(m.ref&&F(m.ref,null,h),u.push(h.ref,h.__c||g,h)),_==null&&g!=null&&(_=g),(y=!!(4&h.__u))||m.__k===h.__k?c=O(h,c,e,y):typeof h.type==`function`&&v!==void 0?c=v:g&&(c=g.nextSibling),h.__u&=-7);return n.__e=_,c}function D(e,t,n,r,i){var a,o,s,c,l,u=n.length,d=u,f=0;for(e.__k=Array(i),a=0;a<i;a++)(o=t[a])!=null&&typeof o!=`boolean`&&typeof o!=`function`?(typeof o==`string`||typeof o==`number`||typeof o==`bigint`||o.constructor==String?o=e.__k[a]=y(null,o,null,null,null):h(o)?o=e.__k[a]=y(b,{children:o},null,null,null):o.constructor===void 0&&o.__b>0?o=e.__k[a]=y(o.type,o.props,o.key,o.ref?o.ref:null,o.__v):e.__k[a]=o,c=a+f,o.__=e,o.__b=e.__b+1,s=null,(l=o.__i=A(o,n,c,d))!=-1&&(d--,(s=n[l])&&(s.__u|=2)),s==null||s.__v==null?(l==-1&&(i>u?f--:i<u&&f++),typeof o.type!=`function`&&(o.__u|=4)):l!=c&&(l==c-1?f--:l==c+1?f++:(l>c?f--:f++,o.__u|=4))):e.__k[a]=null;if(d)for(a=0;a<u;a++)(s=n[a])!=null&&!(2&s.__u)&&(s.__e==r&&(r=S(s)),I(s,s));return r}function O(e,t,n,r){var i,a;if(typeof e.type==`function`){for(i=e.__k,a=0;i&&a<i.length;a++)i[a]&&(i[a].__=e,t=O(i[a],t,n,r));return t}e.__e!=t&&(r&&(t&&e.type&&!t.parentNode&&(t=S(e)),n.insertBefore(e.__e,t||null)),t=e.__e);do t&&=t.nextSibling;while(t!=null&&t.nodeType==8);return t}function k(e,t){return t||=[],e==null||typeof e==`boolean`||(h(e)?e.some(function(e){k(e,t)}):t.push(e)),t}function A(e,t,n,r){var i,a,o,s=e.key,c=e.type,l=t[n],u=l!=null&&(2&l.__u)==0;if(l===null&&s==null||u&&s==l.key&&c==l.type)return n;if(r>(u?1:0)){for(i=n-1,a=n+1;i>=0||a<t.length;)if((l=t[o=i>=0?i--:a++])!=null&&!(2&l.__u)&&s==l.key&&c==l.type)return o}return-1}function j(e,t,n){t[0]==`-`?e.setProperty(t,n??``):e[t]=n==null?``:typeof n!=`number`||m.test(t)?n:n+`px`}function M(e,t,n,r,i){var a,o;n:if(t==`style`)if(typeof n==`string`)e.style.cssText=n;else{if(typeof r==`string`&&(e.style.cssText=r=``),r)for(t in r)n&&t in n||j(e.style,t,``);if(n)for(t in n)r&&n[t]==r[t]||j(e.style,t,n[t])}else if(t[0]==`o`&&t[1]==`n`)a=t!=(t=t.replace(s,`$1`)),o=t.toLowerCase(),t=o in e||t==`onFocusOut`||t==`onFocusIn`?o.slice(2):t.slice(2),e.l||={},e.l[t+a]=n,n?r?n.u=r.u:(n.u=c,e.addEventListener(t,a?u:l,a)):e.removeEventListener(t,a?u:l,a);else{if(i==`http://www.w3.org/2000/svg`)t=t.replace(/xlink(H|:h)/,`h`).replace(/sName$/,`s`);else if(t!=`width`&&t!=`height`&&t!=`href`&&t!=`list`&&t!=`form`&&t!=`tabIndex`&&t!=`download`&&t!=`rowSpan`&&t!=`colSpan`&&t!=`role`&&t!=`popover`&&t in e)try{e[t]=n??``;break n}catch{}typeof n==`function`||(n==null||!1===n&&t[4]!=`-`?e.removeAttribute(t):e.setAttribute(t,t==`popover`&&n==1?``:n))}}function te(e){return function(n){if(this.l){var r=this.l[n.type+e];if(n.t==null)n.t=c++;else if(n.t<r.u)return;return r(t.event?t.event(n):n)}}}function ne(e,n,r,i,a,o,s,c,l,u){var d,f,m,v,y,S,C,w,T,E,D,O,k,A,j,M=n.type;if(n.constructor!==void 0)return null;128&r.__u&&(l=!!(32&r.__u),o=[c=n.__e=r.__e]),(d=t.__b)&&d(n);n:if(typeof M==`function`)try{if(w=n.props,T=M.prototype&&M.prototype.render,E=(d=M.contextType)&&i[d.__c],D=d?E?E.props.value:d.__:i,r.__c?C=(f=n.__c=r.__c).__=f.__E:(T?n.__c=f=new M(w,D):(n.__c=f=new x(w,D),f.constructor=M,f.render=L),E&&E.sub(f),f.state||={},f.__n=i,m=f.__d=!0,f.__h=[],f._sb=[]),T&&f.__s==null&&(f.__s=f.state),T&&M.getDerivedStateFromProps!=null&&(f.__s==f.state&&(f.__s=g({},f.__s)),g(f.__s,M.getDerivedStateFromProps(w,f.__s))),v=f.props,y=f.state,f.__v=n,m)T&&M.getDerivedStateFromProps==null&&f.componentWillMount!=null&&f.componentWillMount(),T&&f.componentDidMount!=null&&f.__h.push(f.componentDidMount);else{if(T&&M.getDerivedStateFromProps==null&&w!==v&&f.componentWillReceiveProps!=null&&f.componentWillReceiveProps(w,D),n.__v==r.__v||!f.__e&&f.shouldComponentUpdate!=null&&!1===f.shouldComponentUpdate(w,f.__s,D)){n.__v!=r.__v&&(f.props=w,f.state=f.__s,f.__d=!1),n.__e=r.__e,n.__k=r.__k,n.__k.some(function(e){e&&(e.__=n)}),p.push.apply(f.__h,f._sb),f._sb=[],f.__h.length&&s.push(f);break n}f.componentWillUpdate!=null&&f.componentWillUpdate(w,f.__s,D),T&&f.componentDidUpdate!=null&&f.__h.push(function(){f.componentDidUpdate(v,y,S)})}if(f.context=D,f.props=w,f.__P=e,f.__e=!1,O=t.__r,k=0,T)f.state=f.__s,f.__d=!1,O&&O(n),d=f.render(f.props,f.state,f.context),p.push.apply(f.__h,f._sb),f._sb=[];else do f.__d=!1,O&&O(n),d=f.render(f.props,f.state,f.context),f.state=f.__s;while(f.__d&&++k<25);f.state=f.__s,f.getChildContext!=null&&(i=g(g({},i),f.getChildContext())),T&&!m&&f.getSnapshotBeforeUpdate!=null&&(S=f.getSnapshotBeforeUpdate(v,y)),A=d!=null&&d.type===b&&d.key==null?ie(d.props.children):d,c=ee(e,h(A)?A:[A],n,r,i,a,o,s,c,l,u),f.base=n.__e,n.__u&=-161,f.__h.length&&s.push(f),C&&(f.__E=f.__=null)}catch(e){if(n.__v=null,l||o!=null)if(e.then){for(n.__u|=l?160:128;c&&c.nodeType==8&&c.nextSibling;)c=c.nextSibling;o[o.indexOf(c)]=null,n.__e=c}else{for(j=o.length;j--;)_(o[j]);N(n)}else n.__e=r.__e,n.__k=r.__k,e.then||N(n);t.__e(e,n,r)}else o==null&&n.__v==r.__v?(n.__k=r.__k,n.__e=r.__e):c=n.__e=P(r.__e,n,r,i,a,o,s,l,u);return(d=t.diffed)&&d(n),128&n.__u?void 0:c}function N(e){e&&(e.__c&&(e.__c.__e=!0),e.__k&&e.__k.some(N))}function re(e,n,r){for(var i=0;i<r.length;i++)F(r[i],r[++i],r[++i]);t.__c&&t.__c(n,e),e.some(function(n){try{e=n.__h,n.__h=[],e.some(function(e){e.call(n)})}catch(e){t.__e(e,n.__v)}})}function ie(e){return typeof e!=`object`||!e||e.__b>0?e:h(e)?e.map(ie):g({},e)}function P(n,r,i,a,o,s,c,l,u){var d,p,m,g,v,y,b,x=i.props||f,C=r.props,w=r.type;if(w==`svg`?o=`http://www.w3.org/2000/svg`:w==`math`?o=`http://www.w3.org/1998/Math/MathML`:o||=`http://www.w3.org/1999/xhtml`,s!=null){for(d=0;d<s.length;d++)if((v=s[d])&&`setAttribute`in v==!!w&&(w?v.localName==w:v.nodeType==3)){n=v,s[d]=null;break}}if(n==null){if(w==null)return document.createTextNode(C);n=document.createElementNS(o,w,C.is&&C),l&&=(t.__m&&t.__m(r,s),!1),s=null}if(w==null)x===C||l&&n.data==C||(n.data=C);else{if(s&&=e.call(n.childNodes),!l&&s!=null)for(x={},d=0;d<n.attributes.length;d++)x[(v=n.attributes[d]).name]=v.value;for(d in x)v=x[d],d==`dangerouslySetInnerHTML`?m=v:d==`children`||d in C||d==`value`&&`defaultValue`in C||d==`checked`&&`defaultChecked`in C||M(n,d,null,v,o);for(d in C)v=C[d],d==`children`?g=v:d==`dangerouslySetInnerHTML`?p=v:d==`value`?y=v:d==`checked`?b=v:l&&typeof v!=`function`||x[d]===v||M(n,d,v,x[d],o);if(p)l||m&&(p.__html==m.__html||p.__html==n.innerHTML)||(n.innerHTML=p.__html),r.__k=[];else if(m&&(n.innerHTML=``),ee(r.type==`template`?n.content:n,h(g)?g:[g],r,i,a,w==`foreignObject`?`http://www.w3.org/1999/xhtml`:o,s,c,s?s[0]:i.__k&&S(i,0),l,u),s!=null)for(d=s.length;d--;)_(s[d]);l||(d=`value`,w==`progress`&&y==null?n.removeAttribute(`value`):y!=null&&(y!==n[d]||w==`progress`&&!y||w==`option`&&y!=x[d])&&M(n,d,y,x[d],o),d=`checked`,b!=null&&b!=n[d]&&M(n,d,b,x[d],o))}return n}function F(e,n,r){try{if(typeof e==`function`){var i=typeof e.__u==`function`;i&&e.__u(),i&&n==null||(e.__u=e(n))}else e.current=n}catch(e){t.__e(e,r)}}function I(e,n,r){var i,a;if(t.unmount&&t.unmount(e),(i=e.ref)&&(i.current&&i.current!=e.__e||F(i,null,n)),(i=e.__c)!=null){if(i.componentWillUnmount)try{i.componentWillUnmount()}catch(e){t.__e(e,n)}i.base=i.__P=null}if(i=e.__k)for(a=0;a<i.length;a++)i[a]&&I(i[a],n,r||typeof e.type!=`function`);r||_(e.__e),e.__c=e.__=e.__e=void 0}function L(e,t,n){return this.constructor(e,n)}function ae(n,r,i){var a,o,s,c;r==document&&(r=document.documentElement),t.__&&t.__(n,r),o=(a=typeof i==`function`)?null:i&&i.__k||r.__k,s=[],c=[],ne(r,n=(!a&&i||r).__k=v(b,null,[n]),o||f,f,r.namespaceURI,!a&&i?[i]:o?null:r.firstChild?e.call(r.childNodes):null,s,!a&&i?i:o?o.__e:r.firstChild,a,c),re(s,n,c)}function oe(t,n,r){var i,a,o,s,c=g({},t.props);for(o in t.type&&t.type.defaultProps&&(s=t.type.defaultProps),n)o==`key`?i=n[o]:o==`ref`?a=n[o]:c[o]=n[o]===void 0&&s!=null?s[o]:n[o];return arguments.length>2&&(c.children=arguments.length>3?e.call(arguments,2):r),y(t.type,c,i||t.key,a||t.ref,null)}function R(e){function t(e){var n,r;return this.getChildContext||(n=new Set,(r={})[t.__c]=this,this.getChildContext=function(){return r},this.componentWillUnmount=function(){n=null},this.shouldComponentUpdate=function(e){this.props.value!=e.value&&n.forEach(function(e){e.__e=!0,T(e)})},this.sub=function(e){n.add(e);var t=e.componentWillUnmount;e.componentWillUnmount=function(){n&&n.delete(e),t&&t.call(e)}}),e.children}return t.__c=`__cC`+ d++,t.__=e,t.Provider=t.__l=(t.Consumer=function(e,t){return e.children(t)}).contextType=t,t}e=p.slice,t={__e:function(e,t,n,r){for(var i,a,o;t=t.__;)if((i=t.__c)&&!i.__)try{if((a=i.constructor)&&a.getDerivedStateFromError!=null&&(i.setState(a.getDerivedStateFromError(e)),o=i.__d),i.componentDidCatch!=null&&(i.componentDidCatch(e,r||{}),o=i.__d),o)return i.__E=i}catch(t){e=t}throw e}},n=0,x.prototype.setState=function(e,t){var n=this.__s!=null&&this.__s!=this.state?this.__s:this.__s=g({},this.state);typeof e==`function`&&(e=e(g({},n),this.props)),e&&g(n,e),e!=null&&this.__v&&(t&&this._sb.push(t),T(this))},x.prototype.forceUpdate=function(e){this.__v&&(this.__e=!0,e&&this.__h.push(e),T(this))},x.prototype.render=b,r=[],a=typeof Promise==`function`?Promise.prototype.then.bind(Promise.resolve()):setTimeout,o=function(e,t){return e.__v.__b-t.__v.__b},E.__r=0,s=/(PointerCapture)$|Capture$/i,c=0,l=te(!1),u=te(!0),d=0;var se,z,ce,le,ue=0,de=[],B=t,fe=B.__b,pe=B.__r,me=B.diffed,he=B.__c,ge=B.unmount,_e=B.__;function ve(e,t){B.__h&&B.__h(z,e,ue||t),ue=0;var n=z.__H||={__:[],__h:[]};return e>=n.__.length&&n.__.push({}),n.__[e]}function V(e){return ue=1,ye(je,e)}function ye(e,t,n){var r=ve(se++,2);if(r.t=e,!r.__c&&(r.__=[n?n(t):je(void 0,t),function(e){var t=r.__N?r.__N[0]:r.__[0],n=r.t(t,e);t!==n&&(r.__N=[n,r.__[1]],r.__c.setState({}))}],r.__c=z,!z.__f)){var i=function(e,t,n){if(!r.__c.__H)return!0;var i=r.__c.__H.__.filter(function(e){return e.__c});if(i.every(function(e){return!e.__N}))return!a||a.call(this,e,t,n);var o=r.__c.props!==e;return i.some(function(e){if(e.__N){var t=e.__[0];e.__=e.__N,e.__N=void 0,t!==e.__[0]&&(o=!0)}}),a&&a.call(this,e,t,n)||o};z.__f=!0;var a=z.shouldComponentUpdate,o=z.componentWillUpdate;z.componentWillUpdate=function(e,t,n){if(this.__e){var r=a;a=void 0,i(e,t,n),a=r}o&&o.call(this,e,t,n)},z.shouldComponentUpdate=i}return r.__N||r.__}function H(e,t){var n=ve(se++,3);!B.__s&&Ae(n.__H,t)&&(n.__=e,n.u=t,z.__H.__h.push(n))}function be(e,t){var n=ve(se++,4);!B.__s&&Ae(n.__H,t)&&(n.__=e,n.u=t,z.__h.push(n))}function xe(e){return ue=5,Se(function(){return{current:e}},[])}function Se(e,t){var n=ve(se++,7);return Ae(n.__H,t)&&(n.__=e(),n.__H=t,n.__h=e),n.__}function Ce(e,t){return ue=8,Se(function(){return e},t)}function we(e){var t=z.context[e.__c],n=ve(se++,9);return n.c=e,t?(n.__??(n.__=!0,t.sub(z)),t.props.value):e.__}function Te(){for(var e;e=de.shift();){var t=e.__H;if(e.__P&&t)try{t.__h.some(Oe),t.__h.some(ke),t.__h=[]}catch(n){t.__h=[],B.__e(n,e.__v)}}}B.__b=function(e){z=null,fe&&fe(e)},B.__=function(e,t){e&&t.__k&&t.__k.__m&&(e.__m=t.__k.__m),_e&&_e(e,t)},B.__r=function(e){pe&&pe(e),se=0;var t=(z=e.__c).__H;t&&(ce===z?(t.__h=[],z.__h=[],t.__.some(function(e){e.__N&&(e.__=e.__N),e.u=e.__N=void 0})):(t.__h.some(Oe),t.__h.some(ke),t.__h=[],se=0)),ce=z},B.diffed=function(e){me&&me(e);var t=e.__c;t&&t.__H&&(t.__H.__h.length&&(de.push(t)!==1&&le===B.requestAnimationFrame||((le=B.requestAnimationFrame)||De)(Te)),t.__H.__.some(function(e){e.u&&(e.__H=e.u),e.u=void 0})),ce=z=null},B.__c=function(e,t){t.some(function(e){try{e.__h.some(Oe),e.__h=e.__h.filter(function(e){return!e.__||ke(e)})}catch(n){t.some(function(e){e.__h&&=[]}),t=[],B.__e(n,e.__v)}}),he&&he(e,t)},B.unmount=function(e){ge&&ge(e);var t,n=e.__c;n&&n.__H&&(n.__H.__.some(function(e){try{Oe(e)}catch(e){t=e}}),n.__H=void 0,t&&B.__e(t,n.__v))};var Ee=typeof requestAnimationFrame==`function`;function De(e){var t,n=function(){clearTimeout(r),Ee&&cancelAnimationFrame(t),setTimeout(e)},r=setTimeout(n,35);Ee&&(t=requestAnimationFrame(n))}function Oe(e){var t=z,n=e.__c;typeof n==`function`&&(e.__c=void 0,n()),z=t}function ke(e){var t=z;e.__c=e.__(),z=t}function Ae(e,t){return!e||e.length!==t.length||t.some(function(t,n){return t!==e[n]})}function je(e,t){return typeof t==`function`?t(e):t}var Me={};function Ne(e,t){for(var n in t)e[n]=t[n];return e}function Pe(e,t,n){var r,i=/(?:\?([^#]*))?(#.*)?$/,a=e.match(i),o={};if(a&&a[1])for(var s=a[1].split(`&`),c=0;c<s.length;c++){var l=s[c].split(`=`);o[decodeURIComponent(l[0])]=decodeURIComponent(l.slice(1).join(`=`))}e=Le(e.replace(i,``)),t=Le(t||``);for(var u=Math.max(e.length,t.length),d=0;d<u;d++)if(t[d]&&t[d].charAt(0)===`:`){var f=t[d].replace(/(^:|[+*?]+$)/g,``),p=(t[d].match(/[+*?]+$/)||Me)[0]||``,m=~p.indexOf(`+`),h=~p.indexOf(`*`),g=e[d]||``;if(!g&&!h&&(p.indexOf(`?`)<0||m)){r=!1;break}if(o[f]=decodeURIComponent(g),m||h){o[f]=e.slice(d).map(decodeURIComponent).join(`/`);break}}else if(t[d]!==e[d]){r=!1;break}return(!0===n.default||!1!==r)&&o}function Fe(e,t){return e.rank<t.rank?1:e.rank>t.rank?-1:e.index-t.index}function Ie(e,t){return e.index=t,e.rank=function(e){return e.props.default?0:Le(e.props.path).map(Re).join(``)}(e),e.props}function Le(e){return e.replace(/(^\/+|\/+$)/g,``).split(`/`)}function Re(e){return e.charAt(0)==`:`?1+`*+?`.indexOf(e.charAt(e.length-1))||4:5}var ze={},Be=[],Ve=[],He=null,Ue={url:Ge()},We=R(Ue);function Ge(){var e;return``+((e=He&&He.location?He.location:He&&He.getCurrentLocation?He.getCurrentLocation():typeof location<`u`?location:ze).pathname||``)+(e.search||``)}function U(e,t){return t===void 0&&(t=!1),typeof e!=`string`&&e.url&&(t=e.replace,e=e.url),function(e){for(var t=Be.length;t--;)if(Be[t].canRoute(e))return!0;return!1}(e)&&function(e,t){t===void 0&&(t=`push`),He&&He[t]?He[t](e):typeof history<`u`&&history[t+`State`]&&history[t+`State`](null,null,e)}(e,t?`replace`:`push`),Ke(e)}function Ke(e){for(var t=!1,n=0;n<Be.length;n++)Be[n].routeTo(e)&&(t=!0);return t}function qe(e){if(e&&e.getAttribute){var t=e.getAttribute(`href`),n=e.getAttribute(`target`);if(t&&t.match(/^\//g)&&(!n||n.match(/^_?self$/i)))return U(t)}}function Je(e){return e.stopImmediatePropagation&&e.stopImmediatePropagation(),e.stopPropagation&&e.stopPropagation(),e.preventDefault(),!1}function Ye(e){if(!(e.ctrlKey||e.metaKey||e.altKey||e.shiftKey||e.button)){var t=e.target;do if(t.localName===`a`&&t.getAttribute(`href`)){if(t.hasAttribute(`data-native`)||t.hasAttribute(`native`))return;if(qe(t))return Je(e)}while(t=t.parentNode)}}var Xe=!1;function Ze(e){e.history&&(He=e.history),this.state={url:e.url||Ge()}}Ne(Ze.prototype=new x,{shouldComponentUpdate:function(e){return!0!==e.static||e.url!==this.props.url||e.onChange!==this.props.onChange},canRoute:function(e){var t=k(this.props.children);return this.g(t,e)!==void 0},routeTo:function(e){this.setState({url:e});var t=this.canRoute(e);return this.p||this.forceUpdate(),t},componentWillMount:function(){this.p=!0},componentDidMount:function(){var e=this;Xe||(Xe=!0,He||addEventListener(`popstate`,function(){Ke(Ge())}),addEventListener(`click`,Ye)),Be.push(this),He&&(this.u=He.listen(function(t){var n=t.location||t;e.routeTo(``+(n.pathname||``)+(n.search||``))})),this.p=!1},componentWillUnmount:function(){typeof this.u==`function`&&this.u(),Be.splice(Be.indexOf(this),1)},componentWillUpdate:function(){this.p=!0},componentDidUpdate:function(){this.p=!1},g:function(e,t){e=e.filter(Ie).sort(Fe);for(var n=0;n<e.length;n++){var r=e[n],i=Pe(t,r.props.path,r.props);if(i)return[r,i]}},render:function(e,t){var n,r,i=e.onChange,a=t.url,o=this.c,s=this.g(k(e.children),a);if(s&&(r=oe(s[0],Ne(Ne({url:a,matches:n=s[1]},n),{key:void 0,ref:void 0}))),a!==(o&&o.url)){Ne(Ue,o=this.c={url:a,previous:o&&o.url,current:r,path:r?r.props.path:null,matches:n}),o.router=this,o.active=r?[r]:[];for(var c=Ve.length;c--;)Ve[c]({});typeof i==`function`&&i(o)}return v(We.Provider,{value:o},r)}});var Qe=function(e){return v(e.component,e)},$e=`DE:AD:BE:EF:00:01`;function W(e){let t=e.startsWith(`/`)?e:`/${e}`,n=(void 0)?.trim();return n?`${n.replace(/\/$/,``)}/api${t}`:`/api${t}`}var et=`0.0.0`;async function tt(e){let t=await e.text();try{let e=JSON.parse(t);if(Array.isArray(e.message))return e.message.join(`, `);if(typeof e.message==`string`)return e.message}catch{}return t||e.statusText}async function G(e){if(!e.ok){let t=await tt(e),n=Error(`HTTP ${e.status}: ${t}`);throw n.status=e.status,n.body=t,n}}async function nt(e){let t=await fetch(W(`/devices/lookup?adresse_mac=${encodeURIComponent(e)}`));return await G(t),t.json()}async function rt(e,t){let n=await fetch(W(`/devices/register`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify({adresse_mac:e,pseudot:t})});return await G(n),n.json()}async function it(){let e=await fetch(W(`/quizz/collections`));return await G(e),e.json()}async function at(e,t){let n=new URLSearchParams;t?.qtype!=null&&t.qtype!==`melanger`&&n.set(`qtype`,t.qtype),t?.orders!=null&&t.orders.length>0&&n.set(`order`,t.orders.join(`,`)),t?.userId!=null&&n.set(`userId`,String(t.userId)),t?.infinite&&n.set(`infinite`,`1`),t?.excludeIds!=null&&t.excludeIds.length>0&&n.set(`exclude`,t.excludeIds.join(`,`));let r=n.size>0?`?${n.toString()}`:``,i=await fetch(W(`/quizz/collections/${e}${r}`));return await G(i),i.json()}async function ot(){let e=await fetch(W(`/quizz/modules`));return await G(e),e.json()}async function st(e){let t=await fetch(W(`/quizz/modules`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify({nom:e})});return await G(t),t.json()}async function ct(e){await G(await fetch(W(`/quizz/modules/${e}`),{method:`DELETE`}))}async function lt(e,t){let n=await fetch(W(`/quizz/collections/${e}/modules`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify({moduleId:t})});return await G(n),n.json()}async function ut(e,t){let n=await fetch(W(`/quizz/collections/${e}/modules/${t}`),{method:`DELETE`});return await G(n),n.json()}async function dt(e,t){let n=new URLSearchParams({userId:String(t)});await G(await fetch(W(`/quizz/collections/${e}?${n.toString()}`),{method:`DELETE`}))}async function ft(e){let t=new URLSearchParams;e?.orders!=null&&e.orders.length>0&&t.set(`order`,e.orders.join(`,`)),e?.qtype!=null&&e.qtype!==`melanger`&&t.set(`qtype`,e.qtype),e?.userId!=null&&t.set(`userId`,String(e.userId)),e?.infinite&&t.set(`infinite`,`1`),e?.excludeIds!=null&&e.excludeIds.length>0&&t.set(`exclude`,e.excludeIds.join(`,`));let n=t.size>0?`?${t.toString()}`:``,r=await fetch(W(`/quizz/random${n}`));return await G(r),r.json()}async function pt(){let e=await fetch(W(`/quizz/categories`));return await G(e),e.json()}async function mt(e){let t=await fetch(W(`/quizz/questions/${e}`));return await G(t),t.json()}async function ht(e){let t=e===`none`?`?collectionId=none`:e===void 0?``:`?collectionId=${e}`,n=await fetch(W(`/quizz/questions${t}`));return await G(n),n.json()}async function gt(e,t){let n=await fetch(W(`/quizz/questions/${e}`),{method:`PATCH`,headers:{"Content-Type":`application/json`},body:JSON.stringify(t)});return await G(n),n.json()}async function _t(e){let t=await fetch(W(`/quizz/questions`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify(e)});return await G(t),t.json()}async function vt(e,t){let n=await fetch(W(`/quizz/reponses/${e}`),{method:`PATCH`,headers:{"Content-Type":`application/json`},body:JSON.stringify(t)});return await G(n),n.json()}async function yt(e){let t=await fetch(W(`/quizz/collections`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify(e)});return await G(t),t.json()}async function bt(e,t){let n=new URLSearchParams;t?.collectionId!=null&&n.set(`collectionId`,String(t.collectionId)),t?.moduleId!=null&&n.set(`moduleId`,String(t.moduleId)),t?.categorie!=null&&n.set(`categorie`,t.categorie);let r=n.size>0?`?${n.toString()}`:``,i=await fetch(W(`/quizz/questions/import${r}`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify(e)});return await G(i),i.json()}async function xt(e,t){let n=new URLSearchParams({collectionId:String(t.collectionId)}),r=await fetch(W(`/quizz/collections/questions/import-app?${n.toString()}`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify(e)});return await G(r),r.json()}async function St(e){await G(await fetch(W(`/quizz/questions/${e}`),{method:`DELETE`}))}async function Ct(e){let t=await fetch(W(`/stats/kpis?userId=${e}`));return await G(t),t.json()}async function wt(e){let t=await fetch(W(`/stats/kpi`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify({userId:e.userId,questionId:e.questionId,reponseId:e.reponseId,dureeSecondes:e.dureeSecondes})});return await G(t),t.json()}async function Tt(e){let t=await fetch(W(`/stats/sessions?userId=${e}`));return await G(t),t.json()}async function Et(e,t){let n=await fetch(W(`/stats/sessions/${encodeURIComponent(e)}?userId=${t}`));return n.status===404?null:(await G(n),n.json())}async function Dt(e,t){let n=await fetch(W(e));return await G(n),{blob:await n.blob(),filename:(n.headers.get(`content-disposition`)??``).match(/filename="([^"]+)"/i)?.[1]??t}}async function Ot(){return Dt(`/admin/database/export.sql`,`quizz-export.sql`)}async function kt(){return Dt(`/admin/database/export.json`,`quizz-export.json`)}async function At(e){let t=await fetch(W(`/admin/database/import.json`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify(e)});return await G(t),t.json()}async function jt(e){let t=await fetch(W(`/admin/database/import.sql`),{method:`POST`,headers:{"Content-Type":`text/plain; charset=utf-8`,"X-Quizz-Confirm":`REMPLACE_TOUT`},body:e});return await G(t),t.json()}var Mt=R(null);function Nt(){let e=we(Mt);if(e==null)throw Error(`useUserSession doit être utilisé sous une session active`);return e}function Pt(e){var t,n,r=``;if(typeof e==`string`||typeof e==`number`)r+=e;else if(typeof e==`object`)if(Array.isArray(e)){var i=e.length;for(t=0;t<i;t++)e[t]&&(n=Pt(e[t]))&&(r&&(r+=` `),r+=n)}else for(n in e)e[n]&&(r&&(r+=` `),r+=n);return r}function Ft(){for(var e,t,n=0,r=``,i=arguments.length;n<i;n++)(e=arguments[n])&&(t=Pt(e))&&(r&&(r+=` `),r+=t);return r}var It=(e,t)=>{let n=Array(e.length+t.length);for(let t=0;t<e.length;t++)n[t]=e[t];for(let r=0;r<t.length;r++)n[e.length+r]=t[r];return n},Lt=(e,t)=>({classGroupId:e,validator:t}),Rt=(e=new Map,t=null,n)=>({nextPart:e,validators:t,classGroupId:n}),zt=`-`,Bt=[],Vt=`arbitrary..`,Ht=e=>{let t=Gt(e),{conflictingClassGroups:n,conflictingClassGroupModifiers:r}=e;return{getClassGroupId:e=>{if(e.startsWith(`[`)&&e.endsWith(`]`))return Wt(e);let n=e.split(zt);return Ut(n,n[0]===``&&n.length>1?1:0,t)},getConflictingClassGroupIds:(e,t)=>{if(t){let t=r[e],i=n[e];return t?i?It(i,t):t:i||Bt}return n[e]||Bt}}},Ut=(e,t,n)=>{if(e.length-t===0)return n.classGroupId;let r=e[t],i=n.nextPart.get(r);if(i){let n=Ut(e,t+1,i);if(n)return n}let a=n.validators;if(a===null)return;let o=t===0?e.join(zt):e.slice(t).join(zt),s=a.length;for(let e=0;e<s;e++){let t=a[e];if(t.validator(o))return t.classGroupId}},Wt=e=>e.slice(1,-1).indexOf(`:`)===-1?void 0:(()=>{let t=e.slice(1,-1),n=t.indexOf(`:`),r=t.slice(0,n);return r?Vt+r:void 0})(),Gt=e=>{let{theme:t,classGroups:n}=e;return Kt(n,t)},Kt=(e,t)=>{let n=Rt();for(let r in e){let i=e[r];qt(i,n,r,t)}return n},qt=(e,t,n,r)=>{let i=e.length;for(let a=0;a<i;a++){let i=e[a];Jt(i,t,n,r)}},Jt=(e,t,n,r)=>{if(typeof e==`string`){Yt(e,t,n);return}if(typeof e==`function`){Xt(e,t,n,r);return}Zt(e,t,n,r)},Yt=(e,t,n)=>{let r=e===``?t:Qt(t,e);r.classGroupId=n},Xt=(e,t,n,r)=>{if($t(e)){qt(e(r),t,n,r);return}t.validators===null&&(t.validators=[]),t.validators.push(Lt(n,e))},Zt=(e,t,n,r)=>{let i=Object.entries(e),a=i.length;for(let e=0;e<a;e++){let[a,o]=i[e];qt(o,Qt(t,a),n,r)}},Qt=(e,t)=>{let n=e,r=t.split(zt),i=r.length;for(let e=0;e<i;e++){let t=r[e],i=n.nextPart.get(t);i||(i=Rt(),n.nextPart.set(t,i)),n=i}return n},$t=e=>`isThemeGetter`in e&&e.isThemeGetter===!0,en=e=>{if(e<1)return{get:()=>void 0,set:()=>{}};let t=0,n=Object.create(null),r=Object.create(null),i=(i,a)=>{n[i]=a,t++,t>e&&(t=0,r=n,n=Object.create(null))};return{get(e){let t=n[e];if(t!==void 0)return t;if((t=r[e])!==void 0)return i(e,t),t},set(e,t){e in n?n[e]=t:i(e,t)}}},tn=`!`,nn=`:`,rn=[],an=(e,t,n,r,i)=>({modifiers:e,hasImportantModifier:t,baseClassName:n,maybePostfixModifierPosition:r,isExternal:i}),on=e=>{let{prefix:t,experimentalParseClassName:n}=e,r=e=>{let t=[],n=0,r=0,i=0,a,o=e.length;for(let s=0;s<o;s++){let o=e[s];if(n===0&&r===0){if(o===nn){t.push(e.slice(i,s)),i=s+1;continue}if(o===`/`){a=s;continue}}o===`[`?n++:o===`]`?n--:o===`(`?r++:o===`)`&&r--}let s=t.length===0?e:e.slice(i),c=s,l=!1;s.endsWith(tn)?(c=s.slice(0,-1),l=!0):s.startsWith(tn)&&(c=s.slice(1),l=!0);let u=a&&a>i?a-i:void 0;return an(t,l,c,u)};if(t){let e=t+nn,n=r;r=t=>t.startsWith(e)?n(t.slice(e.length)):an(rn,!1,t,void 0,!0)}if(n){let e=r;r=t=>n({className:t,parseClassName:e})}return r},sn=e=>{let t=new Map;return e.orderSensitiveModifiers.forEach((e,n)=>{t.set(e,1e6+n)}),e=>{let n=[],r=[];for(let i=0;i<e.length;i++){let a=e[i],o=a[0]===`[`,s=t.has(a);o||s?(r.length>0&&(r.sort(),n.push(...r),r=[]),n.push(a)):r.push(a)}return r.length>0&&(r.sort(),n.push(...r)),n}},cn=e=>({cache:en(e.cacheSize),parseClassName:on(e),sortModifiers:sn(e),...Ht(e)}),ln=/\s+/,un=(e,t)=>{let{parseClassName:n,getClassGroupId:r,getConflictingClassGroupIds:i,sortModifiers:a}=t,o=[],s=e.trim().split(ln),c=``;for(let e=s.length-1;e>=0;--e){let t=s[e],{isExternal:l,modifiers:u,hasImportantModifier:d,baseClassName:f,maybePostfixModifierPosition:p}=n(t);if(l){c=t+(c.length>0?` `+c:c);continue}let m=!!p,h=r(m?f.substring(0,p):f);if(!h){if(!m){c=t+(c.length>0?` `+c:c);continue}if(h=r(f),!h){c=t+(c.length>0?` `+c:c);continue}m=!1}let g=u.length===0?``:u.length===1?u[0]:a(u).join(`:`),_=d?g+tn:g,v=_+h;if(o.indexOf(v)>-1)continue;o.push(v);let y=i(h,m);for(let e=0;e<y.length;++e){let t=y[e];o.push(_+t)}c=t+(c.length>0?` `+c:c)}return c},dn=(...e)=>{let t=0,n,r,i=``;for(;t<e.length;)(n=e[t++])&&(r=fn(n))&&(i&&(i+=` `),i+=r);return i},fn=e=>{if(typeof e==`string`)return e;let t,n=``;for(let r=0;r<e.length;r++)e[r]&&(t=fn(e[r]))&&(n&&(n+=` `),n+=t);return n},pn=(e,...t)=>{let n,r,i,a,o=o=>(n=cn(t.reduce((e,t)=>t(e),e())),r=n.cache.get,i=n.cache.set,a=s,s(o)),s=e=>{let t=r(e);if(t)return t;let a=un(e,n);return i(e,a),a};return a=o,(...e)=>a(dn(...e))},mn=[],K=e=>{let t=t=>t[e]||mn;return t.isThemeGetter=!0,t},hn=/^\[(?:(\w[\w-]*):)?(.+)\]$/i,gn=/^\((?:(\w[\w-]*):)?(.+)\)$/i,_n=/^\d+(?:\.\d+)?\/\d+(?:\.\d+)?$/,vn=/^(\d+(\.\d+)?)?(xs|sm|md|lg|xl)$/,yn=/\d+(%|px|r?em|[sdl]?v([hwib]|min|max)|pt|pc|in|cm|mm|cap|ch|ex|r?lh|cq(w|h|i|b|min|max))|\b(calc|min|max|clamp)\(.+\)|^0$/,bn=/^(rgba?|hsla?|hwb|(ok)?(lab|lch)|color-mix)\(.+\)$/,xn=/^(inset_)?-?((\d+)?\.?(\d+)[a-z]+|0)_-?((\d+)?\.?(\d+)[a-z]+|0)/,Sn=/^(url|image|image-set|cross-fade|element|(repeating-)?(linear|radial|conic)-gradient)\(.+\)$/,Cn=e=>_n.test(e),q=e=>!!e&&!Number.isNaN(Number(e)),wn=e=>!!e&&Number.isInteger(Number(e)),Tn=e=>e.endsWith(`%`)&&q(e.slice(0,-1)),En=e=>vn.test(e),Dn=()=>!0,On=e=>yn.test(e)&&!bn.test(e),kn=()=>!1,An=e=>xn.test(e),jn=e=>Sn.test(e),Mn=e=>!J(e)&&!Y(e),Nn=e=>Jn(e,Qn,kn),J=e=>hn.test(e),Pn=e=>Jn(e,$n,On),Fn=e=>Jn(e,er,q),In=e=>Jn(e,nr,Dn),Ln=e=>Jn(e,tr,kn),Rn=e=>Jn(e,Xn,kn),zn=e=>Jn(e,Zn,jn),Bn=e=>Jn(e,rr,An),Y=e=>gn.test(e),Vn=e=>Yn(e,$n),Hn=e=>Yn(e,tr),Un=e=>Yn(e,Xn),Wn=e=>Yn(e,Qn),Gn=e=>Yn(e,Zn),Kn=e=>Yn(e,rr,!0),qn=e=>Yn(e,nr,!0),Jn=(e,t,n)=>{let r=hn.exec(e);return r?r[1]?t(r[1]):n(r[2]):!1},Yn=(e,t,n=!1)=>{let r=gn.exec(e);return r?r[1]?t(r[1]):n:!1},Xn=e=>e===`position`||e===`percentage`,Zn=e=>e===`image`||e===`url`,Qn=e=>e===`length`||e===`size`||e===`bg-size`,$n=e=>e===`length`,er=e=>e===`number`,tr=e=>e===`family-name`,nr=e=>e===`number`||e===`weight`,rr=e=>e===`shadow`,ir=pn(()=>{let e=K(`color`),t=K(`font`),n=K(`text`),r=K(`font-weight`),i=K(`tracking`),a=K(`leading`),o=K(`breakpoint`),s=K(`container`),c=K(`spacing`),l=K(`radius`),u=K(`shadow`),d=K(`inset-shadow`),f=K(`text-shadow`),p=K(`drop-shadow`),m=K(`blur`),h=K(`perspective`),g=K(`aspect`),_=K(`ease`),v=K(`animate`),y=()=>[`auto`,`avoid`,`all`,`avoid-page`,`page`,`left`,`right`,`column`],b=()=>[`center`,`top`,`bottom`,`left`,`right`,`top-left`,`left-top`,`top-right`,`right-top`,`bottom-right`,`right-bottom`,`bottom-left`,`left-bottom`],x=()=>[...b(),Y,J],S=()=>[`auto`,`hidden`,`clip`,`visible`,`scroll`],C=()=>[`auto`,`contain`,`none`],w=()=>[Y,J,c],T=()=>[Cn,`full`,`auto`,...w()],E=()=>[wn,`none`,`subgrid`,Y,J],ee=()=>[`auto`,{span:[`full`,wn,Y,J]},wn,Y,J],D=()=>[wn,`auto`,Y,J],O=()=>[`auto`,`min`,`max`,`fr`,Y,J],k=()=>[`start`,`end`,`center`,`between`,`around`,`evenly`,`stretch`,`baseline`,`center-safe`,`end-safe`],A=()=>[`start`,`end`,`center`,`stretch`,`center-safe`,`end-safe`],j=()=>[`auto`,...w()],M=()=>[Cn,`auto`,`full`,`dvw`,`dvh`,`lvw`,`lvh`,`svw`,`svh`,`min`,`max`,`fit`,...w()],te=()=>[Cn,`screen`,`full`,`dvw`,`lvw`,`svw`,`min`,`max`,`fit`,...w()],ne=()=>[Cn,`screen`,`full`,`lh`,`dvh`,`lvh`,`svh`,`min`,`max`,`fit`,...w()],N=()=>[e,Y,J],re=()=>[...b(),Un,Rn,{position:[Y,J]}],ie=()=>[`no-repeat`,{repeat:[``,`x`,`y`,`space`,`round`]}],P=()=>[`auto`,`cover`,`contain`,Wn,Nn,{size:[Y,J]}],F=()=>[Tn,Vn,Pn],I=()=>[``,`none`,`full`,l,Y,J],L=()=>[``,q,Vn,Pn],ae=()=>[`solid`,`dashed`,`dotted`,`double`],oe=()=>[`normal`,`multiply`,`screen`,`overlay`,`darken`,`lighten`,`color-dodge`,`color-burn`,`hard-light`,`soft-light`,`difference`,`exclusion`,`hue`,`saturation`,`color`,`luminosity`],R=()=>[q,Tn,Un,Rn],se=()=>[``,`none`,m,Y,J],z=()=>[`none`,q,Y,J],ce=()=>[`none`,q,Y,J],le=()=>[q,Y,J],ue=()=>[Cn,`full`,...w()];return{cacheSize:500,theme:{animate:[`spin`,`ping`,`pulse`,`bounce`],aspect:[`video`],blur:[En],breakpoint:[En],color:[Dn],container:[En],"drop-shadow":[En],ease:[`in`,`out`,`in-out`],font:[Mn],"font-weight":[`thin`,`extralight`,`light`,`normal`,`medium`,`semibold`,`bold`,`extrabold`,`black`],"inset-shadow":[En],leading:[`none`,`tight`,`snug`,`normal`,`relaxed`,`loose`],perspective:[`dramatic`,`near`,`normal`,`midrange`,`distant`,`none`],radius:[En],shadow:[En],spacing:[`px`,q],text:[En],"text-shadow":[En],tracking:[`tighter`,`tight`,`normal`,`wide`,`wider`,`widest`]},classGroups:{aspect:[{aspect:[`auto`,`square`,Cn,J,Y,g]}],container:[`container`],columns:[{columns:[q,J,Y,s]}],"break-after":[{"break-after":y()}],"break-before":[{"break-before":y()}],"break-inside":[{"break-inside":[`auto`,`avoid`,`avoid-page`,`avoid-column`]}],"box-decoration":[{"box-decoration":[`slice`,`clone`]}],box:[{box:[`border`,`content`]}],display:[`block`,`inline-block`,`inline`,`flex`,`inline-flex`,`table`,`inline-table`,`table-caption`,`table-cell`,`table-column`,`table-column-group`,`table-footer-group`,`table-header-group`,`table-row-group`,`table-row`,`flow-root`,`grid`,`inline-grid`,`contents`,`list-item`,`hidden`],sr:[`sr-only`,`not-sr-only`],float:[{float:[`right`,`left`,`none`,`start`,`end`]}],clear:[{clear:[`left`,`right`,`both`,`none`,`start`,`end`]}],isolation:[`isolate`,`isolation-auto`],"object-fit":[{object:[`contain`,`cover`,`fill`,`none`,`scale-down`]}],"object-position":[{object:x()}],overflow:[{overflow:S()}],"overflow-x":[{"overflow-x":S()}],"overflow-y":[{"overflow-y":S()}],overscroll:[{overscroll:C()}],"overscroll-x":[{"overscroll-x":C()}],"overscroll-y":[{"overscroll-y":C()}],position:[`static`,`fixed`,`absolute`,`relative`,`sticky`],inset:[{inset:T()}],"inset-x":[{"inset-x":T()}],"inset-y":[{"inset-y":T()}],start:[{"inset-s":T(),start:T()}],end:[{"inset-e":T(),end:T()}],"inset-bs":[{"inset-bs":T()}],"inset-be":[{"inset-be":T()}],top:[{top:T()}],right:[{right:T()}],bottom:[{bottom:T()}],left:[{left:T()}],visibility:[`visible`,`invisible`,`collapse`],z:[{z:[wn,`auto`,Y,J]}],basis:[{basis:[Cn,`full`,`auto`,s,...w()]}],"flex-direction":[{flex:[`row`,`row-reverse`,`col`,`col-reverse`]}],"flex-wrap":[{flex:[`nowrap`,`wrap`,`wrap-reverse`]}],flex:[{flex:[q,Cn,`auto`,`initial`,`none`,J]}],grow:[{grow:[``,q,Y,J]}],shrink:[{shrink:[``,q,Y,J]}],order:[{order:[wn,`first`,`last`,`none`,Y,J]}],"grid-cols":[{"grid-cols":E()}],"col-start-end":[{col:ee()}],"col-start":[{"col-start":D()}],"col-end":[{"col-end":D()}],"grid-rows":[{"grid-rows":E()}],"row-start-end":[{row:ee()}],"row-start":[{"row-start":D()}],"row-end":[{"row-end":D()}],"grid-flow":[{"grid-flow":[`row`,`col`,`dense`,`row-dense`,`col-dense`]}],"auto-cols":[{"auto-cols":O()}],"auto-rows":[{"auto-rows":O()}],gap:[{gap:w()}],"gap-x":[{"gap-x":w()}],"gap-y":[{"gap-y":w()}],"justify-content":[{justify:[...k(),`normal`]}],"justify-items":[{"justify-items":[...A(),`normal`]}],"justify-self":[{"justify-self":[`auto`,...A()]}],"align-content":[{content:[`normal`,...k()]}],"align-items":[{items:[...A(),{baseline:[``,`last`]}]}],"align-self":[{self:[`auto`,...A(),{baseline:[``,`last`]}]}],"place-content":[{"place-content":k()}],"place-items":[{"place-items":[...A(),`baseline`]}],"place-self":[{"place-self":[`auto`,...A()]}],p:[{p:w()}],px:[{px:w()}],py:[{py:w()}],ps:[{ps:w()}],pe:[{pe:w()}],pbs:[{pbs:w()}],pbe:[{pbe:w()}],pt:[{pt:w()}],pr:[{pr:w()}],pb:[{pb:w()}],pl:[{pl:w()}],m:[{m:j()}],mx:[{mx:j()}],my:[{my:j()}],ms:[{ms:j()}],me:[{me:j()}],mbs:[{mbs:j()}],mbe:[{mbe:j()}],mt:[{mt:j()}],mr:[{mr:j()}],mb:[{mb:j()}],ml:[{ml:j()}],"space-x":[{"space-x":w()}],"space-x-reverse":[`space-x-reverse`],"space-y":[{"space-y":w()}],"space-y-reverse":[`space-y-reverse`],size:[{size:M()}],"inline-size":[{inline:[`auto`,...te()]}],"min-inline-size":[{"min-inline":[`auto`,...te()]}],"max-inline-size":[{"max-inline":[`none`,...te()]}],"block-size":[{block:[`auto`,...ne()]}],"min-block-size":[{"min-block":[`auto`,...ne()]}],"max-block-size":[{"max-block":[`none`,...ne()]}],w:[{w:[s,`screen`,...M()]}],"min-w":[{"min-w":[s,`screen`,`none`,...M()]}],"max-w":[{"max-w":[s,`screen`,`none`,`prose`,{screen:[o]},...M()]}],h:[{h:[`screen`,`lh`,...M()]}],"min-h":[{"min-h":[`screen`,`lh`,`none`,...M()]}],"max-h":[{"max-h":[`screen`,`lh`,...M()]}],"font-size":[{text:[`base`,n,Vn,Pn]}],"font-smoothing":[`antialiased`,`subpixel-antialiased`],"font-style":[`italic`,`not-italic`],"font-weight":[{font:[r,qn,In]}],"font-stretch":[{"font-stretch":[`ultra-condensed`,`extra-condensed`,`condensed`,`semi-condensed`,`normal`,`semi-expanded`,`expanded`,`extra-expanded`,`ultra-expanded`,Tn,J]}],"font-family":[{font:[Hn,Ln,t]}],"font-features":[{"font-features":[J]}],"fvn-normal":[`normal-nums`],"fvn-ordinal":[`ordinal`],"fvn-slashed-zero":[`slashed-zero`],"fvn-figure":[`lining-nums`,`oldstyle-nums`],"fvn-spacing":[`proportional-nums`,`tabular-nums`],"fvn-fraction":[`diagonal-fractions`,`stacked-fractions`],tracking:[{tracking:[i,Y,J]}],"line-clamp":[{"line-clamp":[q,`none`,Y,Fn]}],leading:[{leading:[a,...w()]}],"list-image":[{"list-image":[`none`,Y,J]}],"list-style-position":[{list:[`inside`,`outside`]}],"list-style-type":[{list:[`disc`,`decimal`,`none`,Y,J]}],"text-alignment":[{text:[`left`,`center`,`right`,`justify`,`start`,`end`]}],"placeholder-color":[{placeholder:N()}],"text-color":[{text:N()}],"text-decoration":[`underline`,`overline`,`line-through`,`no-underline`],"text-decoration-style":[{decoration:[...ae(),`wavy`]}],"text-decoration-thickness":[{decoration:[q,`from-font`,`auto`,Y,Pn]}],"text-decoration-color":[{decoration:N()}],"underline-offset":[{"underline-offset":[q,`auto`,Y,J]}],"text-transform":[`uppercase`,`lowercase`,`capitalize`,`normal-case`],"text-overflow":[`truncate`,`text-ellipsis`,`text-clip`],"text-wrap":[{text:[`wrap`,`nowrap`,`balance`,`pretty`]}],indent:[{indent:w()}],"vertical-align":[{align:[`baseline`,`top`,`middle`,`bottom`,`text-top`,`text-bottom`,`sub`,`super`,Y,J]}],whitespace:[{whitespace:[`normal`,`nowrap`,`pre`,`pre-line`,`pre-wrap`,`break-spaces`]}],break:[{break:[`normal`,`words`,`all`,`keep`]}],wrap:[{wrap:[`break-word`,`anywhere`,`normal`]}],hyphens:[{hyphens:[`none`,`manual`,`auto`]}],content:[{content:[`none`,Y,J]}],"bg-attachment":[{bg:[`fixed`,`local`,`scroll`]}],"bg-clip":[{"bg-clip":[`border`,`padding`,`content`,`text`]}],"bg-origin":[{"bg-origin":[`border`,`padding`,`content`]}],"bg-position":[{bg:re()}],"bg-repeat":[{bg:ie()}],"bg-size":[{bg:P()}],"bg-image":[{bg:[`none`,{linear:[{to:[`t`,`tr`,`r`,`br`,`b`,`bl`,`l`,`tl`]},wn,Y,J],radial:[``,Y,J],conic:[wn,Y,J]},Gn,zn]}],"bg-color":[{bg:N()}],"gradient-from-pos":[{from:F()}],"gradient-via-pos":[{via:F()}],"gradient-to-pos":[{to:F()}],"gradient-from":[{from:N()}],"gradient-via":[{via:N()}],"gradient-to":[{to:N()}],rounded:[{rounded:I()}],"rounded-s":[{"rounded-s":I()}],"rounded-e":[{"rounded-e":I()}],"rounded-t":[{"rounded-t":I()}],"rounded-r":[{"rounded-r":I()}],"rounded-b":[{"rounded-b":I()}],"rounded-l":[{"rounded-l":I()}],"rounded-ss":[{"rounded-ss":I()}],"rounded-se":[{"rounded-se":I()}],"rounded-ee":[{"rounded-ee":I()}],"rounded-es":[{"rounded-es":I()}],"rounded-tl":[{"rounded-tl":I()}],"rounded-tr":[{"rounded-tr":I()}],"rounded-br":[{"rounded-br":I()}],"rounded-bl":[{"rounded-bl":I()}],"border-w":[{border:L()}],"border-w-x":[{"border-x":L()}],"border-w-y":[{"border-y":L()}],"border-w-s":[{"border-s":L()}],"border-w-e":[{"border-e":L()}],"border-w-bs":[{"border-bs":L()}],"border-w-be":[{"border-be":L()}],"border-w-t":[{"border-t":L()}],"border-w-r":[{"border-r":L()}],"border-w-b":[{"border-b":L()}],"border-w-l":[{"border-l":L()}],"divide-x":[{"divide-x":L()}],"divide-x-reverse":[`divide-x-reverse`],"divide-y":[{"divide-y":L()}],"divide-y-reverse":[`divide-y-reverse`],"border-style":[{border:[...ae(),`hidden`,`none`]}],"divide-style":[{divide:[...ae(),`hidden`,`none`]}],"border-color":[{border:N()}],"border-color-x":[{"border-x":N()}],"border-color-y":[{"border-y":N()}],"border-color-s":[{"border-s":N()}],"border-color-e":[{"border-e":N()}],"border-color-bs":[{"border-bs":N()}],"border-color-be":[{"border-be":N()}],"border-color-t":[{"border-t":N()}],"border-color-r":[{"border-r":N()}],"border-color-b":[{"border-b":N()}],"border-color-l":[{"border-l":N()}],"divide-color":[{divide:N()}],"outline-style":[{outline:[...ae(),`none`,`hidden`]}],"outline-offset":[{"outline-offset":[q,Y,J]}],"outline-w":[{outline:[``,q,Vn,Pn]}],"outline-color":[{outline:N()}],shadow:[{shadow:[``,`none`,u,Kn,Bn]}],"shadow-color":[{shadow:N()}],"inset-shadow":[{"inset-shadow":[`none`,d,Kn,Bn]}],"inset-shadow-color":[{"inset-shadow":N()}],"ring-w":[{ring:L()}],"ring-w-inset":[`ring-inset`],"ring-color":[{ring:N()}],"ring-offset-w":[{"ring-offset":[q,Pn]}],"ring-offset-color":[{"ring-offset":N()}],"inset-ring-w":[{"inset-ring":L()}],"inset-ring-color":[{"inset-ring":N()}],"text-shadow":[{"text-shadow":[`none`,f,Kn,Bn]}],"text-shadow-color":[{"text-shadow":N()}],opacity:[{opacity:[q,Y,J]}],"mix-blend":[{"mix-blend":[...oe(),`plus-darker`,`plus-lighter`]}],"bg-blend":[{"bg-blend":oe()}],"mask-clip":[{"mask-clip":[`border`,`padding`,`content`,`fill`,`stroke`,`view`]},`mask-no-clip`],"mask-composite":[{mask:[`add`,`subtract`,`intersect`,`exclude`]}],"mask-image-linear-pos":[{"mask-linear":[q]}],"mask-image-linear-from-pos":[{"mask-linear-from":R()}],"mask-image-linear-to-pos":[{"mask-linear-to":R()}],"mask-image-linear-from-color":[{"mask-linear-from":N()}],"mask-image-linear-to-color":[{"mask-linear-to":N()}],"mask-image-t-from-pos":[{"mask-t-from":R()}],"mask-image-t-to-pos":[{"mask-t-to":R()}],"mask-image-t-from-color":[{"mask-t-from":N()}],"mask-image-t-to-color":[{"mask-t-to":N()}],"mask-image-r-from-pos":[{"mask-r-from":R()}],"mask-image-r-to-pos":[{"mask-r-to":R()}],"mask-image-r-from-color":[{"mask-r-from":N()}],"mask-image-r-to-color":[{"mask-r-to":N()}],"mask-image-b-from-pos":[{"mask-b-from":R()}],"mask-image-b-to-pos":[{"mask-b-to":R()}],"mask-image-b-from-color":[{"mask-b-from":N()}],"mask-image-b-to-color":[{"mask-b-to":N()}],"mask-image-l-from-pos":[{"mask-l-from":R()}],"mask-image-l-to-pos":[{"mask-l-to":R()}],"mask-image-l-from-color":[{"mask-l-from":N()}],"mask-image-l-to-color":[{"mask-l-to":N()}],"mask-image-x-from-pos":[{"mask-x-from":R()}],"mask-image-x-to-pos":[{"mask-x-to":R()}],"mask-image-x-from-color":[{"mask-x-from":N()}],"mask-image-x-to-color":[{"mask-x-to":N()}],"mask-image-y-from-pos":[{"mask-y-from":R()}],"mask-image-y-to-pos":[{"mask-y-to":R()}],"mask-image-y-from-color":[{"mask-y-from":N()}],"mask-image-y-to-color":[{"mask-y-to":N()}],"mask-image-radial":[{"mask-radial":[Y,J]}],"mask-image-radial-from-pos":[{"mask-radial-from":R()}],"mask-image-radial-to-pos":[{"mask-radial-to":R()}],"mask-image-radial-from-color":[{"mask-radial-from":N()}],"mask-image-radial-to-color":[{"mask-radial-to":N()}],"mask-image-radial-shape":[{"mask-radial":[`circle`,`ellipse`]}],"mask-image-radial-size":[{"mask-radial":[{closest:[`side`,`corner`],farthest:[`side`,`corner`]}]}],"mask-image-radial-pos":[{"mask-radial-at":b()}],"mask-image-conic-pos":[{"mask-conic":[q]}],"mask-image-conic-from-pos":[{"mask-conic-from":R()}],"mask-image-conic-to-pos":[{"mask-conic-to":R()}],"mask-image-conic-from-color":[{"mask-conic-from":N()}],"mask-image-conic-to-color":[{"mask-conic-to":N()}],"mask-mode":[{mask:[`alpha`,`luminance`,`match`]}],"mask-origin":[{"mask-origin":[`border`,`padding`,`content`,`fill`,`stroke`,`view`]}],"mask-position":[{mask:re()}],"mask-repeat":[{mask:ie()}],"mask-size":[{mask:P()}],"mask-type":[{"mask-type":[`alpha`,`luminance`]}],"mask-image":[{mask:[`none`,Y,J]}],filter:[{filter:[``,`none`,Y,J]}],blur:[{blur:se()}],brightness:[{brightness:[q,Y,J]}],contrast:[{contrast:[q,Y,J]}],"drop-shadow":[{"drop-shadow":[``,`none`,p,Kn,Bn]}],"drop-shadow-color":[{"drop-shadow":N()}],grayscale:[{grayscale:[``,q,Y,J]}],"hue-rotate":[{"hue-rotate":[q,Y,J]}],invert:[{invert:[``,q,Y,J]}],saturate:[{saturate:[q,Y,J]}],sepia:[{sepia:[``,q,Y,J]}],"backdrop-filter":[{"backdrop-filter":[``,`none`,Y,J]}],"backdrop-blur":[{"backdrop-blur":se()}],"backdrop-brightness":[{"backdrop-brightness":[q,Y,J]}],"backdrop-contrast":[{"backdrop-contrast":[q,Y,J]}],"backdrop-grayscale":[{"backdrop-grayscale":[``,q,Y,J]}],"backdrop-hue-rotate":[{"backdrop-hue-rotate":[q,Y,J]}],"backdrop-invert":[{"backdrop-invert":[``,q,Y,J]}],"backdrop-opacity":[{"backdrop-opacity":[q,Y,J]}],"backdrop-saturate":[{"backdrop-saturate":[q,Y,J]}],"backdrop-sepia":[{"backdrop-sepia":[``,q,Y,J]}],"border-collapse":[{border:[`collapse`,`separate`]}],"border-spacing":[{"border-spacing":w()}],"border-spacing-x":[{"border-spacing-x":w()}],"border-spacing-y":[{"border-spacing-y":w()}],"table-layout":[{table:[`auto`,`fixed`]}],caption:[{caption:[`top`,`bottom`]}],transition:[{transition:[``,`all`,`colors`,`opacity`,`shadow`,`transform`,`none`,Y,J]}],"transition-behavior":[{transition:[`normal`,`discrete`]}],duration:[{duration:[q,`initial`,Y,J]}],ease:[{ease:[`linear`,`initial`,_,Y,J]}],delay:[{delay:[q,Y,J]}],animate:[{animate:[`none`,v,Y,J]}],backface:[{backface:[`hidden`,`visible`]}],perspective:[{perspective:[h,Y,J]}],"perspective-origin":[{"perspective-origin":x()}],rotate:[{rotate:z()}],"rotate-x":[{"rotate-x":z()}],"rotate-y":[{"rotate-y":z()}],"rotate-z":[{"rotate-z":z()}],scale:[{scale:ce()}],"scale-x":[{"scale-x":ce()}],"scale-y":[{"scale-y":ce()}],"scale-z":[{"scale-z":ce()}],"scale-3d":[`scale-3d`],skew:[{skew:le()}],"skew-x":[{"skew-x":le()}],"skew-y":[{"skew-y":le()}],transform:[{transform:[Y,J,``,`none`,`gpu`,`cpu`]}],"transform-origin":[{origin:x()}],"transform-style":[{transform:[`3d`,`flat`]}],translate:[{translate:ue()}],"translate-x":[{"translate-x":ue()}],"translate-y":[{"translate-y":ue()}],"translate-z":[{"translate-z":ue()}],"translate-none":[`translate-none`],accent:[{accent:N()}],appearance:[{appearance:[`none`,`auto`]}],"caret-color":[{caret:N()}],"color-scheme":[{scheme:[`normal`,`dark`,`light`,`light-dark`,`only-dark`,`only-light`]}],cursor:[{cursor:[`auto`,`default`,`pointer`,`wait`,`text`,`move`,`help`,`not-allowed`,`none`,`context-menu`,`progress`,`cell`,`crosshair`,`vertical-text`,`alias`,`copy`,`no-drop`,`grab`,`grabbing`,`all-scroll`,`col-resize`,`row-resize`,`n-resize`,`e-resize`,`s-resize`,`w-resize`,`ne-resize`,`nw-resize`,`se-resize`,`sw-resize`,`ew-resize`,`ns-resize`,`nesw-resize`,`nwse-resize`,`zoom-in`,`zoom-out`,Y,J]}],"field-sizing":[{"field-sizing":[`fixed`,`content`]}],"pointer-events":[{"pointer-events":[`auto`,`none`]}],resize:[{resize:[`none`,``,`y`,`x`]}],"scroll-behavior":[{scroll:[`auto`,`smooth`]}],"scroll-m":[{"scroll-m":w()}],"scroll-mx":[{"scroll-mx":w()}],"scroll-my":[{"scroll-my":w()}],"scroll-ms":[{"scroll-ms":w()}],"scroll-me":[{"scroll-me":w()}],"scroll-mbs":[{"scroll-mbs":w()}],"scroll-mbe":[{"scroll-mbe":w()}],"scroll-mt":[{"scroll-mt":w()}],"scroll-mr":[{"scroll-mr":w()}],"scroll-mb":[{"scroll-mb":w()}],"scroll-ml":[{"scroll-ml":w()}],"scroll-p":[{"scroll-p":w()}],"scroll-px":[{"scroll-px":w()}],"scroll-py":[{"scroll-py":w()}],"scroll-ps":[{"scroll-ps":w()}],"scroll-pe":[{"scroll-pe":w()}],"scroll-pbs":[{"scroll-pbs":w()}],"scroll-pbe":[{"scroll-pbe":w()}],"scroll-pt":[{"scroll-pt":w()}],"scroll-pr":[{"scroll-pr":w()}],"scroll-pb":[{"scroll-pb":w()}],"scroll-pl":[{"scroll-pl":w()}],"snap-align":[{snap:[`start`,`end`,`center`,`align-none`]}],"snap-stop":[{snap:[`normal`,`always`]}],"snap-type":[{snap:[`none`,`x`,`y`,`both`]}],"snap-strictness":[{snap:[`mandatory`,`proximity`]}],touch:[{touch:[`auto`,`none`,`manipulation`]}],"touch-x":[{"touch-pan":[`x`,`left`,`right`]}],"touch-y":[{"touch-pan":[`y`,`up`,`down`]}],"touch-pz":[`touch-pinch-zoom`],select:[{select:[`none`,`text`,`all`,`auto`]}],"will-change":[{"will-change":[`auto`,`scroll`,`contents`,`transform`,Y,J]}],fill:[{fill:[`none`,...N()]}],"stroke-w":[{stroke:[q,Vn,Pn,Fn]}],stroke:[{stroke:[`none`,...N()]}],"forced-color-adjust":[{"forced-color-adjust":[`auto`,`none`]}]},conflictingClassGroups:{overflow:[`overflow-x`,`overflow-y`],overscroll:[`overscroll-x`,`overscroll-y`],inset:[`inset-x`,`inset-y`,`inset-bs`,`inset-be`,`start`,`end`,`top`,`right`,`bottom`,`left`],"inset-x":[`right`,`left`],"inset-y":[`top`,`bottom`],flex:[`basis`,`grow`,`shrink`],gap:[`gap-x`,`gap-y`],p:[`px`,`py`,`ps`,`pe`,`pbs`,`pbe`,`pt`,`pr`,`pb`,`pl`],px:[`pr`,`pl`],py:[`pt`,`pb`],m:[`mx`,`my`,`ms`,`me`,`mbs`,`mbe`,`mt`,`mr`,`mb`,`ml`],mx:[`mr`,`ml`],my:[`mt`,`mb`],size:[`w`,`h`],"font-size":[`leading`],"fvn-normal":[`fvn-ordinal`,`fvn-slashed-zero`,`fvn-figure`,`fvn-spacing`,`fvn-fraction`],"fvn-ordinal":[`fvn-normal`],"fvn-slashed-zero":[`fvn-normal`],"fvn-figure":[`fvn-normal`],"fvn-spacing":[`fvn-normal`],"fvn-fraction":[`fvn-normal`],"line-clamp":[`display`,`overflow`],rounded:[`rounded-s`,`rounded-e`,`rounded-t`,`rounded-r`,`rounded-b`,`rounded-l`,`rounded-ss`,`rounded-se`,`rounded-ee`,`rounded-es`,`rounded-tl`,`rounded-tr`,`rounded-br`,`rounded-bl`],"rounded-s":[`rounded-ss`,`rounded-es`],"rounded-e":[`rounded-se`,`rounded-ee`],"rounded-t":[`rounded-tl`,`rounded-tr`],"rounded-r":[`rounded-tr`,`rounded-br`],"rounded-b":[`rounded-br`,`rounded-bl`],"rounded-l":[`rounded-tl`,`rounded-bl`],"border-spacing":[`border-spacing-x`,`border-spacing-y`],"border-w":[`border-w-x`,`border-w-y`,`border-w-s`,`border-w-e`,`border-w-bs`,`border-w-be`,`border-w-t`,`border-w-r`,`border-w-b`,`border-w-l`],"border-w-x":[`border-w-r`,`border-w-l`],"border-w-y":[`border-w-t`,`border-w-b`],"border-color":[`border-color-x`,`border-color-y`,`border-color-s`,`border-color-e`,`border-color-bs`,`border-color-be`,`border-color-t`,`border-color-r`,`border-color-b`,`border-color-l`],"border-color-x":[`border-color-r`,`border-color-l`],"border-color-y":[`border-color-t`,`border-color-b`],translate:[`translate-x`,`translate-y`,`translate-none`],"translate-none":[`translate`,`translate-x`,`translate-y`,`translate-z`],"scroll-m":[`scroll-mx`,`scroll-my`,`scroll-ms`,`scroll-me`,`scroll-mbs`,`scroll-mbe`,`scroll-mt`,`scroll-mr`,`scroll-mb`,`scroll-ml`],"scroll-mx":[`scroll-mr`,`scroll-ml`],"scroll-my":[`scroll-mt`,`scroll-mb`],"scroll-p":[`scroll-px`,`scroll-py`,`scroll-ps`,`scroll-pe`,`scroll-pbs`,`scroll-pbe`,`scroll-pt`,`scroll-pr`,`scroll-pb`,`scroll-pl`],"scroll-px":[`scroll-pr`,`scroll-pl`],"scroll-py":[`scroll-pt`,`scroll-pb`],touch:[`touch-x`,`touch-y`,`touch-pz`],"touch-x":[`touch`],"touch-y":[`touch`],"touch-pz":[`touch`]},conflictingClassGroupModifiers:{"font-size":[`leading`]},orderSensitiveModifiers:[`*`,`**`,`after`,`backdrop`,`before`,`details-content`,`file`,`first-letter`,`first-line`,`marker`,`placeholder`,`selection`]}});function ar(...e){return ir(Ft(e))}var or={base:`btn rounded-full border-0 shadow-md transition-all duration-300 ease-out hover:shadow-lg active:scale-[0.97] disabled:opacity-50 disabled:pointer-events-none`,variants:{flow:`bg-flow text-white hover:brightness-110`,learn:`bg-learn text-white hover:brightness-110`,ghost:`bg-base-100/80 text-flow shadow-sm hover:bg-base-100`,outline:`bg-transparent text-flow border-2 border-flow/40 hover:border-flow hover:bg-flow/5`}},sr=0;Array.isArray;function X(e,n,r,i,a,o){n||={};var s,c,l=n;if(`ref`in l)for(c in l={},n)c==`ref`?s=n[c]:l[c]=n[c];var u={type:e,props:l,key:r,ref:s,__k:null,__:null,__b:0,__e:null,__c:null,constructor:void 0,__v:--sr,__i:-1,__u:0,__source:a,__self:o};if(typeof e==`function`&&(s=e.defaultProps))for(c in s)l[c]===void 0&&(l[c]=s[c]);return t.vnode&&t.vnode(u),u}function Z({children:e,class:t,variant:n=`flow`,disabled:r,type:i=`button`,onClick:a,title:o,"aria-label":s}){return X(`button`,{type:i,disabled:r,onClick:a,class:ar(or.base,or.variants[n],t),title:o,"aria-label":s,children:e})}var cr={base:`rounded-[var(--radius-box)] bg-base-100/90 backdrop-blur-sm shadow-lg shadow-flow/5 border border-base-content/5 transition-all duration-300 ease-out`,paddings:{sm:`p-4`,md:`p-6`,lg:`p-8`}};function lr({children:e,class:t,padding:n=`md`,...r}){return X(`div`,{...r,class:ar(cr.base,cr.paddings[n],t),children:e})}var ur={fullscreenCenter:`fixed inset-0 z-[100] flex items-center justify-center bg-gradient-to-b from-base-100 via-base-100/95 to-base-200/80 p-4`};function dr({pseudot:e,onDone:t}){let[n,r]=V(!1),i=xe(!1),a=()=>{i.current||(i.current=!0,t())};return H(()=>{if(window.matchMedia(`(prefers-reduced-motion: reduce)`).matches){let e=window.setTimeout(a,450);return()=>clearTimeout(e)}let e=window.setTimeout(()=>r(!0),1400);return()=>clearTimeout(e)},[]),X(`div`,{class:`fixed inset-0 z-[100] flex items-center justify-center bg-gradient-to-b from-base-100 via-base-100/98 to-base-200/90 backdrop-blur-[2px] transition-opacity duration-[900ms] ease-out ${n?`pointer-events-none opacity-0`:`opacity-100`}`,onTransitionEnd:e=>{e.target===e.currentTarget&&e.propertyName===`opacity`&&n&&a()},children:X(`p`,{class:`fl-welcome-text px-6 text-center text-2xl font-semibold tracking-tight text-base-content sm:text-3xl`,children:[`Bienvenue, `,X(`span`,{class:`text-flow`,children:e}),X(`span`,{class:`text-learn`,children:` !`})]})})}function fr({onSubmit:e,busy:t,error:n}){let[r,i]=V(``);return X(`div`,{class:ur.fullscreenCenter,children:X(lr,{class:`w-full max-w-md border-base-content/10 shadow-xl shadow-flow/5`,children:[X(`h1`,{class:`mb-1 text-xl font-bold text-base-content`,children:`Premiere visite`}),X(`p`,{class:`mb-4 text-sm text-base-content/65`,children:[`Cet appareil n est pas reconnu en base (ou la base est vide). Indique ton pseudonyme pour creer ton profil (demo : MAC simulee `,X(`code`,{class:`rounded bg-base-200 px-1.5 py-0.5 text-xs`,children:$e}),`).`]}),X(`label`,{class:`mb-2 block text-sm font-medium text-base-content/80`,for:`gate-pseudot`,children:`Pseudonyme`}),X(`input`,{id:`gate-pseudot`,class:`input input-bordered mb-3 w-full rounded-xl border-base-content/15 bg-base-100`,type:`text`,autoComplete:`username`,placeholder:`ex. zelda_quizz`,value:r,disabled:t,onInput:e=>i(e.target.value)}),n?X(`p`,{class:`mb-3 text-sm text-error`,children:n}):null,X(Z,{variant:`flow`,class:`w-full`,disabled:t||!r.trim(),onClick:()=>e(r.trim()),children:t?`Enregistrement...`:`Continuer`})]})})}function pr(){return X(`div`,{class:`fixed inset-0 z-[100] flex flex-col items-center justify-center gap-3 bg-base-100/95`,children:[X(`span`,{class:`loading loading-spinner loading-md text-flow`,"aria-hidden":!0}),X(`p`,{class:`text-sm text-base-content/60`,children:`Reconnaissance de l appareil...`})]})}function mr({message:e,onRetry:t}){return X(`div`,{class:ur.fullscreenCenter,children:X(lr,{class:`w-full max-w-md border-base-content/10 shadow-xl shadow-flow/5`,children:[X(`h1`,{class:`mb-1 text-xl font-bold text-base-content`,children:`Connexion au serveur`}),X(`p`,{class:`mb-4 text-sm text-base-content/65`,children:e}),X(Z,{variant:`flow`,class:`w-full`,onClick:t,children:`Reessayer`})]})})}function hr({children:e}){let[t,n]=V(null),[r,i]=V(`checking`),[a,o]=V(null),[s,c]=V(null),[l,u]=V(!1),[d,f]=V(0);return H(()=>{let e=!1;return c(null),(async()=>{try{let t=await nt($e);if(e)return;t.known?(n({userId:t.user.id,pseudot:t.user.pseudot}),i(`welcome`)):i(`need-pseudot`)}catch(t){e||(c(t instanceof Error?t.message:`Erreur inconnue.`),i(`api-error`))}})(),()=>{e=!0}},[d]),X(Mt.Provider,{value:r===`ready`?t:null,children:[r===`checking`?X(pr,{}):null,r===`api-error`&&s?X(mr,{message:s,onRetry:()=>f(e=>e+1)}):null,r===`need-pseudot`?X(fr,{onSubmit:async e=>{o(null),u(!0);try{let t=await rt($e,e);n({userId:t.userId,pseudot:t.pseudot}),i(`welcome`)}catch(e){o(e instanceof Error?e.message:`Echec de l enregistrement.`)}finally{u(!1)}},busy:l,error:a}):null,r===`welcome`&&t?X(dr,{pseudot:t.pseudot,onDone:()=>i(`ready`)}):null,r===`ready`&&t?e:null]})}var gr=R(`/`);function _r(){return we(gr)}function vr(e){return typeof e==`object`&&!!e&&!Array.isArray(e)}function yr(e,t){if(!Array.isArray(e)||e.length!==4)throw Error(`${t} : il faut exactement 4 réponses.`);let n=[],r=0;for(let i=0;i<e.length;i++){let a=e[i];if(!a||typeof a!=`object`||Array.isArray(a))throw Error(`${t}[${i}] : objet réponse attendu.`);let o=a,s=o.texte??o.reponse;if(typeof s!=`string`||s.trim()===``)throw Error(`${t}[${i}] : "texte" non vide requis.`);let c=o.correcte===!0||o.bonne_reponse===1;c&&(r+=1),n.push({texte:s.trim(),correcte:c})}if(r!==1)throw Error(`${t} : exactement une réponse doit être correcte.`);return n}function br(e,t){if(!e||typeof e!=`object`||Array.isArray(e))throw Error(`${t} : objet question attendu.`);let n=e,r=n.categorie_id;if(typeof r!=`number`||!Number.isInteger(r)||r<1)throw Error(`${t} : categorie_id entier ≥ 1 requis.`);let i=n.categorie_type;if(typeof i!=`string`||i.trim()===``)throw Error(`${t} : categorie_type non vide requis.`);if(typeof n.question!=`string`||n.question.trim()===``)throw Error(`${t} : "question" non vide requise.`);let a=typeof n.commentaire==`string`?n.commentaire.trim():``,o=n.fakechecker??n.verifier,s=o===!0||o===1||o===`1`||typeof o==`string`&&o.trim().toLowerCase()===`true`;return{categorie_id:r,categorie_type:i.trim(),question:n.question.trim(),commentaire:a,fakechecker:s,reponses:yr(n.reponses,`${t}.reponses`)}}function xr(e){let t=JSON.parse(e);if(!vr(t))throw Error(`JSON invalide : objet racine attendu.`);if(t.format!==`flowlearn-app-collection-export`)throw Error(`JSON invalide : format "${String(t.format)}" — attendu "flowlearn-app-collection-export".`);if(t.version!==1)throw Error(`JSON invalide : version ${String(t.version)} — attendu 1.`);let n=t.exportedAt;if(n!=null&&typeof n!=`string`)throw Error(`JSON invalide : exportedAt doit être une chaîne ISO si présent.`);let r=t.user_id,i;if(r!=null)if(typeof r==`number`&&Number.isInteger(r)&&r>0)i=r;else if(typeof r==`string`&&/^\d+$/.test(r.trim())&&Number(r.trim())>0)i=Number(r.trim());else throw Error(`user_id : entier positif attendu si fourni.`);let a=t.collection;if(!vr(a))throw Error(`JSON invalide : objet collection attendu.`);let o=a.nom;if(typeof o!=`string`||o.trim()===``)throw Error(`collection.nom : chaîne non vide requise.`);let s,c=a.id;if(c!=null)if(typeof c==`number`&&Number.isInteger(c)&&c>=1)s=c;else if(typeof c==`string`&&/^\d+$/.test(c.trim())&&Number(c.trim())>=1)s=Number(c.trim());else throw Error(`collection.id : entier ≥ 1 attendu si fourni.`);let l,u=a.user_id;if(u!=null)if(typeof u==`number`&&Number.isInteger(u)&&u>=1)l=u;else if(typeof u==`string`&&/^\d+$/.test(u.trim())&&Number(u.trim())>=1)l=Number(u.trim());else throw Error(`collection.user_id : entier ≥ 1 attendu si fourni.`);let d=t.questions;if(!Array.isArray(d)||d.length===0)throw Error(`questions : tableau non vide requis.`);let f=d.map((e,t)=>br(e,`questions[${t}]`)),p={nom:o.trim()};return s!=null&&(p.id=s),l!=null&&(p.user_id=l),{format:`flowlearn-app-collection-export`,version:1,exportedAt:typeof n==`string`?n:void 0,user_id:i,collection:p,questions:f}}function Sr(e,t){if(!Array.isArray(e)||e.length!==4)throw Error(`${t} : il faut exactement 4 réponses.`);let n=[],r=0;for(let i=0;i<e.length;i++){let a=e[i];if(!a||typeof a!=`object`||Array.isArray(a))throw Error(`${t}[${i}] : objet réponse attendu.`);let o=a,s=o.texte??o.reponse;if(typeof s!=`string`||s.trim()===``)throw Error(`${t}[${i}] : "texte" non vide requis.`);let c=o.correcte===!0||o.bonne_reponse===1;c&&(r+=1),n.push({texte:s.trim(),correcte:c})}if(r!==1)throw Error(`${t} : exactement une réponse doit être correcte.`);return n}function Cr(e,t){if(!e||typeof e!=`object`||Array.isArray(e))throw Error(`${t} : objet question attendu.`);let n=e;if(typeof n.question!=`string`||n.question.trim()===``)throw Error(`${t} : "question" non vide requise.`);let r=typeof n.commentaire==`string`?n.commentaire.trim():``;return{question:n.question.trim(),commentaire:r,reponses:Sr(n.reponses,`${t}.reponses`)}}function wr(e,t){if(e==null)return[];if(!Array.isArray(e))throw Error(`${t} : tableau attendu.`);return e.map((e,n)=>Cr(e,`${t}[${n}]`))}function Tr(e){if(e==null)return[];if(!Array.isArray(e))throw Error(`collections : tableau attendu.`);return e.map((e,t)=>{if(!e||typeof e!=`object`||Array.isArray(e))throw Error(`collections[${t}] : objet attendu.`);let n=e;if(typeof n.nom!=`string`||n.nom.trim()===``)throw Error(`collections[${t}] : "nom" non vide requis.`);let r=wr(n.questions,`collections[${t}].questions`);if(r.length===0)throw Error(`collections[${t}] : au moins une question requise.`);return{nom:n.nom.trim(),questions:r}})}function Er(e){let t=JSON.parse(e);if(!t||typeof t!=`object`||Array.isArray(t))throw Error(`JSON invalide : objet racine attendu.`);let n=t,r=n.user_id,i;if(r!=null)if(typeof r==`number`&&Number.isInteger(r)&&r>0)i=r;else if(typeof r==`string`&&/^\d+$/.test(r.trim())&&Number(r.trim())>0)i=Number(r.trim());else throw Error(`user_id : entier positif attendu.`);let a=wr(n.questions,`questions`),o=Tr(n.collections),s=wr(n.questions_sans_collection,`questions_sans_collection`);if(a.length>0&&(o.length>0||s.length>0))throw Error(`Format mixte interdit : utilise soit "questions", soit "collections"/"questions_sans_collection".`);let c={user_id:i,collections:o,questions_sans_collection:a.length>0?a:s};if(c.collections.length===0&&c.questions_sans_collection.length===0)throw Error(`Aucune question détectée dans le JSON.`);return c}var Dr=(...e)=>e.filter((e,t,n)=>!!e&&e.trim()!==``&&n.indexOf(e)===t).join(` `).trim(),Or=e=>e.replace(/([a-z0-9])([A-Z])/g,`$1-$2`).toLowerCase(),kr=e=>e.replace(/^([A-Z])|[\s-_]+(\w)/g,(e,t,n)=>n?n.toUpperCase():t.toLowerCase()),Ar=e=>{let t=kr(e);return t.charAt(0).toUpperCase()+t.slice(1)},jr={xmlns:`http://www.w3.org/2000/svg`,width:24,height:24,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,"stroke-width":`2`,"stroke-linecap":`round`,"stroke-linejoin":`round`},Mr=R({size:24,color:`currentColor`,strokeWidth:2,absoluteStrokeWidth:!1,class:``}),Nr=()=>we(Mr),Pr=e=>{for(let t in e)if(t.startsWith(`aria-`)||t===`role`||t===`title`)return!0;return!1},Fr=({color:e,size:t,strokeWidth:n,absoluteStrokeWidth:r,children:i,iconNode:a,class:o=``,...s})=>{let{size:c=24,strokeWidth:l=2,absoluteStrokeWidth:u=!1,color:d=`currentColor`,class:f=``}=Nr()??{},p=r??u?Number(n??l)*24/Number(t??c):n??l;return v(`svg`,{...jr,width:t??c??24,height:t??c??24,stroke:e??d,"stroke-width":p,class:Dr(`lucide`,f,o),...!i&&!Pr(s)&&{"aria-hidden":`true`},...s},[...a.map(([e,t])=>v(e,t)),...k(i)])},Q=(e,t)=>{let n=({class:n=``,className:r=``,children:i,...a})=>v(Fr,{...a,iconNode:t,class:Dr(`lucide-${Or(Ar(e))}`,`lucide-${Or(e)}`,n,r)},i);return n.displayName=Ar(e),n},Ir=Q(`arrow-left`,[[`path`,{d:`m12 19-7-7 7-7`,key:`1l729n`}],[`path`,{d:`M19 12H5`,key:`x3x0zl`}]]),Lr=Q(`chart-column`,[[`path`,{d:`M3 3v16a2 2 0 0 0 2 2h16`,key:`c24i48`}],[`path`,{d:`M18 17V9`,key:`2bz60n`}],[`path`,{d:`M13 17V5`,key:`1frdt8`}],[`path`,{d:`M8 17v-3`,key:`17ska0`}]]),Rr=Q(`chevron-right`,[[`path`,{d:`m9 18 6-6-6-6`,key:`mthhwq`}]]),zr=Q(`clipboard-copy`,[[`rect`,{width:`8`,height:`4`,x:`8`,y:`2`,rx:`1`,ry:`1`,key:`tgr4d6`}],[`path`,{d:`M8 4H6a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2v-2`,key:`4jdomd`}],[`path`,{d:`M16 4h2a2 2 0 0 1 2 2v4`,key:`3hqy98`}],[`path`,{d:`M21 14H11`,key:`1bme5i`}],[`path`,{d:`m15 10-4 4 4 4`,key:`5dvupr`}]]),Br=Q(`clock`,[[`circle`,{cx:`12`,cy:`12`,r:`10`,key:`1mglay`}],[`path`,{d:`M12 6v6l4 2`,key:`mmk7yg`}]]),Vr=Q(`download`,[[`path`,{d:`M12 15V3`,key:`m9g1x1`}],[`path`,{d:`M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4`,key:`ih7n3h`}],[`path`,{d:`m7 10 5 5 5-5`,key:`brsn70`}]]),Hr=Q(`file-braces`,[[`path`,{d:`M6 22a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h8a2.4 2.4 0 0 1 1.704.706l3.588 3.588A2.4 2.4 0 0 1 20 8v12a2 2 0 0 1-2 2z`,key:`1oefj6`}],[`path`,{d:`M14 2v5a1 1 0 0 0 1 1h5`,key:`wfsgrz`}],[`path`,{d:`M10 12a1 1 0 0 0-1 1v1a1 1 0 0 1-1 1 1 1 0 0 1 1 1v1a1 1 0 0 0 1 1`,key:`1oajmo`}],[`path`,{d:`M14 18a1 1 0 0 0 1-1v-1a1 1 0 0 1 1-1 1 1 0 0 1-1-1v-1a1 1 0 0 0-1-1`,key:`mpwhp6`}]]),Ur=Q(`file-up`,[[`path`,{d:`M6 22a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h8a2.4 2.4 0 0 1 1.704.706l3.588 3.588A2.4 2.4 0 0 1 20 8v12a2 2 0 0 1-2 2z`,key:`1oefj6`}],[`path`,{d:`M14 2v5a1 1 0 0 0 1 1h5`,key:`wfsgrz`}],[`path`,{d:`M12 12v6`,key:`3ahymv`}],[`path`,{d:`m15 15-3-3-3 3`,key:`15xj92`}]]),Wr=Q(`folder-tree`,[[`path`,{d:`M20 10a1 1 0 0 0 1-1V6a1 1 0 0 0-1-1h-2.5a1 1 0 0 1-.8-.4l-.9-1.2A1 1 0 0 0 15 3h-2a1 1 0 0 0-1 1v5a1 1 0 0 0 1 1Z`,key:`hod4my`}],[`path`,{d:`M20 21a1 1 0 0 0 1-1v-3a1 1 0 0 0-1-1h-2.9a1 1 0 0 1-.88-.55l-.42-.85a1 1 0 0 0-.92-.6H13a1 1 0 0 0-1 1v5a1 1 0 0 0 1 1Z`,key:`w4yl2u`}],[`path`,{d:`M3 5a2 2 0 0 0 2 2h3`,key:`f2jnh7`}],[`path`,{d:`M3 3v13a2 2 0 0 0 2 2h3`,key:`k8epm1`}]]),Gr=Q(`graduation-cap`,[[`path`,{d:`M21.42 10.922a1 1 0 0 0-.019-1.838L12.83 5.18a2 2 0 0 0-1.66 0L2.6 9.08a1 1 0 0 0 0 1.832l8.57 3.908a2 2 0 0 0 1.66 0z`,key:`j76jl0`}],[`path`,{d:`M22 10v6`,key:`1lu8f3`}],[`path`,{d:`M6 12.5V16a6 3 0 0 0 12 0v-3.5`,key:`1r8lef`}]]),Kr=Q(`hard-drive-download`,[[`path`,{d:`M12 2v8`,key:`1q4o3n`}],[`path`,{d:`m16 6-4 4-4-4`,key:`6wukr`}],[`rect`,{width:`20`,height:`8`,x:`2`,y:`14`,rx:`2`,key:`w68u3i`}],[`path`,{d:`M6 18h.01`,key:`uhywen`}],[`path`,{d:`M10 18h.01`,key:`h775k`}]]),qr=Q(`layers`,[[`path`,{d:`M12.83 2.18a2 2 0 0 0-1.66 0L2.6 6.08a1 1 0 0 0 0 1.83l8.58 3.91a2 2 0 0 0 1.66 0l8.58-3.9a1 1 0 0 0 0-1.83z`,key:`zw3jo`}],[`path`,{d:`M2 12a1 1 0 0 0 .58.91l8.6 3.91a2 2 0 0 0 1.65 0l8.58-3.9A1 1 0 0 0 22 12`,key:`1wduqc`}],[`path`,{d:`M2 17a1 1 0 0 0 .58.91l8.6 3.91a2 2 0 0 0 1.65 0l8.58-3.9A1 1 0 0 0 22 17`,key:`kqbvx6`}]]),Jr=Q(`list-tree`,[[`path`,{d:`M8 5h13`,key:`1pao27`}],[`path`,{d:`M13 12h8`,key:`h98zly`}],[`path`,{d:`M13 19h8`,key:`c3s6r1`}],[`path`,{d:`M3 10a2 2 0 0 0 2 2h3`,key:`1npucw`}],[`path`,{d:`M3 5v12a2 2 0 0 0 2 2h3`,key:`x1gjn2`}]]),Yr=Q(`pencil`,[[`path`,{d:`M21.174 6.812a1 1 0 0 0-3.986-3.987L3.842 16.174a2 2 0 0 0-.5.83l-1.321 4.352a.5.5 0 0 0 .623.622l4.353-1.32a2 2 0 0 0 .83-.497z`,key:`1a8usu`}],[`path`,{d:`m15 5 4 4`,key:`1mk7zo`}]]),Xr=Q(`plus`,[[`path`,{d:`M5 12h14`,key:`1ays0h`}],[`path`,{d:`M12 5v14`,key:`s699le`}]]),Zr=Q(`replace`,[[`path`,{d:`M14 4a1 1 0 0 1 1-1`,key:`dhj8ez`}],[`path`,{d:`M15 10a1 1 0 0 1-1-1`,key:`1mnyi5`}],[`path`,{d:`M21 4a1 1 0 0 0-1-1`,key:`sfs9ap`}],[`path`,{d:`M21 9a1 1 0 0 1-1 1`,key:`mp6qeo`}],[`path`,{d:`m3 7 3 3 3-3`,key:`x25e72`}],[`path`,{d:`M6 10V5a2 2 0 0 1 2-2h2`,key:`15xut4`}],[`rect`,{x:`3`,y:`14`,width:`7`,height:`7`,rx:`1`,key:`1bkyp8`}]]),Qr=Q(`shuffle`,[[`path`,{d:`m18 14 4 4-4 4`,key:`10pe0f`}],[`path`,{d:`m18 2 4 4-4 4`,key:`pucp1d`}],[`path`,{d:`M2 18h1.973a4 4 0 0 0 3.3-1.7l5.454-8.6a4 4 0 0 1 3.3-1.7H22`,key:`1ailkh`}],[`path`,{d:`M2 6h1.972a4 4 0 0 1 3.6 2.2`,key:`km57vx`}],[`path`,{d:`M22 18h-6.041a4 4 0 0 1-3.3-1.8l-.359-.45`,key:`os18l9`}]]),$r=Q(`sparkles`,[[`path`,{d:`M11.017 2.814a1 1 0 0 1 1.966 0l1.051 5.558a2 2 0 0 0 1.594 1.594l5.558 1.051a1 1 0 0 1 0 1.966l-5.558 1.051a2 2 0 0 0-1.594 1.594l-1.051 5.558a1 1 0 0 1-1.966 0l-1.051-5.558a2 2 0 0 0-1.594-1.594l-5.558-1.051a1 1 0 0 1 0-1.966l5.558-1.051a2 2 0 0 0 1.594-1.594z`,key:`1s2grr`}],[`path`,{d:`M20 2v4`,key:`1rf3ol`}],[`path`,{d:`M22 4h-4`,key:`gwowj6`}],[`circle`,{cx:`4`,cy:`20`,r:`2`,key:`6kqj1y`}]]),ei=Q(`target`,[[`circle`,{cx:`12`,cy:`12`,r:`10`,key:`1mglay`}],[`circle`,{cx:`12`,cy:`12`,r:`6`,key:`1vlfrh`}],[`circle`,{cx:`12`,cy:`12`,r:`2`,key:`1c9p78`}]]),ti=Q(`trash-2`,[[`path`,{d:`M10 11v6`,key:`nco0om`}],[`path`,{d:`M14 11v6`,key:`outv1u`}],[`path`,{d:`M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6`,key:`miytrc`}],[`path`,{d:`M3 6h18`,key:`d0wm0j`}],[`path`,{d:`M8 6V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2`,key:`e791ji`}]]),ni=Q(`trophy`,[[`path`,{d:`M10 14.66v1.626a2 2 0 0 1-.976 1.696A5 5 0 0 0 7 21.978`,key:`1n3hpd`}],[`path`,{d:`M14 14.66v1.626a2 2 0 0 0 .976 1.696A5 5 0 0 1 17 21.978`,key:`rfe1zi`}],[`path`,{d:`M18 9h1.5a1 1 0 0 0 0-5H18`,key:`7xy6bh`}],[`path`,{d:`M4 22h16`,key:`57wxv0`}],[`path`,{d:`M6 9a6 6 0 0 0 12 0V3a1 1 0 0 0-1-1H7a1 1 0 0 0-1 1z`,key:`1mhfuq`}],[`path`,{d:`M6 9H4.5a1 1 0 0 1 0-5H6`,key:`tex48p`}]]),ri=Q(`wifi`,[[`path`,{d:`M12 20h.01`,key:`zekei9`}],[`path`,{d:`M2 8.82a15 15 0 0 1 20 0`,key:`dnpr2z`}],[`path`,{d:`M5 12.859a10 10 0 0 1 14 0`,key:`1x1e6c`}],[`path`,{d:`M8.5 16.429a5 5 0 0 1 7 0`,key:`1bycff`}]]),ii=Q(`x`,[[`path`,{d:`M18 6 6 18`,key:`1bl5f8`}],[`path`,{d:`m6 6 12 12`,key:`d8bk6v`}]]);function ai(e){let t=e.indexOf(`?`);return t>=0?e.slice(0,t):e}function oi(e,t){let n=ai(e);return t===`/`?n===`/`||n===``:n===t||n.startsWith(`${t}/`)}var si=[{href:`/`,label:`Accueil`},{href:`/collections`,label:`Collection`},{href:`/questions`,label:`Question`},{href:`/dashboard`,label:`Dashboard`},{href:`/database`,label:`Export/Import`}],ci={header:`sticky top-0 z-40 border-b border-base-content/5 bg-base-100/80 backdrop-blur-lg transition-shadow duration-300`,container:`mx-auto flex max-w-4xl flex-col gap-3 px-4 py-3 sm:flex-row sm:items-center sm:justify-between sm:gap-4`,brandButton:`flex items-center gap-2.5 rounded-full px-1 py-1 text-left transition duration-300 ease-out hover:opacity-90 active:scale-[0.98]`,brandIcon:`flex h-11 w-11 shrink-0 items-center justify-center rounded-full bg-linear-to-br from-flow to-learn text-white shadow-lg shadow-flow/20`,nav:`flex flex-wrap items-center justify-center gap-1.5 sm:justify-end`,navLinkBase:`rounded-full px-4 py-2 text-sm font-medium transition-all duration-300 ease-out active:scale-[0.97]`,navLinkActive:`bg-flow text-white shadow-xl shadow-flow/25 hover:scale-[1.03] hover:brightness-110 hover:shadow-xl hover:shadow-flow/35`,navLinkIdle:`text-base-content/70 hover:bg-learn/14 hover:text-learn hover:scale-[1.02]`};function li(){let e=_r();return X(`header`,{class:ci.header,children:X(`div`,{class:ci.container,children:[X(`button`,{type:`button`,onClick:()=>U(`/`),class:ci.brandButton,children:[X(`span`,{class:ci.brandIcon,children:X(Gr,{class:`h-5 w-5`,"aria-hidden":!0})}),X(`div`,{class:`leading-tight`,children:[X(`p`,{class:`text-lg font-semibold tracking-tight text-base-content`,children:`FlowLearn`}),X(`p`,{class:`text-xs text-base-content/50`,children:`Flow · Learn`})]})]}),X(`nav`,{class:ci.nav,"aria-label":`Navigation principale`,children:si.map(({href:t,label:n})=>{let r=oi(e,t);return X(`button`,{type:`button`,onClick:()=>U(t),class:ar(ci.navLinkBase,r?ci.navLinkActive:ci.navLinkIdle),children:n},t)})})]})})}var ui={footer:`mt-auto border-t border-base-content/5 bg-base-100/50 py-4 text-center text-xs text-base-content/55`,container:`mx-auto flex max-w-3xl flex-wrap items-center justify-center gap-x-4 gap-y-2 px-4`,separator:`hidden sm:inline text-base-content/25`,badge:`inline-flex items-center gap-1.5 rounded-full bg-flow/10 px-2.5 py-0.5 text-flow`};function di(){return X(`footer`,{class:ui.footer,children:X(`div`,{class:ui.container,children:[X(`span`,{children:[`FlowLearn v`,et]}),X(`span`,{class:ui.separator,children:`·`}),X(`span`,{class:ui.badge,children:[X(ri,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),`ESP32 (mock) — connecté`]})]})})}var fi={base:`fl-page-enter mx-auto w-full flex-1 px-4 py-6 md:py-8`,narrow:`max-w-2xl`,wide:`max-w-3xl`};function pi({children:e,class:t,narrow:n}){return X(`main`,{class:ar(fi.base,n?fi.narrow:fi.wide,t),children:e})}var mi={overlay:`fixed inset-0 z-[110] flex items-center justify-center p-4`,backdrop:`absolute inset-0 border-0 bg-base-content/45 p-0`,dialogBase:`relative z-10 w-full max-w-md rounded-2xl border bg-base-100 p-6 shadow-2xl`,dialogDanger:`border-error/25 shadow-error/10`,dialogInfo:`border-base-content/10`};function hi({open:e,title:t,message:n,variant:r=`info`,confirmLabel:i=`Confirmer`,cancelLabel:a=`Annuler`,busy:o=!1,onConfirm:s,onCancel:c}){if(!e)return null;let l=r===`danger`;return X(`div`,{class:mi.overlay,role:`presentation`,children:[X(`button`,{type:`button`,class:mi.backdrop,"aria-label":`Fermer`,disabled:o,onClick:()=>!o&&c()}),X(`div`,{class:`${mi.dialogBase} ${l?mi.dialogDanger:mi.dialogInfo}`,role:`dialog`,"aria-modal":`true`,"aria-labelledby":`pop-up-information-title`,onClick:e=>e.stopPropagation(),children:[X(`h2`,{id:`pop-up-information-title`,class:`mb-3 text-lg font-semibold tracking-tight ${l?`text-error`:`text-base-content`}`,children:t}),X(`p`,{class:`mb-6 whitespace-pre-line text-sm leading-relaxed text-base-content/80`,children:n}),X(`div`,{class:`flex flex-wrap justify-end gap-2`,children:[X(Z,{variant:`ghost`,class:`btn-sm`,disabled:o,onClick:c,children:a}),X(Z,{variant:l?`outline`:`learn`,class:l?`btn-sm border-2 border-error/50 text-error hover:border-error hover:bg-error/10 shadow-none`:`btn-sm`,disabled:o,onClick:s,children:o?`…`:i})]})]})]})}function gi(e){if(e==null)return null;if(e.kind===`collection`){let t=e.data,n=t.questions.length;return{title:`Supprimer la collection « ${t.nom} » ?`,message:`Cette action est definitive :
+(function(){let e=document.createElement(`link`).relList;if(e&&e.supports&&e.supports(`modulepreload`))return;for(let e of document.querySelectorAll(`link[rel="modulepreload"]`))n(e);new MutationObserver(e=>{for(let t of e)if(t.type===`childList`)for(let e of t.addedNodes)e.tagName===`LINK`&&e.rel===`modulepreload`&&n(e)}).observe(document,{childList:!0,subtree:!0});function t(e){let t={};return e.integrity&&(t.integrity=e.integrity),e.referrerPolicy&&(t.referrerPolicy=e.referrerPolicy),e.crossOrigin===`use-credentials`?t.credentials=`include`:e.crossOrigin===`anonymous`?t.credentials=`omit`:t.credentials=`same-origin`,t}function n(e){if(e.ep)return;e.ep=!0;let n=t(e);fetch(e.href,n)}})();var e,t,n,r,i,a,o,s,c,l,u,d,f={},p=[],m=/acit|ex(?:s|g|n|p|$)|rph|grid|ows|mnc|ntw|ine[ch]|zoo|^ord|itera/i,h=Array.isArray;function g(e,t){for(var n in t)e[n]=t[n];return e}function _(e){e&&e.parentNode&&e.parentNode.removeChild(e)}function v(t,n,r){var i,a,o,s={};for(o in n)o==`key`?i=n[o]:o==`ref`?a=n[o]:s[o]=n[o];if(arguments.length>2&&(s.children=arguments.length>3?e.call(arguments,2):r),typeof t==`function`&&t.defaultProps!=null)for(o in t.defaultProps)s[o]===void 0&&(s[o]=t.defaultProps[o]);return y(t,s,i,a,null)}function y(e,r,i,a,o){var s={type:e,props:r,key:i,ref:a,__k:null,__:null,__b:0,__e:null,__c:null,constructor:void 0,__v:o??++n,__i:-1,__u:0};return o==null&&t.vnode!=null&&t.vnode(s),s}function b(e){return e.children}function x(e,t){this.props=e,this.context=t}function S(e,t){if(t==null)return e.__?S(e.__,e.__i+1):null;for(var n;t<e.__k.length;t++)if((n=e.__k[t])!=null&&n.__e!=null)return n.__e;return typeof e.type==`function`?S(e):null}function C(e){if(e.__P&&e.__d){var n=e.__v,r=n.__e,i=[],a=[],o=g({},n);o.__v=n.__v+1,t.vnode&&t.vnode(o),ne(e.__P,o,n,e.__n,e.__P.namespaceURI,32&n.__u?[r]:null,i,r??S(n),!!(32&n.__u),a),o.__v=n.__v,o.__.__k[o.__i]=o,re(i,o,a),n.__e=n.__=null,o.__e!=r&&w(o)}}function w(e){if((e=e.__)!=null&&e.__c!=null)return e.__e=e.__c.base=null,e.__k.some(function(t){if(t!=null&&t.__e!=null)return e.__e=e.__c.base=t.__e}),w(e)}function T(e){(!e.__d&&(e.__d=!0)&&r.push(e)&&!E.__r++||i!=t.debounceRendering)&&((i=t.debounceRendering)||a)(E)}function E(){try{for(var e,t=1;r.length;)r.length>t&&r.sort(o),e=r.shift(),t=r.length,C(e)}finally{r.length=E.__r=0}}function ee(e,t,n,r,i,a,o,s,c,l,u){var d,m,h,g,_,v,y,b=r&&r.__k||p,x=t.length;for(c=D(n,t,b,c,x),d=0;d<x;d++)(h=n.__k[d])!=null&&(m=h.__i!=-1&&b[h.__i]||f,h.__i=d,v=ne(e,h,m,i,a,o,s,c,l,u),g=h.__e,h.ref&&m.ref!=h.ref&&(m.ref&&P(m.ref,null,h),u.push(h.ref,h.__c||g,h)),_==null&&g!=null&&(_=g),(y=!!(4&h.__u))||m.__k===h.__k?c=O(h,c,e,y):typeof h.type==`function`&&v!==void 0?c=v:g&&(c=g.nextSibling),h.__u&=-7);return n.__e=_,c}function D(e,t,n,r,i){var a,o,s,c,l,u=n.length,d=u,f=0;for(e.__k=Array(i),a=0;a<i;a++)(o=t[a])!=null&&typeof o!=`boolean`&&typeof o!=`function`?(typeof o==`string`||typeof o==`number`||typeof o==`bigint`||o.constructor==String?o=e.__k[a]=y(null,o,null,null,null):h(o)?o=e.__k[a]=y(b,{children:o},null,null,null):o.constructor===void 0&&o.__b>0?o=e.__k[a]=y(o.type,o.props,o.key,o.ref?o.ref:null,o.__v):e.__k[a]=o,c=a+f,o.__=e,o.__b=e.__b+1,s=null,(l=o.__i=A(o,n,c,d))!=-1&&(d--,(s=n[l])&&(s.__u|=2)),s==null||s.__v==null?(l==-1&&(i>u?f--:i<u&&f++),typeof o.type!=`function`&&(o.__u|=4)):l!=c&&(l==c-1?f--:l==c+1?f++:(l>c?f--:f++,o.__u|=4))):e.__k[a]=null;if(d)for(a=0;a<u;a++)(s=n[a])!=null&&!(2&s.__u)&&(s.__e==r&&(r=S(s)),F(s,s));return r}function O(e,t,n,r){var i,a;if(typeof e.type==`function`){for(i=e.__k,a=0;i&&a<i.length;a++)i[a]&&(i[a].__=e,t=O(i[a],t,n,r));return t}e.__e!=t&&(r&&(t&&e.type&&!t.parentNode&&(t=S(e)),n.insertBefore(e.__e,t||null)),t=e.__e);do t&&=t.nextSibling;while(t!=null&&t.nodeType==8);return t}function k(e,t){return t||=[],e==null||typeof e==`boolean`||(h(e)?e.some(function(e){k(e,t)}):t.push(e)),t}function A(e,t,n,r){var i,a,o,s=e.key,c=e.type,l=t[n],u=l!=null&&(2&l.__u)==0;if(l===null&&s==null||u&&s==l.key&&c==l.type)return n;if(r>(u?1:0)){for(i=n-1,a=n+1;i>=0||a<t.length;)if((l=t[o=i>=0?i--:a++])!=null&&!(2&l.__u)&&s==l.key&&c==l.type)return o}return-1}function j(e,t,n){t[0]==`-`?e.setProperty(t,n??``):e[t]=n==null?``:typeof n!=`number`||m.test(t)?n:n+`px`}function M(e,t,n,r,i){var a,o;n:if(t==`style`)if(typeof n==`string`)e.style.cssText=n;else{if(typeof r==`string`&&(e.style.cssText=r=``),r)for(t in r)n&&t in n||j(e.style,t,``);if(n)for(t in n)r&&n[t]==r[t]||j(e.style,t,n[t])}else if(t[0]==`o`&&t[1]==`n`)a=t!=(t=t.replace(s,`$1`)),o=t.toLowerCase(),t=o in e||t==`onFocusOut`||t==`onFocusIn`?o.slice(2):t.slice(2),e.l||={},e.l[t+a]=n,n?r?n.u=r.u:(n.u=c,e.addEventListener(t,a?u:l,a)):e.removeEventListener(t,a?u:l,a);else{if(i==`http://www.w3.org/2000/svg`)t=t.replace(/xlink(H|:h)/,`h`).replace(/sName$/,`s`);else if(t!=`width`&&t!=`height`&&t!=`href`&&t!=`list`&&t!=`form`&&t!=`tabIndex`&&t!=`download`&&t!=`rowSpan`&&t!=`colSpan`&&t!=`role`&&t!=`popover`&&t in e)try{e[t]=n??``;break n}catch{}typeof n==`function`||(n==null||!1===n&&t[4]!=`-`?e.removeAttribute(t):e.setAttribute(t,t==`popover`&&n==1?``:n))}}function te(e){return function(n){if(this.l){var r=this.l[n.type+e];if(n.t==null)n.t=c++;else if(n.t<r.u)return;return r(t.event?t.event(n):n)}}}function ne(e,n,r,i,a,o,s,c,l,u){var d,f,m,v,y,S,C,w,T,E,D,O,k,A,j,M=n.type;if(n.constructor!==void 0)return null;128&r.__u&&(l=!!(32&r.__u),o=[c=n.__e=r.__e]),(d=t.__b)&&d(n);n:if(typeof M==`function`)try{if(w=n.props,T=M.prototype&&M.prototype.render,E=(d=M.contextType)&&i[d.__c],D=d?E?E.props.value:d.__:i,r.__c?C=(f=n.__c=r.__c).__=f.__E:(T?n.__c=f=new M(w,D):(n.__c=f=new x(w,D),f.constructor=M,f.render=I),E&&E.sub(f),f.state||={},f.__n=i,m=f.__d=!0,f.__h=[],f._sb=[]),T&&f.__s==null&&(f.__s=f.state),T&&M.getDerivedStateFromProps!=null&&(f.__s==f.state&&(f.__s=g({},f.__s)),g(f.__s,M.getDerivedStateFromProps(w,f.__s))),v=f.props,y=f.state,f.__v=n,m)T&&M.getDerivedStateFromProps==null&&f.componentWillMount!=null&&f.componentWillMount(),T&&f.componentDidMount!=null&&f.__h.push(f.componentDidMount);else{if(T&&M.getDerivedStateFromProps==null&&w!==v&&f.componentWillReceiveProps!=null&&f.componentWillReceiveProps(w,D),n.__v==r.__v||!f.__e&&f.shouldComponentUpdate!=null&&!1===f.shouldComponentUpdate(w,f.__s,D)){n.__v!=r.__v&&(f.props=w,f.state=f.__s,f.__d=!1),n.__e=r.__e,n.__k=r.__k,n.__k.some(function(e){e&&(e.__=n)}),p.push.apply(f.__h,f._sb),f._sb=[],f.__h.length&&s.push(f);break n}f.componentWillUpdate!=null&&f.componentWillUpdate(w,f.__s,D),T&&f.componentDidUpdate!=null&&f.__h.push(function(){f.componentDidUpdate(v,y,S)})}if(f.context=D,f.props=w,f.__P=e,f.__e=!1,O=t.__r,k=0,T)f.state=f.__s,f.__d=!1,O&&O(n),d=f.render(f.props,f.state,f.context),p.push.apply(f.__h,f._sb),f._sb=[];else do f.__d=!1,O&&O(n),d=f.render(f.props,f.state,f.context),f.state=f.__s;while(f.__d&&++k<25);f.state=f.__s,f.getChildContext!=null&&(i=g(g({},i),f.getChildContext())),T&&!m&&f.getSnapshotBeforeUpdate!=null&&(S=f.getSnapshotBeforeUpdate(v,y)),A=d!=null&&d.type===b&&d.key==null?ie(d.props.children):d,c=ee(e,h(A)?A:[A],n,r,i,a,o,s,c,l,u),f.base=n.__e,n.__u&=-161,f.__h.length&&s.push(f),C&&(f.__E=f.__=null)}catch(e){if(n.__v=null,l||o!=null)if(e.then){for(n.__u|=l?160:128;c&&c.nodeType==8&&c.nextSibling;)c=c.nextSibling;o[o.indexOf(c)]=null,n.__e=c}else{for(j=o.length;j--;)_(o[j]);N(n)}else n.__e=r.__e,n.__k=r.__k,e.then||N(n);t.__e(e,n,r)}else o==null&&n.__v==r.__v?(n.__k=r.__k,n.__e=r.__e):c=n.__e=ae(r.__e,n,r,i,a,o,s,l,u);return(d=t.diffed)&&d(n),128&n.__u?void 0:c}function N(e){e&&(e.__c&&(e.__c.__e=!0),e.__k&&e.__k.some(N))}function re(e,n,r){for(var i=0;i<r.length;i++)P(r[i],r[++i],r[++i]);t.__c&&t.__c(n,e),e.some(function(n){try{e=n.__h,n.__h=[],e.some(function(e){e.call(n)})}catch(e){t.__e(e,n.__v)}})}function ie(e){return typeof e!=`object`||!e||e.__b>0?e:h(e)?e.map(ie):g({},e)}function ae(n,r,i,a,o,s,c,l,u){var d,p,m,g,v,y,b,x=i.props||f,C=r.props,w=r.type;if(w==`svg`?o=`http://www.w3.org/2000/svg`:w==`math`?o=`http://www.w3.org/1998/Math/MathML`:o||=`http://www.w3.org/1999/xhtml`,s!=null){for(d=0;d<s.length;d++)if((v=s[d])&&`setAttribute`in v==!!w&&(w?v.localName==w:v.nodeType==3)){n=v,s[d]=null;break}}if(n==null){if(w==null)return document.createTextNode(C);n=document.createElementNS(o,w,C.is&&C),l&&=(t.__m&&t.__m(r,s),!1),s=null}if(w==null)x===C||l&&n.data==C||(n.data=C);else{if(s&&=e.call(n.childNodes),!l&&s!=null)for(x={},d=0;d<n.attributes.length;d++)x[(v=n.attributes[d]).name]=v.value;for(d in x)v=x[d],d==`dangerouslySetInnerHTML`?m=v:d==`children`||d in C||d==`value`&&`defaultValue`in C||d==`checked`&&`defaultChecked`in C||M(n,d,null,v,o);for(d in C)v=C[d],d==`children`?g=v:d==`dangerouslySetInnerHTML`?p=v:d==`value`?y=v:d==`checked`?b=v:l&&typeof v!=`function`||x[d]===v||M(n,d,v,x[d],o);if(p)l||m&&(p.__html==m.__html||p.__html==n.innerHTML)||(n.innerHTML=p.__html),r.__k=[];else if(m&&(n.innerHTML=``),ee(r.type==`template`?n.content:n,h(g)?g:[g],r,i,a,w==`foreignObject`?`http://www.w3.org/1999/xhtml`:o,s,c,s?s[0]:i.__k&&S(i,0),l,u),s!=null)for(d=s.length;d--;)_(s[d]);l||(d=`value`,w==`progress`&&y==null?n.removeAttribute(`value`):y!=null&&(y!==n[d]||w==`progress`&&!y||w==`option`&&y!=x[d])&&M(n,d,y,x[d],o),d=`checked`,b!=null&&b!=n[d]&&M(n,d,b,x[d],o))}return n}function P(e,n,r){try{if(typeof e==`function`){var i=typeof e.__u==`function`;i&&e.__u(),i&&n==null||(e.__u=e(n))}else e.current=n}catch(e){t.__e(e,r)}}function F(e,n,r){var i,a;if(t.unmount&&t.unmount(e),(i=e.ref)&&(i.current&&i.current!=e.__e||P(i,null,n)),(i=e.__c)!=null){if(i.componentWillUnmount)try{i.componentWillUnmount()}catch(e){t.__e(e,n)}i.base=i.__P=null}if(i=e.__k)for(a=0;a<i.length;a++)i[a]&&F(i[a],n,r||typeof e.type!=`function`);r||_(e.__e),e.__c=e.__=e.__e=void 0}function I(e,t,n){return this.constructor(e,n)}function oe(n,r,i){var a,o,s,c;r==document&&(r=document.documentElement),t.__&&t.__(n,r),o=(a=typeof i==`function`)?null:i&&i.__k||r.__k,s=[],c=[],ne(r,n=(!a&&i||r).__k=v(b,null,[n]),o||f,f,r.namespaceURI,!a&&i?[i]:o?null:r.firstChild?e.call(r.childNodes):null,s,!a&&i?i:o?o.__e:r.firstChild,a,c),re(s,n,c)}function se(t,n,r){var i,a,o,s,c=g({},t.props);for(o in t.type&&t.type.defaultProps&&(s=t.type.defaultProps),n)o==`key`?i=n[o]:o==`ref`?a=n[o]:c[o]=n[o]===void 0&&s!=null?s[o]:n[o];return arguments.length>2&&(c.children=arguments.length>3?e.call(arguments,2):r),y(t.type,c,i||t.key,a||t.ref,null)}function L(e){function t(e){var n,r;return this.getChildContext||(n=new Set,(r={})[t.__c]=this,this.getChildContext=function(){return r},this.componentWillUnmount=function(){n=null},this.shouldComponentUpdate=function(e){this.props.value!=e.value&&n.forEach(function(e){e.__e=!0,T(e)})},this.sub=function(e){n.add(e);var t=e.componentWillUnmount;e.componentWillUnmount=function(){n&&n.delete(e),t&&t.call(e)}}),e.children}return t.__c=`__cC`+ d++,t.__=e,t.Provider=t.__l=(t.Consumer=function(e,t){return e.children(t)}).contextType=t,t}e=p.slice,t={__e:function(e,t,n,r){for(var i,a,o;t=t.__;)if((i=t.__c)&&!i.__)try{if((a=i.constructor)&&a.getDerivedStateFromError!=null&&(i.setState(a.getDerivedStateFromError(e)),o=i.__d),i.componentDidCatch!=null&&(i.componentDidCatch(e,r||{}),o=i.__d),o)return i.__E=i}catch(t){e=t}throw e}},n=0,x.prototype.setState=function(e,t){var n=this.__s!=null&&this.__s!=this.state?this.__s:this.__s=g({},this.state);typeof e==`function`&&(e=e(g({},n),this.props)),e&&g(n,e),e!=null&&this.__v&&(t&&this._sb.push(t),T(this))},x.prototype.forceUpdate=function(e){this.__v&&(this.__e=!0,e&&this.__h.push(e),T(this))},x.prototype.render=b,r=[],a=typeof Promise==`function`?Promise.prototype.then.bind(Promise.resolve()):setTimeout,o=function(e,t){return e.__v.__b-t.__v.__b},E.__r=0,s=/(PointerCapture)$|Capture$/i,c=0,l=te(!1),u=te(!0),d=0;var R,z,ce,le,ue=0,de=[],B=t,fe=B.__b,pe=B.__r,me=B.diffed,he=B.__c,ge=B.unmount,_e=B.__;function ve(e,t){B.__h&&B.__h(z,e,ue||t),ue=0;var n=z.__H||={__:[],__h:[]};return e>=n.__.length&&n.__.push({}),n.__[e]}function V(e){return ue=1,ye(Me,e)}function ye(e,t,n){var r=ve(R++,2);if(r.t=e,!r.__c&&(r.__=[n?n(t):Me(void 0,t),function(e){var t=r.__N?r.__N[0]:r.__[0],n=r.t(t,e);t!==n&&(r.__N=[n,r.__[1]],r.__c.setState({}))}],r.__c=z,!z.__f)){var i=function(e,t,n){if(!r.__c.__H)return!0;var i=r.__c.__H.__.filter(function(e){return e.__c});if(i.every(function(e){return!e.__N}))return!a||a.call(this,e,t,n);var o=r.__c.props!==e;return i.some(function(e){if(e.__N){var t=e.__[0];e.__=e.__N,e.__N=void 0,t!==e.__[0]&&(o=!0)}}),a&&a.call(this,e,t,n)||o};z.__f=!0;var a=z.shouldComponentUpdate,o=z.componentWillUpdate;z.componentWillUpdate=function(e,t,n){if(this.__e){var r=a;a=void 0,i(e,t,n),a=r}o&&o.call(this,e,t,n)},z.shouldComponentUpdate=i}return r.__N||r.__}function H(e,t){var n=ve(R++,3);!B.__s&&je(n.__H,t)&&(n.__=e,n.u=t,z.__H.__h.push(n))}function be(e,t){var n=ve(R++,4);!B.__s&&je(n.__H,t)&&(n.__=e,n.u=t,z.__h.push(n))}function xe(e){return ue=5,Se(function(){return{current:e}},[])}function Se(e,t){var n=ve(R++,7);return je(n.__H,t)&&(n.__=e(),n.__H=t,n.__h=e),n.__}function Ce(e,t){return ue=8,Se(function(){return e},t)}function we(e){var t=z.context[e.__c],n=ve(R++,9);return n.c=e,t?(n.__??(n.__=!0,t.sub(z)),t.props.value):e.__}function Te(){var e=ve(R++,11);if(!e.__){for(var t=z.__v;t!==null&&!t.__m&&t.__!==null;)t=t.__;var n=t.__m||=[0,0];e.__=`P`+n[0]+`-`+ n[1]++}return e.__}function Ee(){for(var e;e=de.shift();){var t=e.__H;if(e.__P&&t)try{t.__h.some(ke),t.__h.some(Ae),t.__h=[]}catch(n){t.__h=[],B.__e(n,e.__v)}}}B.__b=function(e){z=null,fe&&fe(e)},B.__=function(e,t){e&&t.__k&&t.__k.__m&&(e.__m=t.__k.__m),_e&&_e(e,t)},B.__r=function(e){pe&&pe(e),R=0;var t=(z=e.__c).__H;t&&(ce===z?(t.__h=[],z.__h=[],t.__.some(function(e){e.__N&&(e.__=e.__N),e.u=e.__N=void 0})):(t.__h.some(ke),t.__h.some(Ae),t.__h=[],R=0)),ce=z},B.diffed=function(e){me&&me(e);var t=e.__c;t&&t.__H&&(t.__H.__h.length&&(de.push(t)!==1&&le===B.requestAnimationFrame||((le=B.requestAnimationFrame)||Oe)(Ee)),t.__H.__.some(function(e){e.u&&(e.__H=e.u),e.u=void 0})),ce=z=null},B.__c=function(e,t){t.some(function(e){try{e.__h.some(ke),e.__h=e.__h.filter(function(e){return!e.__||Ae(e)})}catch(n){t.some(function(e){e.__h&&=[]}),t=[],B.__e(n,e.__v)}}),he&&he(e,t)},B.unmount=function(e){ge&&ge(e);var t,n=e.__c;n&&n.__H&&(n.__H.__.some(function(e){try{ke(e)}catch(e){t=e}}),n.__H=void 0,t&&B.__e(t,n.__v))};var De=typeof requestAnimationFrame==`function`;function Oe(e){var t,n=function(){clearTimeout(r),De&&cancelAnimationFrame(t),setTimeout(e)},r=setTimeout(n,35);De&&(t=requestAnimationFrame(n))}function ke(e){var t=z,n=e.__c;typeof n==`function`&&(e.__c=void 0,n()),z=t}function Ae(e){var t=z;e.__c=e.__(),z=t}function je(e,t){return!e||e.length!==t.length||t.some(function(t,n){return t!==e[n]})}function Me(e,t){return typeof t==`function`?t(e):t}var Ne={};function Pe(e,t){for(var n in t)e[n]=t[n];return e}function Fe(e,t,n){var r,i=/(?:\?([^#]*))?(#.*)?$/,a=e.match(i),o={};if(a&&a[1])for(var s=a[1].split(`&`),c=0;c<s.length;c++){var l=s[c].split(`=`);o[decodeURIComponent(l[0])]=decodeURIComponent(l.slice(1).join(`=`))}e=Re(e.replace(i,``)),t=Re(t||``);for(var u=Math.max(e.length,t.length),d=0;d<u;d++)if(t[d]&&t[d].charAt(0)===`:`){var f=t[d].replace(/(^:|[+*?]+$)/g,``),p=(t[d].match(/[+*?]+$/)||Ne)[0]||``,m=~p.indexOf(`+`),h=~p.indexOf(`*`),g=e[d]||``;if(!g&&!h&&(p.indexOf(`?`)<0||m)){r=!1;break}if(o[f]=decodeURIComponent(g),m||h){o[f]=e.slice(d).map(decodeURIComponent).join(`/`);break}}else if(t[d]!==e[d]){r=!1;break}return(!0===n.default||!1!==r)&&o}function Ie(e,t){return e.rank<t.rank?1:e.rank>t.rank?-1:e.index-t.index}function Le(e,t){return e.index=t,e.rank=function(e){return e.props.default?0:Re(e.props.path).map(ze).join(``)}(e),e.props}function Re(e){return e.replace(/(^\/+|\/+$)/g,``).split(`/`)}function ze(e){return e.charAt(0)==`:`?1+`*+?`.indexOf(e.charAt(e.length-1))||4:5}var Be={},Ve=[],He=[],Ue=null,We={url:Ke()},Ge=L(We);function Ke(){var e;return``+((e=Ue&&Ue.location?Ue.location:Ue&&Ue.getCurrentLocation?Ue.getCurrentLocation():typeof location<`u`?location:Be).pathname||``)+(e.search||``)}function U(e,t){return t===void 0&&(t=!1),typeof e!=`string`&&e.url&&(t=e.replace,e=e.url),function(e){for(var t=Ve.length;t--;)if(Ve[t].canRoute(e))return!0;return!1}(e)&&function(e,t){t===void 0&&(t=`push`),Ue&&Ue[t]?Ue[t](e):typeof history<`u`&&history[t+`State`]&&history[t+`State`](null,null,e)}(e,t?`replace`:`push`),qe(e)}function qe(e){for(var t=!1,n=0;n<Ve.length;n++)Ve[n].routeTo(e)&&(t=!0);return t}function Je(e){if(e&&e.getAttribute){var t=e.getAttribute(`href`),n=e.getAttribute(`target`);if(t&&t.match(/^\//g)&&(!n||n.match(/^_?self$/i)))return U(t)}}function Ye(e){return e.stopImmediatePropagation&&e.stopImmediatePropagation(),e.stopPropagation&&e.stopPropagation(),e.preventDefault(),!1}function Xe(e){if(!(e.ctrlKey||e.metaKey||e.altKey||e.shiftKey||e.button)){var t=e.target;do if(t.localName===`a`&&t.getAttribute(`href`)){if(t.hasAttribute(`data-native`)||t.hasAttribute(`native`))return;if(Je(t))return Ye(e)}while(t=t.parentNode)}}var Ze=!1;function Qe(e){e.history&&(Ue=e.history),this.state={url:e.url||Ke()}}Pe(Qe.prototype=new x,{shouldComponentUpdate:function(e){return!0!==e.static||e.url!==this.props.url||e.onChange!==this.props.onChange},canRoute:function(e){var t=k(this.props.children);return this.g(t,e)!==void 0},routeTo:function(e){this.setState({url:e});var t=this.canRoute(e);return this.p||this.forceUpdate(),t},componentWillMount:function(){this.p=!0},componentDidMount:function(){var e=this;Ze||(Ze=!0,Ue||addEventListener(`popstate`,function(){qe(Ke())}),addEventListener(`click`,Xe)),Ve.push(this),Ue&&(this.u=Ue.listen(function(t){var n=t.location||t;e.routeTo(``+(n.pathname||``)+(n.search||``))})),this.p=!1},componentWillUnmount:function(){typeof this.u==`function`&&this.u(),Ve.splice(Ve.indexOf(this),1)},componentWillUpdate:function(){this.p=!0},componentDidUpdate:function(){this.p=!1},g:function(e,t){e=e.filter(Le).sort(Ie);for(var n=0;n<e.length;n++){var r=e[n],i=Fe(t,r.props.path,r.props);if(i)return[r,i]}},render:function(e,t){var n,r,i=e.onChange,a=t.url,o=this.c,s=this.g(k(e.children),a);if(s&&(r=se(s[0],Pe(Pe({url:a,matches:n=s[1]},n),{key:void 0,ref:void 0}))),a!==(o&&o.url)){Pe(We,o=this.c={url:a,previous:o&&o.url,current:r,path:r?r.props.path:null,matches:n}),o.router=this,o.active=r?[r]:[];for(var c=He.length;c--;)He[c]({});typeof i==`function`&&i(o)}return v(Ge.Provider,{value:o},r)}});var $e=function(e){return v(e.component,e)},et=`DE:AD:BE:EF:00:01`;function W(e){let t=e.startsWith(`/`)?e:`/${e}`,n=(void 0)?.trim();return n?`${n.replace(/\/$/,``)}/api${t}`:`/api${t}`}var tt=`0.0.0`;async function nt(e){let t=await e.text();try{let e=JSON.parse(t);if(Array.isArray(e.message))return e.message.join(`, `);if(typeof e.message==`string`)return e.message}catch{}return t||e.statusText}async function G(e){if(!e.ok){let t=await nt(e),n=Error(`HTTP ${e.status}: ${t}`);throw n.status=e.status,n.body=t,n}}async function rt(e){let t=await fetch(W(`/devices/lookup?adresse_mac=${encodeURIComponent(e)}`));return await G(t),t.json()}async function it(e,t){let n=await fetch(W(`/devices/register`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify({adresse_mac:e,pseudot:t})});return await G(n),n.json()}async function at(){let e=await fetch(W(`/quizz/collections`));return await G(e),e.json()}async function ot(e,t){let n=new URLSearchParams;t?.qtype!=null&&t.qtype!==`melanger`&&n.set(`qtype`,t.qtype),t?.orders!=null&&t.orders.length>0&&n.set(`order`,t.orders.join(`,`)),t?.userId!=null&&n.set(`userId`,String(t.userId)),t?.infinite&&n.set(`infinite`,`1`),t?.excludeIds!=null&&t.excludeIds.length>0&&n.set(`exclude`,t.excludeIds.join(`,`));let r=n.size>0?`?${n.toString()}`:``,i=await fetch(W(`/quizz/collections/${e}${r}`));return await G(i),i.json()}async function st(){let e=await fetch(W(`/quizz/modules`));return await G(e),e.json()}async function ct(e){let t=await fetch(W(`/quizz/modules`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify({nom:e})});return await G(t),t.json()}async function lt(e){await G(await fetch(W(`/quizz/modules/${e}`),{method:`DELETE`}))}async function ut(e,t){let n=await fetch(W(`/quizz/collections/${e}/modules`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify({moduleId:t})});return await G(n),n.json()}async function dt(e,t){let n=await fetch(W(`/quizz/collections/${e}/modules/${t}`),{method:`DELETE`});return await G(n),n.json()}async function ft(e,t){let n=new URLSearchParams({userId:String(t)});await G(await fetch(W(`/quizz/collections/${e}?${n.toString()}`),{method:`DELETE`}))}async function pt(e){let t=new URLSearchParams;e?.orders!=null&&e.orders.length>0&&t.set(`order`,e.orders.join(`,`)),e?.qtype!=null&&e.qtype!==`melanger`&&t.set(`qtype`,e.qtype),e?.userId!=null&&t.set(`userId`,String(e.userId)),e?.infinite&&t.set(`infinite`,`1`),e?.excludeIds!=null&&e.excludeIds.length>0&&t.set(`exclude`,e.excludeIds.join(`,`));let n=t.size>0?`?${t.toString()}`:``,r=await fetch(W(`/quizz/random${n}`));return await G(r),r.json()}async function mt(){let e=await fetch(W(`/quizz/categories`));return await G(e),e.json()}async function ht(e){let t=await fetch(W(`/quizz/questions/${e}`));return await G(t),t.json()}async function gt(e){let t=e===`none`?`?collectionId=none`:e===void 0?``:`?collectionId=${e}`,n=await fetch(W(`/quizz/questions${t}`));return await G(n),n.json()}async function _t(e,t){let n=await fetch(W(`/quizz/questions/${e}`),{method:`PATCH`,headers:{"Content-Type":`application/json`},body:JSON.stringify(t)});return await G(n),n.json()}async function vt(e){let t=await fetch(W(`/quizz/questions`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify(e)});return await G(t),t.json()}async function yt(e,t){let n=await fetch(W(`/quizz/reponses/${e}`),{method:`PATCH`,headers:{"Content-Type":`application/json`},body:JSON.stringify(t)});return await G(n),n.json()}async function bt(e){let t=await fetch(W(`/quizz/collections`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify(e)});return await G(t),t.json()}async function xt(e,t){let n=new URLSearchParams;t?.collectionId!=null&&n.set(`collectionId`,String(t.collectionId)),t?.moduleId!=null&&n.set(`moduleId`,String(t.moduleId)),t?.categorie!=null&&n.set(`categorie`,t.categorie);let r=n.size>0?`?${n.toString()}`:``,i=await fetch(W(`/quizz/questions/import${r}`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify(e)});return await G(i),i.json()}async function St(e,t){let n=new URLSearchParams({collectionId:String(t.collectionId)}),r=await fetch(W(`/quizz/collections/questions/import-app?${n.toString()}`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify(e)});return await G(r),r.json()}async function Ct(e){await G(await fetch(W(`/quizz/questions/${e}`),{method:`DELETE`}))}async function wt(e){let t=await fetch(W(`/stats/kpis?userId=${e}`));return await G(t),t.json()}async function Tt(e){let t=await fetch(W(`/stats/kpi`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify({userId:e.userId,questionId:e.questionId,reponseId:e.reponseId,dureeSecondes:e.dureeSecondes})});return await G(t),t.json()}async function Et(e){let t=await fetch(W(`/stats/sessions?userId=${e}`));return await G(t),t.json()}async function Dt(e,t){let n=await fetch(W(`/stats/sessions/${encodeURIComponent(e)}?userId=${t}`));return n.status===404?null:(await G(n),n.json())}async function Ot(e,t){let n=await fetch(W(e));return await G(n),{blob:await n.blob(),filename:(n.headers.get(`content-disposition`)??``).match(/filename="([^"]+)"/i)?.[1]??t}}async function kt(){return Ot(`/admin/database/export.sql`,`quizz-export.sql`)}async function At(){return Ot(`/admin/database/export.json`,`quizz-export.json`)}async function jt(e){let t=await fetch(W(`/admin/database/import.json`),{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify(e)});return await G(t),t.json()}async function Mt(e){let t=await fetch(W(`/admin/database/import.sql`),{method:`POST`,headers:{"Content-Type":`text/plain; charset=utf-8`,"X-Quizz-Confirm":`REMPLACE_TOUT`},body:e});return await G(t),t.json()}var Nt=L(null);function Pt(){let e=we(Nt);if(e==null)throw Error(`useUserSession doit être utilisé sous une session active`);return e}function Ft(e){var t,n,r=``;if(typeof e==`string`||typeof e==`number`)r+=e;else if(typeof e==`object`)if(Array.isArray(e)){var i=e.length;for(t=0;t<i;t++)e[t]&&(n=Ft(e[t]))&&(r&&(r+=` `),r+=n)}else for(n in e)e[n]&&(r&&(r+=` `),r+=n);return r}function It(){for(var e,t,n=0,r=``,i=arguments.length;n<i;n++)(e=arguments[n])&&(t=Ft(e))&&(r&&(r+=` `),r+=t);return r}var Lt=(e,t)=>{let n=Array(e.length+t.length);for(let t=0;t<e.length;t++)n[t]=e[t];for(let r=0;r<t.length;r++)n[e.length+r]=t[r];return n},Rt=(e,t)=>({classGroupId:e,validator:t}),zt=(e=new Map,t=null,n)=>({nextPart:e,validators:t,classGroupId:n}),Bt=`-`,Vt=[],Ht=`arbitrary..`,Ut=e=>{let t=Kt(e),{conflictingClassGroups:n,conflictingClassGroupModifiers:r}=e;return{getClassGroupId:e=>{if(e.startsWith(`[`)&&e.endsWith(`]`))return Gt(e);let n=e.split(Bt);return Wt(n,n[0]===``&&n.length>1?1:0,t)},getConflictingClassGroupIds:(e,t)=>{if(t){let t=r[e],i=n[e];return t?i?Lt(i,t):t:i||Vt}return n[e]||Vt}}},Wt=(e,t,n)=>{if(e.length-t===0)return n.classGroupId;let r=e[t],i=n.nextPart.get(r);if(i){let n=Wt(e,t+1,i);if(n)return n}let a=n.validators;if(a===null)return;let o=t===0?e.join(Bt):e.slice(t).join(Bt),s=a.length;for(let e=0;e<s;e++){let t=a[e];if(t.validator(o))return t.classGroupId}},Gt=e=>e.slice(1,-1).indexOf(`:`)===-1?void 0:(()=>{let t=e.slice(1,-1),n=t.indexOf(`:`),r=t.slice(0,n);return r?Ht+r:void 0})(),Kt=e=>{let{theme:t,classGroups:n}=e;return qt(n,t)},qt=(e,t)=>{let n=zt();for(let r in e){let i=e[r];Jt(i,n,r,t)}return n},Jt=(e,t,n,r)=>{let i=e.length;for(let a=0;a<i;a++){let i=e[a];Yt(i,t,n,r)}},Yt=(e,t,n,r)=>{if(typeof e==`string`){Xt(e,t,n);return}if(typeof e==`function`){Zt(e,t,n,r);return}Qt(e,t,n,r)},Xt=(e,t,n)=>{let r=e===``?t:$t(t,e);r.classGroupId=n},Zt=(e,t,n,r)=>{if(en(e)){Jt(e(r),t,n,r);return}t.validators===null&&(t.validators=[]),t.validators.push(Rt(n,e))},Qt=(e,t,n,r)=>{let i=Object.entries(e),a=i.length;for(let e=0;e<a;e++){let[a,o]=i[e];Jt(o,$t(t,a),n,r)}},$t=(e,t)=>{let n=e,r=t.split(Bt),i=r.length;for(let e=0;e<i;e++){let t=r[e],i=n.nextPart.get(t);i||(i=zt(),n.nextPart.set(t,i)),n=i}return n},en=e=>`isThemeGetter`in e&&e.isThemeGetter===!0,tn=e=>{if(e<1)return{get:()=>void 0,set:()=>{}};let t=0,n=Object.create(null),r=Object.create(null),i=(i,a)=>{n[i]=a,t++,t>e&&(t=0,r=n,n=Object.create(null))};return{get(e){let t=n[e];if(t!==void 0)return t;if((t=r[e])!==void 0)return i(e,t),t},set(e,t){e in n?n[e]=t:i(e,t)}}},nn=`!`,rn=`:`,an=[],on=(e,t,n,r,i)=>({modifiers:e,hasImportantModifier:t,baseClassName:n,maybePostfixModifierPosition:r,isExternal:i}),sn=e=>{let{prefix:t,experimentalParseClassName:n}=e,r=e=>{let t=[],n=0,r=0,i=0,a,o=e.length;for(let s=0;s<o;s++){let o=e[s];if(n===0&&r===0){if(o===rn){t.push(e.slice(i,s)),i=s+1;continue}if(o===`/`){a=s;continue}}o===`[`?n++:o===`]`?n--:o===`(`?r++:o===`)`&&r--}let s=t.length===0?e:e.slice(i),c=s,l=!1;s.endsWith(nn)?(c=s.slice(0,-1),l=!0):s.startsWith(nn)&&(c=s.slice(1),l=!0);let u=a&&a>i?a-i:void 0;return on(t,l,c,u)};if(t){let e=t+rn,n=r;r=t=>t.startsWith(e)?n(t.slice(e.length)):on(an,!1,t,void 0,!0)}if(n){let e=r;r=t=>n({className:t,parseClassName:e})}return r},cn=e=>{let t=new Map;return e.orderSensitiveModifiers.forEach((e,n)=>{t.set(e,1e6+n)}),e=>{let n=[],r=[];for(let i=0;i<e.length;i++){let a=e[i],o=a[0]===`[`,s=t.has(a);o||s?(r.length>0&&(r.sort(),n.push(...r),r=[]),n.push(a)):r.push(a)}return r.length>0&&(r.sort(),n.push(...r)),n}},ln=e=>({cache:tn(e.cacheSize),parseClassName:sn(e),sortModifiers:cn(e),...Ut(e)}),un=/\s+/,dn=(e,t)=>{let{parseClassName:n,getClassGroupId:r,getConflictingClassGroupIds:i,sortModifiers:a}=t,o=[],s=e.trim().split(un),c=``;for(let e=s.length-1;e>=0;--e){let t=s[e],{isExternal:l,modifiers:u,hasImportantModifier:d,baseClassName:f,maybePostfixModifierPosition:p}=n(t);if(l){c=t+(c.length>0?` `+c:c);continue}let m=!!p,h=r(m?f.substring(0,p):f);if(!h){if(!m){c=t+(c.length>0?` `+c:c);continue}if(h=r(f),!h){c=t+(c.length>0?` `+c:c);continue}m=!1}let g=u.length===0?``:u.length===1?u[0]:a(u).join(`:`),_=d?g+nn:g,v=_+h;if(o.indexOf(v)>-1)continue;o.push(v);let y=i(h,m);for(let e=0;e<y.length;++e){let t=y[e];o.push(_+t)}c=t+(c.length>0?` `+c:c)}return c},fn=(...e)=>{let t=0,n,r,i=``;for(;t<e.length;)(n=e[t++])&&(r=pn(n))&&(i&&(i+=` `),i+=r);return i},pn=e=>{if(typeof e==`string`)return e;let t,n=``;for(let r=0;r<e.length;r++)e[r]&&(t=pn(e[r]))&&(n&&(n+=` `),n+=t);return n},mn=(e,...t)=>{let n,r,i,a,o=o=>(n=ln(t.reduce((e,t)=>t(e),e())),r=n.cache.get,i=n.cache.set,a=s,s(o)),s=e=>{let t=r(e);if(t)return t;let a=dn(e,n);return i(e,a),a};return a=o,(...e)=>a(fn(...e))},hn=[],K=e=>{let t=t=>t[e]||hn;return t.isThemeGetter=!0,t},gn=/^\[(?:(\w[\w-]*):)?(.+)\]$/i,_n=/^\((?:(\w[\w-]*):)?(.+)\)$/i,vn=/^\d+(?:\.\d+)?\/\d+(?:\.\d+)?$/,yn=/^(\d+(\.\d+)?)?(xs|sm|md|lg|xl)$/,bn=/\d+(%|px|r?em|[sdl]?v([hwib]|min|max)|pt|pc|in|cm|mm|cap|ch|ex|r?lh|cq(w|h|i|b|min|max))|\b(calc|min|max|clamp)\(.+\)|^0$/,xn=/^(rgba?|hsla?|hwb|(ok)?(lab|lch)|color-mix)\(.+\)$/,Sn=/^(inset_)?-?((\d+)?\.?(\d+)[a-z]+|0)_-?((\d+)?\.?(\d+)[a-z]+|0)/,Cn=/^(url|image|image-set|cross-fade|element|(repeating-)?(linear|radial|conic)-gradient)\(.+\)$/,wn=e=>vn.test(e),q=e=>!!e&&!Number.isNaN(Number(e)),Tn=e=>!!e&&Number.isInteger(Number(e)),En=e=>e.endsWith(`%`)&&q(e.slice(0,-1)),Dn=e=>yn.test(e),On=()=>!0,kn=e=>bn.test(e)&&!xn.test(e),An=()=>!1,jn=e=>Sn.test(e),Mn=e=>Cn.test(e),Nn=e=>!J(e)&&!Y(e),Pn=e=>Yn(e,$n,An),J=e=>gn.test(e),Fn=e=>Yn(e,er,kn),In=e=>Yn(e,tr,q),Ln=e=>Yn(e,rr,On),Rn=e=>Yn(e,nr,An),zn=e=>Yn(e,Zn,An),Bn=e=>Yn(e,Qn,Mn),Vn=e=>Yn(e,ir,jn),Y=e=>_n.test(e),Hn=e=>Xn(e,er),Un=e=>Xn(e,nr),Wn=e=>Xn(e,Zn),Gn=e=>Xn(e,$n),Kn=e=>Xn(e,Qn),qn=e=>Xn(e,ir,!0),Jn=e=>Xn(e,rr,!0),Yn=(e,t,n)=>{let r=gn.exec(e);return r?r[1]?t(r[1]):n(r[2]):!1},Xn=(e,t,n=!1)=>{let r=_n.exec(e);return r?r[1]?t(r[1]):n:!1},Zn=e=>e===`position`||e===`percentage`,Qn=e=>e===`image`||e===`url`,$n=e=>e===`length`||e===`size`||e===`bg-size`,er=e=>e===`length`,tr=e=>e===`number`,nr=e=>e===`family-name`,rr=e=>e===`number`||e===`weight`,ir=e=>e===`shadow`,ar=mn(()=>{let e=K(`color`),t=K(`font`),n=K(`text`),r=K(`font-weight`),i=K(`tracking`),a=K(`leading`),o=K(`breakpoint`),s=K(`container`),c=K(`spacing`),l=K(`radius`),u=K(`shadow`),d=K(`inset-shadow`),f=K(`text-shadow`),p=K(`drop-shadow`),m=K(`blur`),h=K(`perspective`),g=K(`aspect`),_=K(`ease`),v=K(`animate`),y=()=>[`auto`,`avoid`,`all`,`avoid-page`,`page`,`left`,`right`,`column`],b=()=>[`center`,`top`,`bottom`,`left`,`right`,`top-left`,`left-top`,`top-right`,`right-top`,`bottom-right`,`right-bottom`,`bottom-left`,`left-bottom`],x=()=>[...b(),Y,J],S=()=>[`auto`,`hidden`,`clip`,`visible`,`scroll`],C=()=>[`auto`,`contain`,`none`],w=()=>[Y,J,c],T=()=>[wn,`full`,`auto`,...w()],E=()=>[Tn,`none`,`subgrid`,Y,J],ee=()=>[`auto`,{span:[`full`,Tn,Y,J]},Tn,Y,J],D=()=>[Tn,`auto`,Y,J],O=()=>[`auto`,`min`,`max`,`fr`,Y,J],k=()=>[`start`,`end`,`center`,`between`,`around`,`evenly`,`stretch`,`baseline`,`center-safe`,`end-safe`],A=()=>[`start`,`end`,`center`,`stretch`,`center-safe`,`end-safe`],j=()=>[`auto`,...w()],M=()=>[wn,`auto`,`full`,`dvw`,`dvh`,`lvw`,`lvh`,`svw`,`svh`,`min`,`max`,`fit`,...w()],te=()=>[wn,`screen`,`full`,`dvw`,`lvw`,`svw`,`min`,`max`,`fit`,...w()],ne=()=>[wn,`screen`,`full`,`lh`,`dvh`,`lvh`,`svh`,`min`,`max`,`fit`,...w()],N=()=>[e,Y,J],re=()=>[...b(),Wn,zn,{position:[Y,J]}],ie=()=>[`no-repeat`,{repeat:[``,`x`,`y`,`space`,`round`]}],ae=()=>[`auto`,`cover`,`contain`,Gn,Pn,{size:[Y,J]}],P=()=>[En,Hn,Fn],F=()=>[``,`none`,`full`,l,Y,J],I=()=>[``,q,Hn,Fn],oe=()=>[`solid`,`dashed`,`dotted`,`double`],se=()=>[`normal`,`multiply`,`screen`,`overlay`,`darken`,`lighten`,`color-dodge`,`color-burn`,`hard-light`,`soft-light`,`difference`,`exclusion`,`hue`,`saturation`,`color`,`luminosity`],L=()=>[q,En,Wn,zn],R=()=>[``,`none`,m,Y,J],z=()=>[`none`,q,Y,J],ce=()=>[`none`,q,Y,J],le=()=>[q,Y,J],ue=()=>[wn,`full`,...w()];return{cacheSize:500,theme:{animate:[`spin`,`ping`,`pulse`,`bounce`],aspect:[`video`],blur:[Dn],breakpoint:[Dn],color:[On],container:[Dn],"drop-shadow":[Dn],ease:[`in`,`out`,`in-out`],font:[Nn],"font-weight":[`thin`,`extralight`,`light`,`normal`,`medium`,`semibold`,`bold`,`extrabold`,`black`],"inset-shadow":[Dn],leading:[`none`,`tight`,`snug`,`normal`,`relaxed`,`loose`],perspective:[`dramatic`,`near`,`normal`,`midrange`,`distant`,`none`],radius:[Dn],shadow:[Dn],spacing:[`px`,q],text:[Dn],"text-shadow":[Dn],tracking:[`tighter`,`tight`,`normal`,`wide`,`wider`,`widest`]},classGroups:{aspect:[{aspect:[`auto`,`square`,wn,J,Y,g]}],container:[`container`],columns:[{columns:[q,J,Y,s]}],"break-after":[{"break-after":y()}],"break-before":[{"break-before":y()}],"break-inside":[{"break-inside":[`auto`,`avoid`,`avoid-page`,`avoid-column`]}],"box-decoration":[{"box-decoration":[`slice`,`clone`]}],box:[{box:[`border`,`content`]}],display:[`block`,`inline-block`,`inline`,`flex`,`inline-flex`,`table`,`inline-table`,`table-caption`,`table-cell`,`table-column`,`table-column-group`,`table-footer-group`,`table-header-group`,`table-row-group`,`table-row`,`flow-root`,`grid`,`inline-grid`,`contents`,`list-item`,`hidden`],sr:[`sr-only`,`not-sr-only`],float:[{float:[`right`,`left`,`none`,`start`,`end`]}],clear:[{clear:[`left`,`right`,`both`,`none`,`start`,`end`]}],isolation:[`isolate`,`isolation-auto`],"object-fit":[{object:[`contain`,`cover`,`fill`,`none`,`scale-down`]}],"object-position":[{object:x()}],overflow:[{overflow:S()}],"overflow-x":[{"overflow-x":S()}],"overflow-y":[{"overflow-y":S()}],overscroll:[{overscroll:C()}],"overscroll-x":[{"overscroll-x":C()}],"overscroll-y":[{"overscroll-y":C()}],position:[`static`,`fixed`,`absolute`,`relative`,`sticky`],inset:[{inset:T()}],"inset-x":[{"inset-x":T()}],"inset-y":[{"inset-y":T()}],start:[{"inset-s":T(),start:T()}],end:[{"inset-e":T(),end:T()}],"inset-bs":[{"inset-bs":T()}],"inset-be":[{"inset-be":T()}],top:[{top:T()}],right:[{right:T()}],bottom:[{bottom:T()}],left:[{left:T()}],visibility:[`visible`,`invisible`,`collapse`],z:[{z:[Tn,`auto`,Y,J]}],basis:[{basis:[wn,`full`,`auto`,s,...w()]}],"flex-direction":[{flex:[`row`,`row-reverse`,`col`,`col-reverse`]}],"flex-wrap":[{flex:[`nowrap`,`wrap`,`wrap-reverse`]}],flex:[{flex:[q,wn,`auto`,`initial`,`none`,J]}],grow:[{grow:[``,q,Y,J]}],shrink:[{shrink:[``,q,Y,J]}],order:[{order:[Tn,`first`,`last`,`none`,Y,J]}],"grid-cols":[{"grid-cols":E()}],"col-start-end":[{col:ee()}],"col-start":[{"col-start":D()}],"col-end":[{"col-end":D()}],"grid-rows":[{"grid-rows":E()}],"row-start-end":[{row:ee()}],"row-start":[{"row-start":D()}],"row-end":[{"row-end":D()}],"grid-flow":[{"grid-flow":[`row`,`col`,`dense`,`row-dense`,`col-dense`]}],"auto-cols":[{"auto-cols":O()}],"auto-rows":[{"auto-rows":O()}],gap:[{gap:w()}],"gap-x":[{"gap-x":w()}],"gap-y":[{"gap-y":w()}],"justify-content":[{justify:[...k(),`normal`]}],"justify-items":[{"justify-items":[...A(),`normal`]}],"justify-self":[{"justify-self":[`auto`,...A()]}],"align-content":[{content:[`normal`,...k()]}],"align-items":[{items:[...A(),{baseline:[``,`last`]}]}],"align-self":[{self:[`auto`,...A(),{baseline:[``,`last`]}]}],"place-content":[{"place-content":k()}],"place-items":[{"place-items":[...A(),`baseline`]}],"place-self":[{"place-self":[`auto`,...A()]}],p:[{p:w()}],px:[{px:w()}],py:[{py:w()}],ps:[{ps:w()}],pe:[{pe:w()}],pbs:[{pbs:w()}],pbe:[{pbe:w()}],pt:[{pt:w()}],pr:[{pr:w()}],pb:[{pb:w()}],pl:[{pl:w()}],m:[{m:j()}],mx:[{mx:j()}],my:[{my:j()}],ms:[{ms:j()}],me:[{me:j()}],mbs:[{mbs:j()}],mbe:[{mbe:j()}],mt:[{mt:j()}],mr:[{mr:j()}],mb:[{mb:j()}],ml:[{ml:j()}],"space-x":[{"space-x":w()}],"space-x-reverse":[`space-x-reverse`],"space-y":[{"space-y":w()}],"space-y-reverse":[`space-y-reverse`],size:[{size:M()}],"inline-size":[{inline:[`auto`,...te()]}],"min-inline-size":[{"min-inline":[`auto`,...te()]}],"max-inline-size":[{"max-inline":[`none`,...te()]}],"block-size":[{block:[`auto`,...ne()]}],"min-block-size":[{"min-block":[`auto`,...ne()]}],"max-block-size":[{"max-block":[`none`,...ne()]}],w:[{w:[s,`screen`,...M()]}],"min-w":[{"min-w":[s,`screen`,`none`,...M()]}],"max-w":[{"max-w":[s,`screen`,`none`,`prose`,{screen:[o]},...M()]}],h:[{h:[`screen`,`lh`,...M()]}],"min-h":[{"min-h":[`screen`,`lh`,`none`,...M()]}],"max-h":[{"max-h":[`screen`,`lh`,...M()]}],"font-size":[{text:[`base`,n,Hn,Fn]}],"font-smoothing":[`antialiased`,`subpixel-antialiased`],"font-style":[`italic`,`not-italic`],"font-weight":[{font:[r,Jn,Ln]}],"font-stretch":[{"font-stretch":[`ultra-condensed`,`extra-condensed`,`condensed`,`semi-condensed`,`normal`,`semi-expanded`,`expanded`,`extra-expanded`,`ultra-expanded`,En,J]}],"font-family":[{font:[Un,Rn,t]}],"font-features":[{"font-features":[J]}],"fvn-normal":[`normal-nums`],"fvn-ordinal":[`ordinal`],"fvn-slashed-zero":[`slashed-zero`],"fvn-figure":[`lining-nums`,`oldstyle-nums`],"fvn-spacing":[`proportional-nums`,`tabular-nums`],"fvn-fraction":[`diagonal-fractions`,`stacked-fractions`],tracking:[{tracking:[i,Y,J]}],"line-clamp":[{"line-clamp":[q,`none`,Y,In]}],leading:[{leading:[a,...w()]}],"list-image":[{"list-image":[`none`,Y,J]}],"list-style-position":[{list:[`inside`,`outside`]}],"list-style-type":[{list:[`disc`,`decimal`,`none`,Y,J]}],"text-alignment":[{text:[`left`,`center`,`right`,`justify`,`start`,`end`]}],"placeholder-color":[{placeholder:N()}],"text-color":[{text:N()}],"text-decoration":[`underline`,`overline`,`line-through`,`no-underline`],"text-decoration-style":[{decoration:[...oe(),`wavy`]}],"text-decoration-thickness":[{decoration:[q,`from-font`,`auto`,Y,Fn]}],"text-decoration-color":[{decoration:N()}],"underline-offset":[{"underline-offset":[q,`auto`,Y,J]}],"text-transform":[`uppercase`,`lowercase`,`capitalize`,`normal-case`],"text-overflow":[`truncate`,`text-ellipsis`,`text-clip`],"text-wrap":[{text:[`wrap`,`nowrap`,`balance`,`pretty`]}],indent:[{indent:w()}],"vertical-align":[{align:[`baseline`,`top`,`middle`,`bottom`,`text-top`,`text-bottom`,`sub`,`super`,Y,J]}],whitespace:[{whitespace:[`normal`,`nowrap`,`pre`,`pre-line`,`pre-wrap`,`break-spaces`]}],break:[{break:[`normal`,`words`,`all`,`keep`]}],wrap:[{wrap:[`break-word`,`anywhere`,`normal`]}],hyphens:[{hyphens:[`none`,`manual`,`auto`]}],content:[{content:[`none`,Y,J]}],"bg-attachment":[{bg:[`fixed`,`local`,`scroll`]}],"bg-clip":[{"bg-clip":[`border`,`padding`,`content`,`text`]}],"bg-origin":[{"bg-origin":[`border`,`padding`,`content`]}],"bg-position":[{bg:re()}],"bg-repeat":[{bg:ie()}],"bg-size":[{bg:ae()}],"bg-image":[{bg:[`none`,{linear:[{to:[`t`,`tr`,`r`,`br`,`b`,`bl`,`l`,`tl`]},Tn,Y,J],radial:[``,Y,J],conic:[Tn,Y,J]},Kn,Bn]}],"bg-color":[{bg:N()}],"gradient-from-pos":[{from:P()}],"gradient-via-pos":[{via:P()}],"gradient-to-pos":[{to:P()}],"gradient-from":[{from:N()}],"gradient-via":[{via:N()}],"gradient-to":[{to:N()}],rounded:[{rounded:F()}],"rounded-s":[{"rounded-s":F()}],"rounded-e":[{"rounded-e":F()}],"rounded-t":[{"rounded-t":F()}],"rounded-r":[{"rounded-r":F()}],"rounded-b":[{"rounded-b":F()}],"rounded-l":[{"rounded-l":F()}],"rounded-ss":[{"rounded-ss":F()}],"rounded-se":[{"rounded-se":F()}],"rounded-ee":[{"rounded-ee":F()}],"rounded-es":[{"rounded-es":F()}],"rounded-tl":[{"rounded-tl":F()}],"rounded-tr":[{"rounded-tr":F()}],"rounded-br":[{"rounded-br":F()}],"rounded-bl":[{"rounded-bl":F()}],"border-w":[{border:I()}],"border-w-x":[{"border-x":I()}],"border-w-y":[{"border-y":I()}],"border-w-s":[{"border-s":I()}],"border-w-e":[{"border-e":I()}],"border-w-bs":[{"border-bs":I()}],"border-w-be":[{"border-be":I()}],"border-w-t":[{"border-t":I()}],"border-w-r":[{"border-r":I()}],"border-w-b":[{"border-b":I()}],"border-w-l":[{"border-l":I()}],"divide-x":[{"divide-x":I()}],"divide-x-reverse":[`divide-x-reverse`],"divide-y":[{"divide-y":I()}],"divide-y-reverse":[`divide-y-reverse`],"border-style":[{border:[...oe(),`hidden`,`none`]}],"divide-style":[{divide:[...oe(),`hidden`,`none`]}],"border-color":[{border:N()}],"border-color-x":[{"border-x":N()}],"border-color-y":[{"border-y":N()}],"border-color-s":[{"border-s":N()}],"border-color-e":[{"border-e":N()}],"border-color-bs":[{"border-bs":N()}],"border-color-be":[{"border-be":N()}],"border-color-t":[{"border-t":N()}],"border-color-r":[{"border-r":N()}],"border-color-b":[{"border-b":N()}],"border-color-l":[{"border-l":N()}],"divide-color":[{divide:N()}],"outline-style":[{outline:[...oe(),`none`,`hidden`]}],"outline-offset":[{"outline-offset":[q,Y,J]}],"outline-w":[{outline:[``,q,Hn,Fn]}],"outline-color":[{outline:N()}],shadow:[{shadow:[``,`none`,u,qn,Vn]}],"shadow-color":[{shadow:N()}],"inset-shadow":[{"inset-shadow":[`none`,d,qn,Vn]}],"inset-shadow-color":[{"inset-shadow":N()}],"ring-w":[{ring:I()}],"ring-w-inset":[`ring-inset`],"ring-color":[{ring:N()}],"ring-offset-w":[{"ring-offset":[q,Fn]}],"ring-offset-color":[{"ring-offset":N()}],"inset-ring-w":[{"inset-ring":I()}],"inset-ring-color":[{"inset-ring":N()}],"text-shadow":[{"text-shadow":[`none`,f,qn,Vn]}],"text-shadow-color":[{"text-shadow":N()}],opacity:[{opacity:[q,Y,J]}],"mix-blend":[{"mix-blend":[...se(),`plus-darker`,`plus-lighter`]}],"bg-blend":[{"bg-blend":se()}],"mask-clip":[{"mask-clip":[`border`,`padding`,`content`,`fill`,`stroke`,`view`]},`mask-no-clip`],"mask-composite":[{mask:[`add`,`subtract`,`intersect`,`exclude`]}],"mask-image-linear-pos":[{"mask-linear":[q]}],"mask-image-linear-from-pos":[{"mask-linear-from":L()}],"mask-image-linear-to-pos":[{"mask-linear-to":L()}],"mask-image-linear-from-color":[{"mask-linear-from":N()}],"mask-image-linear-to-color":[{"mask-linear-to":N()}],"mask-image-t-from-pos":[{"mask-t-from":L()}],"mask-image-t-to-pos":[{"mask-t-to":L()}],"mask-image-t-from-color":[{"mask-t-from":N()}],"mask-image-t-to-color":[{"mask-t-to":N()}],"mask-image-r-from-pos":[{"mask-r-from":L()}],"mask-image-r-to-pos":[{"mask-r-to":L()}],"mask-image-r-from-color":[{"mask-r-from":N()}],"mask-image-r-to-color":[{"mask-r-to":N()}],"mask-image-b-from-pos":[{"mask-b-from":L()}],"mask-image-b-to-pos":[{"mask-b-to":L()}],"mask-image-b-from-color":[{"mask-b-from":N()}],"mask-image-b-to-color":[{"mask-b-to":N()}],"mask-image-l-from-pos":[{"mask-l-from":L()}],"mask-image-l-to-pos":[{"mask-l-to":L()}],"mask-image-l-from-color":[{"mask-l-from":N()}],"mask-image-l-to-color":[{"mask-l-to":N()}],"mask-image-x-from-pos":[{"mask-x-from":L()}],"mask-image-x-to-pos":[{"mask-x-to":L()}],"mask-image-x-from-color":[{"mask-x-from":N()}],"mask-image-x-to-color":[{"mask-x-to":N()}],"mask-image-y-from-pos":[{"mask-y-from":L()}],"mask-image-y-to-pos":[{"mask-y-to":L()}],"mask-image-y-from-color":[{"mask-y-from":N()}],"mask-image-y-to-color":[{"mask-y-to":N()}],"mask-image-radial":[{"mask-radial":[Y,J]}],"mask-image-radial-from-pos":[{"mask-radial-from":L()}],"mask-image-radial-to-pos":[{"mask-radial-to":L()}],"mask-image-radial-from-color":[{"mask-radial-from":N()}],"mask-image-radial-to-color":[{"mask-radial-to":N()}],"mask-image-radial-shape":[{"mask-radial":[`circle`,`ellipse`]}],"mask-image-radial-size":[{"mask-radial":[{closest:[`side`,`corner`],farthest:[`side`,`corner`]}]}],"mask-image-radial-pos":[{"mask-radial-at":b()}],"mask-image-conic-pos":[{"mask-conic":[q]}],"mask-image-conic-from-pos":[{"mask-conic-from":L()}],"mask-image-conic-to-pos":[{"mask-conic-to":L()}],"mask-image-conic-from-color":[{"mask-conic-from":N()}],"mask-image-conic-to-color":[{"mask-conic-to":N()}],"mask-mode":[{mask:[`alpha`,`luminance`,`match`]}],"mask-origin":[{"mask-origin":[`border`,`padding`,`content`,`fill`,`stroke`,`view`]}],"mask-position":[{mask:re()}],"mask-repeat":[{mask:ie()}],"mask-size":[{mask:ae()}],"mask-type":[{"mask-type":[`alpha`,`luminance`]}],"mask-image":[{mask:[`none`,Y,J]}],filter:[{filter:[``,`none`,Y,J]}],blur:[{blur:R()}],brightness:[{brightness:[q,Y,J]}],contrast:[{contrast:[q,Y,J]}],"drop-shadow":[{"drop-shadow":[``,`none`,p,qn,Vn]}],"drop-shadow-color":[{"drop-shadow":N()}],grayscale:[{grayscale:[``,q,Y,J]}],"hue-rotate":[{"hue-rotate":[q,Y,J]}],invert:[{invert:[``,q,Y,J]}],saturate:[{saturate:[q,Y,J]}],sepia:[{sepia:[``,q,Y,J]}],"backdrop-filter":[{"backdrop-filter":[``,`none`,Y,J]}],"backdrop-blur":[{"backdrop-blur":R()}],"backdrop-brightness":[{"backdrop-brightness":[q,Y,J]}],"backdrop-contrast":[{"backdrop-contrast":[q,Y,J]}],"backdrop-grayscale":[{"backdrop-grayscale":[``,q,Y,J]}],"backdrop-hue-rotate":[{"backdrop-hue-rotate":[q,Y,J]}],"backdrop-invert":[{"backdrop-invert":[``,q,Y,J]}],"backdrop-opacity":[{"backdrop-opacity":[q,Y,J]}],"backdrop-saturate":[{"backdrop-saturate":[q,Y,J]}],"backdrop-sepia":[{"backdrop-sepia":[``,q,Y,J]}],"border-collapse":[{border:[`collapse`,`separate`]}],"border-spacing":[{"border-spacing":w()}],"border-spacing-x":[{"border-spacing-x":w()}],"border-spacing-y":[{"border-spacing-y":w()}],"table-layout":[{table:[`auto`,`fixed`]}],caption:[{caption:[`top`,`bottom`]}],transition:[{transition:[``,`all`,`colors`,`opacity`,`shadow`,`transform`,`none`,Y,J]}],"transition-behavior":[{transition:[`normal`,`discrete`]}],duration:[{duration:[q,`initial`,Y,J]}],ease:[{ease:[`linear`,`initial`,_,Y,J]}],delay:[{delay:[q,Y,J]}],animate:[{animate:[`none`,v,Y,J]}],backface:[{backface:[`hidden`,`visible`]}],perspective:[{perspective:[h,Y,J]}],"perspective-origin":[{"perspective-origin":x()}],rotate:[{rotate:z()}],"rotate-x":[{"rotate-x":z()}],"rotate-y":[{"rotate-y":z()}],"rotate-z":[{"rotate-z":z()}],scale:[{scale:ce()}],"scale-x":[{"scale-x":ce()}],"scale-y":[{"scale-y":ce()}],"scale-z":[{"scale-z":ce()}],"scale-3d":[`scale-3d`],skew:[{skew:le()}],"skew-x":[{"skew-x":le()}],"skew-y":[{"skew-y":le()}],transform:[{transform:[Y,J,``,`none`,`gpu`,`cpu`]}],"transform-origin":[{origin:x()}],"transform-style":[{transform:[`3d`,`flat`]}],translate:[{translate:ue()}],"translate-x":[{"translate-x":ue()}],"translate-y":[{"translate-y":ue()}],"translate-z":[{"translate-z":ue()}],"translate-none":[`translate-none`],accent:[{accent:N()}],appearance:[{appearance:[`none`,`auto`]}],"caret-color":[{caret:N()}],"color-scheme":[{scheme:[`normal`,`dark`,`light`,`light-dark`,`only-dark`,`only-light`]}],cursor:[{cursor:[`auto`,`default`,`pointer`,`wait`,`text`,`move`,`help`,`not-allowed`,`none`,`context-menu`,`progress`,`cell`,`crosshair`,`vertical-text`,`alias`,`copy`,`no-drop`,`grab`,`grabbing`,`all-scroll`,`col-resize`,`row-resize`,`n-resize`,`e-resize`,`s-resize`,`w-resize`,`ne-resize`,`nw-resize`,`se-resize`,`sw-resize`,`ew-resize`,`ns-resize`,`nesw-resize`,`nwse-resize`,`zoom-in`,`zoom-out`,Y,J]}],"field-sizing":[{"field-sizing":[`fixed`,`content`]}],"pointer-events":[{"pointer-events":[`auto`,`none`]}],resize:[{resize:[`none`,``,`y`,`x`]}],"scroll-behavior":[{scroll:[`auto`,`smooth`]}],"scroll-m":[{"scroll-m":w()}],"scroll-mx":[{"scroll-mx":w()}],"scroll-my":[{"scroll-my":w()}],"scroll-ms":[{"scroll-ms":w()}],"scroll-me":[{"scroll-me":w()}],"scroll-mbs":[{"scroll-mbs":w()}],"scroll-mbe":[{"scroll-mbe":w()}],"scroll-mt":[{"scroll-mt":w()}],"scroll-mr":[{"scroll-mr":w()}],"scroll-mb":[{"scroll-mb":w()}],"scroll-ml":[{"scroll-ml":w()}],"scroll-p":[{"scroll-p":w()}],"scroll-px":[{"scroll-px":w()}],"scroll-py":[{"scroll-py":w()}],"scroll-ps":[{"scroll-ps":w()}],"scroll-pe":[{"scroll-pe":w()}],"scroll-pbs":[{"scroll-pbs":w()}],"scroll-pbe":[{"scroll-pbe":w()}],"scroll-pt":[{"scroll-pt":w()}],"scroll-pr":[{"scroll-pr":w()}],"scroll-pb":[{"scroll-pb":w()}],"scroll-pl":[{"scroll-pl":w()}],"snap-align":[{snap:[`start`,`end`,`center`,`align-none`]}],"snap-stop":[{snap:[`normal`,`always`]}],"snap-type":[{snap:[`none`,`x`,`y`,`both`]}],"snap-strictness":[{snap:[`mandatory`,`proximity`]}],touch:[{touch:[`auto`,`none`,`manipulation`]}],"touch-x":[{"touch-pan":[`x`,`left`,`right`]}],"touch-y":[{"touch-pan":[`y`,`up`,`down`]}],"touch-pz":[`touch-pinch-zoom`],select:[{select:[`none`,`text`,`all`,`auto`]}],"will-change":[{"will-change":[`auto`,`scroll`,`contents`,`transform`,Y,J]}],fill:[{fill:[`none`,...N()]}],"stroke-w":[{stroke:[q,Hn,Fn,In]}],stroke:[{stroke:[`none`,...N()]}],"forced-color-adjust":[{"forced-color-adjust":[`auto`,`none`]}]},conflictingClassGroups:{overflow:[`overflow-x`,`overflow-y`],overscroll:[`overscroll-x`,`overscroll-y`],inset:[`inset-x`,`inset-y`,`inset-bs`,`inset-be`,`start`,`end`,`top`,`right`,`bottom`,`left`],"inset-x":[`right`,`left`],"inset-y":[`top`,`bottom`],flex:[`basis`,`grow`,`shrink`],gap:[`gap-x`,`gap-y`],p:[`px`,`py`,`ps`,`pe`,`pbs`,`pbe`,`pt`,`pr`,`pb`,`pl`],px:[`pr`,`pl`],py:[`pt`,`pb`],m:[`mx`,`my`,`ms`,`me`,`mbs`,`mbe`,`mt`,`mr`,`mb`,`ml`],mx:[`mr`,`ml`],my:[`mt`,`mb`],size:[`w`,`h`],"font-size":[`leading`],"fvn-normal":[`fvn-ordinal`,`fvn-slashed-zero`,`fvn-figure`,`fvn-spacing`,`fvn-fraction`],"fvn-ordinal":[`fvn-normal`],"fvn-slashed-zero":[`fvn-normal`],"fvn-figure":[`fvn-normal`],"fvn-spacing":[`fvn-normal`],"fvn-fraction":[`fvn-normal`],"line-clamp":[`display`,`overflow`],rounded:[`rounded-s`,`rounded-e`,`rounded-t`,`rounded-r`,`rounded-b`,`rounded-l`,`rounded-ss`,`rounded-se`,`rounded-ee`,`rounded-es`,`rounded-tl`,`rounded-tr`,`rounded-br`,`rounded-bl`],"rounded-s":[`rounded-ss`,`rounded-es`],"rounded-e":[`rounded-se`,`rounded-ee`],"rounded-t":[`rounded-tl`,`rounded-tr`],"rounded-r":[`rounded-tr`,`rounded-br`],"rounded-b":[`rounded-br`,`rounded-bl`],"rounded-l":[`rounded-tl`,`rounded-bl`],"border-spacing":[`border-spacing-x`,`border-spacing-y`],"border-w":[`border-w-x`,`border-w-y`,`border-w-s`,`border-w-e`,`border-w-bs`,`border-w-be`,`border-w-t`,`border-w-r`,`border-w-b`,`border-w-l`],"border-w-x":[`border-w-r`,`border-w-l`],"border-w-y":[`border-w-t`,`border-w-b`],"border-color":[`border-color-x`,`border-color-y`,`border-color-s`,`border-color-e`,`border-color-bs`,`border-color-be`,`border-color-t`,`border-color-r`,`border-color-b`,`border-color-l`],"border-color-x":[`border-color-r`,`border-color-l`],"border-color-y":[`border-color-t`,`border-color-b`],translate:[`translate-x`,`translate-y`,`translate-none`],"translate-none":[`translate`,`translate-x`,`translate-y`,`translate-z`],"scroll-m":[`scroll-mx`,`scroll-my`,`scroll-ms`,`scroll-me`,`scroll-mbs`,`scroll-mbe`,`scroll-mt`,`scroll-mr`,`scroll-mb`,`scroll-ml`],"scroll-mx":[`scroll-mr`,`scroll-ml`],"scroll-my":[`scroll-mt`,`scroll-mb`],"scroll-p":[`scroll-px`,`scroll-py`,`scroll-ps`,`scroll-pe`,`scroll-pbs`,`scroll-pbe`,`scroll-pt`,`scroll-pr`,`scroll-pb`,`scroll-pl`],"scroll-px":[`scroll-pr`,`scroll-pl`],"scroll-py":[`scroll-pt`,`scroll-pb`],touch:[`touch-x`,`touch-y`,`touch-pz`],"touch-x":[`touch`],"touch-y":[`touch`],"touch-pz":[`touch`]},conflictingClassGroupModifiers:{"font-size":[`leading`]},orderSensitiveModifiers:[`*`,`**`,`after`,`backdrop`,`before`,`details-content`,`file`,`first-letter`,`first-line`,`marker`,`placeholder`,`selection`]}});function or(...e){return ar(It(e))}var sr={base:`btn rounded-full border-0 shadow-md transition-all duration-300 ease-out hover:shadow-lg active:scale-[0.97] disabled:opacity-50 disabled:pointer-events-none`,variants:{flow:`bg-flow text-white hover:brightness-110`,learn:`bg-learn text-white hover:brightness-110`,ghost:`bg-base-100/80 text-flow shadow-sm hover:bg-base-100`,outline:`bg-transparent text-flow border-2 border-flow/40 hover:border-flow hover:bg-flow/5`}},cr=0;Array.isArray;function X(e,n,r,i,a,o){n||={};var s,c,l=n;if(`ref`in l)for(c in l={},n)c==`ref`?s=n[c]:l[c]=n[c];var u={type:e,props:l,key:r,ref:s,__k:null,__:null,__b:0,__e:null,__c:null,constructor:void 0,__v:--cr,__i:-1,__u:0,__source:a,__self:o};if(typeof e==`function`&&(s=e.defaultProps))for(c in s)l[c]===void 0&&(l[c]=s[c]);return t.vnode&&t.vnode(u),u}function Z({children:e,class:t,variant:n=`flow`,disabled:r,type:i=`button`,onClick:a,title:o,"aria-label":s}){return X(`button`,{type:i,disabled:r,onClick:a,class:or(sr.base,sr.variants[n],t),title:o,"aria-label":s,children:e})}var lr={base:`rounded-[var(--radius-box)] bg-base-100/90 backdrop-blur-sm shadow-lg shadow-flow/5 border border-base-content/5 transition-all duration-300 ease-out`,paddings:{sm:`p-4`,md:`p-6`,lg:`p-8`}};function ur({children:e,class:t,padding:n=`md`,...r}){return X(`div`,{...r,class:or(lr.base,lr.paddings[n],t),children:e})}var dr={fullscreenCenter:`fixed inset-0 z-[100] flex items-center justify-center bg-gradient-to-b from-base-100 via-base-100/95 to-base-200/80 p-4`};function fr({pseudot:e,onDone:t}){let[n,r]=V(!1),i=xe(!1),a=()=>{i.current||(i.current=!0,t())};return H(()=>{if(window.matchMedia(`(prefers-reduced-motion: reduce)`).matches){let e=window.setTimeout(a,450);return()=>clearTimeout(e)}let e=window.setTimeout(()=>r(!0),1400);return()=>clearTimeout(e)},[]),X(`div`,{class:`fixed inset-0 z-[100] flex items-center justify-center bg-gradient-to-b from-base-100 via-base-100/98 to-base-200/90 backdrop-blur-[2px] transition-opacity duration-[900ms] ease-out ${n?`pointer-events-none opacity-0`:`opacity-100`}`,onTransitionEnd:e=>{e.target===e.currentTarget&&e.propertyName===`opacity`&&n&&a()},children:X(`p`,{class:`fl-welcome-text px-6 text-center text-2xl font-semibold tracking-tight text-base-content sm:text-3xl`,children:[`Bienvenue, `,X(`span`,{class:`text-flow`,children:e}),X(`span`,{class:`text-learn`,children:` !`})]})})}function pr({onSubmit:e,busy:t,error:n}){let[r,i]=V(``);return X(`div`,{class:dr.fullscreenCenter,children:X(ur,{class:`w-full max-w-md border-base-content/10 shadow-xl shadow-flow/5`,children:[X(`h1`,{class:`mb-1 text-xl font-bold text-base-content`,children:`Premiere visite`}),X(`p`,{class:`mb-4 text-sm text-base-content/65`,children:[`Cet appareil n est pas reconnu en base (ou la base est vide). Indique ton pseudonyme pour creer ton profil (demo : MAC simulee `,X(`code`,{class:`rounded bg-base-200 px-1.5 py-0.5 text-xs`,children:et}),`).`]}),X(`label`,{class:`mb-2 block text-sm font-medium text-base-content/80`,for:`gate-pseudot`,children:`Pseudonyme`}),X(`input`,{id:`gate-pseudot`,class:`input input-bordered mb-3 w-full rounded-xl border-base-content/15 bg-base-100`,type:`text`,autoComplete:`username`,placeholder:`ex. zelda_quizz`,value:r,disabled:t,onInput:e=>i(e.target.value)}),n?X(`p`,{class:`mb-3 text-sm text-error`,children:n}):null,X(Z,{variant:`flow`,class:`w-full`,disabled:t||!r.trim(),onClick:()=>e(r.trim()),children:t?`Enregistrement...`:`Continuer`})]})})}function mr(){return X(`div`,{class:`fixed inset-0 z-[100] flex flex-col items-center justify-center gap-3 bg-base-100/95`,children:[X(`span`,{class:`loading loading-spinner loading-md text-flow`,"aria-hidden":!0}),X(`p`,{class:`text-sm text-base-content/60`,children:`Reconnaissance de l appareil...`})]})}function hr({message:e,onRetry:t}){return X(`div`,{class:dr.fullscreenCenter,children:X(ur,{class:`w-full max-w-md border-base-content/10 shadow-xl shadow-flow/5`,children:[X(`h1`,{class:`mb-1 text-xl font-bold text-base-content`,children:`Connexion au serveur`}),X(`p`,{class:`mb-4 text-sm text-base-content/65`,children:e}),X(Z,{variant:`flow`,class:`w-full`,onClick:t,children:`Reessayer`})]})})}function gr({children:e}){let[t,n]=V(null),[r,i]=V(`checking`),[a,o]=V(null),[s,c]=V(null),[l,u]=V(!1),[d,f]=V(0);return H(()=>{let e=!1;return c(null),(async()=>{try{let t=await rt(et);if(e)return;t.known?(n({userId:t.user.id,pseudot:t.user.pseudot}),i(`welcome`)):i(`need-pseudot`)}catch(t){e||(c(t instanceof Error?t.message:`Erreur inconnue.`),i(`api-error`))}})(),()=>{e=!0}},[d]),X(Nt.Provider,{value:r===`ready`?t:null,children:[r===`checking`?X(mr,{}):null,r===`api-error`&&s?X(hr,{message:s,onRetry:()=>f(e=>e+1)}):null,r===`need-pseudot`?X(pr,{onSubmit:async e=>{o(null),u(!0);try{let t=await it(et,e);n({userId:t.userId,pseudot:t.pseudot}),i(`welcome`)}catch(e){o(e instanceof Error?e.message:`Echec de l enregistrement.`)}finally{u(!1)}},busy:l,error:a}):null,r===`welcome`&&t?X(fr,{pseudot:t.pseudot,onDone:()=>i(`ready`)}):null,r===`ready`&&t?e:null]})}var _r=L(`/`);function vr(){return we(_r)}var yr=(...e)=>e.filter((e,t,n)=>!!e&&e.trim()!==``&&n.indexOf(e)===t).join(` `).trim(),br=e=>e.replace(/([a-z0-9])([A-Z])/g,`$1-$2`).toLowerCase(),xr=e=>e.replace(/^([A-Z])|[\s-_]+(\w)/g,(e,t,n)=>n?n.toUpperCase():t.toLowerCase()),Sr=e=>{let t=xr(e);return t.charAt(0).toUpperCase()+t.slice(1)},Cr={xmlns:`http://www.w3.org/2000/svg`,width:24,height:24,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,"stroke-width":`2`,"stroke-linecap":`round`,"stroke-linejoin":`round`},wr=L({size:24,color:`currentColor`,strokeWidth:2,absoluteStrokeWidth:!1,class:``}),Tr=()=>we(wr),Er=e=>{for(let t in e)if(t.startsWith(`aria-`)||t===`role`||t===`title`)return!0;return!1},Dr=({color:e,size:t,strokeWidth:n,absoluteStrokeWidth:r,children:i,iconNode:a,class:o=``,...s})=>{let{size:c=24,strokeWidth:l=2,absoluteStrokeWidth:u=!1,color:d=`currentColor`,class:f=``}=Tr()??{},p=r??u?Number(n??l)*24/Number(t??c):n??l;return v(`svg`,{...Cr,width:t??c??24,height:t??c??24,stroke:e??d,"stroke-width":p,class:yr(`lucide`,f,o),...!i&&!Er(s)&&{"aria-hidden":`true`},...s},[...a.map(([e,t])=>v(e,t)),...k(i)])},Q=(e,t)=>{let n=({class:n=``,className:r=``,children:i,...a})=>v(Dr,{...a,iconNode:t,class:yr(`lucide-${br(Sr(e))}`,`lucide-${br(e)}`,n,r)},i);return n.displayName=Sr(e),n},Or=Q(`arrow-left`,[[`path`,{d:`m12 19-7-7 7-7`,key:`1l729n`}],[`path`,{d:`M19 12H5`,key:`x3x0zl`}]]),kr=Q(`chart-column`,[[`path`,{d:`M3 3v16a2 2 0 0 0 2 2h16`,key:`c24i48`}],[`path`,{d:`M18 17V9`,key:`2bz60n`}],[`path`,{d:`M13 17V5`,key:`1frdt8`}],[`path`,{d:`M8 17v-3`,key:`17ska0`}]]),Ar=Q(`chevron-right`,[[`path`,{d:`m9 18 6-6-6-6`,key:`mthhwq`}]]),jr=Q(`clipboard-copy`,[[`rect`,{width:`8`,height:`4`,x:`8`,y:`2`,rx:`1`,ry:`1`,key:`tgr4d6`}],[`path`,{d:`M8 4H6a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2v-2`,key:`4jdomd`}],[`path`,{d:`M16 4h2a2 2 0 0 1 2 2v4`,key:`3hqy98`}],[`path`,{d:`M21 14H11`,key:`1bme5i`}],[`path`,{d:`m15 10-4 4 4 4`,key:`5dvupr`}]]),Mr=Q(`clock`,[[`circle`,{cx:`12`,cy:`12`,r:`10`,key:`1mglay`}],[`path`,{d:`M12 6v6l4 2`,key:`mmk7yg`}]]),Nr=Q(`download`,[[`path`,{d:`M12 15V3`,key:`m9g1x1`}],[`path`,{d:`M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4`,key:`ih7n3h`}],[`path`,{d:`m7 10 5 5 5-5`,key:`brsn70`}]]),Pr=Q(`file-braces`,[[`path`,{d:`M6 22a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h8a2.4 2.4 0 0 1 1.704.706l3.588 3.588A2.4 2.4 0 0 1 20 8v12a2 2 0 0 1-2 2z`,key:`1oefj6`}],[`path`,{d:`M14 2v5a1 1 0 0 0 1 1h5`,key:`wfsgrz`}],[`path`,{d:`M10 12a1 1 0 0 0-1 1v1a1 1 0 0 1-1 1 1 1 0 0 1 1 1v1a1 1 0 0 0 1 1`,key:`1oajmo`}],[`path`,{d:`M14 18a1 1 0 0 0 1-1v-1a1 1 0 0 1 1-1 1 1 0 0 1-1-1v-1a1 1 0 0 0-1-1`,key:`mpwhp6`}]]),Fr=Q(`file-up`,[[`path`,{d:`M6 22a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h8a2.4 2.4 0 0 1 1.704.706l3.588 3.588A2.4 2.4 0 0 1 20 8v12a2 2 0 0 1-2 2z`,key:`1oefj6`}],[`path`,{d:`M14 2v5a1 1 0 0 0 1 1h5`,key:`wfsgrz`}],[`path`,{d:`M12 12v6`,key:`3ahymv`}],[`path`,{d:`m15 15-3-3-3 3`,key:`15xj92`}]]),Ir=Q(`folder-tree`,[[`path`,{d:`M20 10a1 1 0 0 0 1-1V6a1 1 0 0 0-1-1h-2.5a1 1 0 0 1-.8-.4l-.9-1.2A1 1 0 0 0 15 3h-2a1 1 0 0 0-1 1v5a1 1 0 0 0 1 1Z`,key:`hod4my`}],[`path`,{d:`M20 21a1 1 0 0 0 1-1v-3a1 1 0 0 0-1-1h-2.9a1 1 0 0 1-.88-.55l-.42-.85a1 1 0 0 0-.92-.6H13a1 1 0 0 0-1 1v5a1 1 0 0 0 1 1Z`,key:`w4yl2u`}],[`path`,{d:`M3 5a2 2 0 0 0 2 2h3`,key:`f2jnh7`}],[`path`,{d:`M3 3v13a2 2 0 0 0 2 2h3`,key:`k8epm1`}]]),Lr=Q(`graduation-cap`,[[`path`,{d:`M21.42 10.922a1 1 0 0 0-.019-1.838L12.83 5.18a2 2 0 0 0-1.66 0L2.6 9.08a1 1 0 0 0 0 1.832l8.57 3.908a2 2 0 0 0 1.66 0z`,key:`j76jl0`}],[`path`,{d:`M22 10v6`,key:`1lu8f3`}],[`path`,{d:`M6 12.5V16a6 3 0 0 0 12 0v-3.5`,key:`1r8lef`}]]),Rr=Q(`hard-drive-download`,[[`path`,{d:`M12 2v8`,key:`1q4o3n`}],[`path`,{d:`m16 6-4 4-4-4`,key:`6wukr`}],[`rect`,{width:`20`,height:`8`,x:`2`,y:`14`,rx:`2`,key:`w68u3i`}],[`path`,{d:`M6 18h.01`,key:`uhywen`}],[`path`,{d:`M10 18h.01`,key:`h775k`}]]),zr=Q(`layers`,[[`path`,{d:`M12.83 2.18a2 2 0 0 0-1.66 0L2.6 6.08a1 1 0 0 0 0 1.83l8.58 3.91a2 2 0 0 0 1.66 0l8.58-3.9a1 1 0 0 0 0-1.83z`,key:`zw3jo`}],[`path`,{d:`M2 12a1 1 0 0 0 .58.91l8.6 3.91a2 2 0 0 0 1.65 0l8.58-3.9A1 1 0 0 0 22 12`,key:`1wduqc`}],[`path`,{d:`M2 17a1 1 0 0 0 .58.91l8.6 3.91a2 2 0 0 0 1.65 0l8.58-3.9A1 1 0 0 0 22 17`,key:`kqbvx6`}]]),Br=Q(`list-tree`,[[`path`,{d:`M8 5h13`,key:`1pao27`}],[`path`,{d:`M13 12h8`,key:`h98zly`}],[`path`,{d:`M13 19h8`,key:`c3s6r1`}],[`path`,{d:`M3 10a2 2 0 0 0 2 2h3`,key:`1npucw`}],[`path`,{d:`M3 5v12a2 2 0 0 0 2 2h3`,key:`x1gjn2`}]]),Vr=Q(`pencil`,[[`path`,{d:`M21.174 6.812a1 1 0 0 0-3.986-3.987L3.842 16.174a2 2 0 0 0-.5.83l-1.321 4.352a.5.5 0 0 0 .623.622l4.353-1.32a2 2 0 0 0 .83-.497z`,key:`1a8usu`}],[`path`,{d:`m15 5 4 4`,key:`1mk7zo`}]]),Hr=Q(`plus`,[[`path`,{d:`M5 12h14`,key:`1ays0h`}],[`path`,{d:`M12 5v14`,key:`s699le`}]]),Ur=Q(`replace`,[[`path`,{d:`M14 4a1 1 0 0 1 1-1`,key:`dhj8ez`}],[`path`,{d:`M15 10a1 1 0 0 1-1-1`,key:`1mnyi5`}],[`path`,{d:`M21 4a1 1 0 0 0-1-1`,key:`sfs9ap`}],[`path`,{d:`M21 9a1 1 0 0 1-1 1`,key:`mp6qeo`}],[`path`,{d:`m3 7 3 3 3-3`,key:`x25e72`}],[`path`,{d:`M6 10V5a2 2 0 0 1 2-2h2`,key:`15xut4`}],[`rect`,{x:`3`,y:`14`,width:`7`,height:`7`,rx:`1`,key:`1bkyp8`}]]),Wr=Q(`shuffle`,[[`path`,{d:`m18 14 4 4-4 4`,key:`10pe0f`}],[`path`,{d:`m18 2 4 4-4 4`,key:`pucp1d`}],[`path`,{d:`M2 18h1.973a4 4 0 0 0 3.3-1.7l5.454-8.6a4 4 0 0 1 3.3-1.7H22`,key:`1ailkh`}],[`path`,{d:`M2 6h1.972a4 4 0 0 1 3.6 2.2`,key:`km57vx`}],[`path`,{d:`M22 18h-6.041a4 4 0 0 1-3.3-1.8l-.359-.45`,key:`os18l9`}]]),Gr=Q(`sparkles`,[[`path`,{d:`M11.017 2.814a1 1 0 0 1 1.966 0l1.051 5.558a2 2 0 0 0 1.594 1.594l5.558 1.051a1 1 0 0 1 0 1.966l-5.558 1.051a2 2 0 0 0-1.594 1.594l-1.051 5.558a1 1 0 0 1-1.966 0l-1.051-5.558a2 2 0 0 0-1.594-1.594l-5.558-1.051a1 1 0 0 1 0-1.966l5.558-1.051a2 2 0 0 0 1.594-1.594z`,key:`1s2grr`}],[`path`,{d:`M20 2v4`,key:`1rf3ol`}],[`path`,{d:`M22 4h-4`,key:`gwowj6`}],[`circle`,{cx:`4`,cy:`20`,r:`2`,key:`6kqj1y`}]]),Kr=Q(`target`,[[`circle`,{cx:`12`,cy:`12`,r:`10`,key:`1mglay`}],[`circle`,{cx:`12`,cy:`12`,r:`6`,key:`1vlfrh`}],[`circle`,{cx:`12`,cy:`12`,r:`2`,key:`1c9p78`}]]),qr=Q(`trash-2`,[[`path`,{d:`M10 11v6`,key:`nco0om`}],[`path`,{d:`M14 11v6`,key:`outv1u`}],[`path`,{d:`M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6`,key:`miytrc`}],[`path`,{d:`M3 6h18`,key:`d0wm0j`}],[`path`,{d:`M8 6V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2`,key:`e791ji`}]]),Jr=Q(`trophy`,[[`path`,{d:`M10 14.66v1.626a2 2 0 0 1-.976 1.696A5 5 0 0 0 7 21.978`,key:`1n3hpd`}],[`path`,{d:`M14 14.66v1.626a2 2 0 0 0 .976 1.696A5 5 0 0 1 17 21.978`,key:`rfe1zi`}],[`path`,{d:`M18 9h1.5a1 1 0 0 0 0-5H18`,key:`7xy6bh`}],[`path`,{d:`M4 22h16`,key:`57wxv0`}],[`path`,{d:`M6 9a6 6 0 0 0 12 0V3a1 1 0 0 0-1-1H7a1 1 0 0 0-1 1z`,key:`1mhfuq`}],[`path`,{d:`M6 9H4.5a1 1 0 0 1 0-5H6`,key:`tex48p`}]]),Yr=Q(`wifi`,[[`path`,{d:`M12 20h.01`,key:`zekei9`}],[`path`,{d:`M2 8.82a15 15 0 0 1 20 0`,key:`dnpr2z`}],[`path`,{d:`M5 12.859a10 10 0 0 1 14 0`,key:`1x1e6c`}],[`path`,{d:`M8.5 16.429a5 5 0 0 1 7 0`,key:`1bycff`}]]),Xr=Q(`x`,[[`path`,{d:`M18 6 6 18`,key:`1bl5f8`}],[`path`,{d:`m6 6 12 12`,key:`d8bk6v`}]]);function Zr(e){let t=e.indexOf(`?`);return t>=0?e.slice(0,t):e}function Qr(e,t){let n=Zr(e);return t===`/`?n===`/`||n===``:n===t||n.startsWith(`${t}/`)}var $r=[{href:`/`,label:`Accueil`},{href:`/collections`,label:`Collection`},{href:`/questions`,label:`Question`},{href:`/dashboard`,label:`Dashboard`},{href:`/database`,label:`Export/Import`}],ei={header:`sticky top-0 z-40 border-b border-base-content/5 bg-base-100/80 backdrop-blur-lg transition-shadow duration-300`,container:`mx-auto flex max-w-4xl flex-col gap-3 px-4 py-3 sm:flex-row sm:items-center sm:justify-between sm:gap-4`,brandButton:`flex items-center gap-2.5 rounded-full px-1 py-1 text-left transition duration-300 ease-out hover:opacity-90 active:scale-[0.98]`,brandIcon:`flex h-11 w-11 shrink-0 items-center justify-center rounded-full bg-linear-to-br from-flow to-learn text-white shadow-lg shadow-flow/20`,nav:`flex flex-wrap items-center justify-center gap-1.5 sm:justify-end`,navLinkBase:`rounded-full px-4 py-2 text-sm font-medium transition-all duration-300 ease-out active:scale-[0.97]`,navLinkActive:`bg-flow text-white shadow-xl shadow-flow/25 hover:scale-[1.03] hover:brightness-110 hover:shadow-xl hover:shadow-flow/35`,navLinkIdle:`text-base-content/70 hover:bg-learn/14 hover:text-learn hover:scale-[1.02]`};function ti(){let e=vr();return X(`header`,{class:ei.header,children:X(`div`,{class:ei.container,children:[X(`button`,{type:`button`,onClick:()=>U(`/`),class:ei.brandButton,children:[X(`span`,{class:ei.brandIcon,children:X(Lr,{class:`h-5 w-5`,"aria-hidden":!0})}),X(`div`,{class:`leading-tight`,children:[X(`p`,{class:`text-lg font-semibold tracking-tight text-base-content`,children:`FlowLearn`}),X(`p`,{class:`text-xs text-base-content/50`,children:`Flow · Learn`})]})]}),X(`nav`,{class:ei.nav,"aria-label":`Navigation principale`,children:$r.map(({href:t,label:n})=>{let r=Qr(e,t);return X(`button`,{type:`button`,onClick:()=>U(t),class:or(ei.navLinkBase,r?ei.navLinkActive:ei.navLinkIdle),children:n},t)})})]})})}var ni={footer:`mt-auto border-t border-base-content/5 bg-base-100/50 py-4 text-center text-xs text-base-content/55`,container:`mx-auto flex max-w-3xl flex-wrap items-center justify-center gap-x-4 gap-y-2 px-4`,separator:`hidden sm:inline text-base-content/25`,badge:`inline-flex items-center gap-1.5 rounded-full bg-flow/10 px-2.5 py-0.5 text-flow`};function ri(){return X(`footer`,{class:ni.footer,children:X(`div`,{class:ni.container,children:[X(`span`,{children:[`FlowLearn v`,tt]}),X(`span`,{class:ni.separator,children:`·`}),X(`span`,{class:ni.badge,children:[X(Yr,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),`ESP32 (mock) — connecté`]})]})})}var ii={base:`fl-page-enter mx-auto w-full flex-1 px-4 py-6 md:py-8`,narrow:`max-w-2xl`,wide:`max-w-3xl`};function ai({children:e,class:t,narrow:n}){return X(`main`,{class:or(ii.base,n?ii.narrow:ii.wide,t),children:e})}var oi={overlay:`fixed inset-0 z-[110] flex items-center justify-center p-4`,backdrop:`absolute inset-0 border-0 bg-base-content/45 p-0`,dialogBase:`relative z-10 w-full max-w-md rounded-2xl border bg-base-100 p-6 shadow-2xl`,dialogDanger:`border-error/25 shadow-error/10`,dialogInfo:`border-base-content/10`};function si({open:e,title:t,message:n,variant:r=`info`,confirmLabel:i=`Confirmer`,cancelLabel:a=`Annuler`,busy:o=!1,onConfirm:s,onCancel:c}){if(!e)return null;let l=r===`danger`;return X(`div`,{class:oi.overlay,role:`presentation`,children:[X(`button`,{type:`button`,class:oi.backdrop,"aria-label":`Fermer`,disabled:o,onClick:()=>!o&&c()}),X(`div`,{class:`${oi.dialogBase} ${l?oi.dialogDanger:oi.dialogInfo}`,role:`dialog`,"aria-modal":`true`,"aria-labelledby":`pop-up-information-title`,onClick:e=>e.stopPropagation(),children:[X(`h2`,{id:`pop-up-information-title`,class:`mb-3 text-lg font-semibold tracking-tight ${l?`text-error`:`text-base-content`}`,children:t}),X(`p`,{class:`mb-6 whitespace-pre-line text-sm leading-relaxed text-base-content/80`,children:n}),X(`div`,{class:`flex flex-wrap justify-end gap-2`,children:[X(Z,{variant:`ghost`,class:`btn-sm`,disabled:o,onClick:c,children:a}),X(Z,{variant:l?`outline`:`learn`,class:l?`btn-sm border-2 border-error/50 text-error hover:border-error hover:bg-error/10 shadow-none`:`btn-sm`,disabled:o,onClick:s,children:o?`…`:i})]})]})]})}function ci(e){return typeof e==`object`&&!!e&&!Array.isArray(e)}function li(e,t){if(!Array.isArray(e)||e.length!==4)throw Error(`${t} : il faut exactement 4 réponses.`);let n=[],r=0;for(let i=0;i<e.length;i++){let a=e[i];if(!a||typeof a!=`object`||Array.isArray(a))throw Error(`${t}[${i}] : objet réponse attendu.`);let o=a,s=o.texte??o.reponse;if(typeof s!=`string`||s.trim()===``)throw Error(`${t}[${i}] : "texte" non vide requis.`);let c=o.correcte===!0||o.bonne_reponse===1;c&&(r+=1),n.push({texte:s.trim(),correcte:c})}if(r!==1)throw Error(`${t} : exactement une réponse doit être correcte.`);return n}function ui(e,t){if(!e||typeof e!=`object`||Array.isArray(e))throw Error(`${t} : objet question attendu.`);let n=e,r=n.categorie_id;if(typeof r!=`number`||!Number.isInteger(r)||r<1)throw Error(`${t} : categorie_id entier ≥ 1 requis.`);let i=n.categorie_type;if(typeof i!=`string`||i.trim()===``)throw Error(`${t} : categorie_type non vide requis.`);if(typeof n.question!=`string`||n.question.trim()===``)throw Error(`${t} : "question" non vide requise.`);let a=typeof n.commentaire==`string`?n.commentaire.trim():``,o=n.fakechecker??n.verifier,s=o===!0||o===1||o===`1`||typeof o==`string`&&o.trim().toLowerCase()===`true`;return{categorie_id:r,categorie_type:i.trim(),question:n.question.trim(),commentaire:a,fakechecker:s,reponses:li(n.reponses,`${t}.reponses`)}}function di(e){let t=JSON.parse(e);if(!ci(t))throw Error(`JSON invalide : objet racine attendu.`);if(t.format!==`flowlearn-app-collection-export`)throw Error(`JSON invalide : format "${String(t.format)}" — attendu "flowlearn-app-collection-export".`);if(t.version!==1)throw Error(`JSON invalide : version ${String(t.version)} — attendu 1.`);let n=t.exportedAt;if(n!=null&&typeof n!=`string`)throw Error(`JSON invalide : exportedAt doit être une chaîne ISO si présent.`);let r=t.user_id,i;if(r!=null)if(typeof r==`number`&&Number.isInteger(r)&&r>0)i=r;else if(typeof r==`string`&&/^\d+$/.test(r.trim())&&Number(r.trim())>0)i=Number(r.trim());else throw Error(`user_id : entier positif attendu si fourni.`);let a=t.collection;if(!ci(a))throw Error(`JSON invalide : objet collection attendu.`);let o=a.nom;if(typeof o!=`string`||o.trim()===``)throw Error(`collection.nom : chaîne non vide requise.`);let s,c=a.id;if(c!=null)if(typeof c==`number`&&Number.isInteger(c)&&c>=1)s=c;else if(typeof c==`string`&&/^\d+$/.test(c.trim())&&Number(c.trim())>=1)s=Number(c.trim());else throw Error(`collection.id : entier ≥ 1 attendu si fourni.`);let l,u=a.user_id;if(u!=null)if(typeof u==`number`&&Number.isInteger(u)&&u>=1)l=u;else if(typeof u==`string`&&/^\d+$/.test(u.trim())&&Number(u.trim())>=1)l=Number(u.trim());else throw Error(`collection.user_id : entier ≥ 1 attendu si fourni.`);let d=t.questions;if(!Array.isArray(d)||d.length===0)throw Error(`questions : tableau non vide requis.`);let f=d.map((e,t)=>ui(e,`questions[${t}]`)),p={nom:o.trim()};return s!=null&&(p.id=s),l!=null&&(p.user_id=l),{format:`flowlearn-app-collection-export`,version:1,exportedAt:typeof n==`string`?n:void 0,user_id:i,collection:p,questions:f}}function fi(e,t){if(!Array.isArray(e)||e.length!==4)throw Error(`${t} : il faut exactement 4 réponses.`);let n=[],r=0;for(let i=0;i<e.length;i++){let a=e[i];if(!a||typeof a!=`object`||Array.isArray(a))throw Error(`${t}[${i}] : objet réponse attendu.`);let o=a,s=o.texte??o.reponse;if(typeof s!=`string`||s.trim()===``)throw Error(`${t}[${i}] : "texte" non vide requis.`);let c=o.correcte===!0||o.bonne_reponse===1;c&&(r+=1),n.push({texte:s.trim(),correcte:c})}if(r!==1)throw Error(`${t} : exactement une réponse doit être correcte.`);return n}function pi(e,t){if(!e||typeof e!=`object`||Array.isArray(e))throw Error(`${t} : objet question attendu.`);let n=e;if(typeof n.question!=`string`||n.question.trim()===``)throw Error(`${t} : "question" non vide requise.`);let r=typeof n.commentaire==`string`?n.commentaire.trim():``;return{question:n.question.trim(),commentaire:r,reponses:fi(n.reponses,`${t}.reponses`)}}function mi(e,t){if(e==null)return[];if(!Array.isArray(e))throw Error(`${t} : tableau attendu.`);return e.map((e,n)=>pi(e,`${t}[${n}]`))}function hi(e){if(e==null)return[];if(!Array.isArray(e))throw Error(`collections : tableau attendu.`);return e.map((e,t)=>{if(!e||typeof e!=`object`||Array.isArray(e))throw Error(`collections[${t}] : objet attendu.`);let n=e;if(typeof n.nom!=`string`||n.nom.trim()===``)throw Error(`collections[${t}] : "nom" non vide requis.`);let r=mi(n.questions,`collections[${t}].questions`);if(r.length===0)throw Error(`collections[${t}] : au moins une question requise.`);return{nom:n.nom.trim(),questions:r}})}function gi(e){let t=JSON.parse(e);if(!t||typeof t!=`object`||Array.isArray(t))throw Error(`JSON invalide : objet racine attendu.`);let n=t,r=n.user_id,i;if(r!=null)if(typeof r==`number`&&Number.isInteger(r)&&r>0)i=r;else if(typeof r==`string`&&/^\d+$/.test(r.trim())&&Number(r.trim())>0)i=Number(r.trim());else throw Error(`user_id : entier positif attendu.`);let a=mi(n.questions,`questions`),o=hi(n.collections),s=mi(n.questions_sans_collection,`questions_sans_collection`);if(a.length>0&&(o.length>0||s.length>0))throw Error(`Format mixte interdit : utilise soit "questions", soit "collections"/"questions_sans_collection".`);let c={user_id:i,collections:o,questions_sans_collection:a.length>0?a:s};if(c.collections.length===0&&c.questions_sans_collection.length===0)throw Error(`Aucune question détectée dans le JSON.`);return c}function _i(e){if(e==null)return null;if(e.kind===`collection`){let t=e.data,n=t.questions.length;return{title:`Supprimer la collection « ${t.nom} » ?`,message:`Cette action est definitive :
 · la collection et ses liens vers les supercollections seront supprimes ;
-· les ${n} question${n>1?`s`:``} qui ne sont liees qu a cette collection seront supprimees (reponses et scores inclus) ;\n· une question encore presente dans une autre collection sera seulement detachee de celle-ci.`}}return{title:`Supprimer la supercollection « ${e.data.nom} » ?`,message:`Les liens avec les collections seront retires. Les collections elles-memes ne sont pas supprimees.`}}function _i(e,t,n){if(t===`all`)return e;if(t===`mine`)return e.filter(e=>e.user_id===n);if(t.startsWith(`user-`)){let n=Number(t.slice(5));if(Number.isFinite(n))return e.filter(e=>e.user_id===n)}return e}function vi(e,t){return t===`all`?e:e.filter(e=>(e.modules??[]).some(e=>e.id===t))}var yi=[`random`,`linear`,`jamais_repondu`,`recent`,`ancien`,`mal_repondu`];function bi(e){return yi.includes(e)}function xi(e){let t=e.split(`,`).map(e=>e.trim()).filter(e=>e.length>0),n=[];for(let e of t)bi(e)&&n.push(e);return n.length>0?n:[`random`]}function Si(e){return e.some(e=>e===`jamais_repondu`||e===`mal_repondu`)}function Ci(e){let t=[];return e.neverAnswered&&t.push(`jamais_repondu`),e.sortBase===`linear`?t.push(`linear`):e.sortBase===`recent`?t.push(`recent`):e.sortBase===`ancien`&&t.push(`ancien`),e.errorPriority&&t.push(`mal_repondu`),e.shuffleExtra&&t.push(`random`),t.length===0&&t.push(`random`),t}function wi(){if(typeof window>`u`)return[`random`];let e=new URLSearchParams(window.location.search).get(`order`);return e==null||e===``?[`random`]:xi(e)}function Ti(){if(typeof window>`u`)return`melanger`;let e=new URLSearchParams(window.location.search).get(`qtype`);return e===`histoire`||e===`pratique`||e===`melanger`?e:`melanger`}function Ei(){if(typeof window>`u`)return!1;let e=new URLSearchParams(window.location.search).get(`infinite`);return e===`1`||e===`true`}function Di(e){let t=e.get(`exclude`);return t==null||t===``?[]:t.split(`,`).map(e=>Number(e.trim())).filter(e=>Number.isInteger(e)&&e>0)}function Oi(){if(typeof window>`u`)return{orders:[`random`],qtype:`melanger`,infinite:!1,excludeIds:[],useServerPlayModes:!1};let e=new URLSearchParams(window.location.search),t=e.has(`order`)||e.has(`qtype`)||e.has(`infinite`)||e.has(`userId`)||e.has(`exclude`),n=e.get(`userId`),r=n!=null&&n!==``?Number(n):NaN;return{orders:wi(),qtype:Ti(),infinite:Ei(),userId:Number.isInteger(r)&&r>=1?r:void 0,excludeIds:Di(e),useServerPlayModes:t}}function ki(e){let t=new URLSearchParams,n=e.orders!=null&&e.orders.length>0?e.orders:[`random`];t.set(`order`,n.join(`,`));let r=e.qtype??`melanger`;r!==`melanger`&&t.set(`qtype`,r),e.infinite&&t.set(`infinite`,`1`),e.userId!=null&&Si(n)&&t.set(`userId`,String(e.userId)),e.excludeIds!=null&&e.excludeIds.length>0&&t.set(`exclude`,e.excludeIds.join(`,`));let i=t.toString();return i?`?${i}`:``}function Ai(e){let t=[...e];for(let e=t.length-1;e>0;e--){let n=Math.floor(Math.random()*(e+1));[t[e],t[n]]=[t[n],t[e]]}return t}function ji(e){return e===`histoire`?`Histoire`:e===`pratique`?`Pratique`:`Mélanger`}function Mi(e){switch(e){case`linear`:return`Ordre linéaire`;case`jamais_repondu`:return`Jamais répondues`;case`recent`:return`Ajout récent d’abord`;case`ancien`:return`Plus anciennes d’abord`;case`mal_repondu`:return`Priorité aux erreurs`;default:return`Aléatoire`}}function Ni(e){return e.map(Mi).join(` → `)}var Pi={base:`inline-flex items-center rounded-full border px-3 py-0.5 text-xs font-medium`,tones:{flow:`bg-flow/15 text-flow border-flow/20`,learn:`bg-learn/15 text-learn border-learn/20`,neutral:`bg-base-200/80 text-base-content/70 border-base-content/10`}};function Fi({children:e,class:t,tone:n=`neutral`}){return X(`span`,{class:ar(Pi.base,Pi.tones[n],t),children:e})}var Ii=[[`none`,`Par défaut (ordre collection / id)`],[`linear`,`Linéaire (ordre stable)`],[`recent`,`Ajout récent d’abord`],[`ancien`,`Plus anciennes d’abord`]],Li={wrapper:`grid w-full gap-3 text-left`,optionLabelBase:`mb-1 block text-xs font-semibold uppercase tracking-wide text-base-content/45`,checkboxLabel:`flex cursor-pointer items-center gap-2 text-sm text-base-content/80`,radioGroup:`flex flex-col gap-1.5 text-sm text-base-content/85`,radioLabel:`flex cursor-pointer items-center gap-2`,helper:`text-xs text-base-content/45`};function Ri({idPrefix:e,settings:t,onChange:n,labelAlignClass:r=`text-center`}){let{neverAnswered:i,sortBase:a,errorPriority:o,shuffleExtra:s}=t,c=`${Li.optionLabelBase} ${r}`;return X(`div`,{class:Li.wrapper,children:[X(`label`,{class:c,for:`${e}-never`,children:`Filtre`}),X(`label`,{class:Li.checkboxLabel,children:[X(`input`,{id:`${e}-never`,type:`checkbox`,class:`checkbox checkbox-sm checkbox-primary`,checked:i,onChange:e=>n({neverAnswered:e.target.checked})}),`Jamais répondues (par moi)`]}),X(`p`,{class:`${c} mt-1`,children:`Tri des questions`}),X(`div`,{class:Li.radioGroup,children:Ii.map(([t,r])=>X(`label`,{class:Li.radioLabel,children:[X(`input`,{type:`radio`,name:`${e}-sort`,class:`radio radio-sm radio-primary`,checked:a===t,onChange:()=>n({sortBase:t})}),r]},t))}),X(`label`,{class:`${c} mt-1`,for:`${e}-err`,children:`Pondération`}),X(`label`,{class:Li.checkboxLabel,children:[X(`input`,{id:`${e}-err`,type:`checkbox`,class:`checkbox checkbox-sm checkbox-primary`,checked:o,onChange:e=>n({errorPriority:e.target.checked})}),`Priorité aux erreurs (KPI)`]}),X(`label`,{class:`${c} mt-1`,for:`${e}-shuf`,children:`Mélange`}),X(`label`,{class:Li.checkboxLabel,children:[X(`input`,{id:`${e}-shuf`,type:`checkbox`,class:`checkbox checkbox-sm checkbox-primary`,checked:s,onChange:e=>n({shuffleExtra:e.target.checked})}),`Mélange aléatoire en fin de chaîne`]}),X(`p`,{class:Li.helper,children:`Ex. « Plus anciennes » + « Priorité erreurs » : tri par date puis tirage pondéré selon tes fautes.`})]})}function zi(e,t){let n=t[0];return`/questions/${e}${n==null?``:`?module=${n.id}`}`}var Bi={root:`group cursor-pointer transition-all duration-300 ease-out hover:-translate-y-1 hover:shadow-xl hover:shadow-flow/15`,headerLayout:`flex flex-col gap-4 sm:flex-row sm:items-start sm:justify-between`};function Vi({collection:e,myUserId:t,allModules:n,assignBusyCollectionId:r,deleteBusyCollectionId:i,interactionLocked:a=!1,onAssign:o,onUnassign:s,onDeleteCollection:c}){let l=e.questions.length,u=a||r!==null||i!==null,d=e.question_counts_by_type,f=e.user_id===t,[p,m]=V(``),[h,g]=V({neverAnswered:!1,sortBase:`none`,errorPriority:!1,shuffleExtra:!1}),[_,v]=V(`melanger`),[y,b]=V(!1),x=e.modules??[],S=()=>U(zi(e.id,x)),C=e=>{e.target?.closest(`button, a, input, select, textarea, label`)||S()},w=n.filter(e=>!x.some(t=>t.id===e.id)),T=x.map(e=>e.id).join(`,`);return H(()=>m(``),[e.id,T]),X(lr,{class:Bi.root,onClick:C,role:`button`,tabIndex:0,onKeyDown:e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),S())},children:X(`div`,{class:Bi.headerLayout,children:[X(`div`,{class:`space-y-2`,children:[X(`div`,{class:`flex flex-wrap items-center gap-2`,children:[X(Fi,{tone:`flow`,children:[l,` question`,l>1?`s`:``]}),X(Fi,{tone:`learn`,class:`font-normal opacity-90`,children:[`Histoire `,d.histoire]}),X(Fi,{tone:`flow`,class:`border border-flow/25 bg-flow/10 font-normal opacity-95`,children:[`Pratique `,d.pratique]})]}),X(`h2`,{class:`text-xl font-semibold tracking-tight text-base-content`,children:e.nom}),X(`p`,{class:`text-sm text-base-content/60`,children:[`Collection · mise à jour `,e.update_at.slice(0,10)]}),X(`p`,{class:`text-sm text-base-content/60`,children:[`Créé par `,e.createur_pseudot]}),x.length>0?X(`div`,{class:`flex flex-wrap items-center gap-2 pt-1`,children:[X(`span`,{class:`inline-flex items-center gap-1 text-xs font-medium text-base-content/50`,children:[X(Wr,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),`Supercollections`]}),x.map(t=>X(`span`,{class:`inline-flex max-w-full items-center gap-0.5 rounded-full border border-learn/30 bg-learn/10 pl-2.5 pr-0.5 text-xs font-medium text-learn`,children:[X(`span`,{class:`truncate py-1`,children:t.nom}),f?X(`button`,{type:`button`,class:`btn btn-ghost btn-xs min-h-0 h-7 w-7 shrink-0 rounded-full p-0 text-base-content/60 hover:bg-error/15 hover:text-error`,title:`Retirer « ${t.nom} »`,"aria-label":`Retirer la supercollection ${t.nom}`,disabled:u,onClick:()=>void s(e.id,t.id),children:X(ii,{class:`h-3.5 w-3.5`,"aria-hidden":!0})}):null]},t.id))]}):null,f&&n.length>0?X(`div`,{class:`flex flex-col gap-2 border-t border-base-content/10 pt-3 sm:max-w-md`,children:[X(`p`,{class:`text-xs font-medium text-base-content/55`,children:`Rattacher à une supercollection`}),w.length===0?X(`p`,{class:`text-xs text-base-content/50`,children:`Cette collection est déjà liée à toutes les supercollections.`}):X(`div`,{class:`flex flex-col gap-2 sm:flex-row sm:items-center`,children:[X(`select`,{class:`select select-bordered select-sm w-full rounded-xl border-base-content/15 bg-base-100 sm:min-w-[12rem]`,value:p===``?``:String(p),disabled:u,onChange:e=>{let t=e.target.value;m(t===``?``:Number(t))},children:[X(`option`,{value:``,children:`Choisir…`}),w.map(e=>X(`option`,{value:e.id,children:e.nom},e.id))]}),X(Z,{variant:`learn`,class:`btn-sm shrink-0`,disabled:u||p===``,onClick:()=>{p!==``&&o(e.id,p)},children:r===e.id?`…`:`Assigner`})]})]}):null,f&&n.length===0?X(`p`,{class:`border-t border-base-content/10 pt-3 text-xs text-base-content/50`,children:`Crée d’abord une supercollection (bloc ci-dessus) pour pouvoir rattacher cette collection.`}):null]}),X(`div`,{class:`flex shrink-0 flex-col gap-2 self-start sm:self-center sm:items-end`,children:[X(`p`,{class:`w-full text-xs font-medium text-base-content/55 sm:text-end`,children:`Mode de jeu`}),X(`div`,{class:`w-full rounded-xl border border-base-content/10 bg-base-100/50 p-2 sm:max-w-[14rem]`,children:X(Ri,{idPrefix:`col-${e.id}`,settings:h,onChange:e=>g(t=>({...t,...e})),labelAlignClass:`sm:text-end`})}),X(`label`,{class:`w-full text-xs font-medium text-base-content/55 sm:text-end`,for:`play-qtype-${e.id}`,children:`Type de questions`}),X(`select`,{id:`play-qtype-${e.id}`,class:`select select-bordered select-sm w-full rounded-xl border-base-content/15 bg-base-100 sm:max-w-[11rem]`,value:_,onChange:e=>{let t=e.target.value;(t===`histoire`||t===`pratique`||t===`melanger`)&&v(t)},children:[X(`option`,{value:`melanger`,children:`Mélanger`}),X(`option`,{value:`histoire`,children:`Histoire`}),X(`option`,{value:`pratique`,children:`Pratique`})]}),X(`label`,{class:`flex cursor-pointer items-center justify-end gap-2 text-xs text-base-content/70`,children:[X(`input`,{type:`checkbox`,class:`checkbox checkbox-xs checkbox-primary`,checked:y,onChange:e=>b(e.target.checked)}),`Session infinie (15)`]}),X(Z,{variant:`outline`,class:`btn-sm gap-1`,onClick:S,children:[X(Jr,{class:`h-4 w-4`,"aria-hidden":!0}),`Questions`]}),X(Z,{variant:`flow`,class:`btn-sm gap-1`,onClick:()=>{let n=Ci(h);U(`/play/${e.id}${ki({orders:n,qtype:_,infinite:y,userId:Si(n)?t:void 0})}`)},children:[`Jouer`,X(Rr,{class:`h-4 w-4 transition group-hover:translate-x-0.5`,"aria-hidden":!0})]}),f&&c!=null?X(`button`,{type:`button`,class:`btn btn-outline btn-sm gap-1 border-error/40 text-error hover:bg-error/10`,"aria-label":`Supprimer la collection ${e.nom}`,disabled:u,onClick:()=>c(e),children:[i===e.id?X(`span`,{class:`loading loading-spinner loading-xs`,"aria-hidden":!0}):X(ti,{class:`h-4 w-4`,"aria-hidden":!0}),`Supprimer`]}):null]})]})})}function Hi({jsonImportOpen:e,jsonImportMode:t,onOpenJsonImport:n}){return X(`div`,{class:`mb-6 flex flex-col gap-4 sm:flex-row sm:items-end sm:justify-between`,children:[X(`div`,{class:`space-y-2`,children:[X(`p`,{class:`inline-flex items-center gap-2 rounded-full bg-learn/10 px-3 py-1 text-xs font-medium text-learn`,children:[X(qr,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),`Tes collections`]}),X(`h1`,{class:`text-2xl font-bold tracking-tight text-base-content sm:text-3xl`,children:`Choisir une collection`}),X(`p`,{class:`max-w-xl text-sm text-base-content/60`,children:`Lance un quiz cible : uniquement les questions liees a la collection selectionnee.`})]}),X(`div`,{class:`flex flex-col gap-2 self-start sm:flex-row sm:items-center sm:self-auto`,children:X(Z,{variant:e&&t===`app`?`learn`:`outline`,class:`gap-2`,onClick:n,children:[X(Hr,{class:`h-4 w-4`,"aria-hidden":!0}),`JSON FlowLearn`]})})]})}function Ui({jsonImportOpen:e,jsonImportMode:t,jsonImportCategorie:n,jsonImportBusy:r,jsonImportText:i,jsonImportError:a,jsonImportMessage:o,onChangeCategorie:s,onOpenFilePicker:c,onChangeText:l,onRun:u}){return e?X(lr,{class:`fl-reveal-enter mb-6 border-learn/15 bg-learn/6`,children:X(`div`,{class:`flex flex-col gap-3`,children:[X(`div`,{class:`flex flex-col gap-2 sm:flex-row sm:items-end sm:justify-between`,children:[X(`div`,{children:X(`p`,{class:`text-sm font-medium text-base-content`,children:[`Importer des questions — `,t===`app`?`FlowLearn`:`LLM`]})}),t===`llm`?X(`div`,{class:`flex flex-col gap-2 sm:flex-row sm:items-center`,children:[X(`label`,{class:`text-xs font-medium text-base-content/60`,for:`collections-json-categorie`,children:`Categorie enregistree`}),X(`select`,{id:`collections-json-categorie`,class:`select select-bordered select-sm w-full rounded-xl border-base-content/15 bg-base-100 sm:w-44`,value:n,disabled:r,onChange:e=>s(e.target.value===`pratique`?`pratique`:`histoire`),children:[X(`option`,{value:`histoire`,children:`Histoire`}),X(`option`,{value:`pratique`,children:`Pratique`})]})]}):null]}),X(`div`,{class:`flex flex-wrap gap-2`,children:X(`button`,{type:`button`,class:`btn btn-outline btn-sm rounded-full border-base-content/15`,disabled:r,onClick:c,children:`Choisir un fichier...`})}),X(`textarea`,{class:`textarea textarea-bordered w-full min-h-32 rounded-2xl border-dashed border-learn/35 bg-base-100/60 font-mono text-xs leading-relaxed`,placeholder:t===`app`?`{ "format": "flowlearn-app-collection-export", ... }`:`{ "collections": [ ... ], "questions_sans_collection": [] }`,value:i,disabled:r,onInput:e=>l(e.target.value)}),a?X(`p`,{class:`text-xs text-error`,children:a}):null,o?X(`p`,{class:`text-sm text-base-content/80`,children:o}):null,X(Z,{variant:`flow`,disabled:r||!i.trim(),onClick:u,children:r?`Import...`:`Importer en base`})]})}):null}function Wi({modules:e,pendingDelete:t,deleteModuleBusyId:n,assignBusyCollectionId:r,deleteCollectionBusyId:i,deleteModuleError:a,newModuleName:o,createModuleBusy:s,createModuleError:c,onChangeNewModuleName:l,onCreateModule:u,onRequestDeleteModule:d,newCollName:f,newCollModuleId:p,createCollBusy:m,createCollError:h,onChangeNewCollName:g,onChangeNewCollModuleId:_,onCreateCollection:v,assignError:y,deleteCollectionError:x,filter:S,onChangeFilter:C,autresCreateurs:w,moduleFilter:T,onChangeModuleFilter:E,filtered:ee,userId:D,onAssign:O,onUnassign:k,onRequestDeleteCollection:A}){return X(b,{children:[X(`section`,{class:`mb-8 rounded-box border border-base-content/10 bg-base-200/30 p-4 sm:p-5`,children:[X(`h2`,{class:`text-sm font-semibold tracking-tight text-base-content`,children:`Supercollections`}),e.length>0?X(`ul`,{class:`mt-3 flex flex-col gap-2`,children:e.map(e=>X(`li`,{class:`flex items-center justify-between gap-3 rounded-xl border border-learn/25 bg-learn/10 px-3 py-2`,children:[X(`span`,{class:`min-w-0 flex-1 text-xs font-medium text-learn`,children:e.nom}),X(`button`,{type:`button`,class:`btn btn-ghost btn-xs shrink-0 gap-1 text-error hover:bg-error/10`,"aria-label":`Supprimer la supercollection ${e.nom}`,disabled:n!==null||r!==null||i!==null||t!==null,onClick:()=>d(e),children:n===e.id?X(`span`,{class:`loading loading-spinner loading-xs`,"aria-hidden":!0}):X(ti,{class:`h-4 w-4`,"aria-hidden":!0})})]},e.id))}):X(`p`,{class:`mt-3 text-xs text-base-content/50`,children:`Aucune supercollection pour l instant.`}),a?X(`p`,{class:`mt-2 text-xs text-error`,children:a}):null,X(`div`,{class:`mt-4 flex flex-col gap-2 sm:flex-row sm:items-end`,children:[X(`div`,{class:`flex-1`,children:[X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`new-supercollection-name`,children:`Nouvelle supercollection`}),X(`input`,{id:`new-supercollection-name`,class:`input input-bordered input-sm w-full rounded-xl border-base-content/15 bg-base-100`,type:`text`,value:o,disabled:s,onInput:e=>l(e.target.value)})]}),X(Z,{variant:`learn`,class:`btn-sm shrink-0`,disabled:s||!o.trim(),onClick:u,children:s?`Creation...`:`Creer`})]}),c?X(`p`,{class:`mt-2 text-xs text-error`,children:c}):null]}),X(`section`,{class:`mb-8 rounded-box border border-base-content/10 bg-base-200/30 p-4 sm:p-5`,children:[X(`h2`,{class:`text-sm font-semibold tracking-tight text-base-content`,children:`Nouvelle collection`}),X(`div`,{class:`mt-4 flex flex-col gap-3 sm:flex-row sm:flex-wrap sm:items-end`,children:[X(`div`,{class:`min-w-0 flex-1 sm:min-w-48`,children:[X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`new-collection-name`,children:`Nom`}),X(`input`,{id:`new-collection-name`,class:`input input-bordered input-sm w-full rounded-xl border-base-content/15 bg-base-100`,type:`text`,value:f,disabled:m,onInput:e=>g(e.target.value)})]}),X(`div`,{class:`w-full sm:w-auto sm:min-w-40`,children:[X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`new-collection-module`,children:`Supercollection (optionnel)`}),X(`select`,{id:`new-collection-module`,class:`select select-bordered select-sm w-full rounded-xl border-base-content/15 bg-base-100`,value:p===``?``:String(p),disabled:m||e.length===0,onChange:e=>_(e.target.value===``?``:Number(e.target.value)),children:[X(`option`,{value:``,children:`—`}),e.map(e=>X(`option`,{value:e.id,children:e.nom},e.id))]})]}),X(Z,{variant:`flow`,class:`btn-sm shrink-0`,disabled:m||!f.trim(),onClick:v,children:m?`Creation...`:`Creer collection`})]}),h?X(`p`,{class:`mt-2 text-xs text-error`,children:h}):null]}),y?X(`p`,{class:`mb-4 rounded-lg border border-error/30 bg-error/10 px-3 py-2 text-xs text-error`,children:y}):null,x?X(`p`,{class:`mb-4 rounded-lg border border-error/30 bg-error/10 px-3 py-2 text-xs text-error`,children:x}):null,X(`fieldset`,{class:`mb-6`,children:[X(`legend`,{class:`mb-3 text-xs font-medium uppercase tracking-wide text-base-content/45`,children:`Filtrer par createur`}),X(`div`,{class:`filter`,children:[X(`input`,{class:`btn btn-sm rounded-full border-0 filter-reset`,type:`radio`,name:`flowlearn-collections-filter`,"aria-label":`Toutes les collections`,checked:S===`all`,onChange:()=>C(`all`)}),X(`input`,{class:`btn btn-sm rounded-full border-0`,type:`radio`,name:`flowlearn-collections-filter`,"aria-label":`Mes collections`,checked:S===`mine`,onChange:()=>C(`mine`)}),w.map(([e,t])=>X(`input`,{class:`btn btn-sm rounded-full border-0`,type:`radio`,name:`flowlearn-collections-filter`,"aria-label":t,checked:S===`user-${e}`,onChange:()=>C(`user-${e}`)},e))]})]}),X(`fieldset`,{class:`mb-8`,children:[X(`legend`,{class:`mb-3 text-xs font-medium uppercase tracking-wide text-base-content/45`,children:`Par supercollection`}),e.length===0?X(`p`,{class:`text-xs text-base-content/50`,children:`Aucune supercollection : cree-en une plus haut pour filtrer.`}):X(`select`,{class:`select select-bordered select-sm w-full max-w-md rounded-xl border-base-content/15 bg-base-100`,value:T===`all`?``:String(T),onChange:e=>E(e.target.value===``?`all`:Number(e.target.value)),children:[X(`option`,{value:``,children:`Toutes les supercollections`}),e.map(e=>X(`option`,{value:e.id,children:e.nom},e.id))]})]}),ee.length===0?X(`p`,{class:`rounded-box border border-base-content/10 bg-base-200/40 px-4 py-8 text-center text-sm text-base-content/65`,children:`Aucune collection pour ce filtre.`}):X(`ul`,{class:`flex flex-col gap-4`,children:ee.map(n=>X(`li`,{children:X(Vi,{collection:n,myUserId:D,allModules:e,assignBusyCollectionId:r,deleteBusyCollectionId:i,onAssign:O,onUnassign:k,interactionLocked:t!==null,onDeleteCollection:A})},n.id))})]})}function Gi(){let{userId:e}=Nt(),[t,n]=V([]),[r,i]=V([]),[a,o]=V(!0),[s,c]=V(null),[l,u]=V(`all`),[d,f]=V(`all`),[p,m]=V(``),[h,g]=V(!1),[_,v]=V(null),[y,b]=V(null),[x,S]=V(null),[C,w]=V(null),[T,E]=V(null),[ee,D]=V(null),[O,k]=V(null),[A,j]=V(``),[M,te]=V(``),[ne,N]=V(!1),[re,ie]=V(null),[P,F]=V(null),I=xe(null),[L,ae]=V(!1),[oe,R]=V(`app`),[se,z]=V(``),[ce,le]=V(!1),[ue,de]=V(null),[B,fe]=V(null),[pe,me]=V(`histoire`),he=Ce(async()=>{let[e,t]=await Promise.all([it(),ot()]);return{list:e,mods:t}},[]);H(()=>{let e=!1;return(async()=>{o(!0),c(null);try{let{list:t,mods:r}=await he();e||(n(t),i(r))}catch{e||c(`fetch`)}finally{e||o(!1)}})(),()=>{e=!0}},[he]),H(()=>{d!==`all`&&!r.some(e=>e.id===d)&&f(`all`)},[r,d]);let ge=async()=>{let e=p.trim();if(e){g(!0),v(null);try{let t=await st(e);i(e=>[...e,t].sort((e,t)=>e.id-t.id)),m(``)}catch(e){v(e instanceof Error?e.message:`Creation impossible.`)}finally{g(!1)}}},_e=async(e,t)=>{b(e),S(null);try{let r=await lt(e,t);n(t=>t.map(t=>t.id===e?r:t))}catch(e){S(e instanceof Error?e.message:`Assignation impossible.`)}finally{b(null)}},ve=async(e,t)=>{b(e),S(null);try{let r=await ut(e,t);n(t=>t.map(t=>t.id===e?r:t))}catch(e){S(e instanceof Error?e.message:`Retrait impossible.`)}finally{b(null)}},ye=async()=>{let t=A.trim();if(t){N(!0),ie(null);try{let r={userId:e,nom:t};M!==``&&(r.moduleId=Number(M));let i=await yt(r);n(e=>[...e,i].sort((e,t)=>e.id-t.id)),j(``),te(``);let a=r.moduleId==null?``:`?module=${r.moduleId}`;U(`/questions/${i.id}${a}`)}catch(e){ie(e instanceof Error?e.message:`Creation impossible.`)}finally{N(!1)}}},be=P?.kind===`collection`?ee!==null:P?.kind===`module`?C!==null:!1,we=async()=>{if(P==null)return;if(P.kind===`collection`){let t=P.data;D(t.id),k(null);try{await dt(t.id,e),n(e=>e.filter(e=>e.id!==t.id)),F(null)}catch(e){k(e instanceof Error?e.message:`Suppression de la collection impossible.`),F(null)}finally{D(null)}return}let t=P.data;E(null),w(t.id);try{await ct(t.id),i(e=>e.filter(e=>e.id!==t.id)),n(e=>e.map(e=>({...e,modules:(e.modules??[]).filter(e=>e.id!==t.id)}))),F(null)}catch(e){E(e instanceof Error?e.message:`Suppression impossible.`),F(null)}finally{w(null)}},Te=gi(P),Ee=Se(()=>{let n=new Map;for(let r of t)r.user_id!==e&&n.set(r.user_id,r.createur_pseudot);return[...n.entries()].sort((e,t)=>e[0]-t[0])},[t,e]),De=Se(()=>vi(_i(t,l,e),d),[t,l,e,d]),Oe=async()=>{le(!0),fe(null),de(null);try{if(oe===`app`){let t=xr(se);t.user_id=e;let r;try{let a=await yt({userId:e,nom:t.collection.nom});r=a.id;let o=await xt(t,{collectionId:a.id}),{list:s,mods:c}=await he();n(s),i(c),z(``),de(`Import reussi (FlowLearn) : collection « ${t.collection.nom} » creee, ${o.createdQuestions} question(s).`)}catch(t){if(r!=null)try{await dt(r,e)}catch{}throw t}return}let t=Er(se);t.user_id=e;let r=await bt(t,{categorie:pe}),{list:a,mods:o}=await he();n(a),i(o),z(``),de(r.createdCollections>0?`Import reussi (LLM) : ${r.createdQuestions} question(s), ${r.createdCollections} nouvelle(s) collection(s).`:`Import reussi (LLM) : ${r.createdQuestions} question(s).`)}catch(e){fe(e instanceof Error?e.message:`Import JSON impossible.`)}finally{le(!1)}};return X(`div`,{class:`flex min-h-dvh flex-col`,children:[X(li,{}),X(pi,{children:[X(`input`,{ref:I,type:`file`,accept:`.json,application/json`,class:`hidden`,onChange:async e=>{let t=e.target.files?.[0];if(t)try{fe(null),de(null),z(await t.text()),ae(!0)}catch{fe(`Lecture du fichier impossible.`)}finally{e.target.value=``}}}),X(Hi,{jsonImportOpen:L,jsonImportMode:oe,onOpenJsonImport:()=>{R(`app`),ae(!0)}}),X(Ui,{jsonImportOpen:L,jsonImportMode:oe,jsonImportCategorie:pe,jsonImportBusy:ce,jsonImportText:se,jsonImportError:B,jsonImportMessage:ue,onChangeCategorie:me,onOpenFilePicker:()=>I.current?.click(),onChangeText:z,onRun:()=>void Oe()}),a?X(`p`,{class:`text-sm text-base-content/60`,children:`Chargement...`}):s?X(`div`,{class:`rounded-box border border-base-content/10 bg-base-200/40 px-4 py-8 text-center text-sm text-base-content/65`,children:[X(`p`,{class:`mb-3`,children:`Impossible de charger les collections (API indisponible ?).`}),X(Z,{variant:`flow`,class:`btn-sm`,onClick:()=>void he().then(({list:e,mods:t})=>{n(e),i(t)}).catch(()=>c(`fetch`)),children:`Reessayer`})]}):X(Wi,{modules:r,pendingDelete:P,deleteModuleBusyId:C,assignBusyCollectionId:y,deleteCollectionBusyId:ee,deleteModuleError:T,newModuleName:p,createModuleBusy:h,createModuleError:_,onChangeNewModuleName:m,onCreateModule:()=>void ge(),onRequestDeleteModule:e=>F({kind:`module`,data:e}),newCollName:A,newCollModuleId:M,createCollBusy:ne,createCollError:re,onChangeNewCollName:j,onChangeNewCollModuleId:te,onCreateCollection:()=>void ye(),assignError:x,deleteCollectionError:O,filter:l,onChangeFilter:u,autresCreateurs:Ee,moduleFilter:d,onChangeModuleFilter:f,filtered:De,userId:e,onAssign:_e,onUnassign:ve,onRequestDeleteCollection:e=>F({kind:`collection`,data:e})})]}),X(di,{}),X(hi,{open:P!=null,title:Te?.title??``,message:Te?.message??``,variant:`danger`,confirmLabel:`Supprimer`,cancelLabel:`Annuler`,busy:be,onCancel:()=>{be||F(null)},onConfirm:()=>void we()})]})}function Ki(e,t){let n=URL.createObjectURL(e),r=document.createElement(`a`);r.href=n,r.download=t,document.body.appendChild(r),r.click(),r.remove(),URL.revokeObjectURL(n)}async function qi(e){return await e.text()}var Ji={root:`flex min-h-dvh flex-col`};function Yi(){let[e,t]=V(!1),[n,r]=V(!1),[i,a]=V(null),[o,s]=V(null),c=xe(null),l=xe(null),[u,d]=V(null),[f,p]=V(null),[m,h]=V(!1),[g,_]=V(!1),[v,y]=V(null),[b,x]=V(null),[S,C]=V(null),w=async()=>{t(!0),a(null);try{let{blob:e,filename:t}=await Ot();Ki(e,t),s(t)}catch(e){a(e instanceof Error?e.message:`Export impossible.`)}finally{t(!1)}},T=async()=>{r(!0),a(null);try{let{blob:e,filename:t}=await kt();Ki(e,t),s(t)}catch(e){a(e instanceof Error?e.message:`Export impossible.`)}finally{r(!1)}},E=async()=>{if(u!=null&&window.confirm(`Remplacement SQL : cette operation peut detruire ou remplacer des donnees.
+· les ${n} question${n>1?`s`:``} qui ne sont liees qu a cette collection seront supprimees (reponses et scores inclus) ;\n· une question encore presente dans une autre collection sera seulement detachee de celle-ci.`}}return{title:`Supprimer la supercollection « ${e.data.nom} » ?`,message:`Les liens avec les collections seront retires. Les collections elles-memes ne sont pas supprimees.`}}function vi(e,t,n){if(t===`all`)return e;if(t===`mine`)return e.filter(e=>e.user_id===n);if(t.startsWith(`user-`)){let n=Number(t.slice(5));if(Number.isFinite(n))return e.filter(e=>e.user_id===n)}return e}function yi(e,t){return t===`all`?e:e.filter(e=>(e.modules??[]).some(e=>e.id===t))}function bi(){let{userId:e}=Pt(),[t,n]=V([]),[r,i]=V([]),[a,o]=V(!0),[s,c]=V(null),[l,u]=V(`all`),[d,f]=V(`all`),[p,m]=V(``),[h,g]=V(!1),[_,v]=V(null),[y,b]=V(null),[x,S]=V(null),[C,w]=V(null),[T,E]=V(null),[ee,D]=V(null),[O,k]=V(null),[A,j]=V(``),[M,te]=V(``),[ne,N]=V(!1),[re,ie]=V(null),[ae,P]=V(null),F=xe(null),[I,oe]=V(!1),[se,L]=V(`app`),[R,z]=V(``),[ce,le]=V(!1),[ue,de]=V(null),[B,fe]=V(null),[pe,me]=V(`histoire`),he=Ce(async()=>{let[e,t]=await Promise.all([at(),st()]);return{list:e,mods:t}},[]);H(()=>{let e=!1;return(async()=>{o(!0),c(null);try{let{list:t,mods:r}=await he();e||(n(t),i(r))}catch{e||c(`fetch`)}finally{e||o(!1)}})(),()=>{e=!0}},[he]),H(()=>{d!==`all`&&!r.some(e=>e.id===d)&&f(`all`)},[r,d]);let ge=async()=>{let e=p.trim();if(e){g(!0),v(null);try{let t=await ct(e);i(e=>[...e,t].sort((e,t)=>e.id-t.id)),m(``)}catch(e){v(e instanceof Error?e.message:`Creation impossible.`)}finally{g(!1)}}},_e=async(e,t)=>{b(e),S(null);try{let r=await ut(e,t);n(t=>t.map(t=>t.id===e?r:t))}catch(e){S(e instanceof Error?e.message:`Assignation impossible.`)}finally{b(null)}},ve=async(e,t)=>{b(e),S(null);try{let r=await dt(e,t);n(t=>t.map(t=>t.id===e?r:t))}catch(e){S(e instanceof Error?e.message:`Retrait impossible.`)}finally{b(null)}},ye=async()=>{let t=A.trim();if(t){N(!0),ie(null);try{let r={userId:e,nom:t};M!==``&&(r.moduleId=Number(M));let i=await bt(r);n(e=>[...e,i].sort((e,t)=>e.id-t.id)),j(``),te(``);let a=r.moduleId==null?``:`?module=${r.moduleId}`;U(`/questions/${i.id}${a}`)}catch(e){ie(e instanceof Error?e.message:`Creation impossible.`)}finally{N(!1)}}},be=ae?.kind===`collection`?ee!==null:ae?.kind===`module`?C!==null:!1,we=async()=>{if(ae==null)return;if(ae.kind===`collection`){let t=ae.data;D(t.id),k(null);try{await ft(t.id,e),n(e=>e.filter(e=>e.id!==t.id)),P(null)}catch(e){k(e instanceof Error?e.message:`Suppression de la collection impossible.`),P(null)}finally{D(null)}return}let t=ae.data;E(null),w(t.id);try{await lt(t.id),i(e=>e.filter(e=>e.id!==t.id)),n(e=>e.map(e=>({...e,modules:(e.modules??[]).filter(e=>e.id!==t.id)}))),P(null)}catch(e){E(e instanceof Error?e.message:`Suppression impossible.`),P(null)}finally{w(null)}},Te=_i(ae),Ee=Se(()=>{let n=new Map;for(let r of t)r.user_id!==e&&n.set(r.user_id,r.createur_pseudot);return[...n.entries()].sort((e,t)=>e[0]-t[0])},[t,e]),De=Se(()=>yi(vi(t,l,e),d),[t,l,e,d]),Oe=async()=>{le(!0),fe(null),de(null);try{if(se===`app`){let t=di(R);t.user_id=e;let r;try{let a=await bt({userId:e,nom:t.collection.nom});r=a.id;let o=await St(t,{collectionId:a.id}),{list:s,mods:c}=await he();n(s),i(c),z(``),de(`Import reussi (FlowLearn) : collection « ${t.collection.nom} » creee, ${o.createdQuestions} question(s).`)}catch(t){if(r!=null)try{await ft(r,e)}catch{}throw t}return}let t=gi(R);t.user_id=e;let r=await xt(t,{categorie:pe}),{list:a,mods:o}=await he();n(a),i(o),z(``),de(r.createdCollections>0?`Import reussi (LLM) : ${r.createdQuestions} question(s), ${r.createdCollections} nouvelle(s) collection(s).`:`Import reussi (LLM) : ${r.createdQuestions} question(s).`)}catch(e){fe(e instanceof Error?e.message:`Import JSON impossible.`)}finally{le(!1)}},ke=async e=>{let t=e.target,n=t.files?.[0];if(n)try{fe(null),de(null),z(await n.text()),oe(!0)}catch{fe(`Lecture du fichier impossible.`)}finally{t.value=``}};return{page:{loading:a,error:s,userId:e},header:{jsonImportOpen:I,jsonImportMode:se,onOpenJsonImport:()=>{L(`app`),oe(!0)}},jsonImport:{inputRef:F,open:I,mode:se,categorie:pe,busy:ce,text:R,error:B,message:ue,onChangeCategorie:me,onOpenFilePicker:()=>F.current?.click(),onChangeText:z,onRun:()=>void Oe(),onFileChange:e=>void ke(e)},content:{modules:r,pendingDelete:ae,deleteModuleBusyId:C,assignBusyCollectionId:y,deleteCollectionBusyId:ee,deleteModuleError:T,newModuleName:p,createModuleBusy:h,createModuleError:_,onChangeNewModuleName:m,onCreateModule:()=>void ge(),onRequestDeleteModule:e=>P({kind:`module`,data:e}),newCollName:A,newCollModuleId:M,createCollBusy:ne,createCollError:re,onChangeNewCollName:j,onChangeNewCollModuleId:te,onCreateCollection:()=>void ye(),assignError:x,deleteCollectionError:O,filter:l,onChangeFilter:u,autresCreateurs:Ee,moduleFilter:d,onChangeModuleFilter:f,filtered:De,userId:e,onAssign:_e,onUnassign:ve,onRequestDeleteCollection:e=>P({kind:`collection`,data:e})},confirmPopup:{open:ae!=null,title:Te?.title??``,message:Te?.message??``,busy:be,onCancel:()=>{be||P(null)},onConfirm:()=>void we()},retry:{onRetryLoad:()=>{he().then(({list:e,mods:t})=>{n(e),i(t)}).catch(()=>c(`fetch`))}}}}var xi=[`random`,`linear`,`jamais_repondu`,`recent`,`ancien`,`mal_repondu`];function Si(e){return xi.includes(e)}function Ci(e){let t=e.split(`,`).map(e=>e.trim()).filter(e=>e.length>0),n=[];for(let e of t)Si(e)&&n.push(e);return n.length>0?n:[`random`]}function wi(e){return e.some(e=>e===`jamais_repondu`||e===`mal_repondu`)}function Ti(e){let t=[];return e.neverAnswered&&t.push(`jamais_repondu`),e.sortBase===`linear`?t.push(`linear`):e.sortBase===`recent`?t.push(`recent`):e.sortBase===`ancien`&&t.push(`ancien`),e.errorPriority&&t.push(`mal_repondu`),e.shuffleExtra&&t.push(`random`),t.length===0&&t.push(`random`),t}function Ei(){if(typeof window>`u`)return[`random`];let e=new URLSearchParams(window.location.search).get(`order`);return e==null||e===``?[`random`]:Ci(e)}function Di(){if(typeof window>`u`)return`melanger`;let e=new URLSearchParams(window.location.search).get(`qtype`);return e===`histoire`||e===`pratique`||e===`melanger`?e:`melanger`}function Oi(){if(typeof window>`u`)return!1;let e=new URLSearchParams(window.location.search).get(`infinite`);return e===`1`||e===`true`}function ki(e){let t=e.get(`exclude`);return t==null||t===``?[]:t.split(`,`).map(e=>Number(e.trim())).filter(e=>Number.isInteger(e)&&e>0)}function Ai(){if(typeof window>`u`)return{orders:[`random`],qtype:`melanger`,infinite:!1,excludeIds:[],useServerPlayModes:!1};let e=new URLSearchParams(window.location.search),t=e.has(`order`)||e.has(`qtype`)||e.has(`infinite`)||e.has(`userId`)||e.has(`exclude`),n=e.get(`userId`),r=n!=null&&n!==``?Number(n):NaN;return{orders:Ei(),qtype:Di(),infinite:Oi(),userId:Number.isInteger(r)&&r>=1?r:void 0,excludeIds:ki(e),useServerPlayModes:t}}function ji(e){let t=new URLSearchParams,n=e.orders!=null&&e.orders.length>0?e.orders:[`random`];t.set(`order`,n.join(`,`));let r=e.qtype??`melanger`;r!==`melanger`&&t.set(`qtype`,r),e.infinite&&t.set(`infinite`,`1`),e.userId!=null&&wi(n)&&t.set(`userId`,String(e.userId)),e.excludeIds!=null&&e.excludeIds.length>0&&t.set(`exclude`,e.excludeIds.join(`,`));let i=t.toString();return i?`?${i}`:``}function Mi(e){let t=[...e];for(let e=t.length-1;e>0;e--){let n=Math.floor(Math.random()*(e+1));[t[e],t[n]]=[t[n],t[e]]}return t}function Ni(e){return e===`histoire`?`Histoire`:e===`pratique`?`Pratique`:`Mélanger`}function Pi(e){switch(e){case`linear`:return`Ordre linéaire`;case`jamais_repondu`:return`Jamais répondues`;case`recent`:return`Ajout récent d’abord`;case`ancien`:return`Plus anciennes d’abord`;case`mal_repondu`:return`Priorité aux erreurs`;default:return`Aléatoire`}}function Fi(e){return e.map(Pi).join(` → `)}var Ii={base:`inline-flex items-center rounded-full border px-3 py-0.5 text-xs font-medium`,tones:{flow:`bg-flow/15 text-flow border-flow/20`,learn:`bg-learn/15 text-learn border-learn/20`,neutral:`bg-base-200/80 text-base-content/70 border-base-content/10`}};function Li({children:e,class:t,tone:n=`neutral`}){return X(`span`,{class:or(Ii.base,Ii.tones[n],t),children:e})}var Ri=[[`none`,`Par défaut (ordre collection / id)`],[`linear`,`Linéaire (ordre stable)`],[`recent`,`Ajout récent d’abord`],[`ancien`,`Plus anciennes d’abord`]],zi={wrapper:`grid w-full gap-3 text-left`,optionLabelBase:`mb-1 block text-xs font-semibold uppercase tracking-wide text-base-content/45`,checkboxLabel:`flex cursor-pointer items-center gap-2 text-sm text-base-content/80`,radioGroup:`flex flex-col gap-1.5 text-sm text-base-content/85`,radioLabel:`flex cursor-pointer items-center gap-2`,helper:`text-xs text-base-content/45`};function Bi({idPrefix:e,settings:t,onChange:n,labelAlignClass:r=`text-center`}){let{neverAnswered:i,sortBase:a,errorPriority:o,shuffleExtra:s}=t,c=`${zi.optionLabelBase} ${r}`;return X(`div`,{class:zi.wrapper,children:[X(`label`,{class:c,for:`${e}-never`,children:`Filtre`}),X(`label`,{class:zi.checkboxLabel,children:[X(`input`,{id:`${e}-never`,type:`checkbox`,class:`checkbox checkbox-sm checkbox-primary`,checked:i,onChange:e=>n({neverAnswered:e.target.checked})}),`Jamais répondues (par moi)`]}),X(`p`,{class:`${c} mt-1`,children:`Tri des questions`}),X(`div`,{class:zi.radioGroup,children:Ri.map(([t,r])=>X(`label`,{class:zi.radioLabel,children:[X(`input`,{type:`radio`,name:`${e}-sort`,class:`radio radio-sm radio-primary`,checked:a===t,onChange:()=>n({sortBase:t})}),r]},t))}),X(`label`,{class:`${c} mt-1`,for:`${e}-err`,children:`Pondération`}),X(`label`,{class:zi.checkboxLabel,children:[X(`input`,{id:`${e}-err`,type:`checkbox`,class:`checkbox checkbox-sm checkbox-primary`,checked:o,onChange:e=>n({errorPriority:e.target.checked})}),`Priorité aux erreurs (KPI)`]}),X(`label`,{class:`${c} mt-1`,for:`${e}-shuf`,children:`Mélange`}),X(`label`,{class:zi.checkboxLabel,children:[X(`input`,{id:`${e}-shuf`,type:`checkbox`,class:`checkbox checkbox-sm checkbox-primary`,checked:s,onChange:e=>n({shuffleExtra:e.target.checked})}),`Mélange aléatoire en fin de chaîne`]}),X(`p`,{class:zi.helper,children:`Ex. « Plus anciennes » + « Priorité erreurs » : tri par date puis tirage pondéré selon tes fautes.`})]})}function Vi(e,t){let n=t[0];return`/questions/${e}${n==null?``:`?module=${n.id}`}`}var Hi={root:`group cursor-pointer transition-all duration-300 ease-out hover:-translate-y-1 hover:shadow-xl hover:shadow-flow/15`,headerLayout:`flex flex-col gap-4 sm:flex-row sm:items-start sm:justify-between`};function Ui({collection:e,myUserId:t,allModules:n,assignBusyCollectionId:r,deleteBusyCollectionId:i,interactionLocked:a=!1,onAssign:o,onUnassign:s,onDeleteCollection:c}){let l=e.questions.length,u=a||r!==null||i!==null,d=e.question_counts_by_type,f=e.user_id===t,[p,m]=V(``),[h,g]=V({neverAnswered:!1,sortBase:`none`,errorPriority:!1,shuffleExtra:!1}),[_,v]=V(`melanger`),[y,b]=V(!1),x=e.modules??[],S=()=>U(Vi(e.id,x)),C=e=>{e.target?.closest(`button, a, input, select, textarea, label`)||S()},w=n.filter(e=>!x.some(t=>t.id===e.id)),T=x.map(e=>e.id).join(`,`);return H(()=>m(``),[e.id,T]),X(ur,{class:Hi.root,onClick:C,role:`button`,tabIndex:0,onKeyDown:e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),S())},children:X(`div`,{class:Hi.headerLayout,children:[X(`div`,{class:`space-y-2`,children:[X(`div`,{class:`flex flex-wrap items-center gap-2`,children:[X(Li,{tone:`flow`,children:[l,` question`,l>1?`s`:``]}),X(Li,{tone:`learn`,class:`font-normal opacity-90`,children:[`Histoire `,d.histoire]}),X(Li,{tone:`flow`,class:`border border-flow/25 bg-flow/10 font-normal opacity-95`,children:[`Pratique `,d.pratique]})]}),X(`h2`,{class:`text-xl font-semibold tracking-tight text-base-content`,children:e.nom}),X(`p`,{class:`text-sm text-base-content/60`,children:[`Collection · mise à jour `,e.update_at.slice(0,10)]}),X(`p`,{class:`text-sm text-base-content/60`,children:[`Créé par `,e.createur_pseudot]}),x.length>0?X(`div`,{class:`flex flex-wrap items-center gap-2 pt-1`,children:[X(`span`,{class:`inline-flex items-center gap-1 text-xs font-medium text-base-content/50`,children:[X(Ir,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),`Supercollections`]}),x.map(t=>X(`span`,{class:`inline-flex max-w-full items-center gap-0.5 rounded-full border border-learn/30 bg-learn/10 pl-2.5 pr-0.5 text-xs font-medium text-learn`,children:[X(`span`,{class:`truncate py-1`,children:t.nom}),f?X(`button`,{type:`button`,class:`btn btn-ghost btn-xs min-h-0 h-7 w-7 shrink-0 rounded-full p-0 text-base-content/60 hover:bg-error/15 hover:text-error`,title:`Retirer « ${t.nom} »`,"aria-label":`Retirer la supercollection ${t.nom}`,disabled:u,onClick:()=>void s(e.id,t.id),children:X(Xr,{class:`h-3.5 w-3.5`,"aria-hidden":!0})}):null]},t.id))]}):null,f&&n.length>0?X(`div`,{class:`flex flex-col gap-2 border-t border-base-content/10 pt-3 sm:max-w-md`,children:[X(`p`,{class:`text-xs font-medium text-base-content/55`,children:`Rattacher à une supercollection`}),w.length===0?X(`p`,{class:`text-xs text-base-content/50`,children:`Cette collection est déjà liée à toutes les supercollections.`}):X(`div`,{class:`flex flex-col gap-2 sm:flex-row sm:items-center`,children:[X(`select`,{class:`select select-bordered select-sm w-full rounded-xl border-base-content/15 bg-base-100 sm:min-w-[12rem]`,value:p===``?``:String(p),disabled:u,onChange:e=>{let t=e.target.value;m(t===``?``:Number(t))},children:[X(`option`,{value:``,children:`Choisir…`}),w.map(e=>X(`option`,{value:e.id,children:e.nom},e.id))]}),X(Z,{variant:`learn`,class:`btn-sm shrink-0`,disabled:u||p===``,onClick:()=>{p!==``&&o(e.id,p)},children:r===e.id?`…`:`Assigner`})]})]}):null,f&&n.length===0?X(`p`,{class:`border-t border-base-content/10 pt-3 text-xs text-base-content/50`,children:`Crée d’abord une supercollection (bloc ci-dessus) pour pouvoir rattacher cette collection.`}):null]}),X(`div`,{class:`flex shrink-0 flex-col gap-2 self-start sm:self-center sm:items-end`,children:[X(`p`,{class:`w-full text-xs font-medium text-base-content/55 sm:text-end`,children:`Mode de jeu`}),X(`div`,{class:`w-full rounded-xl border border-base-content/10 bg-base-100/50 p-2 sm:max-w-[14rem]`,children:X(Bi,{idPrefix:`col-${e.id}`,settings:h,onChange:e=>g(t=>({...t,...e})),labelAlignClass:`sm:text-end`})}),X(`label`,{class:`w-full text-xs font-medium text-base-content/55 sm:text-end`,for:`play-qtype-${e.id}`,children:`Type de questions`}),X(`select`,{id:`play-qtype-${e.id}`,class:`select select-bordered select-sm w-full rounded-xl border-base-content/15 bg-base-100 sm:max-w-[11rem]`,value:_,onChange:e=>{let t=e.target.value;(t===`histoire`||t===`pratique`||t===`melanger`)&&v(t)},children:[X(`option`,{value:`melanger`,children:`Mélanger`}),X(`option`,{value:`histoire`,children:`Histoire`}),X(`option`,{value:`pratique`,children:`Pratique`})]}),X(`label`,{class:`flex cursor-pointer items-center justify-end gap-2 text-xs text-base-content/70`,children:[X(`input`,{type:`checkbox`,class:`checkbox checkbox-xs checkbox-primary`,checked:y,onChange:e=>b(e.target.checked)}),`Session infinie (15)`]}),X(Z,{variant:`outline`,class:`btn-sm gap-1`,onClick:S,children:[X(Br,{class:`h-4 w-4`,"aria-hidden":!0}),`Questions`]}),X(Z,{variant:`flow`,class:`btn-sm gap-1`,onClick:()=>{let n=Ti(h);U(`/play/${e.id}${ji({orders:n,qtype:_,infinite:y,userId:wi(n)?t:void 0})}`)},children:[`Jouer`,X(Ar,{class:`h-4 w-4 transition group-hover:translate-x-0.5`,"aria-hidden":!0})]}),f&&c!=null?X(`button`,{type:`button`,class:`btn btn-outline btn-sm gap-1 border-error/40 text-error hover:bg-error/10`,"aria-label":`Supprimer la collection ${e.nom}`,disabled:u,onClick:()=>c(e),children:[i===e.id?X(`span`,{class:`loading loading-spinner loading-xs`,"aria-hidden":!0}):X(qr,{class:`h-4 w-4`,"aria-hidden":!0}),`Supprimer`]}):null]})]})})}function Wi({jsonImportOpen:e,jsonImportMode:t,onOpenJsonImport:n}){return X(`div`,{class:`mb-6 flex flex-col gap-4 sm:flex-row sm:items-end sm:justify-between`,children:[X(`div`,{class:`space-y-2`,children:[X(`p`,{class:`inline-flex items-center gap-2 rounded-full bg-learn/10 px-3 py-1 text-xs font-medium text-learn`,children:[X(zr,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),`Tes collections`]}),X(`h1`,{class:`text-2xl font-bold tracking-tight text-base-content sm:text-3xl`,children:`Choisir une collection`}),X(`p`,{class:`max-w-xl text-sm text-base-content/60`,children:`Lance un quiz cible : uniquement les questions liees a la collection selectionnee.`})]}),X(`div`,{class:`flex flex-col gap-2 self-start sm:flex-row sm:items-center sm:self-auto`,children:X(Z,{variant:e&&t===`app`?`learn`:`outline`,class:`gap-2`,onClick:n,children:[X(Pr,{class:`h-4 w-4`,"aria-hidden":!0}),`JSON FlowLearn`]})})]})}function Gi({jsonImportOpen:e,jsonImportMode:t,jsonImportCategorie:n,jsonImportBusy:r,jsonImportText:i,jsonImportError:a,jsonImportMessage:o,onChangeCategorie:s,onOpenFilePicker:c,onChangeText:l,onRun:u}){return e?X(ur,{class:`fl-reveal-enter mb-6 border-learn/15 bg-learn/6`,children:X(`div`,{class:`flex flex-col gap-3`,children:[X(`div`,{class:`flex flex-col gap-2 sm:flex-row sm:items-end sm:justify-between`,children:[X(`div`,{children:X(`p`,{class:`text-sm font-medium text-base-content`,children:[`Importer des questions — `,t===`app`?`FlowLearn`:`LLM`]})}),t===`llm`?X(`div`,{class:`flex flex-col gap-2 sm:flex-row sm:items-center`,children:[X(`label`,{class:`text-xs font-medium text-base-content/60`,for:`collections-json-categorie`,children:`Categorie enregistree`}),X(`select`,{id:`collections-json-categorie`,class:`select select-bordered select-sm w-full rounded-xl border-base-content/15 bg-base-100 sm:w-44`,value:n,disabled:r,onChange:e=>s(e.target.value===`pratique`?`pratique`:`histoire`),children:[X(`option`,{value:`histoire`,children:`Histoire`}),X(`option`,{value:`pratique`,children:`Pratique`})]})]}):null]}),X(`div`,{class:`flex flex-wrap gap-2`,children:X(`button`,{type:`button`,class:`btn btn-outline btn-sm rounded-full border-base-content/15`,disabled:r,onClick:c,children:`Choisir un fichier...`})}),X(`textarea`,{class:`textarea textarea-bordered w-full min-h-32 rounded-2xl border-dashed border-learn/35 bg-base-100/60 font-mono text-xs leading-relaxed`,placeholder:t===`app`?`{ "format": "flowlearn-app-collection-export", ... }`:`{ "collections": [ ... ], "questions_sans_collection": [] }`,value:i,disabled:r,onInput:e=>l(e.target.value)}),a?X(`p`,{class:`text-xs text-error`,children:a}):null,o?X(`p`,{class:`text-sm text-base-content/80`,children:o}):null,X(Z,{variant:`flow`,disabled:r||!i.trim(),onClick:u,children:r?`Import...`:`Importer en base`})]})}):null}function Ki({modules:e,pendingDelete:t,deleteModuleBusyId:n,assignBusyCollectionId:r,deleteCollectionBusyId:i,deleteModuleError:a,newModuleName:o,createModuleBusy:s,createModuleError:c,onChangeNewModuleName:l,onCreateModule:u,onRequestDeleteModule:d,newCollName:f,newCollModuleId:p,createCollBusy:m,createCollError:h,onChangeNewCollName:g,onChangeNewCollModuleId:_,onCreateCollection:v,assignError:y,deleteCollectionError:x,filter:S,onChangeFilter:C,autresCreateurs:w,moduleFilter:T,onChangeModuleFilter:E,filtered:ee,userId:D,onAssign:O,onUnassign:k,onRequestDeleteCollection:A}){return X(b,{children:[X(`section`,{class:`mb-8 rounded-box border border-base-content/10 bg-base-200/30 p-4 sm:p-5`,children:[X(`h2`,{class:`text-sm font-semibold tracking-tight text-base-content`,children:`Supercollections`}),e.length>0?X(`ul`,{class:`mt-3 flex flex-col gap-2`,children:e.map(e=>X(`li`,{class:`flex items-center justify-between gap-3 rounded-xl border border-learn/25 bg-learn/10 px-3 py-2`,children:[X(`span`,{class:`min-w-0 flex-1 text-xs font-medium text-learn`,children:e.nom}),X(`button`,{type:`button`,class:`btn btn-ghost btn-xs shrink-0 gap-1 text-error hover:bg-error/10`,"aria-label":`Supprimer la supercollection ${e.nom}`,disabled:n!==null||r!==null||i!==null||t!==null,onClick:()=>d(e),children:n===e.id?X(`span`,{class:`loading loading-spinner loading-xs`,"aria-hidden":!0}):X(qr,{class:`h-4 w-4`,"aria-hidden":!0})})]},e.id))}):X(`p`,{class:`mt-3 text-xs text-base-content/50`,children:`Aucune supercollection pour l instant.`}),a?X(`p`,{class:`mt-2 text-xs text-error`,children:a}):null,X(`div`,{class:`mt-4 flex flex-col gap-2 sm:flex-row sm:items-end`,children:[X(`div`,{class:`flex-1`,children:[X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`new-supercollection-name`,children:`Nouvelle supercollection`}),X(`input`,{id:`new-supercollection-name`,class:`input input-bordered input-sm w-full rounded-xl border-base-content/15 bg-base-100`,type:`text`,value:o,disabled:s,onInput:e=>l(e.target.value)})]}),X(Z,{variant:`learn`,class:`btn-sm shrink-0`,disabled:s||!o.trim(),onClick:u,children:s?`Creation...`:`Creer`})]}),c?X(`p`,{class:`mt-2 text-xs text-error`,children:c}):null]}),X(`section`,{class:`mb-8 rounded-box border border-base-content/10 bg-base-200/30 p-4 sm:p-5`,children:[X(`h2`,{class:`text-sm font-semibold tracking-tight text-base-content`,children:`Nouvelle collection`}),X(`div`,{class:`mt-4 flex flex-col gap-3 sm:flex-row sm:flex-wrap sm:items-end`,children:[X(`div`,{class:`min-w-0 flex-1 sm:min-w-48`,children:[X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`new-collection-name`,children:`Nom`}),X(`input`,{id:`new-collection-name`,class:`input input-bordered input-sm w-full rounded-xl border-base-content/15 bg-base-100`,type:`text`,value:f,disabled:m,onInput:e=>g(e.target.value)})]}),X(`div`,{class:`w-full sm:w-auto sm:min-w-40`,children:[X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`new-collection-module`,children:`Supercollection (optionnel)`}),X(`select`,{id:`new-collection-module`,class:`select select-bordered select-sm w-full rounded-xl border-base-content/15 bg-base-100`,value:p===``?``:String(p),disabled:m||e.length===0,onChange:e=>_(e.target.value===``?``:Number(e.target.value)),children:[X(`option`,{value:``,children:`—`}),e.map(e=>X(`option`,{value:e.id,children:e.nom},e.id))]})]}),X(Z,{variant:`flow`,class:`btn-sm shrink-0`,disabled:m||!f.trim(),onClick:v,children:m?`Creation...`:`Creer collection`})]}),h?X(`p`,{class:`mt-2 text-xs text-error`,children:h}):null]}),y?X(`p`,{class:`mb-4 rounded-lg border border-error/30 bg-error/10 px-3 py-2 text-xs text-error`,children:y}):null,x?X(`p`,{class:`mb-4 rounded-lg border border-error/30 bg-error/10 px-3 py-2 text-xs text-error`,children:x}):null,X(`fieldset`,{class:`mb-6`,children:[X(`legend`,{class:`mb-3 text-xs font-medium uppercase tracking-wide text-base-content/45`,children:`Filtrer par createur`}),X(`div`,{class:`filter`,children:[X(`input`,{class:`btn btn-sm rounded-full border-0 filter-reset`,type:`radio`,name:`flowlearn-collections-filter`,"aria-label":`Toutes les collections`,checked:S===`all`,onChange:()=>C(`all`)}),X(`input`,{class:`btn btn-sm rounded-full border-0`,type:`radio`,name:`flowlearn-collections-filter`,"aria-label":`Mes collections`,checked:S===`mine`,onChange:()=>C(`mine`)}),w.map(([e,t])=>X(`input`,{class:`btn btn-sm rounded-full border-0`,type:`radio`,name:`flowlearn-collections-filter`,"aria-label":t,checked:S===`user-${e}`,onChange:()=>C(`user-${e}`)},e))]})]}),X(`fieldset`,{class:`mb-8`,children:[X(`legend`,{class:`mb-3 text-xs font-medium uppercase tracking-wide text-base-content/45`,children:`Par supercollection`}),e.length===0?X(`p`,{class:`text-xs text-base-content/50`,children:`Aucune supercollection : cree-en une plus haut pour filtrer.`}):X(`select`,{class:`select select-bordered select-sm w-full max-w-md rounded-xl border-base-content/15 bg-base-100`,value:T===`all`?``:String(T),onChange:e=>E(e.target.value===``?`all`:Number(e.target.value)),children:[X(`option`,{value:``,children:`Toutes les supercollections`}),e.map(e=>X(`option`,{value:e.id,children:e.nom},e.id))]})]}),ee.length===0?X(`p`,{class:`rounded-box border border-base-content/10 bg-base-200/40 px-4 py-8 text-center text-sm text-base-content/65`,children:`Aucune collection pour ce filtre.`}):X(`ul`,{class:`flex flex-col gap-4`,children:ee.map(n=>X(`li`,{children:X(Ui,{collection:n,myUserId:D,allModules:e,assignBusyCollectionId:r,deleteBusyCollectionId:i,onAssign:O,onUnassign:k,interactionLocked:t!==null,onDeleteCollection:A})},n.id))})]})}var qi={root:`flex min-h-dvh flex-col`,loading:`text-sm text-base-content/60`,errorBox:`rounded-box border border-base-content/10 bg-base-200/40 px-4 py-8 text-center text-sm text-base-content/65`,errorBoxMessage:`mb-3`,hiddenFileInput:`hidden`};function Ji(){let{page:e,header:t,jsonImport:n,content:r,confirmPopup:i,retry:a}=bi();return X(`div`,{class:qi.root,children:[X(ti,{}),X(ai,{children:[X(`input`,{ref:n.inputRef,type:`file`,accept:`.json,application/json`,class:qi.hiddenFileInput,onChange:n.onFileChange}),X(Wi,{jsonImportOpen:t.jsonImportOpen,jsonImportMode:t.jsonImportMode,onOpenJsonImport:t.onOpenJsonImport}),X(Gi,{jsonImportOpen:n.open,jsonImportMode:n.mode,jsonImportCategorie:n.categorie,jsonImportBusy:n.busy,jsonImportText:n.text,jsonImportError:n.error,jsonImportMessage:n.message,onChangeCategorie:n.onChangeCategorie,onOpenFilePicker:n.onOpenFilePicker,onChangeText:n.onChangeText,onRun:n.onRun}),e.loading?X(`p`,{class:qi.loading,children:`Chargement...`}):e.error?X(`div`,{class:qi.errorBox,children:[X(`p`,{class:qi.errorBoxMessage,children:`Impossible de charger les collections (API indisponible ?).`}),X(Z,{variant:`flow`,class:`btn-sm`,onClick:a.onRetryLoad,children:`Reessayer`})]}):X(Ki,{...r})]}),X(ri,{}),X(si,{open:i.open,title:i.title,message:i.message,variant:`danger`,confirmLabel:`Supprimer`,cancelLabel:`Annuler`,busy:i.busy,onCancel:i.onCancel,onConfirm:i.onConfirm})]})}function Yi(e,t){let n=URL.createObjectURL(e),r=document.createElement(`a`);r.href=n,r.download=t,document.body.appendChild(r),r.click(),r.remove(),URL.revokeObjectURL(n)}async function Xi(e){return await e.text()}function Zi(){let[e,t]=V(!1),[n,r]=V(!1),[i,a]=V(null),[o,s]=V(null),c=xe(null),l=xe(null),[u,d]=V(null),[f,p]=V(null),[m,h]=V(!1),[g,_]=V(!1),[v,y]=V(null),[b,x]=V(null),[S,C]=V(null),w=async(e,t)=>{t(!0),a(null);try{let{blob:t,filename:n}=await e();Yi(t,n),s(n)}catch(e){a(e instanceof Error?e.message:`Export impossible.`)}finally{t(!1)}};return{exporting:{sqlBusy:e,jsonBusy:n,exportError:i,lastFilename:o,handleExportSql:()=>void w(kt,t),handleExportJson:()=>void w(At,r)},importing:{sqlImportFile:u,jsonImportFile:f,sqlImportBusy:m,jsonImportBusy:g,importError:v,sqlImportResult:b,jsonImportResult:S,onPickSqlImportFile:d,onPickJsonImportFile:p,handleSqlImport:()=>{if(u!=null&&window.confirm(`Remplacement SQL : cette operation peut detruire ou remplacer des donnees.
 
-Tape OK seulement si tu viens de choisir un dump .sql fiable.`)){if(window.prompt(`Confirmation requise : saisir exactement REMPLACE_TOUT`)!==`REMPLACE_TOUT`){y(`Import SQL annule : confirmation invalide.`);return}h(!0),y(null),x(null),C(null);try{x(await jt(await qi(u)))}catch(e){y(e instanceof Error?e.message:`Import SQL impossible.`)}finally{h(!1)}}},ee=async()=>{if(f!=null&&window.confirm(`Fusion JSON : les donnees seront ajoutees / reconciliees dans la base courante.
+Tape OK seulement si tu viens de choisir un dump .sql fiable.`)){if(window.prompt(`Confirmation requise : saisir exactement REMPLACE_TOUT`)!==`REMPLACE_TOUT`){y(`Import SQL annule : confirmation invalide.`);return}(async()=>{h(!0),y(null),x(null),C(null);try{x(await Mt(await Xi(u)))}catch(e){y(e instanceof Error?e.message:`Import SQL impossible.`)}finally{h(!1)}})()}},handleJsonImport:()=>{f!=null&&window.confirm(`Fusion JSON : les donnees seront ajoutees / reconciliees dans la base courante.
 
-Les identifiants importes seront remappes pour eviter les collisions.`)){_(!0),y(null),C(null),x(null);try{let e=await qi(f);C(await At(JSON.parse(e)))}catch(e){y(e instanceof Error?e.message:`Import JSON impossible.`)}finally{_(!1)}}};return X(`div`,{class:Ji.root,children:[X(li,{}),X(pi,{narrow:!0,children:[X(`div`,{class:`mb-8 space-y-3 text-center`,children:[X(`p`,{class:`inline-flex items-center gap-2 rounded-full bg-flow/10 px-3 py-1 text-xs font-medium text-flow`,children:[X(Kr,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),`Export / import`]}),X(`h1`,{class:`text-2xl font-bold tracking-tight text-base-content sm:text-3xl`,children:`Sauvegarde de la base`}),X(`p`,{class:`mx-auto max-w-lg text-sm text-base-content/60`,children:[`Choisis le type d export : `,X(`strong`,{children:`remplacement`}),` via un dump `,X(`code`,{children:`.sql`}),`, ou`,` `,X(`strong`,{children:`fusion`}),` via un export `,X(`code`,{children:`.json`}),` pense pour reimporter les donnees sans rejouer un schema complet. L import correspondant est disponible plus bas.`]})]}),X(`section`,{class:`mb-6 rounded-box border border-base-content/10 bg-base-200/30 p-6 sm:p-8`,children:[X(`div`,{class:`mb-6 flex items-start justify-center gap-3 text-center`,children:[X(`span`,{class:`rounded-full bg-flow/15 p-2 text-flow`,children:X(Vr,{class:`h-5 w-5`,"aria-hidden":!0})}),X(`div`,{class:`max-w-xl`,children:[X(`h2`,{class:`text-base font-semibold text-base-content`,children:`Exporter`}),X(`p`,{class:`mt-1 text-sm text-base-content/55`,children:[`Les deux boutons telechargent un fichier : `,X(`code`,{children:`.sql`}),` pour une restauration classique,`,` `,X(`code`,{children:`.json`}),` pour preparer une fusion (meme schema, nouveaux identifiants geres plus tard par l import).`]})]})]}),X(`div`,{class:`mx-auto flex max-w-2xl flex-col items-stretch gap-4 sm:flex-row sm:items-center sm:justify-center sm:gap-5`,children:[X(Z,{variant:`flow`,class:`btn-lg min-h-16 w-full min-w-[240px] justify-center gap-2 px-8 text-base shadow-xl shadow-flow/20 sm:flex-1`,disabled:e,onClick:()=>void w(),children:[X(Zr,{class:`h-5 w-5 shrink-0`,"aria-hidden":!0}),X(`span`,{class:`flex min-w-0 flex-col items-center leading-tight`,children:[X(`span`,{class:`font-semibold`,children:`Remplacement`}),X(`span`,{class:`text-xs font-normal opacity-90`,children:e?`Preparation...`:`fichier .sql`})]})]}),X(Z,{variant:`learn`,class:`btn-lg min-h-16 w-full min-w-[240px] justify-center gap-2 px-8 text-base shadow-xl shadow-learn/20 sm:flex-1`,disabled:n,onClick:()=>void T(),children:[X(Qr,{class:`h-5 w-5 shrink-0`,"aria-hidden":!0}),X(`span`,{class:`flex min-w-0 flex-col items-center leading-tight`,children:[X(`span`,{class:`font-semibold`,children:`Fusion`}),X(`span`,{class:`text-xs font-normal opacity-90`,children:n?`Preparation...`:`fichier .json`})]})]})]}),o?X(`p`,{class:`mt-6 text-center text-xs text-success`,children:[`Dernier export telecharge : `,o]}):null,i?X(`p`,{class:`mt-3 text-center text-xs text-error`,children:i}):null]}),X(`section`,{class:`rounded-box border border-dashed border-base-content/15 bg-base-200/20 p-5`,children:[X(`div`,{class:`mb-5 flex items-start gap-3`,children:[X(`span`,{class:`rounded-full bg-learn/15 p-2 text-learn`,children:X(Ur,{class:`h-5 w-5`,"aria-hidden":!0})}),X(`div`,{children:[X(`h2`,{class:`text-base font-semibold text-base-content`,children:`Importer`}),X(`p`,{class:`mt-1 text-sm text-base-content/55`,children:[X(`strong`,{children:`Remplacement`}),` : rejoue un dump `,X(`code`,{children:`.sql`}),` (dangereux). `,X(`strong`,{children:`Fusion`}),` : importe un export `,X(`code`,{children:`.json`}),` FlowLearn en conservant la base existante autant que possible.`]})]})]}),X(`div`,{class:`grid gap-4 sm:grid-cols-2`,children:[X(`div`,{class:`rounded-box border border-base-content/10 bg-base-100/40 p-4`,children:[X(`p`,{class:`mb-2 text-sm font-semibold text-base-content`,children:`Remplacement (.sql)`}),X(`p`,{class:`mb-3 text-xs text-base-content/50`,children:[`Choisis un fichier puis confirme explicitement. Le backend exige la confirmation `,X(`code`,{children:`REMPLACE_TOUT`}),`.`]}),X(`input`,{ref:c,type:`file`,accept:`.sql,text/sql,application/sql`,class:`hidden`,onChange:e=>{d(e.target.files?.[0]??null)}}),X(`div`,{class:`flex flex-col gap-2`,children:[X(`button`,{type:`button`,class:`btn btn-outline btn-sm rounded-full border-base-content/15`,disabled:m,onClick:()=>c.current?.click(),children:u?u.name:`Choisir un fichier .sql`}),X(Z,{variant:`flow`,class:`btn-sm`,disabled:m||u==null,onClick:()=>void E(),children:m?`Import SQL...`:`Importer (remplacement)`})]})]}),X(`div`,{class:`rounded-box border border-base-content/10 bg-base-100/40 p-4`,children:[X(`p`,{class:`mb-2 text-sm font-semibold text-base-content`,children:`Fusion (.json)`}),X(`p`,{class:`mb-3 text-xs text-base-content/50`,children:[`Utilise un export `,X(`code`,{children:`.json`}),` genere par cette app (format `,X(`code`,{children:`flowlearn-sqlite-dump-json`}),`).`]}),X(`input`,{ref:l,type:`file`,accept:`.json,application/json`,class:`hidden`,onChange:e=>{p(e.target.files?.[0]??null)}}),X(`div`,{class:`flex flex-col gap-2`,children:[X(`button`,{type:`button`,class:`btn btn-outline btn-sm rounded-full border-base-content/15`,disabled:g,onClick:()=>l.current?.click(),children:f?f.name:`Choisir un fichier .json`}),X(Z,{variant:`learn`,class:`btn-sm`,disabled:g||f==null,onClick:()=>void ee(),children:g?`Import JSON...`:`Importer (fusion)`})]})]})]}),v?X(`p`,{class:`mt-4 text-xs text-error`,children:v}):null,b?X(`p`,{class:`mt-3 text-xs text-success`,children:[`SQL : `,b.statementsExecuted,` instruction(s) executee(s). Prisma a ete reconnecte.`]}):null,S?X(`div`,{class:`mt-3 rounded-box border border-base-content/10 bg-base-100/40 p-3 text-xs text-base-content/70`,children:[X(`p`,{class:`font-semibold text-base-content`,children:`Fusion JSON`}),X(`p`,{class:`mt-1`,children:[`Inserees : `,S.insertedRows,` · Ignorees : `,S.skippedRows,` · Remap d ids :`,` `,S.remappedIds]}),S.warnings.length>0?X(`ul`,{class:`mt-2 list-disc space-y-1 pl-4`,children:S.warnings.map(e=>X(`li`,{children:e},e))}):null]}):null]})]}),X(di,{})]})}var Xi={root:`flex min-h-dvh flex-col`,main:`fl-page-enter flex flex-1 flex-col items-center justify-center px-4 py-16`};function Zi(){let{userId:e}=Nt(),[t,n]=V({neverAnswered:!1,sortBase:`none`,errorPriority:!1,shuffleExtra:!1}),[r,i]=V(`melanger`),[a,o]=V(!1),s=()=>{let n=Ci(t);U(`/play/random${ki({orders:n,qtype:r,infinite:a,userId:Si(n)?e:void 0})}`)};return X(`div`,{class:Xi.root,children:[X(li,{}),X(`main`,{class:Xi.main,children:X(`div`,{class:`max-w-md text-center`,children:[X(`p`,{class:`mb-4 inline-flex items-center gap-2 rounded-full bg-flow/10 px-4 py-1.5 text-xs font-medium text-flow transition duration-300`,children:[X($r,{class:`h-4 w-4`,"aria-hidden":!0}),`Mode decouverte`]}),X(`h1`,{class:`mb-3 text-3xl font-bold tracking-tight text-base-content sm:text-4xl`,children:`Pret a apprendre ?`}),X(`p`,{class:`mb-6 text-base leading-relaxed text-base-content/65`,children:`Les questions viennent de toutes les collections. Combine plusieurs modes (ex. anciennes d abord + priorite aux erreurs), choisis le type de questions, et eventuellement une session infinie par paquets de 15.`}),X(`div`,{class:`mx-auto mb-8 w-full max-w-xs space-y-5 text-left`,children:[X(Ri,{idPrefix:`home`,settings:t,onChange:e=>n(t=>({...t,...e}))}),X(`div`,{children:[X(`label`,{class:`mb-1 block text-center text-xs font-semibold uppercase tracking-wide text-base-content/45`,for:`home-play-qtype`,children:`Type de questions`}),X(`select`,{id:`home-play-qtype`,class:`select select-bordered select-sm w-full rounded-xl border-base-content/15 bg-base-100 text-sm`,value:r,onChange:e=>{let t=e.target.value;(t===`histoire`||t===`pratique`||t===`melanger`)&&i(t)},children:[X(`option`,{value:`melanger`,children:`Melanger (tout)`}),X(`option`,{value:`histoire`,children:`Histoire`}),X(`option`,{value:`pratique`,children:`Pratique`})]})]}),X(`label`,{class:`flex cursor-pointer items-center justify-center gap-2 text-sm text-base-content/80`,children:[X(`input`,{type:`checkbox`,class:`checkbox checkbox-sm checkbox-primary`,checked:a,onChange:e=>o(e.target.checked)}),`Session infinie (paquets de 15 questions)`]})]}),X(Z,{variant:`flow`,class:`btn-lg min-h-14 min-w-[220px] px-10 text-base shadow-xl shadow-flow/25 transition duration-300 hover:scale-[1.03]`,onClick:()=>s(),children:`Commencer les quiz`})]})}),X(di,{})]})}var Qi={root:`mb-4 rounded-xl border border-flow/20 bg-flow/5 px-4 py-3 text-sm text-base-content/80`};function $i({targetCollectionNumeric:e,collections:t,allModules:n,importTargetModuleId:r,setImportTargetModuleId:i}){return e==null?null:X(`div`,{class:Qi.root,children:[X(`p`,{class:`font-medium text-base-content`,children:[`Collection cible: `,t.find(t=>t.id===e)?.nom??`#${e}`]}),X(`p`,{class:`mt-2 text-xs text-base-content/60`,children:`L import LLM ajoute les questions ici. Tu peux forcer un lien vers une supercollection apres import:`}),X(`label`,{class:`mt-2 block text-xs font-medium text-base-content/55`,for:`import-module-link`,children:`Supercollection (optionnel)`}),X(`select`,{id:`import-module-link`,class:`select select-bordered select-sm mt-1 max-w-md rounded-xl border-base-content/15 bg-base-100`,value:r??``,onChange:e=>{let t=e.target.value;i(t===``?null:Number(t))},children:[X(`option`,{value:``,children:`(aucune - pas de lien force)`}),n.map(e=>X(`option`,{value:e.id,children:e.nom},e.id))]})]})}var ea={item:`overflow-hidden rounded-[1.35rem] border border-base-content/10 bg-base-200/35 transition duration-300 hover:border-flow/20 hover:shadow-md hover:shadow-flow/5`};function ta({questions:e,saving:t,onEdit:n,onRemove:r}){return X(lr,{children:[X(`p`,{class:`mb-4 text-sm text-base-content/60`,children:[e.length,` question`,e.length===1?``:`s`,` affichee`,e.length===1?``:`s`,`.`]}),X(`ul`,{class:`space-y-3`,children:e.map(e=>X(`li`,{class:ea.item,children:X(`div`,{class:`flex flex-col gap-3 p-4 sm:flex-row sm:items-start sm:justify-between`,children:[X(`div`,{class:`min-w-0 flex-1`,children:[X(`div`,{class:`mb-2 flex flex-wrap items-center gap-2`,children:[X(Fi,{tone:`neutral`,children:[`id `,e.id]}),X(Fi,{tone:`learn`,children:e.categorie_type}),e.collections.length===0?X(Fi,{tone:`learn`,children:`Sans collection`}):e.collections.map(e=>X(Fi,{tone:`flow`,children:e.nom},e.id))]}),e.commentaire.trim()?X(`p`,{class:`mb-2 text-xs leading-relaxed text-base-content/55 line-clamp-3`,children:e.commentaire}):null,X(`p`,{class:`text-sm font-medium leading-relaxed text-base-content`,children:e.question})]}),X(`div`,{class:`flex shrink-0 flex-wrap gap-2`,children:[X(Z,{variant:`outline`,class:`btn-sm gap-1`,onClick:()=>n(e),disabled:t,children:[X(Yr,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),`Modifier`]}),X(Z,{variant:`learn`,class:`btn-sm gap-1`,onClick:()=>r(e.id),disabled:t,children:[X(ti,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),`Supprimer`]})]})]})},e.id))})]})}function na(e){return String(e).padStart(2,`0`)}function ra(e=new Date){return`${e.getFullYear()}-${na(e.getMonth()+1)}-${na(e.getDate())}_${na(e.getHours())}-${na(e.getMinutes())}-${na(e.getSeconds())}`}function ia(e){let t=e.trim();return t?t.normalize(`NFKD`).replace(/[\u0300-\u036f]/g,``).replace(/[^\w\-]+/g,`-`).replace(/-+/g,`-`).replace(/^-|-$/g,``).slice(0,60):`collection`}function aa(e,t){let n=new Blob([`${JSON.stringify(t,null,2)}\n`],{type:`application/json;charset=utf-8`}),r=URL.createObjectURL(n),i=document.createElement(`a`);i.href=r,i.download=e,document.body.appendChild(i),i.click(),i.remove(),URL.revokeObjectURL(r)}function oa(e){let t=[...e.reponses];if(t.length!==4)throw Error(`Question ${e.id} : 4 réponses attendues pour l’export (reçu ${t.length}).`);let n=0;for(let e of t)e.bonne_reponse&&(n+=1);if(n!==1)throw Error(`Question ${e.id} : exactement une bonne réponse attendue pour l’export.`);return{categorie_id:e.categorie_id,categorie_type:e.categorie_type,question:e.question,commentaire:e.commentaire??``,fakechecker:e.verifier,reponses:t.map(e=>({texte:e.reponse,correcte:e.bonne_reponse}))}}function sa(e){let t=e.questions.map(oa),n={format:`flowlearn-app-collection-export`,version:1,exportedAt:new Date().toISOString(),collection:{nom:e.nom},questions:t};aa(`${ia(e.nom)}-${ra()}.json`,n)}var ca=`Tu produis un JSON STRICT (sans markdown, sans commentaires) pour l’app FlowLearn.
+Les identifiants importes seront remappes pour eviter les collisions.`)&&(async()=>{_(!0),y(null),C(null),x(null);try{let e=await Xi(f);C(await jt(JSON.parse(e)))}catch(e){y(e instanceof Error?e.message:`Import JSON impossible.`)}finally{_(!1)}})()},openSqlFilePicker:()=>c.current?.click(),openJsonFilePicker:()=>l.current?.click(),sqlImportInputRef:c,jsonImportInputRef:l}}}var Qi={root:`flex min-h-dvh flex-col`};function $i(){let{exporting:e,importing:t}=Zi();return X(`div`,{class:Qi.root,children:[X(ti,{}),X(ai,{narrow:!0,children:[X(`div`,{class:`mb-8 space-y-3 text-center`,children:[X(`p`,{class:`inline-flex items-center gap-2 rounded-full bg-flow/10 px-3 py-1 text-xs font-medium text-flow`,children:[X(Rr,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),`Export / import`]}),X(`h1`,{class:`text-2xl font-bold tracking-tight text-base-content sm:text-3xl`,children:`Sauvegarde de la base`}),X(`p`,{class:`mx-auto max-w-lg text-sm text-base-content/60`,children:[`Choisis le type d export : `,X(`strong`,{children:`remplacement`}),` via un dump `,X(`code`,{children:`.sql`}),`, ou`,` `,X(`strong`,{children:`fusion`}),` via un export `,X(`code`,{children:`.json`}),` pense pour reimporter les donnees sans rejouer un schema complet. L import correspondant est disponible plus bas.`]})]}),X(`section`,{class:`mb-6 rounded-box border border-base-content/10 bg-base-200/30 p-6 sm:p-8`,children:[X(`div`,{class:`mb-6 flex items-start justify-center gap-3 text-center`,children:[X(`span`,{class:`rounded-full bg-flow/15 p-2 text-flow`,children:X(Nr,{class:`h-5 w-5`,"aria-hidden":!0})}),X(`div`,{class:`max-w-xl`,children:[X(`h2`,{class:`text-base font-semibold text-base-content`,children:`Exporter`}),X(`p`,{class:`mt-1 text-sm text-base-content/55`,children:[`Les deux boutons telechargent un fichier : `,X(`code`,{children:`.sql`}),` pour une restauration classique,`,` `,X(`code`,{children:`.json`}),` pour preparer une fusion (meme schema, nouveaux identifiants geres plus tard par l import).`]})]})]}),X(`div`,{class:`mx-auto flex max-w-2xl flex-col items-stretch gap-4 sm:flex-row sm:items-center sm:justify-center sm:gap-5`,children:[X(Z,{variant:`flow`,class:`btn-lg min-h-16 w-full min-w-[240px] justify-center gap-2 px-8 text-base shadow-xl shadow-flow/20 sm:flex-1`,disabled:e.sqlBusy,onClick:e.handleExportSql,children:[X(Ur,{class:`h-5 w-5 shrink-0`,"aria-hidden":!0}),X(`span`,{class:`flex min-w-0 flex-col items-center leading-tight`,children:[X(`span`,{class:`font-semibold`,children:`Remplacement`}),X(`span`,{class:`text-xs font-normal opacity-90`,children:e.sqlBusy?`Preparation...`:`fichier .sql`})]})]}),X(Z,{variant:`learn`,class:`btn-lg min-h-16 w-full min-w-[240px] justify-center gap-2 px-8 text-base shadow-xl shadow-learn/20 sm:flex-1`,disabled:e.jsonBusy,onClick:e.handleExportJson,children:[X(Wr,{class:`h-5 w-5 shrink-0`,"aria-hidden":!0}),X(`span`,{class:`flex min-w-0 flex-col items-center leading-tight`,children:[X(`span`,{class:`font-semibold`,children:`Fusion`}),X(`span`,{class:`text-xs font-normal opacity-90`,children:e.jsonBusy?`Preparation...`:`fichier .json`})]})]})]}),e.lastFilename?X(`p`,{class:`mt-6 text-center text-xs text-success`,children:[`Dernier export telecharge : `,e.lastFilename]}):null,e.exportError?X(`p`,{class:`mt-3 text-center text-xs text-error`,children:e.exportError}):null]}),X(`section`,{class:`rounded-box border border-dashed border-base-content/15 bg-base-200/20 p-5`,children:[X(`div`,{class:`mb-5 flex items-start gap-3`,children:[X(`span`,{class:`rounded-full bg-learn/15 p-2 text-learn`,children:X(Fr,{class:`h-5 w-5`,"aria-hidden":!0})}),X(`div`,{children:[X(`h2`,{class:`text-base font-semibold text-base-content`,children:`Importer`}),X(`p`,{class:`mt-1 text-sm text-base-content/55`,children:[X(`strong`,{children:`Remplacement`}),` : rejoue un dump `,X(`code`,{children:`.sql`}),` (dangereux). `,X(`strong`,{children:`Fusion`}),` : importe un export `,X(`code`,{children:`.json`}),` FlowLearn en conservant la base existante autant que possible.`]})]})]}),X(`div`,{class:`grid gap-4 sm:grid-cols-2`,children:[X(`div`,{class:`rounded-box border border-base-content/10 bg-base-100/40 p-4`,children:[X(`p`,{class:`mb-2 text-sm font-semibold text-base-content`,children:`Remplacement (.sql)`}),X(`p`,{class:`mb-3 text-xs text-base-content/50`,children:[`Choisis un fichier puis confirme explicitement. Le backend exige la confirmation `,X(`code`,{children:`REMPLACE_TOUT`}),`.`]}),X(`input`,{ref:t.sqlImportInputRef,type:`file`,accept:`.sql,text/sql,application/sql`,class:`hidden`,onChange:e=>t.onPickSqlImportFile(e.target.files?.[0]??null)}),X(`div`,{class:`flex flex-col gap-2`,children:[X(`button`,{type:`button`,class:`btn btn-outline btn-sm rounded-full border-base-content/15`,disabled:t.sqlImportBusy,onClick:t.openSqlFilePicker,children:t.sqlImportFile?t.sqlImportFile.name:`Choisir un fichier .sql`}),X(Z,{variant:`flow`,class:`btn-sm`,disabled:t.sqlImportBusy||t.sqlImportFile==null,onClick:t.handleSqlImport,children:t.sqlImportBusy?`Import SQL...`:`Importer (remplacement)`})]})]}),X(`div`,{class:`rounded-box border border-base-content/10 bg-base-100/40 p-4`,children:[X(`p`,{class:`mb-2 text-sm font-semibold text-base-content`,children:`Fusion (.json)`}),X(`p`,{class:`mb-3 text-xs text-base-content/50`,children:[`Utilise un export `,X(`code`,{children:`.json`}),` genere par cette app (format `,X(`code`,{children:`flowlearn-sqlite-dump-json`}),`).`]}),X(`input`,{ref:t.jsonImportInputRef,type:`file`,accept:`.json,application/json`,class:`hidden`,onChange:e=>t.onPickJsonImportFile(e.target.files?.[0]??null)}),X(`div`,{class:`flex flex-col gap-2`,children:[X(`button`,{type:`button`,class:`btn btn-outline btn-sm rounded-full border-base-content/15`,disabled:t.jsonImportBusy,onClick:t.openJsonFilePicker,children:t.jsonImportFile?t.jsonImportFile.name:`Choisir un fichier .json`}),X(Z,{variant:`learn`,class:`btn-sm`,disabled:t.jsonImportBusy||t.jsonImportFile==null,onClick:t.handleJsonImport,children:t.jsonImportBusy?`Import JSON...`:`Importer (fusion)`})]})]})]}),t.importError?X(`p`,{class:`mt-4 text-xs text-error`,children:t.importError}):null,t.sqlImportResult?X(`p`,{class:`mt-3 text-xs text-success`,children:[`SQL : `,t.sqlImportResult.statementsExecuted,` instruction(s) executee(s). Prisma a ete reconnecte.`]}):null,t.jsonImportResult?X(`div`,{class:`mt-3 rounded-box border border-base-content/10 bg-base-100/40 p-3 text-xs text-base-content/70`,children:[X(`p`,{class:`font-semibold text-base-content`,children:`Fusion JSON`}),X(`p`,{class:`mt-1`,children:[`Inserees : `,t.jsonImportResult.insertedRows,` · Ignorees : `,t.jsonImportResult.skippedRows,` · Remap d ids :`,` `,t.jsonImportResult.remappedIds]}),t.jsonImportResult.warnings.length>0?X(`ul`,{class:`mt-2 list-disc space-y-1 pl-4`,children:t.jsonImportResult.warnings.map(e=>X(`li`,{children:e},e))}):null]}):null]})]}),X(ri,{})]})}var ea={root:`flex min-h-dvh flex-col`,main:`fl-page-enter flex flex-1 flex-col items-center justify-center px-4 py-16`};function ta(){let{userId:e}=Pt(),[t,n]=V({neverAnswered:!1,sortBase:`none`,errorPriority:!1,shuffleExtra:!1}),[r,i]=V(`melanger`),[a,o]=V(!1),s=()=>{let n=Ti(t);U(`/play/random${ji({orders:n,qtype:r,infinite:a,userId:wi(n)?e:void 0})}`)};return X(`div`,{class:ea.root,children:[X(ti,{}),X(`main`,{class:ea.main,children:X(`div`,{class:`max-w-md text-center`,children:[X(`p`,{class:`mb-4 inline-flex items-center gap-2 rounded-full bg-flow/10 px-4 py-1.5 text-xs font-medium text-flow transition duration-300`,children:[X(Gr,{class:`h-4 w-4`,"aria-hidden":!0}),`Mode decouverte`]}),X(`h1`,{class:`mb-3 text-3xl font-bold tracking-tight text-base-content sm:text-4xl`,children:`Pret a apprendre ?`}),X(`p`,{class:`mb-6 text-base leading-relaxed text-base-content/65`,children:`Les questions viennent de toutes les collections. Combine plusieurs modes (ex. anciennes d abord + priorite aux erreurs), choisis le type de questions, et eventuellement une session infinie par paquets de 15.`}),X(`div`,{class:`mx-auto mb-8 w-full max-w-xs space-y-5 text-left`,children:[X(Bi,{idPrefix:`home`,settings:t,onChange:e=>n(t=>({...t,...e}))}),X(`div`,{children:[X(`label`,{class:`mb-1 block text-center text-xs font-semibold uppercase tracking-wide text-base-content/45`,for:`home-play-qtype`,children:`Type de questions`}),X(`select`,{id:`home-play-qtype`,class:`select select-bordered select-sm w-full rounded-xl border-base-content/15 bg-base-100 text-sm`,value:r,onChange:e=>{let t=e.target.value;(t===`histoire`||t===`pratique`||t===`melanger`)&&i(t)},children:[X(`option`,{value:`melanger`,children:`Melanger (tout)`}),X(`option`,{value:`histoire`,children:`Histoire`}),X(`option`,{value:`pratique`,children:`Pratique`})]})]}),X(`label`,{class:`flex cursor-pointer items-center justify-center gap-2 text-sm text-base-content/80`,children:[X(`input`,{type:`checkbox`,class:`checkbox checkbox-sm checkbox-primary`,checked:a,onChange:e=>o(e.target.checked)}),`Session infinie (paquets de 15 questions)`]})]}),X(Z,{variant:`flow`,class:`btn-lg min-h-14 min-w-[220px] px-10 text-base shadow-xl shadow-flow/25 transition duration-300 hover:scale-[1.03]`,onClick:()=>s(),children:`Commencer les quiz`})]})}),X(ri,{})]})}var na={root:`mb-4 rounded-xl border border-flow/20 bg-flow/5 px-4 py-3 text-sm text-base-content/80`};function ra({targetCollectionNumeric:e,collections:t,allModules:n,importTargetModuleId:r,setImportTargetModuleId:i}){return e==null?null:X(`div`,{class:na.root,children:[X(`p`,{class:`font-medium text-base-content`,children:[`Collection cible: `,t.find(t=>t.id===e)?.nom??`#${e}`]}),X(`p`,{class:`mt-2 text-xs text-base-content/60`,children:`L import LLM ajoute les questions ici. Tu peux forcer un lien vers une supercollection apres import:`}),X(`label`,{class:`mt-2 block text-xs font-medium text-base-content/55`,for:`import-module-link`,children:`Supercollection (optionnel)`}),X(`select`,{id:`import-module-link`,class:`select select-bordered select-sm mt-1 max-w-md rounded-xl border-base-content/15 bg-base-100`,value:r??``,onChange:e=>{let t=e.target.value;i(t===``?null:Number(t))},children:[X(`option`,{value:``,children:`(aucune - pas de lien force)`}),n.map(e=>X(`option`,{value:e.id,children:e.nom},e.id))]})]})}var ia=`categorie`,aa=[`histoire`,`pratique`],oa={histoire:`Questions plutôt « information » : contexte, repères, culture — le « pourquoi / qu’est-ce que ».`,pratique:`Questions sur comment appliquer l’information et dans quels cas l’utiliser — mise en situation et usage.`};function sa(e){return aa.includes(e)}function ca(e){let t=Array.from(new Set(e.map(e=>e.type.trim()).filter(Boolean))),n=[...aa],r=t.filter(e=>!sa(e));return t.length!==n.length||r.length>0?`Attention : la base expose ${t.length} catégorie(s) (${t.join(`, `)||`aucune`}) alors que le frontend en gère ${n.length} (${n.join(`, `)}).`:null}function la(e){let t=Array.from(new Set(e.map(e=>e.type).filter(sa)));return t.length>0?t:[...aa]}function ua(e,t){let n=t[0]??aa[0],r=typeof e==`string`&&t.includes(e)?e:n;return{id:ia,titre:`Catégorie (enregistrée en base)`,type:`liste_selection`,liste_choix:[...t],value:r}}function da(e){let{data:t,actions:n}=e,{options:r}=t,{onOptionsChange:i}=n,[a,o]=V(aa),[s,c]=V(null);H(()=>{mt().then(e=>{o(la(e)),c(ca(e))}).catch(()=>o(aa))},[]),H(()=>{let e=r.find(e=>e.id===ia),t=ua(e?.value??null,a);if(e==null)return void i([t,...r]);if(e.type!==`liste_selection`)return;let n=e.value===t.value,o=e.liste_choix.length===t.liste_choix.length&&e.liste_choix.every((e,n)=>e===t.liste_choix[n]);n&&o||i(r.map(e=>e.id===`categorie`?{...e,...t}:e))},[r,i,a]);let l=(e,t)=>i(r.map(n=>n.id===e?t(n):n));return{bandeauCategories:{messageAvertissement:s},champs:{options:r,updateOption:l,runCheckboxOption:async(e,t)=>{if(!t)return void l(e.id,e=>({...e,value:!1}));if(!e.action)return void l(e.id,e=>({...e,value:!0}));let n=await e.action();l(e.id,e=>({...e,value:n}))}}}}var fa={aside:`flex w-full shrink-0 flex-col gap-4 rounded-xl border border-base-content/10 bg-base-100/60 p-3 lg:min-w-70 lg:max-w-xs`,titreSection:`text-[0.65rem] font-semibold uppercase tracking-wide text-base-content/45`,alerteCategories:`rounded-lg border border-warning/30 bg-warning/10 px-3 py-2 text-xs leading-snug text-warning-content`,etiquetteChamp:`mb-1 block text-xs font-medium text-base-content/70`,select:`select select-bordered select-sm w-full rounded-lg border-base-content/15 bg-base-100 text-sm`,textarea:`textarea textarea-bordered w-full min-h-20 rounded-lg border-base-content/15 bg-base-100 text-sm`,inputTexte:`input input-bordered input-sm w-full rounded-lg border-base-content/15 bg-base-100 text-sm`,etiquetteCaseBase:`flex items-start gap-2 rounded-lg border border-base-content/10 p-2 text-xs leading-snug`,etiquetteCaseActive:`cursor-pointer`,etiquetteCaseDesactivee:`cursor-not-allowed opacity-50`,caseACocher:`checkbox checkbox-sm mt-0.5 shrink-0 border-base-content/30`,etiquetteCaseTitre:`mb-1 block font-medium text-base-content/70`,piedAide:`text-[0.65rem] leading-snug text-base-content/50`};function pa(e){let{bandeauCategories:t,champs:n}=da(e);return X(`aside`,{class:fa.aside,children:[X(`p`,{class:fa.titreSection,children:`Options`}),t.messageAvertissement?X(`div`,{class:fa.alerteCategories,children:t.messageAvertissement}):null,n.options.map(e=>e.type===`liste_selection`?X(`div`,{children:[X(`label`,{class:fa.etiquetteChamp,for:e.id,children:e.titre}),X(`select`,{id:e.id,class:fa.select,value:String(e.value??``),onChange:t=>n.updateOption(e.id,e=>({...e,value:t.target.value})),children:e.liste_choix.map(e=>X(`option`,{value:e,children:e},e))})]},e.id):e.type===`texte`?X(`div`,{children:[X(`label`,{class:fa.etiquetteChamp,for:e.id,children:e.titre}),e.multiline?X(`textarea`,{id:e.id,class:fa.textarea,placeholder:e.contenus,value:String(e.value??``),onInput:t=>n.updateOption(e.id,e=>({...e,value:t.target.value}))}):X(`input`,{id:e.id,type:`text`,class:fa.inputTexte,placeholder:e.contenus,value:String(e.value??``),onInput:t=>n.updateOption(e.id,e=>({...e,value:t.target.value}))})]},e.id):X(`label`,{class:`${fa.etiquetteCaseBase} ${e.disabled?fa.etiquetteCaseDesactivee:fa.etiquetteCaseActive}`,children:[X(`input`,{type:`checkbox`,class:fa.caseACocher,checked:!!e.value,disabled:e.disabled,onChange:t=>void n.runCheckboxOption(e,t.target.checked)}),X(`span`,{children:[X(`span`,{class:fa.etiquetteCaseTitre,children:e.titre}),e.description]})]},e.id)),X(`p`,{class:fa.piedAide,children:`Ces champs sont injectés dans le prompt copié pour le LLM.`})]})}function ma(e){let{data:t,actions:n,settings:r}=e,{prompt:i}=t,{importFromJson:a}=n,o=Te(),s=`llm-import-prompt-heading-${o}`,c=`llm-import-prompt-text-${o}`,[l,u]=V(``),[d,f]=V(null),[p,m]=V(!1),[h,g]=V(!1),_=!!r?.disabled,v=r?.submitLabel??`Importer en base`,y=r?.submitBusyLabel??`Import…`,b=r?.pasteAreaInstruction,x=r?.jsonPastePlaceholder??ga(i),S=ha(i)?`questions`:`collections`;return{header:{headingId:s,onCopyPrompt:async()=>{if(!_)try{await navigator.clipboard.writeText(i),g(!0),window.setTimeout(()=>g(!1),2e3)}catch{f(`Impossible de copier (permissions du navigateur).`)}},promptCopied:h,disabled:_},promptArea:{promptTextareaId:c,prompt:i,disabled:_},pasteHelp:{customInstruction:b,defaultVariant:S},importArea:{importText:l,setImportText:u,placeholder:x,disabled:_,importMessage:d},submit:{label:p?y:v,disabled:_||p||l.trim().length===0,onRunImport:async()=>{m(!0),f(null);try{f(await a(l)),u(``)}catch(e){f(e instanceof Error?e.message:`JSON ou réseau invalide.`)}finally{m(!1)}}}}}function ha(e){return e.includes(`"questions"`)&&!e.includes(`"questions_sans_collection"`)}function ga(e){return ha(e)?`{ "user_id": 1, "questions": [ { "question": "…", "commentaire": "…", "reponses": [ ... ] } ] }`:`{ "collections": [ ... ], "questions_sans_collection": [] }`}var _a={wrapper:`min-w-0 flex-1`,sectionPrompt:`mb-3 rounded-2xl border border-base-content/10 bg-base-200/25 p-3`,sectionPromptLegend:`mb-2 block text-xs font-medium uppercase tracking-wide text-base-content/55`,promptTextarea:`textarea textarea-bordered w-full min-h-48 resize-y rounded-xl border-base-content/15 bg-base-100/90 font-mono text-xs leading-relaxed text-base-content`,toolbar:`mb-2 flex flex-wrap items-center justify-between gap-2`,toolbarTitle:`text-sm font-medium text-base-content`,pasteHelp:`mb-2 text-xs text-base-content/55`,jsonTextarea:`textarea textarea-bordered mb-3 w-full min-h-32 rounded-2xl border-dashed border-learn/35 bg-base-100/60 font-mono text-xs leading-relaxed`,feedback:`mb-3 text-sm text-base-content/80`};function va(e){let{class:t}=e,{header:n,promptArea:r,pasteHelp:i,importArea:a,submit:o}=ma(e);return X(`div`,{class:or(_a.wrapper,t),children:[X(`div`,{class:_a.toolbar,children:[X(`h2`,{id:n.headingId,class:_a.toolbarTitle,children:`Prompt à copier pour ton LLM`}),X(Z,{variant:`outline`,class:`btn-sm gap-1`,type:`button`,disabled:n.disabled,onClick:n.onCopyPrompt,children:[X(jr,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),n.promptCopied?`Copié`:`Copier`]})]}),X(`section`,{class:_a.sectionPrompt,"aria-labelledby":n.headingId,children:[X(`label`,{class:_a.sectionPromptLegend,for:r.promptTextareaId,children:`Texte du prompt (lecture seule)`}),X(`textarea`,{id:r.promptTextareaId,class:_a.promptTextarea,readOnly:!0,rows:14,value:r.prompt,disabled:r.disabled})]}),X(ya,{customInstruction:i.customInstruction,defaultVariant:i.defaultVariant}),X(`textarea`,{class:_a.jsonTextarea,placeholder:a.placeholder,value:a.importText,disabled:a.disabled,onInput:e=>a.setImportText(e.target.value)}),a.importMessage?X(`p`,{class:_a.feedback,children:a.importMessage}):null,X(Z,{variant:`flow`,disabled:o.disabled,onClick:o.onRunImport,children:o.label})]})}function ya({customInstruction:e,defaultVariant:t}){return e==null?t===`questions`?X(`p`,{class:_a.pasteHelp,children:[`Colle le JSON au format `,X(`code`,{class:`text-xs`,children:`questions`}),` uniquement : chaque entrée sera liée à la collection affichée (aucun autre format requis).`]}):X(`p`,{class:_a.pasteHelp,children:[`Colle le JSON généré ci-dessous (champs `,X(`code`,{class:`text-xs`,children:`collections`}),` et/ou`,` `,X(`code`,{class:`text-xs`,children:`questions_sans_collection`}),`, ou tableau racine`,` `,X(`code`,{class:`text-xs`,children:`questions`}),`).`]}):X(`p`,{class:_a.pasteHelp,children:e})}var ba={card:`fl-reveal-enter mb-6 border-learn/15 bg-learn/[0.06]`,layout:`flex flex-col gap-4 lg:flex-row lg:items-stretch`};function xa(e){let{data:t,actions:n}=e,{options:r,llmImportWorkflow:i}=t,a=i.buildPrompt(r);return X(ur,{class:ba.card,children:X(`div`,{class:ba.layout,children:[X(pa,{data:{options:r},actions:{onOptionsChange:n.onOptionsChange}}),X(va,{data:{prompt:a},actions:{importFromJson:i.importFromJson}})]})})}var Sa={header:`mb-6 flex flex-col gap-4 sm:flex-row sm:items-end sm:justify-between`};function Ca(e){return String(e).padStart(2,`0`)}function wa(e=new Date){return`${e.getFullYear()}-${Ca(e.getMonth()+1)}-${Ca(e.getDate())}_${Ca(e.getHours())}-${Ca(e.getMinutes())}-${Ca(e.getSeconds())}`}function Ta(e){let t=e.trim();return t?t.normalize(`NFKD`).replace(/[\u0300-\u036f]/g,``).replace(/[^\w\-]+/g,`-`).replace(/-+/g,`-`).replace(/^-|-$/g,``).slice(0,60):`collection`}function Ea(e,t){let n=new Blob([`${JSON.stringify(t,null,2)}\n`],{type:`application/json;charset=utf-8`}),r=URL.createObjectURL(n),i=document.createElement(`a`);i.href=r,i.download=e,document.body.appendChild(i),i.click(),i.remove(),URL.revokeObjectURL(r)}function Da(e){let t=[...e.reponses];if(t.length!==4)throw Error(`Question ${e.id} : 4 réponses attendues pour l’export (reçu ${t.length}).`);let n=0;for(let e of t)e.bonne_reponse&&(n+=1);if(n!==1)throw Error(`Question ${e.id} : exactement une bonne réponse attendue pour l’export.`);return{categorie_id:e.categorie_id,categorie_type:e.categorie_type,question:e.question,commentaire:e.commentaire??``,fakechecker:e.verifier,reponses:t.map(e=>({texte:e.reponse,correcte:e.bonne_reponse}))}}function Oa(e){let t=e.questions.map(Da),n={format:`flowlearn-app-collection-export`,version:1,exportedAt:new Date().toISOString(),collection:{nom:e.nom},questions:t};Ea(`${Ta(e.nom)}-${wa()}.json`,n)}var ka=`Tu produis un JSON STRICT (sans markdown, sans commentaires) pour l’app FlowLearn.
 
 Structure racine :
 - "user_id" (optionnel) : entier — si absent, le premier utilisateur en base est utilisé.
@@ -40,7 +40,7 @@ Exemple minimal :
     }
   ],
   "questions_sans_collection": []
-}`,la=[1,2,3,4,5,6,7,8,9,10,12,15,20,25,30,40,50],ua=`Tu produis UN SEUL JSON valide (sans markdown, sans texte avant ou après).
+}`,Aa=[1,2,3,4,5,6,7,8,9,10,12,15,20,25,30,40,50],ja=`Tu produis UN SEUL JSON valide (sans markdown, sans texte avant ou après).
 
 L’application est déjà positionnée sur UNE collection précise : exporte uniquement une liste de questions. Elles seront toutes enregistrées dans cette collection — n’inclus aucun nom de collection, aucun tableau "collections", aucun "questions_sans_collection".
 
@@ -65,8 +65,8 @@ Exemple minimal :
       ]
     }
   ]
-}`;function da(e){return e.length===0?``:e.map((e,t)=>{let n=e.question.replace(/\s+/g,` `).trim()||`(intitulé vide)`;return`${t+1}. ${n}`}).join(`
-`)}var fa=[`histoire`,`pratique`],pa={histoire:`Questions plutôt « information » : contexte, repères, culture — le « pourquoi / qu’est-ce que ».`,pratique:`Questions sur comment appliquer l’information et dans quels cas l’utiliser — mise en situation et usage.`};function ma(e){return fa.includes(e)}function ha(e){let t=Array.from(new Set(e.map(e=>e.type.trim()).filter(Boolean))),n=[...fa],r=t.filter(e=>!ma(e));return t.length!==n.length||r.length>0?`Attention : la base expose ${t.length} catégorie(s) (${t.join(`, `)||`aucune`}) alors que le frontend en gère ${n.length} (${n.join(`, `)}).`:null}function ga(e){let t=Array.from(new Set(e.map(e=>e.type).filter(ma)));return t.length>0?t:[...fa]}var _a=`categorie`;function va(e,t){let n=t[0]??fa[0],r=typeof e==`string`&&t.includes(e)?e:n;return{id:_a,titre:`Catégorie (enregistrée en base)`,type:`liste_selection`,liste_choix:[...t],value:r}}function ya(e){let t=e.data.options,n=e.actions.onOptionsChange,[r,i]=V(fa),[a,o]=V(null);H(()=>{pt().then(e=>{i(ga(e)),o(ha(e))}).catch(()=>i(fa))},[]),H(()=>{ba(t,r,n)},[t,n,r]);let s=(e,r)=>n(t.map(t=>t.id===e?r(t):t));return{alertes:{avertissementCategories:a},champs:{options:t,onSelectChange:(e,t)=>{s(e,e=>({...e,value:t}))},onTextInput:(e,t)=>{s(e,e=>({...e,value:t}))},onCheckboxToggle:async(e,t)=>{if(!t)return void s(e.id,e=>({...e,value:!1}));let n=await e.function();s(e.id,e=>({...e,value:n}))}}}}function ba(e,t,n){let r=e.find(e=>e.id===_a),i=va(r?.value??null,t);if(r==null)return void n([i,...e]);if(r.type!==`liste_selection`)return;let a=r.value===i.value,o=r.liste_choix.length===i.liste_choix.length&&r.liste_choix.every((e,t)=>e===i.liste_choix[t]);a&&o||n(e.map(e=>e.id===`categorie`?{...e,...i}:e))}var xa={aside:`flex w-full shrink-0 flex-col gap-4 rounded-xl border border-base-content/10 bg-base-100/60 p-3 lg:min-w-70 lg:max-w-xs`,titreSection:`text-[0.65rem] font-semibold uppercase tracking-wide text-base-content/45`,alerteSync:`rounded-lg border border-warning/30 bg-warning/10 px-3 py-2 text-xs leading-snug text-warning-content`,etiquetteChamp:`mb-1 block text-xs font-medium text-base-content/70`,select:`select select-bordered select-sm w-full rounded-lg border-base-content/15 bg-base-100 text-sm`,textarea:`textarea textarea-bordered w-full min-h-20 rounded-lg border-base-content/15 bg-base-100 text-sm`,inputTexte:`input input-bordered input-sm w-full rounded-lg border-base-content/15 bg-base-100 text-sm`,caseACocherLigneBase:`flex items-start gap-2 rounded-lg border border-base-content/10 p-2 text-xs leading-snug`,caseACocherLigneActive:`cursor-pointer`,caseACocherLigneDesactivee:`cursor-not-allowed opacity-50`,caseACocherInput:`checkbox checkbox-sm mt-0.5 shrink-0 border-base-content/30`,caseACocherTitre:`mb-1 block font-medium text-base-content/70`,piedDePanneau:`text-[0.65rem] leading-snug text-base-content/50`};function Sa(e){let t=ya(e),n=t.alertes,r=t.champs;return X(`aside`,{class:xa.aside,children:[X(`p`,{class:xa.titreSection,children:`Options`}),n.avertissementCategories?X(`div`,{class:xa.alerteSync,children:n.avertissementCategories}):null,r.options.map(e=>e.type===`liste_selection`?X(`div`,{children:[X(`label`,{class:xa.etiquetteChamp,for:e.id,children:e.titre}),X(`select`,{id:e.id,class:xa.select,value:String(e.value??``),onChange:t=>r.onSelectChange(e.id,t.target.value),children:e.liste_choix.map(e=>X(`option`,{value:e,children:e},e))})]},e.id):e.type===`texte`?X(`div`,{children:[X(`label`,{class:xa.etiquetteChamp,for:e.id,children:e.titre}),e.multiline?X(`textarea`,{id:e.id,class:xa.textarea,placeholder:e.contenus,value:String(e.value??``),onInput:t=>r.onTextInput(e.id,t.target.value)}):X(`input`,{id:e.id,type:`text`,class:xa.inputTexte,placeholder:e.contenus,value:String(e.value??``),onInput:t=>r.onTextInput(e.id,t.target.value)})]},e.id):X(`label`,{class:ar(xa.caseACocherLigneBase,e.disabled?xa.caseACocherLigneDesactivee:xa.caseACocherLigneActive),children:[X(`input`,{type:`checkbox`,class:xa.caseACocherInput,checked:!!e.value,disabled:e.disabled,onChange:t=>void r.onCheckboxToggle(e,t.target.checked)}),X(`span`,{children:[X(`span`,{class:xa.caseACocherTitre,children:e.titre}),e.description]})]},e.id)),X(`p`,{class:xa.piedDePanneau,children:`Ces champs sont injectés dans le prompt copié pour le LLM.`})]})}var Ca={wrapper:`min-w-0 flex-1`};function wa({prompt:e,importFromJson:t}){let[n,r]=V(``),[i,a]=V(null),[o,s]=V(!1),[c,l]=V(!1),u=async()=>{try{await navigator.clipboard.writeText(e),l(!0),window.setTimeout(()=>l(!1),2e3)}catch{a(`Impossible de copier (permissions du navigateur).`)}},d=async()=>{s(!0),a(null);try{a(await t(n)),r(``)}catch(e){a(e instanceof Error?e.message:`JSON ou réseau invalide.`)}finally{s(!1)}};return X(`div`,{class:Ca.wrapper,children:[X(`div`,{class:`mb-2 flex flex-wrap items-center justify-between gap-2`,children:[X(`p`,{class:`text-sm font-medium text-base-content`,children:`Prompt à copier pour ton LLM`}),X(Z,{variant:`outline`,class:`btn-sm gap-1`,type:`button`,onClick:u,children:[X(zr,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),c?`Copié`:`Copier`]})]}),X(`textarea`,{class:`textarea textarea-bordered mb-3 w-full min-h-28 rounded-2xl border-base-content/15 bg-base-100/80 text-sm transition duration-300`,readOnly:!0,value:e}),X(`p`,{class:`mb-2 text-xs text-base-content/55`,children:e.includes(`"questions"`)&&!e.includes(`"questions_sans_collection"`)?X(b,{children:[`Colle le JSON au format `,X(`code`,{class:`text-xs`,children:`questions`}),` uniquement : chaque entrée sera liée à la collection affichée (aucun autre format requis).`]}):X(b,{children:[`Colle le JSON généré ci-dessous (champs `,X(`code`,{class:`text-xs`,children:`collections`}),` et/ou `,X(`code`,{class:`text-xs`,children:`questions_sans_collection`}),`, ou tableau racine `,X(`code`,{class:`text-xs`,children:`questions`}),`).`]})}),X(`textarea`,{class:`textarea textarea-bordered mb-3 w-full min-h-32 rounded-2xl border-dashed border-learn/35 bg-base-100/60 font-mono text-xs leading-relaxed`,placeholder:e.includes(`"questions"`)&&!e.includes(`"questions_sans_collection"`)?`{ "user_id": 1, "questions": [ { "question": "…", "commentaire": "…", "reponses": [ ... ] } ] }`:`{ "collections": [ ... ], "questions_sans_collection": [] }`,value:n,onInput:e=>r(e.target.value)}),i?X(`p`,{class:`mb-3 text-sm text-base-content/80`,children:i}):null,X(Z,{variant:`flow`,disabled:o||!n.trim(),onClick:d,children:o?`Import…`:`Importer en base`})]})}var Ta={card:`fl-reveal-enter mb-6 border-learn/15 bg-learn/[0.06]`,layout:`flex flex-col gap-4 lg:flex-row lg:items-stretch`};function Ea({options:e,setOptions:t,llmImportWorkflow:n}){let r=n.buildPrompt(e);return X(lr,{class:Ta.card,children:X(`div`,{class:Ta.layout,children:[X(Sa,{data:{options:e},actions:{onOptionsChange:t}}),X(wa,{prompt:r,importFromJson:n.importFromJson})]})})}var Da={header:`mb-6 flex flex-col gap-4 sm:flex-row sm:items-end sm:justify-between`},Oa=`question_count`,ka=`collection_name`,Aa=`subject`,ja=`existing_stems`;function Ma(e,t){let n=e.find(e=>e.id===t);return typeof n?.value==`string`||typeof n?.value==`number`?String(n.value):``}function Na({targetCollectionNumeric:e,collections:t,allModules:n,importTargetModuleId:r,questions:i,onImportSuccess:a}){let[o,s]=V(!1),[c,l]=V(!1),[u,d]=V(null),[f,p]=V([]),m=xe(null),h=xe([]);H(()=>{h.current=f},[f]),H(()=>{let n=Ma(f,ka),r=Ma(f,Aa),a=Ma(f,Oa),o=f.find(e=>e.id===ja)?.value,s=e==null?null:t.find(t=>t.id===e),c=m.current!==e;e==null?m.current=null:c&&(m.current=e),p([{id:Oa,titre:`Nombre de questions`,type:`liste_selection`,liste_choix:la.map(String),value:a||`5`},{id:ka,titre:`Nom de la collection`,type:`texte`,contenus:e==null?`Ex. Ma thematique`:`Prerempli depuis la collection`,value:c&&e!=null?s?.nom??``:n||(e==null?``:s?.nom??``)},{id:Aa,titre:`Sujet des questions`,type:`texte`,contenus:`Ex. verbes irreguliers...`,multiline:!0,value:r},{id:ja,titre:`Eviter les repetitions`,type:`case_a_cocher`,description:e==null?`Disponible lorsque tu filtres sur une collection.`:`Lister les intitules deja en base.`,disabled:e==null,value:e==null?!1:o??!1,function:()=>{let e=Ma(h.current,`categorie`)||`histoire`,t=i.filter(t=>t.categorie_type===e);return t.length>0?da(t):``}}])},[e,t,i]);let g=Se(()=>i=>{let a=Ma(i,Oa)||`5`,o=Ma(i,`categorie`)||`histoire`,s=Ma(i,ka).trim(),c=Ma(i,Aa).trim(),l=Ma(i,ja).trim(),u=o===`pratique`?`pratique`:`histoire`,d=e==null?`\n\n- Quantite: JSON doit representer exactement ${a} questions au total.`:`\n\n- Quantite: "questions" doit contenir exactement ${a}.`,f=s.length>0?`\n\n- Nom de collection: ${s}`:``,p=c.length>0?`\n\n- Sujet:\n${c}`:``,m=`\n\n- Categorie: ${u} - ${pa[u]}`,h=e!=null&&l.length>0?`\n\n- Eviter doublons:\n${l}`:``;if(e==null)return ca+d+f+p+m;let g=t.find(t=>t.id===e)?.nom??`id ${e}`,_=r==null?void 0:n.find(e=>e.id===r),v=`\n\n- Collection active: ${g} (${e}).`;return _&&(v+=`\n- Lien supercollection: ${_.nom} (${_.id}).`),ua+d+f+p+m+h+v},[e,t,r,n]),_=Se(()=>({buildPrompt:g,importFromJson:async t=>{let n=Er(t),i=e??void 0,o=await bt(n,{collectionId:i,moduleId:i!=null&&r!=null?r:void 0,categorie:Ma(f,`categorie`)===`pratique`?`pratique`:`histoire`});return a(),i==null?o.createdCollections>0?`Import reussi: ${o.createdQuestions} question(s), ${o.createdCollections} collection(s).`:`Import reussi: ${o.createdQuestions} question(s).`:`Import reussi: ${o.createdQuestions} question(s) creee(s).`}}),[g,f,e,r,a]),v=()=>{if(e!=null){l(!0),d(null);try{let n=t.find(t=>t.id===e);if(!n)throw Error(`Collection introuvable dans la liste chargee.`);sa(n)}catch(e){d(e instanceof Error?e.message:`Export JSON impossible.`)}finally{l(!1)}}};return X(b,{children:[X(`div`,{class:Da.header,children:[X(`div`,{children:[X(`h1`,{class:`text-2xl font-bold tracking-tight text-base-content sm:text-3xl`,children:`Questions`}),X(`p`,{class:`mt-1 text-sm text-base-content/60`,children:`Modifier ou supprimer via l API backend (Prisma / SQLite).`})]}),X(`div`,{class:`flex flex-col gap-2 self-start sm:flex-row sm:items-center sm:self-auto`,children:[X(Z,{variant:`outline`,class:`gap-2`,disabled:c||e==null,onClick:()=>v(),children:[X(Vr,{class:`h-4 w-4`,"aria-hidden":!0}),c?`Export...`:`Export JSON`]}),X(Z,{variant:`learn`,class:`gap-2`,onClick:()=>s(e=>!e),children:[X(Hr,{class:`h-4 w-4`,"aria-hidden":!0}),`Import LLM`]})]})]}),u?X(`p`,{class:`mb-3 text-xs text-error`,children:u}):null,o?X(Ea,{options:f,setOptions:p,llmImportWorkflow:_}):null]})}function Pa(e){let t=e.trim(),n=t.match(/^```(?:json)?\s*([\s\S]*?)\s*```$/i);return n?n[1].trim():t}function Fa(e){if(typeof e!=`object`||!e)return null;let t=e;return typeof t.texte==`string`&&typeof t.correcte==`boolean`?{texte:t.texte,correcte:t.correcte}:typeof t.reponse==`string`&&typeof t.bonne_reponse==`boolean`?{texte:t.reponse,correcte:t.bonne_reponse}:typeof t.reponse==`string`&&typeof t.correcte==`boolean`?{texte:t.reponse,correcte:t.correcte}:null}function Ia(e,t,n){let r=e.categorie_id;if(typeof r==`number`&&Number.isInteger(r)&&r>=1&&t.some(e=>e.id===r))return r;if(typeof r==`string`&&/^\d+$/.test(r.trim())){let e=Number(r.trim());if(t.some(t=>t.id===e))return e}let i=e.categorie_type;if(typeof i==`string`&&i.trim().length>0){let e=i.trim().toLowerCase(),n=t.find(t=>t.type.trim().toLowerCase()===e);if(n)return n.id}return n!=null&&t.some(e=>e.id===n)?n:t[0]?.id??null}function La(e,t){let n=t.categorieOptions;if(n.length===0)return{ok:!1,error:`Aucune catégorie disponible.`};let r;try{r=JSON.parse(Pa(e))}catch{return{ok:!1,error:`JSON invalide (syntaxe).`}}if(typeof r!=`object`||!r)return{ok:!1,error:`Le JSON doit être un objet.`};let i=r,a=typeof i.question==`string`?i.question.trim():typeof i.enonce==`string`?i.enonce.trim():``;if(a.length===0)return{ok:!1,error:`Champ "question" (ou "enonce") obligatoire et non vide.`};let o=``;typeof i.commentaire==`string`?o=i.commentaire:typeof i.anecdote==`string`&&(o=i.anecdote),o=o.trim();let s=Ia(i,n,t.fallbackCategorieId);if(s==null)return{ok:!1,error:`Impossible de résoudre categorie_id / categorie_type.`};let c=i.reponses;if(!Array.isArray(c))return{ok:!1,error:`Tableau "reponses" obligatoire.`};if(c.length!==4)return{ok:!1,error:`Exactement 4 réponses sont attendues.`};let l=[];for(let e=0;e<4;e+=1){let t=Fa(c[e]);if(t==null)return{ok:!1,error:`Réponse ${e+1} : format attendu { "texte", "correcte" } ou { "reponse", "bonne_reponse" }.`};let n=t.texte.trim();if(n.length===0)return{ok:!1,error:`Réponse ${e+1} : texte vide.`};l.push({texte:n,correcte:t.correcte})}return l.filter(e=>e.correcte).length===1?{ok:!0,value:{question:a,commentaire:o,categorie_id:s,reponses:l}}:{ok:!1,error:`Exactement une réponse doit avoir "correcte" / "bonne_reponse" à true.`}}function Ra(e,t){let n=t.map(e=>`"${e.type}"`).join(`, `),r=t.map(e=>`${e.id} = ${e.type}`).join(` ; `),i=e.trim();return`Tu complètes une question de quiz FlowLearn.
+}`;function Ma(e){return e.length===0?``:e.map((e,t)=>{let n=e.question.replace(/\s+/g,` `).trim()||`(intitulé vide)`;return`${t+1}. ${n}`}).join(`
+`)}var Na=`question_count`,Pa=`collection_name`,Fa=`subject`,Ia=`existing_stems`;function La(e){let{questionCount:t,category:n,collectionName:r,subject:i,existingStems:a,targetCollectionNumeric:o,collections:s,importTargetModuleId:c,allModules:l}=e,u=n===`pratique`?`pratique`:`histoire`,d={countBlock:o==null?`\n\n- Quantite: JSON doit representer exactement ${t} questions au total.`:`\n\n- Quantite: "questions" doit contenir exactement ${t}.`,nameBlock:r.length>0?`\n\n- Nom de collection: ${r}`:``,subjectBlock:i.length>0?`\n\n- Sujet:\n${i}`:``,categorieBlock:`\n\n- Categorie: ${u} - ${oa[u]}`,existingBlock:o!=null&&a.length>0?`\n\n- Eviter doublons:\n${a}`:``};if(o==null)return ka+d.countBlock+d.nameBlock+d.subjectBlock+d.categorieBlock;let f=s.find(e=>e.id===o)?.nom??`id ${o}`,p=c==null?void 0:l.find(e=>e.id===c),m=`\n\n- Collection active: ${f} (${o}).`;return p&&(m+=`\n- Lien supercollection: ${p.nom} (${p.id}).`),ja+d.countBlock+d.nameBlock+d.subjectBlock+d.categorieBlock+d.existingBlock+m}function Ra(e,t){return t?`Import reussi: ${e.createdQuestions} question(s) creee(s).`:e.createdCollections>0?`Import reussi: ${e.createdQuestions} question(s), ${e.createdCollections} collection(s).`:`Import reussi: ${e.createdQuestions} question(s).`}function za(e,t){let n=e.find(e=>e.id===t);return typeof n?.value==`string`||typeof n?.value==`number`?String(n.value):``}function Ba(e){let{data:t,actions:n}=e,{targetCollectionNumeric:r,collections:i,allModules:a,importTargetModuleId:o,questions:s}=t,[c,l]=V(!1),[u,d]=V(!1),[f,p]=V(null),[m,h]=V([]),g=xe(null),_=xe([]);H(()=>{_.current=m},[m]),H(()=>{let e=za(m,Pa),t=za(m,Fa),n=za(m,Na),a=m.find(e=>e.id===Ia)?.value,o=r==null?null:i.find(e=>e.id===r),c=g.current!==r;r==null?g.current=null:c&&(g.current=r),h([{id:Na,titre:`Nombre de questions`,type:`liste_selection`,liste_choix:Aa.map(String),value:n||`5`},{id:Pa,titre:`Nom de la collection`,type:`texte`,contenus:r==null?`Ex. Ma thematique`:`Prerempli depuis la collection`,value:c&&r!=null?o?.nom??``:e||(r==null?``:o?.nom??``)},{id:Fa,titre:`Sujet des questions`,type:`texte`,contenus:`Ex. verbes irreguliers...`,multiline:!0,value:t},{id:Ia,titre:`Eviter les repetitions`,type:`case_a_cocher`,description:r==null?`Disponible lorsque tu filtres sur une collection.`:`Lister les intitules deja en base.`,disabled:r==null,value:r==null?!1:a??!1,action:()=>{let e=za(_.current,`categorie`)||`histoire`,t=s.filter(t=>t.categorie_type===e);return t.length>0?Ma(t):``}}])},[r,i,s]);let v=Se(()=>e=>La({questionCount:za(e,`question_count`)||`5`,category:za(e,`categorie`)||`histoire`,collectionName:za(e,Pa).trim(),subject:za(e,Fa).trim(),existingStems:za(e,Ia).trim(),targetCollectionNumeric:r,collections:i,importTargetModuleId:o,allModules:a}),[r,i,o,a]);return{state:{importOpen:c,exportBusy:u,exportError:f,options:m,llmImportWorkflow:Se(()=>({buildPrompt:v,importFromJson:async e=>{let t=gi(e),i=r??void 0,a=await xt(t,{collectionId:i,moduleId:i!=null&&o!=null?o:void 0,categorie:za(m,`categorie`)===`pratique`?`pratique`:`histoire`});return n.onImportSuccess(),Ra(a,i!=null)}}),[v,m,r,o,n])},actions:{setOptions:h,handleExportCollectionJson:()=>{if(r!=null){d(!0),p(null);try{let e=i.find(e=>e.id===r);if(!e)throw Error(`Collection introuvable dans la liste chargee.`);Oa(e)}catch(e){p(e instanceof Error?e.message:`Export JSON impossible.`)}finally{d(!1)}}},toggleImport:()=>l(e=>!e)}}}function Va(e){let{data:t}=e,{state:n,actions:r}=Ba(e);return X(b,{children:[X(`div`,{class:Sa.header,children:[X(`div`,{children:[X(`h1`,{class:`text-2xl font-bold tracking-tight text-base-content sm:text-3xl`,children:`Questions`}),X(`p`,{class:`mt-1 text-sm text-base-content/60`,children:`Modifier ou supprimer via l API backend (Prisma / SQLite).`})]}),X(`div`,{class:`flex flex-col gap-2 self-start sm:flex-row sm:items-center sm:self-auto`,children:[X(Z,{variant:`outline`,class:`gap-2`,disabled:n.exportBusy||t.targetCollectionNumeric==null,onClick:r.handleExportCollectionJson,children:[X(Nr,{class:`h-4 w-4`,"aria-hidden":!0}),n.exportBusy?`Export...`:`Export JSON`]}),X(Z,{variant:`learn`,class:`gap-2`,onClick:r.toggleImport,children:[X(Pr,{class:`h-4 w-4`,"aria-hidden":!0}),`Import LLM`]})]})]}),n.exportError&&X(`p`,{class:`mb-3 text-xs text-error`,children:n.exportError}),n.importOpen&&X(xa,{data:{options:n.options,llmImportWorkflow:n.llmImportWorkflow},actions:{onOptionsChange:r.setOptions}})]})}function Ha(e){let t=e.trim(),n=t.match(/^```(?:json)?\s*([\s\S]*?)\s*```$/i);return n?n[1].trim():t}function Ua(e){if(typeof e!=`object`||!e)return null;let t=e;return typeof t.texte==`string`&&typeof t.correcte==`boolean`?{texte:t.texte,correcte:t.correcte}:typeof t.reponse==`string`&&typeof t.bonne_reponse==`boolean`?{texte:t.reponse,correcte:t.bonne_reponse}:typeof t.reponse==`string`&&typeof t.correcte==`boolean`?{texte:t.reponse,correcte:t.correcte}:null}function Wa(e,t,n){let r=e.categorie_id;if(typeof r==`number`&&Number.isInteger(r)&&r>=1&&t.some(e=>e.id===r))return r;if(typeof r==`string`&&/^\d+$/.test(r.trim())){let e=Number(r.trim());if(t.some(t=>t.id===e))return e}let i=e.categorie_type;if(typeof i==`string`&&i.trim().length>0){let e=i.trim().toLowerCase(),n=t.find(t=>t.type.trim().toLowerCase()===e);if(n)return n.id}return n!=null&&t.some(e=>e.id===n)?n:t[0]?.id??null}function Ga(e,t){let n=t.categorieOptions;if(n.length===0)return{ok:!1,error:`Aucune catégorie disponible.`};let r;try{r=JSON.parse(Ha(e))}catch{return{ok:!1,error:`JSON invalide (syntaxe).`}}if(typeof r!=`object`||!r)return{ok:!1,error:`Le JSON doit être un objet.`};let i=r,a=typeof i.question==`string`?i.question.trim():typeof i.enonce==`string`?i.enonce.trim():``;if(a.length===0)return{ok:!1,error:`Champ "question" (ou "enonce") obligatoire et non vide.`};let o=``;typeof i.commentaire==`string`?o=i.commentaire:typeof i.anecdote==`string`&&(o=i.anecdote),o=o.trim();let s=Wa(i,n,t.fallbackCategorieId);if(s==null)return{ok:!1,error:`Impossible de résoudre categorie_id / categorie_type.`};let c=i.reponses;if(!Array.isArray(c))return{ok:!1,error:`Tableau "reponses" obligatoire.`};if(c.length!==4)return{ok:!1,error:`Exactement 4 réponses sont attendues.`};let l=[];for(let e=0;e<4;e+=1){let t=Ua(c[e]);if(t==null)return{ok:!1,error:`Réponse ${e+1} : format attendu { "texte", "correcte" } ou { "reponse", "bonne_reponse" }.`};let n=t.texte.trim();if(n.length===0)return{ok:!1,error:`Réponse ${e+1} : texte vide.`};l.push({texte:n,correcte:t.correcte})}return l.filter(e=>e.correcte).length===1?{ok:!0,value:{question:a,commentaire:o,categorie_id:s,reponses:l}}:{ok:!1,error:`Exactement une réponse doit avoir "correcte" / "bonne_reponse" à true.`}}function Ka(e,t){let n=t.map(e=>`"${e.type}"`).join(`, `),r=t.map(e=>`${e.id} = ${e.type}`).join(` ; `),i=e.trim();return`Tu complètes une question de quiz FlowLearn.
 
 L’auteur a écrit UNE SEULE FOIS un brouillon d’intention (idée, phrase brute ou ébauche). Tu ne dois pas le recopier tel quel dans le JSON : tu le transformes en une vraie question de QCM claire, fluide et correcte grammaticalement.
 
@@ -101,4 +101,4 @@ Contraintes :
 
 Référence des id (tu peux utiliser "categorie_id": nombre à la place de "categorie_type" si tu préfères) : ${r}
 
-Alternative acceptée pour chaque réponse : { "reponse": "...", "bonne_reponse": true|false } au lieu de texte/correcte.`}function za(){return[{texte:``,correcte:!0},{texte:``,correcte:!1},{texte:``,correcte:!1},{texte:``,correcte:!1}]}function Ba(e){let t=e.settings,n=e.actions,r=e.status,i=e.data,a=e.drafts,o=t.variant??`edit`,s=i.categorieOptions,[c,l]=V(null),[u,d]=V(``),[f,p]=V(!1),[m,h]=V(null),[g,_]=V(za),[v,y]=V(null),[b,x]=V(``),[S,C]=V(null),w=o===`create`,T=t.modalTitle??(w?`Nouvelle question`:`Modifier la question`);return H(()=>{t.open&&w&&(_(za()),y(null),x(``),C(null))},[t.open,w]),{dialogue:{title:T,isCreate:w,onBackdropClick:()=>{!r.saving&&!f&&t.onClose()}},creation:{submitCreate:async()=>{if(y(null),a.categorieId==null)return void y(`Choisis une catégorie.`);let e=a.question.trim();if(!e)return void y(`L’énoncé ne peut pas être vide.`);let t=g.map(e=>({...e,texte:e.texte.trim()}));if(t.some(e=>!e.texte))return void y(`Toutes les propositions doivent être renseignées.`);if(t.filter(e=>e.correcte).length!==1)return void y(`Indique exactement une bonne réponse.`);if(!n.onCreateSave)return void y(`Configuration interne : onCreateSave manquant.`);await Promise.resolve(n.onCreateSave({question:e,commentaire:a.commentaire.trim(),categorie_id:a.categorieId,reponses:t}))},createFormError:v,createReponses:g,setCreateReponseCorrectAt:e=>{_(t=>t.map((t,n)=>({...t,correcte:n===e})))},setCreateReponseTexteAt:(e,t)=>{_(n=>n.map((n,r)=>r===e?{...n,texte:t}:n))}},llm:{llmHint:S,llmJsonDraft:b,setLlmJsonDraft:x,applyLlmJsonPaste:()=>{let e=La(b,{categorieOptions:s,fallbackCategorieId:a.categorieId});if(!e.ok)return void C({tone:`err`,text:e.error});n.onDraftQuestion(e.value.question),n.onDraftCommentaire(e.value.commentaire),n.onDraftCategorieId(e.value.categorie_id),_(e.value.reponses),x(``),C({tone:`ok`,text:`Champs mis à jour à partir du JSON.`})},copyLlmPrompt:()=>{navigator.clipboard.writeText(Ra(a.question,s))}},editionReponses:{editingReponseId:c,reponseDraft:u,setReponseDraft:d,reponseBusy:f,reponseError:m,submitReponse:async()=>{if(c==null)return;let e=u.trim();if(!e)return void h(`Le texte ne peut pas être vide.`);p(!0),h(null);try{await vt(c,{reponse:e}),await Promise.resolve(n.onReponseUpdated()),l(null),d(``)}catch{h(`Enregistrement impossible (réseau ou serveur).`)}finally{p(!1)}},beginEditReponse:(e,t)=>{l(e),d(t)}}}}var Va={overlay:`fixed inset-0 z-100 flex items-center justify-center p-4`,backdrop:`absolute inset-0 border-0 bg-base-content/40 p-0`,dialog:`relative z-10 max-h-[90vh] w-full max-w-2xl overflow-y-auto rounded-2xl border border-base-content/10 bg-base-100 p-6 shadow-2xl`};function Ha(e){let t=e.settings,n=e.actions,r=e.status,i=e.data,a=e.drafts,o=Ba(e),s=o.dialogue,c=o.creation,l=o.llm,u=o.editionReponses;return t.open?X(`div`,{class:Va.overlay,role:`presentation`,children:[X(`button`,{type:`button`,class:Va.backdrop,"aria-label":`Fermer`,onClick:s.onBackdropClick}),X(`div`,{class:Va.dialog,role:`dialog`,"aria-modal":`true`,"aria-labelledby":`qm-title`,onClick:e=>e.stopPropagation(),children:[X(`h2`,{id:`qm-title`,class:`mb-4 text-lg font-semibold text-base-content`,children:s.title}),s.isCreate?X(b,{children:[c.createFormError?X(`p`,{class:`mb-3 text-sm text-error`,children:c.createFormError}):null,X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`qm-categorie`,children:`Catégorie`}),X(`select`,{id:`qm-categorie`,class:`select select-bordered select-sm mb-4 w-full max-w-md rounded-xl border-base-content/15 bg-base-100`,value:a.categorieId??``,disabled:r.saving,onChange:e=>n.onDraftCategorieId(Number(e.target.value)),children:i.categorieOptions.map(e=>X(`option`,{value:e.id,children:e.type},e.id))}),X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`qm-question`,children:`Énoncé`}),X(`textarea`,{id:`qm-question`,class:`textarea textarea-bordered mb-4 w-full rounded-xl border-base-content/15 text-sm`,rows:4,value:a.question,onInput:e=>n.onDraftQuestion(e.target.value)}),X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`qm-commentaire`,children:`Commentaire`}),X(`textarea`,{id:`qm-commentaire`,class:`textarea textarea-bordered mb-4 w-full rounded-xl border-base-content/15 text-sm`,rows:3,value:a.commentaire,onInput:e=>n.onDraftCommentaire(e.target.value)}),X(`div`,{class:`mb-4 rounded-xl border border-learn/20 bg-learn/6 p-4`,children:[l.llmHint?X(`p`,{class:`mb-3 text-sm ${l.llmHint.tone===`ok`?`text-flow`:`text-error`}`,children:l.llmHint.text}):null,X(Z,{variant:`outline`,class:`btn-sm gap-1 mb-2`,type:`button`,disabled:r.saving,onClick:l.copyLlmPrompt,children:[X(zr,{class:`h-3.5 w-3.5 shrink-0`,"aria-hidden":!0}),`Copier prompt pour le LLM`]}),X(`textarea`,{id:`qm-llm-json`,class:`textarea textarea-bordered mb-2 w-full rounded-xl border-base-content/15 font-mono text-xs`,rows:5,value:l.llmJsonDraft,onInput:e=>l.setLlmJsonDraft(e.target.value)}),X(Z,{variant:`learn`,class:`btn-sm`,type:`button`,disabled:r.saving,onClick:l.applyLlmJsonPaste,children:`Appliquer le JSON`})]}),X(`ul`,{class:`mb-6 space-y-3 rounded-xl border border-base-content/10 bg-base-200/30 p-3`,children:c.createReponses.map((e,t)=>X(`li`,{children:[X(`div`,{class:`mb-1 flex items-center gap-2`,children:[X(`input`,{type:`radio`,name:`qm-bonne-create`,class:`radio radio-sm radio-primary`,checked:e.correcte,onChange:()=>c.setCreateReponseCorrectAt(t)}),X(`span`,{class:`text-xs text-base-content/60`,children:e.correcte?`Bonne réponse`:`Proposition ${t+1}`})]}),X(`textarea`,{class:`textarea textarea-bordered w-full rounded-xl border-base-content/15 text-sm`,rows:2,value:e.texte,onInput:e=>{c.setCreateReponseTexteAt(t,e.target.value)}})]},t))}),X(`div`,{class:`mt-0 flex flex-wrap justify-end gap-2 border-t border-base-content/10 pt-4`,children:[X(Z,{variant:`ghost`,class:`btn-sm`,disabled:r.saving,onClick:t.onClose,children:`Annuler`}),X(Z,{variant:`flow`,class:`btn-sm`,disabled:r.saving,onClick:()=>void c.submitCreate(),children:r.saving?`Création…`:`Créer`})]})]}):r.error==null?r.loading||i.questionDetail==null?X(`div`,{class:`py-8 text-center`,children:[X(`p`,{class:`mb-4 text-sm text-base-content/60`,children:`Chargement…`}),X(Z,{variant:`ghost`,class:`btn-sm`,onClick:t.onClose,children:`Annuler`})]}):X(b,{children:[X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`qe-question`,children:`Énoncé`}),X(`textarea`,{id:`qe-question`,class:`textarea textarea-bordered mb-4 w-full rounded-xl border-base-content/15 text-sm`,rows:4,value:a.question,onInput:e=>n.onDraftQuestion(e.target.value)}),X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`qe-commentaire`,children:`Commentaire`}),X(`textarea`,{id:`qe-commentaire`,class:`textarea textarea-bordered mb-4 w-full rounded-xl border-base-content/15 text-sm`,rows:3,value:a.commentaire,onInput:e=>n.onDraftCommentaire(e.target.value)}),u.reponseError?X(`p`,{class:`mb-2 text-xs text-error`,children:u.reponseError}):null,X(`ul`,{class:`mb-6 space-y-3 rounded-xl border border-base-content/10 bg-base-200/30 p-3`,children:i.questionDetail.reponses.map(e=>X(`li`,{children:u.editingReponseId===e.id?X(`div`,{class:`flex gap-2`,children:[X(`textarea`,{class:`textarea textarea-bordered w-full rounded-xl border-base-content/15 text-sm`,rows:2,value:u.reponseDraft,onInput:e=>u.setReponseDraft(e.target.value)}),X(Z,{variant:`flow`,class:`btn-xs`,disabled:u.reponseBusy,onClick:()=>void u.submitReponse(),children:u.reponseBusy?`…`:`Valider`})]}):X(`div`,{class:`flex items-start justify-between gap-2`,children:[X(`span`,{children:[e.bonne_reponse?`✓ `:`○ `,e.reponse]}),X(Z,{variant:`outline`,class:`btn-xs`,onClick:()=>u.beginEditReponse(e.id,e.reponse),children:`Modifier`})]})},e.id))}),X(`div`,{class:`mt-0 flex flex-wrap justify-end gap-2 border-t border-base-content/10 pt-4`,children:[X(Z,{variant:`ghost`,class:`btn-sm`,onClick:t.onClose,children:`Annuler`}),X(Z,{variant:`flow`,class:`btn-sm`,disabled:r.saving,onClick:n.onSave,children:r.saving?`Enregistrement…`:`Enregistrer`})]})]}):X(`div`,{class:`py-4 text-center`,children:[X(`p`,{class:`mb-4 text-sm text-error`,children:`Impossible de charger le détail.`}),X(Z,{variant:`ghost`,class:`btn-sm`,onClick:t.onClose,children:`Fermer`})]})]})]}):null}var Ua={root:`flex min-h-dvh flex-col`};function Wa(e){if(e===``)return;if(e===`none`)return`none`;let t=Number(e);return Number.isFinite(t)?t:void 0}function Ga(e){return e&&/^\d+$/.test(e)?e:``}function Ka({collectionId:e}){let[t,n]=V([]),[r,i]=V([]),[a,o]=V(()=>Ga(e)),[s,c]=V(null),[l,u]=V([]),[d,f]=V(!0),[p,m]=V(null),[h,g]=V([]),[_,v]=V(!1),[y,b]=V(!1),[x,S]=V(null),[C,w]=V(null),[T,E]=V(``),[ee,D]=V(``),[O,k]=V(null),[A,j]=V(!1),[M,te]=V(`melanger`),ne=a!==``&&a!==`none`&&/^\d+$/.test(a)?Number(a):null,N=Se(()=>M===`melanger`?l:l.filter(e=>e.categorie_type===M),[l,M]);H(()=>{o(Ga(e))},[e]),H(()=>{if(typeof window>`u`)return;let e=new URLSearchParams(window.location.search).get(`module`);e&&/^\d+$/.test(e)?c(Number(e)):c(null)},[e]);let re=()=>{f(!0),m(null),ht(Wa(a)).then(u).catch(()=>m(`fetch`)).finally(()=>f(!1))};H(()=>{it().then(n).catch(()=>{}),ot().then(i).catch(()=>{}),pt().then(g).catch(()=>{})},[]),H(()=>{re()},[a]);let ie=e=>{if(o(e),e===``||e===`none`)return void U(`/questions`);/^\d+$/.test(e)&&U(`/questions/${e}${s==null?``:`?module=${s}`}`)},P=e=>{v(!0),b(!0),S(null),w(null),mt(e.id).then(e=>{w(e),E(e.question),D(e.commentaire),k(e.categorie_id)}).catch(()=>S(`fetch`)).finally(()=>b(!1))},F=()=>{v(!1),b(!1),S(null),w(null)},I=async()=>{if(C!=null)try{w(await mt(C.id))}catch{}},L=async()=>{if(C!=null){j(!0);try{let e={};if(T!==C.question&&(e.question=T),ee!==C.commentaire&&(e.commentaire=ee),O!=null&&O!==C.categorie_id&&(e.categorie_id=O),Object.keys(e).length===0){F();return}let t=await gt(C.id,e);u(e=>e.map(e=>e.id===t.id?t:e)),F()}catch{m(`save`)}finally{j(!1)}}};return X(`div`,{class:Ua.root,children:[X(li,{}),X(pi,{children:[X(Na,{targetCollectionNumeric:ne,collections:t,allModules:r,importTargetModuleId:s,questions:l,onImportSuccess:()=>{re(),it().then(n).catch(()=>{})}}),p?X(`div`,{class:`mb-6 rounded-box border border-error/20 bg-error/5 px-4 py-3 text-sm text-base-content`,children:[`Une operation a echoue.`,` `,X(`button`,{type:`button`,class:`link link-primary`,onClick:()=>m(null),children:`Fermer`})]}):null,X($i,{targetCollectionNumeric:ne,collections:t,allModules:r,importTargetModuleId:s,setImportTargetModuleId:c}),X(`div`,{class:`mb-4 flex flex-col gap-4 sm:flex-row sm:flex-wrap sm:items-end`,children:[X(`div`,{class:`flex min-w-0 flex-1 flex-col gap-2 sm:max-w-xs`,children:[X(`label`,{class:`text-sm font-medium text-base-content/80`,for:`q-collection-filter`,children:`Filtrer par collection`}),X(`select`,{id:`q-collection-filter`,class:`select select-bordered select-sm w-full rounded-xl border-base-content/15 bg-base-100`,value:a,onChange:e=>ie(e.target.value),children:[X(`option`,{value:``,children:`Toutes les questions`}),X(`option`,{value:`none`,children:`Sans collection`}),t.map(e=>X(`option`,{value:String(e.id),children:e.nom},e.id))]})]}),X(`div`,{class:`flex min-w-0 flex-col gap-2 sm:w-44`,children:[X(`label`,{class:`text-sm font-medium text-base-content/80`,for:`q-list-qtype-filter`,children:`Filtrer par type (affichage)`}),X(`select`,{id:`q-list-qtype-filter`,class:`select select-bordered select-sm w-full rounded-xl border-base-content/15 bg-base-100`,value:M,onChange:e=>{let t=e.target.value;(t===`histoire`||t===`pratique`||t===`melanger`)&&te(t)},children:[X(`option`,{value:`melanger`,children:`Tout (histoire + pratique)`}),X(`option`,{value:`histoire`,children:`Histoire seulement`}),X(`option`,{value:`pratique`,children:`Pratique seulement`})]})]})]}),d?X(`p`,{class:`text-sm text-base-content/60`,children:`Chargement...`}):p===`fetch`?X(lr,{class:`border-base-content/15`,children:[X(`p`,{class:`mb-3 text-sm text-base-content/70`,children:`Impossible de charger les questions.`}),X(Z,{variant:`flow`,class:`btn-sm`,onClick:re,children:`Reessayer`})]}):X(ta,{questions:N,saving:A,onEdit:P,onRemove:async e=>{j(!0);try{await St(e),u(t=>t.filter(t=>t.id!==e)),C?.id===e&&F()}catch{m(`delete`)}finally{j(!1)}}}),X(Ha,{settings:{open:_,onClose:F},actions:{onSave:()=>void L(),onDraftQuestion:E,onDraftCommentaire:D,onDraftCategorieId:k,onReponseUpdated:()=>void I()},status:{loading:y,saving:A,error:x},data:{questionDetail:C,categorieOptions:h},drafts:{question:T,commentaire:ee,categorieId:O}})]}),X(di,{})]})}var qa=`flowlearn_last_result`;function Ja(e){if(e.playOrders!=null&&Array.isArray(e.playOrders)){let t=e.playOrders.filter(e=>typeof e==`string`&&bi(e));if(t.length>0)return t}let t=e.playOrder;if(typeof t==`string`){if(t.includes(`,`))return xi(t);if(bi(t))return[t]}}function Ya(e){try{sessionStorage.setItem(qa,JSON.stringify(e))}catch{}}function Xa(){try{let e=sessionStorage.getItem(qa);if(!e)return null;let t=JSON.parse(e),n=Ja(t);if(t.mode===`random`||t.mode===`collection`){let e=t;return n==null?e:{...e,playOrders:n}}return t.collectionName!=null&&typeof t.good==`number`&&typeof t.total==`number`?{mode:`collection`,collectionId:t.collectionId??null,collectionName:t.collectionName,good:t.good,total:t.total,playOrders:n??void 0,playQtype:t.playQtype,playInfinite:t.playInfinite}:null}catch{return null}}var Za={root:`flex min-h-dvh flex-col`,main:`fl-page-enter mx-auto flex w-full max-w-lg flex-1 flex-col items-center justify-center px-4 py-10`};function Qa(){let{userId:e}=Nt(),t=Se(()=>Xa(),[]);if(!t)return X(`div`,{class:Za.root,children:[X(li,{}),X(`main`,{class:`fl-page-enter mx-auto flex max-w-lg flex-1 flex-col items-center justify-center gap-4 px-4 py-12 text-center`,children:[X(`p`,{class:`text-base text-base-content/70`,children:`Aucun resultat recent. Lance un quiz depuis l accueil ou une collection.`}),X(Z,{variant:`flow`,onClick:()=>U(`/`),children:`Accueil`})]}),X(di,{})]});let n=t.total<=0?0:Math.round(t.good/t.total*100),r=()=>{let n=t.playOrders!=null&&t.playOrders.length>0?t.playOrders:t.playOrder!=null&&bi(t.playOrder)?[t.playOrder]:[`random`],r=ki({orders:n,qtype:t.playQtype??`melanger`,infinite:t.playInfinite===!0,userId:Si(n)?e:void 0});t.mode===`random`?U(`/play/random${r}`):t.collectionId==null?U(`/collections`):U(`/play/${t.collectionId}${r}`)};return X(`div`,{class:Za.root,children:[X(li,{}),X(`main`,{class:Za.main,children:X(lr,{class:`w-full text-center transition duration-300`,children:[X(`div`,{class:`mx-auto mb-4 flex h-14 w-14 items-center justify-center rounded-full bg-gradient-to-br from-flow to-learn text-white shadow-xl shadow-flow/25 transition duration-300 hover:scale-105`,children:X(ni,{class:`h-7 w-7`,"aria-hidden":!0})}),X(Fi,{tone:`flow`,class:`mb-3`,children:`Session terminee`}),X(`h1`,{class:`mb-2 text-2xl font-bold tracking-tight text-base-content`,children:`Resultat`}),X(`p`,{class:`mb-6 text-sm text-base-content/65`,children:t.collectionName}),X(`p`,{class:`mb-1 text-4xl font-semibold text-flow`,children:[t.good,`/`,t.total]}),X(`p`,{class:`mb-8 text-sm text-base-content/55`,children:[n,`% de bonnes reponses`]}),X(`div`,{class:`flex flex-col gap-2 sm:flex-row sm:justify-center`,children:[X(Z,{variant:`flow`,onClick:r,children:`Rejouer`}),X(Z,{variant:`outline`,onClick:()=>U(`/collections`),children:`Collections`}),X(Z,{variant:`ghost`,onClick:()=>U(`/`),children:`Accueil`})]})]})}),X(di,{})]})}function $a(e,t,n){return e[t]?.reponses.some(e=>e.id===n&&e.bonne_reponse)??!1}function eo(e){return JSON.stringify({question:e.question,commentaire:e.commentaire,reponses:e.reponses.map(e=>({reponse:e.reponse,bonne_reponse:e.bonne_reponse}))},null,2)}function to(e,t){return t<=0?0:Math.min(100,Math.round(e/t*100))}var no={wrapper:`w-full`,progress:`flowlearn-progress progress h-3 w-full rounded-full bg-base-200/80`,footer:`mt-1 flex justify-between text-xs text-base-content/60`,value:`font-medium text-flow`};function ro({value:e,max:t,class:n}){let r=to(e,t);return X(`div`,{class:ar(no.wrapper,n),children:[X(`progress`,{class:no.progress,value:e,max:t}),X(`div`,{class:no.footer,children:[X(`span`,{children:`Progression`}),X(`span`,{class:no.value,children:[r,`%`]})]})]})}function io(e){let t=e.pickedId===e.reponseId;return{isPicked:t,showCorrect:e.revealed&&e.isCorrectAnswer,showWrongPick:e.revealed&&t&&!e.isCorrectAnswer,dimmed:e.revealed&&!e.isCorrectAnswer&&!t}}var ao={base:`w-full rounded-full border-2 px-5 py-3.5 text-left text-sm font-medium transition-all duration-300 ease-[cubic-bezier(0.22,1,0.36,1)]`,idle:`border-base-content/10 bg-base-200/45 hover:border-flow/40 hover:bg-flow/[0.07] hover:scale-[1.02]`,dimmed:`scale-[0.98] border-transparent bg-base-200/25 opacity-50`,correct:`border-flow/45 bg-flow/12 text-base-content shadow-md shadow-flow/10 ring-2 ring-flow/20`,wrong:`border-learn/40 bg-learn/10 text-base-content ring-2 ring-learn/15`};function oo({label:e,reponseId:t,pickedId:n,revealed:r,isCorrectAnswer:i,disabled:a,onPick:o}){let{dimmed:s,showCorrect:c,showWrongPick:l}=io({reponseId:t,pickedId:n,revealed:r,isCorrectAnswer:i});return X(`button`,{type:`button`,disabled:a,onClick:o,class:ar(ao.base,!r&&ao.idle,s&&ao.dimmed,c&&ao.correct,l&&ao.wrong),children:e})}var $={pageShell:`flex min-h-dvh flex-col`,centeredMain:`fl-page-enter mx-auto flex max-w-lg flex-1 flex-col items-center justify-center gap-4 px-4 py-12 text-center`,contentMain:`fl-page-enter mx-auto w-full max-w-2xl flex-1 px-4 py-6 md:py-8`,topRow:`mb-6 flex items-center gap-2`,badgesRow:`flex flex-wrap items-center gap-2`,orderBadgeWrap:`max-w-[min(100%,18rem)] truncate`,actionMessage:`mb-3 rounded-xl border border-flow/20 bg-flow/5 px-3 py-2 text-center text-sm text-base-content`,bodyLayout:`flex flex-col gap-4 md:flex-row md:items-stretch`,aside:`flex shrink-0 flex-row flex-wrap gap-2 md:w-36 md:flex-col md:items-stretch`,actionButton:`btn-sm flex-1 gap-1 sm:flex-none`,actionButtonText:`hidden sm:inline`,verifierToggleBase:`btn btn-sm mt-1 flex h-auto min-h-0 w-full cursor-pointer flex-nowrap items-center justify-start gap-2 rounded-full border-2 py-2.5 pl-3 pr-4 text-left text-sm font-medium shadow-md transition-all duration-300 ease-out hover:shadow-lg active:scale-[0.97] md:mt-0`,verifierToggleOn:`border-flow/40 bg-transparent text-flow hover:border-flow hover:bg-flow/5`,verifierToggleOff:`border-error/50 bg-error/10 text-error hover:border-error hover:bg-error/15`,card:`min-w-0 flex-1 transition duration-300`,questionMeta:`mb-4 text-xs font-medium uppercase tracking-wide text-base-content/45`,questionTitle:`mb-6 text-lg font-semibold leading-snug text-base-content sm:text-xl`,answers:`flex flex-col gap-2.5`,revealBox:`fl-reveal-enter mt-8 space-y-5 rounded-[1.75rem] border border-base-content/8 bg-gradient-to-b from-base-100/95 to-base-200/40 p-6 shadow-inner`,revealText:`text-center text-base leading-relaxed text-base-content/85`,revealActions:`flex flex-col items-center justify-center gap-2 sm:flex-row`,nextButton:`min-w-[11rem] px-8`};function so(){return X(`div`,{class:$.pageShell,children:[X(li,{}),X(`main`,{class:$.centeredMain,children:X(`p`,{class:`text-base text-base-content/70`,children:`Chargement du quiz…`})}),X(di,{})]})}function co({loadError:e}){return X(`div`,{class:$.pageShell,children:[X(li,{}),X(`main`,{class:$.centeredMain,children:[X(`p`,{class:`text-lg font-medium text-base-content`,children:`Parcours introuvable`}),e===`fetch`?X(`p`,{class:`text-sm text-base-content/60`,children:`Impossible de joindre l’API. Le backend est-il démarré ?`}):null,e===`empty`?X(`p`,{class:`text-sm text-base-content/60`,children:`Aucune question avec ce filtre (essaie « Mélanger » ou un autre type) ou collection vide.`}):null,X(Z,{variant:`flow`,onClick:()=>U(`/`),children:`Accueil`})]}),X(di,{})]})}function lo({data:e,backTarget:t}){return X(`div`,{class:$.topRow,children:[X(Z,{variant:`ghost`,class:`btn-sm gap-1 px-3`,onClick:()=>U(t),children:[X(Ir,{class:`h-4 w-4`,"aria-hidden":!0}),`Retour`]}),X(`div`,{class:$.badgesRow,children:[X(Fi,{tone:e.mode===`random`?`flow`:`learn`,children:e.nom}),X(`span`,{class:$.orderBadgeWrap,title:Ni(e.playOrders),children:X(Fi,{tone:`learn`,class:`font-normal opacity-90`,children:Ni(e.playOrders)})}),X(Fi,{tone:`flow`,class:`font-normal opacity-90`,children:ji(e.playQtype)}),e.playInfinite?X(Fi,{tone:`learn`,class:`font-normal opacity-90`,children:`Session infinie`}):null]})]})}function uo({data:e,index:t,total:n,q:r,pickedId:i,revealed:a,anecdote:o,correct:s,draftVerifier:c,nextBusy:l,fetchingMore:u,onPick:d,onOpenCreateLinkedQuestionModal:f,onOpenEditQuestionModal:p,onCopyCurrentQuestionJson:m,onDraftVerifier:h,onNext:g,onEndInfiniteSession:_}){return X(`div`,{class:$.bodyLayout,children:[X(`aside`,{class:$.aside,"aria-label":`Actions sur la question`,children:[X(Z,{variant:`outline`,class:$.actionButton,type:`button`,title:`Ajouter une question liée à celle-ci (relation parent → enfant)`,onClick:()=>f(r),children:[X(Xr,{class:`h-4 w-4 shrink-0`,"aria-hidden":!0}),X(`span`,{class:$.actionButtonText,children:`Ajouter`})]}),X(Z,{variant:`outline`,class:$.actionButton,type:`button`,title:`Modifier la question affichée`,onClick:()=>p(r),children:[X(Yr,{class:`h-4 w-4 shrink-0`,"aria-hidden":!0}),X(`span`,{class:$.actionButtonText,children:`Modifier`})]}),X(Z,{variant:`outline`,class:$.actionButton,type:`button`,title:`Copier un JSON (question, commentaire, réponses)`,onClick:()=>void m(r),children:[X(zr,{class:`h-4 w-4 shrink-0`,"aria-hidden":!0}),X(`span`,{class:$.actionButtonText,children:`Copier`})]}),X(`label`,{"aria-label":`Fake-checker, ${c?`oui`:`non`}. Cliquer pour basculer.`,class:ar($.verifierToggleBase,c?$.verifierToggleOn:$.verifierToggleOff,l&&`pointer-events-none opacity-50`),children:[X(`input`,{type:`checkbox`,class:`sr-only`,checked:c,disabled:l,onChange:e=>h(e.target.checked)}),X(`span`,{class:`pointer-events-none select-none`,children:[`verifier : `,c?`Oui`:`Non`]})]})]}),X(lr,{class:$.card,children:[X(`p`,{class:$.questionMeta,children:[`Question `,t+1,` / `,n]}),X(`h2`,{class:$.questionTitle,children:r.question}),X(`div`,{class:$.answers,children:r.reponses.map(e=>X(oo,{label:e.reponse,reponseId:e.id,pickedId:i,revealed:a,isCorrectAnswer:e.bonne_reponse,disabled:i!=null,onPick:()=>d(e.id)},e.id))}),a?X(`div`,{class:$.revealBox,children:[X(`p`,{class:$.revealText,children:o?X(`span`,{class:`block`,children:o}):s?X(b,{children:[`Bien vu — `,X(`span`,{class:`font-semibold text-flow`,children:`c’est la bonne réponse`}),`.`]}):X(b,{children:`Ce n’était pas la bonne proposition.`})}),X(`div`,{class:$.revealActions,children:[X(Z,{variant:`flow`,class:$.nextButton,disabled:l||u,onClick:g,children:l?`Enregistrement…`:u?`Chargement…`:t>=n-1?e.playInfinite?`Nouvelles questions`:`Voir le résultat`:`Suivant`}),e.playInfinite&&t>=n-1?X(Z,{variant:`outline`,class:$.nextButton,disabled:l||u,onClick:_,children:`Terminer la session`}):null]})]}):null]})]})}function fo({playInfinite:e,progressValue:t,total:n}){return e?null:X(ro,{value:t,max:n,class:`mb-6`})}function po({collectionId:e}){let{userId:t}=Nt(),n=_r(),r=xe(0),i=xe([]),a=xe(0),[o,s]=V(null),[c,l]=V(!0),[u,d]=V(null),[f,p]=V(0),[m,h]=V(null),[g,_]=V(0),[v,y]=V([]),[b,x]=V(!1),[S,C]=V(`edit`),[w,T]=V(null),[E,ee]=V(!1),[D,O]=V(null),[k,A]=V(null),[j,M]=V(``),[te,ne]=V(``),[N,re]=V(null),[ie,P]=V(!1),[F,I]=V(null),[L,ae]=V(!1),[oe,R]=V(!1),[se,z]=V(!1);H(()=>{pt().then(y).catch(()=>y([]))},[]),H(()=>{let n=!1;return l(!0),d(null),s(null),i.current=[],a.current=0,(async()=>{try{let r=Oi(),a=r.qtype,o=r.orders,c=r.userId??(Si(o)?t:void 0);if(e===`random`){let e=await ft(r.useServerPlayModes?{orders:r.orders,qtype:r.qtype,userId:c,infinite:r.infinite,excludeIds:r.excludeIds}:{qtype:r.qtype});if(n)return;if(e.length===0){d(`empty`);return}r.infinite&&(i.current=e.map(e=>e.id)),s({mode:`random`,collectionId:null,nom:`Mélange aléatoire`,questions:e,playOrders:o,playQtype:a,playInfinite:r.infinite,playUserId:c,useServerPlayModes:r.useServerPlayModes}),p(0),h(null),_(0);return}let l=Number(e);if(!Number.isFinite(l)){d(`bad`);return}let u=await at(l,r.useServerPlayModes?{qtype:r.qtype,orders:r.orders,userId:c,infinite:r.infinite,excludeIds:r.excludeIds}:{qtype:r.qtype});if(n)return;if(u.questions.length===0){d(`empty`);return}let f=[...u.questions];!r.useServerPlayModes&&o.length===1&&o[0]===`random`&&(f=Ai(f)),r.infinite&&(i.current=f.map(e=>e.id)),s({mode:`collection`,collectionId:l,nom:u.nom,questions:f,playOrders:o,playQtype:a,playInfinite:r.infinite,playUserId:c,useServerPlayModes:r.useServerPlayModes}),p(0),h(null),_(0)}catch(e){n||((typeof e==`object`&&e&&`status`in e?e.status:void 0)===404?d(`empty`):d(`fetch`))}finally{n||l(!1)}})(),()=>{n=!0}},[e,n,t]),be(()=>{o!=null&&(r.current=performance.now())},[o,f]),H(()=>{x(!1)},[f]),H(()=>{if(o==null)return;let e=o.questions[f];e&&ae(e.verifier)},[o,f]),H(()=>{if(F==null)return;let e=window.setTimeout(()=>I(null),3200);return()=>clearTimeout(e)},[F]);let ce=()=>{x(!1),ee(!1),O(null),A(null),T(null)},le=e=>{C(`edit`),x(!0),ee(!0),O(null),A(null),mt(e.id).then(e=>{A(e),M(e.question),ne(e.commentaire),re(e.categorie_id)}).catch(()=>O(`fetch`)).finally(()=>ee(!1))},ue=e=>{if(v.length===0){I(`Catégories indisponibles : impossible de créer une question pour l’instant.`);return}C(`create`),T(e.id),x(!0),ee(!1),O(null),A(null),M(``),ne(``),re(v.some(t=>t.id===e.categorie_id)?e.categorie_id:v[0].id)},de=async()=>{if(k!=null)try{A(await mt(k.id))}catch{}},B=async()=>{if(k!=null){P(!0);try{let e={};if(j!==k.question&&(e.question=j),te!==k.commentaire&&(e.commentaire=te),N!=null&&N!==k.categorie_id&&(e.categorie_id=N),Object.keys(e).length===0){ce();return}let t=await gt(k.id,e);s(e=>{if(!e)return e;let n=e.questions.findIndex(e=>e.id===t.id);if(n<0)return e;let r=[...e.questions],i=r[n];return i?(r[n]={...i,question:t.question,commentaire:t.commentaire,categorie_id:t.categorie_id,categorie_type:t.categorie_type,verifier:t.verifier},{...e,questions:r}):e}),I(`Question mise à jour.`),ce()}catch{I(`Enregistrement impossible.`)}finally{P(!1)}}},fe=async e=>{if(w!=null){P(!0);try{await _t({user_id:t,categorie_id:e.categorie_id,question:e.question,commentaire:e.commentaire,reponses:e.reponses,parent_question_id:w,collection_id:o?.collectionId??void 0}),I(`Question créée et liée à la question affichée (parent).`),ce()}catch{throw Error(`create failed`)}finally{P(!1)}}},pe=async e=>{let t=eo(e);try{await navigator.clipboard.writeText(t),I(`JSON copié dans le presse-papiers.`)}catch{I(`Copie impossible (permissions du navigateur).`)}};if(c)return X(so,{});if(u!=null||!o)return X(co,{loadError:u});let me=o.questions[f],he=o.questions.length,ge=m==null?f:f+1,_e=m!=null,ve=_e&&m!=null&&$a(o.questions,f,m),ye=(me.commentaire??``).trim(),Se=e=>{if(m!=null)return;let n=(performance.now()-r.current)/1e3;wt({userId:t,questionId:me.id,reponseId:e,dureeSecondes:n}).catch(()=>{}),h(e)},Ce=async()=>{let e=o.questions[f];if(L===e.verifier)return!0;try{let t=await gt(e.id,{verifier:L});return s(n=>{if(!n)return n;let r=[...n.questions],i=r.findIndex(t=>t.id===e.id);return i>=0&&r[i]&&(r[i]={...r[i],verifier:t.verifier}),{...n,questions:r}}),!0}catch{return I(`Enregistrement du fake-checker impossible. Réessaie.`),!1}},we=()=>{oe||m==null||o==null||(async()=>{R(!0);try{if(!await Ce())return;let t=g+(m!=null&&$a(o.questions,f,m)?1:0);if(f+1>=he){if(o.playInfinite){z(!0);try{let n=i.current,r=e===`random`?await ft({orders:o.playOrders,qtype:o.playQtype,userId:o.playUserId,infinite:!0,excludeIds:n}):(await at(o.collectionId,{orders:o.playOrders,qtype:o.playQtype,userId:o.playUserId,infinite:!0,excludeIds:n})).questions;if(a.current+=1,r.length===0){Ya({mode:o.mode,collectionId:o.collectionId,collectionName:o.nom,good:t,total:a.current,playOrders:o.playOrders,playQtype:o.playQtype,playInfinite:!0}),U(`/results`);return}i.current=[...i.current,...r.map(e=>e.id)],_(t),s(e=>e==null?e:{...e,questions:r}),p(0),h(null);return}catch{I(`Impossible de charger la suite des questions.`)}finally{z(!1)}return}a.current+=1,Ya({mode:o.mode,collectionId:o.collectionId,collectionName:o.nom,good:t,total:a.current,playOrders:o.playOrders,playQtype:o.playQtype,playInfinite:!1}),U(`/results`);return}a.current+=1,_(t),p(e=>e+1),h(null)}finally{R(!1)}})()},Te=()=>{o==null||m==null||(async()=>{R(!0);try{if(!await Ce())return;let e=g+(m!=null&&$a(o.questions,f,m)?1:0);a.current+=1,Ya({mode:o.mode,collectionId:o.collectionId,collectionName:o.nom,good:e,total:a.current,playOrders:o.playOrders,playQtype:o.playQtype,playInfinite:!0}),U(`/results`)}finally{R(!1)}})()},Ee=o.mode===`random`?`/`:`/collections`;return X(`div`,{class:$.pageShell,children:[X(li,{}),X(`main`,{class:$.contentMain,children:[X(lo,{data:o,backTarget:Ee}),F?X(`p`,{class:$.actionMessage,children:F}):null,X(fo,{playInfinite:o.playInfinite,progressValue:ge,total:he}),X(uo,{data:o,index:f,total:he,q:me,pickedId:m,revealed:_e,anecdote:ye,correct:ve,draftVerifier:L,nextBusy:oe,fetchingMore:se,onPick:Se,onOpenCreateLinkedQuestionModal:ue,onOpenEditQuestionModal:le,onCopyCurrentQuestionJson:pe,onDraftVerifier:ae,onNext:we,onEndInfiniteSession:Te})]}),X(di,{}),X(Ha,{settings:{open:b,onClose:ce,variant:S,modalTitle:S===`create`?`Nouvelle question liée`:void 0},actions:{onSave:()=>void B(),onDraftQuestion:M,onDraftCommentaire:ne,onDraftCategorieId:re,onReponseUpdated:()=>void de(),onCreateSave:e=>fe(e)},status:{loading:E,saving:ie,error:D},data:{questionDetail:k,categorieOptions:v},drafts:{question:j,commentaire:te,categorieId:N}})]})}var mo={root:`flex min-h-dvh flex-col`,centeredMain:`fl-page-enter mx-auto flex max-w-lg flex-1 flex-col items-center justify-center gap-4 px-4 py-12`,contentMain:`fl-page-enter mx-auto w-full max-w-2xl flex-1 px-4 py-6 md:py-8`};function ho({sessionId:e}){let{userId:t}=Nt(),[n,r]=V(null),[i,a]=V(!0),[o,s]=V(!1);return H(()=>{if(e==null||e===``){s(!0),a(!1);return}let n=!1;return(async()=>{a(!0),s(!1);try{let i=await Et(e,t);if(n)return;i==null?s(!0):r(i)}catch{n||s(!0)}finally{n||a(!1)}})(),()=>{n=!0}},[e,t]),i?X(`div`,{class:mo.root,children:[X(li,{}),X(`main`,{class:mo.centeredMain,children:X(`p`,{class:`text-sm text-base-content/60`,children:`Chargement...`})}),X(di,{})]}):o||!n?X(`div`,{class:mo.root,children:[X(li,{}),X(`main`,{class:mo.centeredMain,children:[X(`p`,{class:`text-base text-base-content/70`,children:`Session introuvable.`}),X(Z,{variant:`flow`,onClick:()=>U(`/dashboard`),children:`Retour au dashboard`})]}),X(di,{})]}):X(`div`,{class:mo.root,children:[X(li,{}),X(`main`,{class:mo.contentMain,children:[X(Z,{variant:`ghost`,class:`btn-sm mb-6 gap-1 px-3`,onClick:()=>U(`/dashboard`),children:[X(Ir,{class:`h-4 w-4`,"aria-hidden":!0}),`Dashboard`]}),X(`h1`,{class:`mb-2 text-2xl font-bold tracking-tight text-base-content`,children:`Detail de session`}),X(`p`,{class:`mb-6 text-sm text-base-content/60`,children:n.date}),X(lr,{class:`mb-6`,children:[X(`div`,{class:`flex flex-wrap items-center gap-2`,children:[X(Fi,{tone:`flow`,children:[`id `,n.id]}),X(`span`,{class:`font-semibold text-base-content`,children:n.collectionName})]}),X(`p`,{class:`mt-4 text-3xl font-semibold text-flow`,children:n.scoreLabel}),X(`p`,{class:`mt-2 text-sm text-base-content/55`,children:[`Reponses enregistrees ce jour pour cette collection (agregat `,X(`code`,{class:`text-xs`,children:`user_kpi`}),`).`]})]}),X(lr,{children:[X(`p`,{class:`mb-3 text-sm font-medium text-base-content`,children:`Questions concernees`}),X(`ul`,{class:`space-y-2 text-sm text-base-content/75`,children:n.questionsPreview.map(e=>X(`li`,{class:`rounded-lg bg-base-200/40 px-3 py-2`,children:[X(`span`,{class:`text-xs text-base-content/45`,children:[`#`,e.id]}),` `,e.question]},e.id))})]})]}),X(di,{})]})}var go={wrapper:`relative overflow-hidden`,glow:`pointer-events-none absolute -right-6 -top-6 h-24 w-24 rounded-full opacity-20 blur-2xl`,content:`relative flex items-start gap-3`};function _o({title:e,value:t,hint:n,icon:r,accent:i=`flow`}){return X(lr,{padding:`sm`,class:go.wrapper,children:[X(`div`,{class:ar(go.glow,i===`flow`?`bg-flow`:`bg-learn`)}),X(`div`,{class:go.content,children:[r?X(`div`,{class:i===`flow`?`text-flow`:`text-learn`,children:r}):null,X(`div`,{class:`min-w-0 flex-1 space-y-1`,children:[X(`p`,{class:`text-xs font-medium uppercase tracking-wide text-base-content/50`,children:e}),X(`p`,{class:ar(`text-2xl font-semibold tracking-tight`,i===`flow`?`text-flow`:`text-learn`),children:t}),n?X(`p`,{class:`text-xs text-base-content/55`,children:n}):null]})]})]})}function vo(e){return e.length===0?0:e.reduce((e,t)=>e+t,0)/e.length}function yo(e){let t=e.trim().replace(`,`,`.`);if(!t)return null;let n=Number.parseFloat(t);return Number.isFinite(n)&&n>=0?n:null}function bo(e){return String(e).padStart(2,`0`)}function xo(e){return`${e.getFullYear()}-${bo(e.getMonth()+1)}-${bo(e.getDate())}`}function So(e,t=new Date){let n=new Date(t.getFullYear(),t.getMonth(),t.getDate()),r=[];for(let e=6;e>=0;--e){let t=new Date(n);t.setDate(t.getDate()-e);let i=xo(t),a=t.toLocaleDateString(`fr-FR`,{weekday:`short`}),o=a.length>0?a.charAt(0).toUpperCase()+a.slice(1):i;r.push({key:i,label:o,count:0,h:0})}for(let t of e){let e=new Date(t.create_at);if(Number.isNaN(e.getTime()))continue;let n=xo(e),i=r.find(e=>e.key===n);i&&(i.count+=1)}let i=Math.max(...r.map(e=>e.count),0);for(let e of r)e.h=i===0?0:Math.round(e.count/i*100);return r}function Co(e,t){let n=e.length,r=e.filter(e=>e.correct).length,i=n===0?0:Math.round(r/n*100),a=e.map(e=>yo(e.duree_session)).filter(e=>e!=null),o=vo(a),s=new Set(e.map(e=>e.question_id)).size,c=t.reduce((e,t)=>e+t.good,0),l=t.reduce((e,t)=>e+t.total,0),u=t.length===0?`${s} question${s===1?``:`s`} distincte${s===1?``:`s`} · aucune session dans l historique`:`${s} question${s===1?``:`s`} distincte${s===1?``:`s`} · sessions : ${c}/${l} bonnes reponses`;return{total:n,good:r,ratio:i,avgSecLabel:a.length===0?`—`:`${o.toFixed(1)} s`,uniqueQuestions:s,sessionsHint:u}}var wo={root:`flex min-h-dvh flex-col`};function To(){let{userId:e}=Nt(),[t,n]=V([]),[r,i]=V([]),[a,o]=V(!0),[s,c]=V(null),l=Ce(async()=>{o(!0),c(null);try{let[t,r]=await Promise.all([Ct(e),Tt(e)]);n(t),i(r)}catch{c(`fetch`)}finally{o(!1)}},[e]);H(()=>{l()},[l]);let u=Se(()=>Co(t,r),[t,r]),d=Se(()=>So(t),[t]);return X(`div`,{class:wo.root,children:[X(li,{}),X(pi,{children:[X(`h1`,{class:`mb-2 text-2xl font-bold tracking-tight text-base-content sm:text-3xl`,children:`Dashboard`}),X(`p`,{class:`mb-6 text-sm text-base-content/60`,children:[`Donnees live : `,X(`code`,{class:`text-xs`,children:`/stats/kpis`}),` et `,X(`code`,{class:`text-xs`,children:`/stats/sessions`}),` pour l utilisateur connecte (id `,e,`).`]}),a?X(`p`,{class:`text-sm text-base-content/60`,children:`Chargement...`}):s?X(`div`,{class:`rounded-[var(--radius-box)] border border-base-content/10 bg-base-200/40 px-4 py-8 text-center text-sm text-base-content/65`,children:[X(`p`,{class:`mb-3`,children:`Impossible de charger les statistiques.`}),X(Z,{variant:`flow`,class:`btn-sm`,onClick:()=>void l(),children:`Reessayer`})]}):X(b,{children:[X(`div`,{class:`mb-6 grid gap-3 sm:grid-cols-3`,children:[X(_o,{title:`Ratio correct`,value:`${u.ratio}%`,hint:`${u.good} / ${u.total} reponses`,accent:`flow`,icon:X(ei,{class:`h-5 w-5`,"aria-hidden":!0})}),X(_o,{title:`Temps moyen`,value:u.avgSecLabel,hint:`Moyenne des durees enregistrees (champ duree_session)`,accent:`learn`,icon:X(Br,{class:`h-5 w-5`,"aria-hidden":!0})}),X(_o,{title:`Volume & couverture`,value:`${u.total}`,hint:u.sessionsHint,accent:`flow`,icon:X(Lr,{class:`h-5 w-5`,"aria-hidden":!0})})]}),X(lr,{class:`mb-6 transition duration-300`,children:[X(`p`,{class:`mb-4 text-sm font-medium text-base-content`,children:`Reponses enregistrees (7 derniers jours)`}),X(`p`,{class:`mb-3 text-xs text-base-content/50`,children:[`Agregat des lignes `,X(`code`,{class:`text-[10px]`,children:`user_kpi`}),` par date locale (fuseau du navigateur).`]}),X(`div`,{class:`flex h-36 items-end justify-between gap-2`,children:d.map(e=>{let t=d.some(e=>e.count>0)?Math.max(e.h,e.count>0?14:10):8;return X(`div`,{class:`flex flex-1 flex-col items-center gap-2`,children:[X(`div`,{class:`w-full max-w-10 rounded-t-full bg-gradient-to-t from-flow to-learn/80 transition-all duration-300 ease-out hover:opacity-100 ${e.count>0?`opacity-90`:`opacity-35`}`,style:{height:`${t}%`},title:`${e.count} reponse(s) · ${e.key}`}),X(`span`,{class:`text-[10px] font-medium text-base-content/50`,children:e.label})]},e.key)})})]}),X(lr,{class:`transition duration-300`,children:[X(`p`,{class:`mb-4 text-sm font-medium text-base-content`,children:`Historique des sessions`}),r.length===0?X(`p`,{class:`text-sm text-base-content/55`,children:`Aucune session agregee pour cet utilisateur.`}):X(`ul`,{class:`space-y-3`,children:r.map(e=>X(`li`,{children:X(`button`,{type:`button`,class:`flex w-full flex-col gap-2 rounded-full border border-base-content/10 bg-base-200/30 px-4 py-4 text-left transition-all duration-300 ease-out hover:border-flow/30 hover:bg-base-200/55 sm:flex-row sm:items-center sm:justify-between`,onClick:()=>U(`/dashboard/session/${encodeURIComponent(e.id)}`),children:[X(`div`,{children:[X(`p`,{class:`font-medium text-base-content`,children:e.collectionName}),X(`p`,{class:`text-xs text-base-content/50`,children:e.date})]}),X(`div`,{class:`flex items-center gap-2`,children:[X(Fi,{tone:`learn`,children:e.scoreLabel}),X(`span`,{class:`text-xs text-flow`,children:`Details →`})]})]})},e.id))})]})]})]}),X(di,{})]})}function Eo(){return typeof window>`u`?`/`:`${window.location.pathname||`/`}${window.location.search||``}`}function Do(){let[e,t]=V(Eo);return X(hr,{children:X(gr.Provider,{value:e,children:X(Ze,{onChange:()=>{t(Eo())},children:[X(Qe,{path:`/`,component:Zi}),X(Qe,{path:`/collections`,component:Gi}),X(Qe,{path:`/database`,component:Yi}),X(Qe,{path:`/questions/:collectionId`,component:Ka}),X(Qe,{path:`/questions`,component:Ka}),X(Qe,{path:`/dashboard/session/:sessionId`,component:ho}),X(Qe,{path:`/dashboard`,component:To}),X(Qe,{path:`/play/:collectionId`,component:po}),X(Qe,{path:`/results`,component:Qa})]})})})}ae(X(Do,{}),document.getElementById(`app`));
+Alternative acceptée pour chaque réponse : { "reponse": "...", "bonne_reponse": true|false } au lieu de texte/correcte.`}function qa(){return[{texte:``,correcte:!0},{texte:``,correcte:!1},{texte:``,correcte:!1},{texte:``,correcte:!1}]}function Ja(e){let{settings:t,actions:n,status:r,data:i,drafts:a}=e,o=t.variant??`edit`,s=i.categorieOptions,[c,l]=V(null),[u,d]=V(``),[f,p]=V(!1),[m,h]=V(null),[g,_]=V(qa),[v,y]=V(null),b=o===`create`,x=t.modalTitle??(b?`Nouvelle question`:`Modifier la question`);return H(()=>{t.open&&b&&(_(qa()),y(null))},[t.open,b]),{dialogue:{title:x,isCreate:b,onBackdropClick:()=>{!r.saving&&!f&&t.onClose()}},creation:{submitCreate:async()=>{if(y(null),a.categorieId==null)return void y(`Choisis une catégorie.`);let e=a.question.trim();if(!e)return void y(`L’énoncé ne peut pas être vide.`);let t=g.map(e=>({...e,texte:e.texte.trim()}));if(t.some(e=>!e.texte))return void y(`Toutes les propositions doivent être renseignées.`);if(t.filter(e=>e.correcte).length!==1)return void y(`Indique exactement une bonne réponse.`);if(!n.onCreateSave)return void y(`Configuration interne : onCreateSave manquant.`);await Promise.resolve(n.onCreateSave({question:e,commentaire:a.commentaire.trim(),categorie_id:a.categorieId,reponses:t}))},createFormError:v,createReponses:g,setCreateReponseCorrectAt:e=>{_(t=>t.map((t,n)=>({...t,correcte:n===e})))},setCreateReponseTexteAt:(e,t)=>{_(n=>n.map((n,r)=>r===e?{...n,texte:t}:n))}},composantExterne:{argProp:{class:`w-full min-w-0 max-w-full`,data:{prompt:Ka(a.question,s)},actions:{importFromJson:async e=>{let t=Ga(e,{categorieOptions:s,fallbackCategorieId:a.categorieId});if(!t.ok)throw Error(t.error);return n.onDraftQuestion(t.value.question),n.onDraftCommentaire(t.value.commentaire),n.onDraftCategorieId(t.value.categorie_id),_(t.value.reponses),`Champs mis à jour à partir du JSON.`}},settings:{submitLabel:`Appliquer le JSON`,submitBusyLabel:`Application…`,pasteAreaInstruction:`Colle le JSON renvoyé par le LLM (objet unique, format décrit dans le prompt). L’énoncé, le commentaire, la catégorie et les quatre réponses du formulaire seront remplis.`,jsonPastePlaceholder:`{ "question": "…", "commentaire": "…", "categorie_type": "histoire", "reponses": [ { "texte": "…", "correcte": false }, … ] }`,disabled:r.saving}}},editionReponses:{editingReponseId:c,reponseDraft:u,setReponseDraft:d,reponseBusy:f,reponseError:m,submitReponse:async()=>{if(c==null)return;let e=u.trim();if(!e)return void h(`Le texte ne peut pas être vide.`);p(!0),h(null);try{await yt(c,{reponse:e}),await Promise.resolve(n.onReponseUpdated()),l(null),d(``)}catch{h(`Enregistrement impossible (réseau ou serveur).`)}finally{p(!1)}},beginEditReponse:(e,t)=>{l(e),d(t)}}}}var Ya={overlay:`fixed inset-0 z-100 flex items-center justify-center p-4`,backdrop:`absolute inset-0 border-0 bg-base-content/40 p-0`,dialog:`relative z-10 max-h-[90vh] w-full max-w-2xl overflow-y-auto rounded-2xl border border-base-content/10 bg-base-100 p-6 shadow-2xl`};function Xa(e){let{settings:t,actions:n,status:r,data:i,drafts:a}=e,{dialogue:o,creation:s,composantExterne:c,editionReponses:l}=Ja(e);return t.open?X(`div`,{class:Ya.overlay,role:`presentation`,children:[X(`button`,{type:`button`,class:Ya.backdrop,"aria-label":`Fermer`,onClick:o.onBackdropClick}),X(`div`,{class:Ya.dialog,role:`dialog`,"aria-modal":`true`,"aria-labelledby":`qm-title`,onClick:e=>e.stopPropagation(),children:[X(`h2`,{id:`qm-title`,class:`mb-4 text-lg font-semibold text-base-content`,children:o.title}),o.isCreate?X(b,{children:[s.createFormError?X(`p`,{class:`mb-3 text-sm text-error`,children:s.createFormError}):null,X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`qm-categorie`,children:`Catégorie`}),X(`select`,{id:`qm-categorie`,class:`select select-bordered select-sm mb-4 w-full max-w-md rounded-xl border-base-content/15 bg-base-100`,value:a.categorieId??``,disabled:r.saving,onChange:e=>n.onDraftCategorieId(Number(e.target.value)),children:i.categorieOptions.map(e=>X(`option`,{value:e.id,children:e.type},e.id))}),X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`qm-question`,children:`Énoncé`}),X(`textarea`,{id:`qm-question`,class:`textarea textarea-bordered mb-4 w-full rounded-xl border-base-content/15 text-sm`,rows:4,value:a.question,onInput:e=>n.onDraftQuestion(e.target.value)}),X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`qm-commentaire`,children:`Commentaire`}),X(`textarea`,{id:`qm-commentaire`,class:`textarea textarea-bordered mb-4 w-full rounded-xl border-base-content/15 text-sm`,rows:3,value:a.commentaire,onInput:e=>n.onDraftCommentaire(e.target.value)}),X(`div`,{class:`mb-4 rounded-xl border border-learn/20 bg-learn/6 p-3`,children:X(va,{...c.argProp})}),X(`ul`,{class:`mb-6 space-y-3 rounded-xl border border-base-content/10 bg-base-200/30 p-3`,children:s.createReponses.map((e,t)=>X(`li`,{children:[X(`div`,{class:`mb-1 flex items-center gap-2`,children:[X(`input`,{type:`radio`,name:`qm-bonne-create`,class:`radio radio-sm radio-primary`,checked:e.correcte,onChange:()=>s.setCreateReponseCorrectAt(t)}),X(`span`,{class:`text-xs text-base-content/60`,children:e.correcte?`Bonne réponse`:`Proposition ${t+1}`})]}),X(`textarea`,{class:`textarea textarea-bordered w-full rounded-xl border-base-content/15 text-sm`,rows:2,value:e.texte,onInput:e=>{s.setCreateReponseTexteAt(t,e.target.value)}})]},t))}),X(`div`,{class:`mt-0 flex flex-wrap justify-end gap-2 border-t border-base-content/10 pt-4`,children:[X(Z,{variant:`ghost`,class:`btn-sm`,disabled:r.saving,onClick:t.onClose,children:`Annuler`}),X(Z,{variant:`flow`,class:`btn-sm`,disabled:r.saving,onClick:()=>void s.submitCreate(),children:r.saving?`Création…`:`Créer`})]})]}):r.error==null?r.loading||i.questionDetail==null?X(`div`,{class:`py-8 text-center`,children:[X(`p`,{class:`mb-4 text-sm text-base-content/60`,children:`Chargement…`}),X(Z,{variant:`ghost`,class:`btn-sm`,onClick:t.onClose,children:`Annuler`})]}):X(b,{children:[X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`qe-question`,children:`Énoncé`}),X(`textarea`,{id:`qe-question`,class:`textarea textarea-bordered mb-4 w-full rounded-xl border-base-content/15 text-sm`,rows:4,value:a.question,onInput:e=>n.onDraftQuestion(e.target.value)}),X(`label`,{class:`mb-1 block text-xs font-medium text-base-content/60`,for:`qe-commentaire`,children:`Commentaire`}),X(`textarea`,{id:`qe-commentaire`,class:`textarea textarea-bordered mb-4 w-full rounded-xl border-base-content/15 text-sm`,rows:3,value:a.commentaire,onInput:e=>n.onDraftCommentaire(e.target.value)}),l.reponseError?X(`p`,{class:`mb-2 text-xs text-error`,children:l.reponseError}):null,X(`ul`,{class:`mb-6 space-y-3 rounded-xl border border-base-content/10 bg-base-200/30 p-3`,children:i.questionDetail.reponses.map(e=>X(`li`,{children:l.editingReponseId===e.id?X(`div`,{class:`flex gap-2`,children:[X(`textarea`,{class:`textarea textarea-bordered w-full rounded-xl border-base-content/15 text-sm`,rows:2,value:l.reponseDraft,onInput:e=>l.setReponseDraft(e.target.value)}),X(Z,{variant:`flow`,class:`btn-xs`,disabled:l.reponseBusy,onClick:()=>void l.submitReponse(),children:l.reponseBusy?`…`:`Valider`})]}):X(`div`,{class:`flex items-start justify-between gap-2`,children:[X(`span`,{children:[e.bonne_reponse?`✓ `:`○ `,e.reponse]}),X(Z,{variant:`outline`,class:`btn-xs`,onClick:()=>l.beginEditReponse(e.id,e.reponse),children:`Modifier`})]})},e.id))}),X(`div`,{class:`mt-0 flex flex-wrap justify-end gap-2 border-t border-base-content/10 pt-4`,children:[X(Z,{variant:`ghost`,class:`btn-sm`,onClick:t.onClose,children:`Annuler`}),X(Z,{variant:`flow`,class:`btn-sm`,disabled:r.saving,onClick:n.onSave,children:r.saving?`Enregistrement…`:`Enregistrer`})]})]}):X(`div`,{class:`py-4 text-center`,children:[X(`p`,{class:`mb-4 text-sm text-error`,children:`Impossible de charger le détail.`}),X(Z,{variant:`ghost`,class:`btn-sm`,onClick:t.onClose,children:`Fermer`})]})]})]}):null}function Za(e){if(e===``)return;if(e===`none`)return`none`;let t=Number(e);return Number.isFinite(t)?t:void 0}function Qa(e){return e&&/^\d+$/.test(e)?e:``}function $a(e,t){return t===`melanger`?e:e.filter(e=>e.categorie_type===t)}function eo({collectionId:e}){let[t,n]=V([]),[r,i]=V([]),[a,o]=V(()=>Qa(e)),[s,c]=V(null),[l,u]=V([]),[d,f]=V(!0),[p,m]=V(null),[h,g]=V([]),[_,v]=V(!1),[y,b]=V(!1),[x,S]=V(null),[C,w]=V(null),[T,E]=V(``),[ee,D]=V(``),[O,k]=V(null),[A,j]=V(!1),[M,te]=V(`melanger`),ne=a!==``&&a!==`none`&&/^\d+$/.test(a)?Number(a):null,N=Se(()=>$a(l,M),[l,M]);H(()=>{o(Qa(e))},[e]),H(()=>{if(typeof window>`u`)return;let e=new URLSearchParams(window.location.search).get(`module`);e&&/^\d+$/.test(e)?c(Number(e)):c(null)},[e]);let re=Ce(()=>{f(!0),m(null),gt(Za(a)).then(u).catch(()=>m(`fetch`)).finally(()=>f(!1))},[a]);H(()=>{at().then(n).catch(()=>{}),st().then(i).catch(()=>{}),mt().then(g).catch(()=>{})},[]),H(()=>{re()},[re]);let ie=Ce(e=>{if(o(e),e===``||e===`none`)return void U(`/questions`);/^\d+$/.test(e)&&U(`/questions/${e}${s==null?``:`?module=${s}`}`)},[s]),ae=Ce(e=>{v(!0),b(!0),S(null),w(null),ht(e.id).then(e=>{w(e),E(e.question),D(e.commentaire),k(e.categorie_id)}).catch(()=>S(`fetch`)).finally(()=>b(!1))},[]),P=Ce(()=>{v(!1),b(!1),S(null),w(null)},[]),F=Ce(async()=>{if(C!=null)try{w(await ht(C.id))}catch{}},[C]),I=Ce(async()=>{if(C!=null){j(!0);try{let e={};if(T!==C.question&&(e.question=T),ee!==C.commentaire&&(e.commentaire=ee),O!=null&&O!==C.categorie_id&&(e.categorie_id=O),Object.keys(e).length===0){P();return}let t=await _t(C.id,e);u(e=>e.map(e=>e.id===t.id?t:e)),P()}catch{m(`save`)}finally{j(!1)}}},[P,C,O,ee,T]),oe=Ce(async e=>{j(!0);try{await Ct(e),u(t=>t.filter(t=>t.id!==e)),C?.id===e&&P()}catch{m(`delete`)}finally{j(!1)}},[P,C?.id]),se=Ce(()=>{re(),at().then(n).catch(()=>{})},[re]),L=Ce(e=>{te(e)},[]);return{llmImport:{data:{targetCollectionNumeric:ne,collections:t,allModules:r,importTargetModuleId:s,questions:l},actions:{onImportSuccess:se}},operationError:{visible:!!p,onDismiss:()=>m(null)},contextBar:{targetCollectionNumeric:ne,collections:t,allModules:r,importTargetModuleId:s,setImportTargetModuleId:c},filtres:{collectionFilter:a,collections:t,onCollectionFilterChange:ie,listFilterQtype:M,onListFilterQtypeChange:L},liste:{loading:d,fetchError:p===`fetch`,onReload:re,questionsForTable:N,saving:A,onEdit:ae,onRemove:oe},editModal:{settings:{open:_,onClose:P},actions:{onSave:()=>void I(),onDraftQuestion:E,onDraftCommentaire:D,onDraftCategorieId:k,onReponseUpdated:()=>void F()},status:{loading:y,saving:A,error:x},data:{questionDetail:C,categorieOptions:h},drafts:{question:T,commentaire:ee,categorieId:O}}}}var to={item:`overflow-hidden rounded-[1.35rem] border border-base-content/10 bg-base-200/35 transition duration-300 hover:border-flow/20 hover:shadow-md hover:shadow-flow/5`};function no(e){let{data:t,actions:n,status:r}=e,{questions:i}=t,{onEdit:a,onRemove:o}=n,{saving:s}=r;return X(ur,{children:[X(`p`,{class:`mb-4 text-sm text-base-content/60`,children:[i.length,` question`,i.length===1?``:`s`,` affichee`,i.length===1?``:`s`,`.`]}),X(`ul`,{class:`space-y-3`,children:i.map(e=>X(`li`,{class:to.item,children:X(`div`,{class:`flex flex-col gap-3 p-4 sm:flex-row sm:items-start sm:justify-between`,children:[X(`div`,{class:`min-w-0 flex-1`,children:[X(`div`,{class:`mb-2 flex flex-wrap items-center gap-2`,children:[X(Li,{tone:`neutral`,children:[`id `,e.id]}),X(Li,{tone:`learn`,children:e.categorie_type}),e.collections.length===0?X(Li,{tone:`learn`,children:`Sans collection`}):e.collections.map(e=>X(Li,{tone:`flow`,children:e.nom},e.id))]}),e.commentaire.trim()?X(`p`,{class:`mb-2 text-xs leading-relaxed text-base-content/55 line-clamp-3`,children:e.commentaire}):null,X(`p`,{class:`text-sm font-medium leading-relaxed text-base-content`,children:e.question})]}),X(`div`,{class:`flex shrink-0 flex-wrap gap-2`,children:[X(Z,{variant:`outline`,class:`btn-sm gap-1`,onClick:()=>a(e),disabled:s,children:[X(Vr,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),`Modifier`]}),X(Z,{variant:`learn`,class:`btn-sm gap-1`,onClick:()=>o(e.id),disabled:s,children:[X(qr,{class:`h-3.5 w-3.5`,"aria-hidden":!0}),`Supprimer`]})]})]})},e.id))})]})}var ro={root:`flex min-h-dvh flex-col`,operationError:`mb-6 rounded-box border border-error/20 bg-error/5 px-4 py-3 text-sm text-base-content`,dismissLink:`link link-primary`,filtersRow:`mb-4 flex flex-col gap-4 sm:flex-row sm:flex-wrap sm:items-end`,filterField:`flex min-w-0 flex-1 flex-col gap-2 sm:max-w-xs`,filterFieldNarrow:`flex min-w-0 flex-col gap-2 sm:w-44`,filterLabel:`text-sm font-medium text-base-content/80`,select:`select select-bordered select-sm w-full rounded-xl border-base-content/15 bg-base-100`,loadingHint:`text-sm text-base-content/60`,fetchErrorCard:`border-base-content/15`,fetchErrorText:`mb-3 text-sm text-base-content/70`};function io({onDismiss:e}){return X(`div`,{class:ro.operationError,children:[`Une operation a echoue.`,` `,X(`button`,{type:`button`,class:ro.dismissLink,onClick:e,children:`Fermer`})]})}function ao({collectionFilter:e,collections:t,onCollectionFilterChange:n,listFilterQtype:r,onListFilterQtypeChange:i}){return X(`div`,{class:ro.filtersRow,children:[X(`div`,{class:ro.filterField,children:[X(`label`,{class:ro.filterLabel,for:`q-collection-filter`,children:`Filtrer par collection`}),X(`select`,{id:`q-collection-filter`,class:ro.select,value:e,onChange:e=>n(e.target.value),children:[X(`option`,{value:``,children:`Toutes les questions`}),X(`option`,{value:`none`,children:`Sans collection`}),t.map(e=>X(`option`,{value:String(e.id),children:e.nom},e.id))]})]}),X(`div`,{class:ro.filterFieldNarrow,children:[X(`label`,{class:ro.filterLabel,for:`q-list-qtype-filter`,children:`Filtrer par type (affichage)`}),X(`select`,{id:`q-list-qtype-filter`,class:ro.select,value:r,onChange:e=>{let t=e.target.value;(t===`histoire`||t===`pratique`||t===`melanger`)&&i(t)},children:[X(`option`,{value:`melanger`,children:`Tout (histoire + pratique)`}),X(`option`,{value:`histoire`,children:`Histoire seulement`}),X(`option`,{value:`pratique`,children:`Pratique seulement`})]})]})]})}function oo({loading:e,fetchError:t,onReload:n,questionsForTable:r,saving:i,onEdit:a,onRemove:o}){return e?X(`p`,{class:ro.loadingHint,children:`Chargement...`}):t?X(ur,{class:ro.fetchErrorCard,children:[X(`p`,{class:ro.fetchErrorText,children:`Impossible de charger les questions.`}),X(Z,{variant:`flow`,class:`btn-sm`,onClick:n,children:`Reessayer`})]}):X(no,{data:{questions:r},actions:{onEdit:a,onRemove:o},status:{saving:i}})}function so(e){let{llmImport:t,operationError:n,contextBar:r,filtres:i,liste:a,editModal:o}=eo(e);return X(`div`,{class:ro.root,children:[X(ti,{}),X(ai,{children:[X(Va,{data:t.data,actions:t.actions}),n.visible?X(io,{onDismiss:n.onDismiss}):null,X(ra,{...r}),X(ao,{...i}),X(oo,{...a}),X(Xa,{settings:o.settings,actions:o.actions,status:o.status,data:o.data,drafts:o.drafts})]}),X(ri,{})]})}var co=`flowlearn_last_result`;function lo(e){if(e.playOrders!=null&&Array.isArray(e.playOrders)){let t=e.playOrders.filter(e=>typeof e==`string`&&Si(e));if(t.length>0)return t}let t=e.playOrder;if(typeof t==`string`){if(t.includes(`,`))return Ci(t);if(Si(t))return[t]}}function uo(e){try{sessionStorage.setItem(co,JSON.stringify(e))}catch{}}function fo(){try{let e=sessionStorage.getItem(co);if(!e)return null;let t=JSON.parse(e),n=lo(t);if(t.mode===`random`||t.mode===`collection`){let e=t;return n==null?e:{...e,playOrders:n}}return t.collectionName!=null&&typeof t.good==`number`&&typeof t.total==`number`?{mode:`collection`,collectionId:t.collectionId??null,collectionName:t.collectionName,good:t.good,total:t.total,playOrders:n??void 0,playQtype:t.playQtype,playInfinite:t.playInfinite}:null}catch{return null}}function po(e,t){return t<=0?0:Math.round(e/t*100)}function mo(e){return e.playOrders!=null&&e.playOrders.length>0?e.playOrders:e.playOrder!=null&&Si(e.playOrder)?[e.playOrder]:[`random`]}function ho(e,t){let n=mo(e),r=ji({orders:n,qtype:e.playQtype??`melanger`,infinite:e.playInfinite===!0,userId:wi(n)?t:void 0});return e.mode===`random`?`/play/random${r}`:e.collectionId==null?`/collections`:`/play/${e.collectionId}${r}`}function go(){let{userId:e}=Pt(),t=Se(()=>fo(),[]),n=()=>U(`/`),r=()=>U(`/collections`);return t==null?{result:null,summary:null,navigation:{goHome:n,goCollections:r,replay:n}}:{result:t,summary:{good:t.good,total:t.total,percent:po(t.good,t.total),collectionName:t.collectionName},navigation:{goHome:n,goCollections:r,replay:()=>{U(ho(t,e))}}}}var _o={root:`flex min-h-dvh flex-col`,main:`fl-page-enter mx-auto flex w-full max-w-lg flex-1 flex-col items-center justify-center px-4 py-10`};function vo(){let{result:e,summary:t,navigation:n}=go();return e==null||t==null?X(`div`,{class:_o.root,children:[X(ti,{}),X(`main`,{class:`fl-page-enter mx-auto flex max-w-lg flex-1 flex-col items-center justify-center gap-4 px-4 py-12 text-center`,children:[X(`p`,{class:`text-base text-base-content/70`,children:`Aucun resultat recent. Lance un quiz depuis l accueil ou une collection.`}),X(Z,{variant:`flow`,onClick:n.goHome,children:`Accueil`})]}),X(ri,{})]}):X(`div`,{class:_o.root,children:[X(ti,{}),X(`main`,{class:_o.main,children:X(ur,{class:`w-full text-center transition duration-300`,children:[X(`div`,{class:`mx-auto mb-4 flex h-14 w-14 items-center justify-center rounded-full bg-gradient-to-br from-flow to-learn text-white shadow-xl shadow-flow/25 transition duration-300 hover:scale-105`,children:X(Jr,{class:`h-7 w-7`,"aria-hidden":!0})}),X(Li,{tone:`flow`,class:`mb-3`,children:`Session terminee`}),X(`h1`,{class:`mb-2 text-2xl font-bold tracking-tight text-base-content`,children:`Resultat`}),X(`p`,{class:`mb-6 text-sm text-base-content/65`,children:t.collectionName}),X(`p`,{class:`mb-1 text-4xl font-semibold text-flow`,children:[t.good,`/`,t.total]}),X(`p`,{class:`mb-8 text-sm text-base-content/55`,children:[t.percent,`% de bonnes reponses`]}),X(`div`,{class:`flex flex-col gap-2 sm:flex-row sm:justify-center`,children:[X(Z,{variant:`flow`,onClick:n.replay,children:`Rejouer`}),X(Z,{variant:`outline`,onClick:n.goCollections,children:`Collections`}),X(Z,{variant:`ghost`,onClick:n.goHome,children:`Accueil`})]})]})}),X(ri,{})]})}function yo(e,t,n){return e[t]?.reponses.some(e=>e.id===n&&e.bonne_reponse)??!1}function bo(e){return JSON.stringify({question:e.question,commentaire:e.commentaire,reponses:e.reponses.map(e=>({reponse:e.reponse,bonne_reponse:e.bonne_reponse}))},null,2)}function xo(e){let{collectionId:t}=e,{userId:n}=Pt(),r=vr(),i=xe(0),a=xe([]),o=xe(0),[s,c]=V(null),[l,u]=V(!0),[d,f]=V(null),[p,m]=V(0),[h,g]=V(null),[_,v]=V(0),[y,b]=V([]),[x,S]=V(!1),[C,w]=V(`edit`),[T,E]=V(null),[ee,D]=V(!1),[O,k]=V(null),[A,j]=V(null),[M,te]=V(``),[ne,N]=V(``),[re,ie]=V(null),[ae,P]=V(!1),[F,I]=V(null),[oe,se]=V(!1),[L,R]=V(!1),[z,ce]=V(!1);H(()=>{mt().then(b).catch(()=>b([]))},[]),H(()=>{let e=!1;return u(!0),f(null),c(null),a.current=[],o.current=0,(async()=>{try{let r=Ai(),i=r.qtype,o=r.orders,s=r.userId??(wi(o)?n:void 0);if(t===`random`){let t=await pt(r.useServerPlayModes?{orders:r.orders,qtype:r.qtype,userId:s,infinite:r.infinite,excludeIds:r.excludeIds}:{qtype:r.qtype});if(e)return;if(t.length===0){f(`empty`);return}r.infinite&&(a.current=t.map(e=>e.id)),c({mode:`random`,collectionId:null,nom:`Mélange aléatoire`,questions:t,playOrders:o,playQtype:i,playInfinite:r.infinite,playUserId:s,useServerPlayModes:r.useServerPlayModes}),m(0),g(null),v(0);return}let l=Number(t);if(!Number.isFinite(l)){f(`bad`);return}let u=await ot(l,r.useServerPlayModes?{qtype:r.qtype,orders:r.orders,userId:s,infinite:r.infinite,excludeIds:r.excludeIds}:{qtype:r.qtype});if(e)return;if(u.questions.length===0){f(`empty`);return}let d=[...u.questions];!r.useServerPlayModes&&o.length===1&&o[0]===`random`&&(d=Mi(d)),r.infinite&&(a.current=d.map(e=>e.id)),c({mode:`collection`,collectionId:l,nom:u.nom,questions:d,playOrders:o,playQtype:i,playInfinite:r.infinite,playUserId:s,useServerPlayModes:r.useServerPlayModes}),m(0),g(null),v(0)}catch(t){e||((typeof t==`object`&&t&&`status`in t?t.status:void 0)===404?f(`empty`):f(`fetch`))}finally{e||u(!1)}})(),()=>{e=!0}},[t,r,n]),be(()=>{s!=null&&(i.current=performance.now())},[s,p]),H(()=>{S(!1)},[p]),H(()=>{if(s==null)return;let e=s.questions[p];e&&se(e.verifier)},[s,p]),H(()=>{if(F==null)return;let e=window.setTimeout(()=>I(null),3200);return()=>clearTimeout(e)},[F]);let le=()=>{S(!1),D(!1),k(null),j(null),E(null)},ue=e=>{w(`edit`),S(!0),D(!0),k(null),j(null),ht(e.id).then(e=>{j(e),te(e.question),N(e.commentaire),ie(e.categorie_id)}).catch(()=>k(`fetch`)).finally(()=>D(!1))},de=e=>{if(y.length===0){I(`Catégories indisponibles : impossible de créer une question pour l’instant.`);return}w(`create`),E(e.id),S(!0),D(!1),k(null),j(null),te(``),N(``),ie(y.some(t=>t.id===e.categorie_id)?e.categorie_id:y[0].id)},B=async()=>{if(A!=null)try{j(await ht(A.id))}catch{}},fe=async()=>{if(A!=null){P(!0);try{let e={};if(M!==A.question&&(e.question=M),ne!==A.commentaire&&(e.commentaire=ne),re!=null&&re!==A.categorie_id&&(e.categorie_id=re),Object.keys(e).length===0){le();return}let t=await _t(A.id,e);c(e=>{if(!e)return e;let n=e.questions.findIndex(e=>e.id===t.id);if(n<0)return e;let r=[...e.questions],i=r[n];return i?(r[n]={...i,question:t.question,commentaire:t.commentaire,categorie_id:t.categorie_id,categorie_type:t.categorie_type,verifier:t.verifier},{...e,questions:r}):e}),I(`Question mise à jour.`),le()}catch{I(`Enregistrement impossible.`)}finally{P(!1)}}},pe=async e=>{if(T!=null){P(!0);try{await vt({user_id:n,categorie_id:e.categorie_id,question:e.question,commentaire:e.commentaire,reponses:e.reponses,parent_question_id:T,collection_id:s?.collectionId??void 0}),I(`Question créée et liée à la question affichée (parent).`),le()}catch{throw Error(`create failed`)}finally{P(!1)}}},me=async e=>{let t=bo(e);try{await navigator.clipboard.writeText(t),I(`JSON copié dans le presse-papiers.`)}catch{I(`Copie impossible (permissions du navigateur).`)}},he=e=>{if(s==null||h!=null)return;let t=s.questions[p];if(t==null)return;let r=(performance.now()-i.current)/1e3;Tt({userId:n,questionId:t.id,reponseId:e,dureeSecondes:r}).catch(()=>{}),g(e)},ge=async()=>{if(s==null)return!1;let e=s.questions[p];if(e==null)return!1;if(oe===e.verifier)return!0;try{let t=await _t(e.id,{verifier:oe});return c(n=>{if(!n)return n;let r=[...n.questions],i=r.findIndex(t=>t.id===e.id);return i>=0&&r[i]&&(r[i]={...r[i],verifier:t.verifier}),{...n,questions:r}}),!0}catch{return I(`Enregistrement du fake-checker impossible. Réessaie.`),!1}},_e=()=>{if(L||h==null||s==null)return;let e=s.questions.length;(async()=>{R(!0);try{if(!await ge())return;let n=_+(h!=null&&yo(s.questions,p,h)?1:0);if(p+1>=e){if(s.playInfinite){ce(!0);try{let e=a.current,r=t===`random`?await pt({orders:s.playOrders,qtype:s.playQtype,userId:s.playUserId,infinite:!0,excludeIds:e}):(await ot(s.collectionId,{orders:s.playOrders,qtype:s.playQtype,userId:s.playUserId,infinite:!0,excludeIds:e})).questions;if(o.current+=1,r.length===0){uo({mode:s.mode,collectionId:s.collectionId,collectionName:s.nom,good:n,total:o.current,playOrders:s.playOrders,playQtype:s.playQtype,playInfinite:!0}),U(`/results`);return}a.current=[...a.current,...r.map(e=>e.id)],v(n),c(e=>e==null?e:{...e,questions:r}),m(0),g(null);return}catch{I(`Impossible de charger la suite des questions.`)}finally{ce(!1)}return}o.current+=1,uo({mode:s.mode,collectionId:s.collectionId,collectionName:s.nom,good:n,total:o.current,playOrders:s.playOrders,playQtype:s.playQtype,playInfinite:!1}),U(`/results`);return}o.current+=1,v(n),m(e=>e+1),g(null)}finally{R(!1)}})()},ve=()=>{s==null||h==null||(async()=>{R(!0);try{if(!await ge())return;let e=_+(h!=null&&yo(s.questions,p,h)?1:0);o.current+=1,uo({mode:s.mode,collectionId:s.collectionId,collectionName:s.nom,good:e,total:o.current,playOrders:s.playOrders,playQtype:s.playQtype,playInfinite:!0}),U(`/results`)}finally{R(!1)}})()};if(l)return{kind:`loading`};if(d!=null||s==null)return{kind:`error`,loadError:d};let ye=s.questions.length,Se=s.questions[p],Ce=h==null?p:p+1,we=h!=null,Te=we&&h!=null&&yo(s.questions,p,h),Ee=(Se.commentaire??``).trim();return{kind:`ready`,session:{data:s,backTarget:s.mode===`random`?`/`:`/collections`,actionMessage:F},progress:{playInfinite:s.playInfinite,progressValue:Ce,total:ye},questionCard:{data:s,index:p,total:ye,q:Se,pickedId:h,revealed:we,anecdote:Ee,correct:Te,draftVerifier:oe,nextBusy:L,fetchingMore:z,onPick:he,onOpenCreateLinkedQuestionModal:de,onOpenEditQuestionModal:ue,onCopyCurrentQuestionJson:me,onDraftVerifier:se,onNext:_e,onEndInfiniteSession:ve},editModal:{settings:{open:x,onClose:le,variant:C,modalTitle:C===`create`?`Nouvelle question liée`:void 0},actions:{onSave:()=>void fe(),onDraftQuestion:te,onDraftCommentaire:N,onDraftCategorieId:ie,onReponseUpdated:()=>void B(),onCreateSave:e=>pe(e)},status:{loading:ee,saving:ae,error:O},data:{questionDetail:A,categorieOptions:y},drafts:{question:M,commentaire:ne,categorieId:re}}}}function So(e,t){return t<=0?0:Math.min(100,Math.round(e/t*100))}var Co={wrapper:`w-full`,progress:`flowlearn-progress progress h-3 w-full rounded-full bg-base-200/80`,footer:`mt-1 flex justify-between text-xs text-base-content/60`,value:`font-medium text-flow`};function wo({value:e,max:t,class:n}){let r=So(e,t);return X(`div`,{class:or(Co.wrapper,n),children:[X(`progress`,{class:Co.progress,value:e,max:t}),X(`div`,{class:Co.footer,children:[X(`span`,{children:`Progression`}),X(`span`,{class:Co.value,children:[r,`%`]})]})]})}function To(e){let t=e.pickedId===e.reponseId;return{isPicked:t,showCorrect:e.revealed&&e.isCorrectAnswer,showWrongPick:e.revealed&&t&&!e.isCorrectAnswer,dimmed:e.revealed&&!e.isCorrectAnswer&&!t}}var Eo={base:`w-full rounded-full border-2 px-5 py-3.5 text-left text-sm font-medium transition-all duration-300 ease-[cubic-bezier(0.22,1,0.36,1)]`,idle:`border-base-content/10 bg-base-200/45 hover:border-flow/40 hover:bg-flow/[0.07] hover:scale-[1.02]`,dimmed:`scale-[0.98] border-transparent bg-base-200/25 opacity-50`,correct:`border-flow/45 bg-flow/12 text-base-content shadow-md shadow-flow/10 ring-2 ring-flow/20`,wrong:`border-learn/40 bg-learn/10 text-base-content ring-2 ring-learn/15`};function Do({label:e,reponseId:t,pickedId:n,revealed:r,isCorrectAnswer:i,disabled:a,onPick:o}){let{dimmed:s,showCorrect:c,showWrongPick:l}=To({reponseId:t,pickedId:n,revealed:r,isCorrectAnswer:i});return X(`button`,{type:`button`,disabled:a,onClick:o,class:or(Eo.base,!r&&Eo.idle,s&&Eo.dimmed,c&&Eo.correct,l&&Eo.wrong),children:e})}var $={pageShell:`flex min-h-dvh flex-col`,centeredMain:`fl-page-enter mx-auto flex max-w-lg flex-1 flex-col items-center justify-center gap-4 px-4 py-12 text-center`,contentMain:`fl-page-enter mx-auto w-full max-w-2xl flex-1 px-4 py-6 md:py-8`,topRow:`mb-6 flex items-center gap-2`,badgesRow:`flex flex-wrap items-center gap-2`,orderBadgeWrap:`max-w-[min(100%,18rem)] truncate`,actionMessage:`mb-3 rounded-xl border border-flow/20 bg-flow/5 px-3 py-2 text-center text-sm text-base-content`,bodyLayout:`flex flex-col gap-4 md:flex-row md:items-stretch`,aside:`flex shrink-0 flex-row flex-wrap gap-2 md:w-36 md:flex-col md:items-stretch`,actionButton:`btn-sm flex-1 gap-1 sm:flex-none`,actionButtonText:`hidden sm:inline`,verifierToggleBase:`btn btn-sm mt-1 flex h-auto min-h-0 w-full cursor-pointer flex-nowrap items-center justify-start gap-2 rounded-full border-2 py-2.5 pl-3 pr-4 text-left text-sm font-medium shadow-md transition-all duration-300 ease-out hover:shadow-lg active:scale-[0.97] md:mt-0`,verifierToggleOn:`border-flow/40 bg-transparent text-flow hover:border-flow hover:bg-flow/5`,verifierToggleOff:`border-error/50 bg-error/10 text-error hover:border-error hover:bg-error/15`,card:`min-w-0 flex-1 transition duration-300`,questionMeta:`mb-4 text-xs font-medium uppercase tracking-wide text-base-content/45`,questionTitle:`mb-6 text-lg font-semibold leading-snug text-base-content sm:text-xl`,answers:`flex flex-col gap-2.5`,revealBox:`fl-reveal-enter mt-8 space-y-5 rounded-[1.75rem] border border-base-content/8 bg-gradient-to-b from-base-100/95 to-base-200/40 p-6 shadow-inner`,revealText:`text-center text-base leading-relaxed text-base-content/85`,revealActions:`flex flex-col items-center justify-center gap-2 sm:flex-row`,nextButton:`min-w-[11rem] px-8`};function Oo(){return X(`div`,{class:$.pageShell,children:[X(ti,{}),X(`main`,{class:$.centeredMain,children:X(`p`,{class:`text-base text-base-content/70`,children:`Chargement du quiz…`})}),X(ri,{})]})}function ko({loadError:e}){return X(`div`,{class:$.pageShell,children:[X(ti,{}),X(`main`,{class:$.centeredMain,children:[X(`p`,{class:`text-lg font-medium text-base-content`,children:`Parcours introuvable`}),e===`fetch`?X(`p`,{class:`text-sm text-base-content/60`,children:`Impossible de joindre l’API. Le backend est-il démarré ?`}):null,e===`empty`?X(`p`,{class:`text-sm text-base-content/60`,children:`Aucune question avec ce filtre (essaie « Mélanger » ou un autre type) ou collection vide.`}):null,X(Z,{variant:`flow`,onClick:()=>U(`/`),children:`Accueil`})]}),X(ri,{})]})}function Ao({data:e,backTarget:t}){return X(`div`,{class:$.topRow,children:[X(Z,{variant:`ghost`,class:`btn-sm gap-1 px-3`,onClick:()=>U(t),children:[X(Or,{class:`h-4 w-4`,"aria-hidden":!0}),`Retour`]}),X(`div`,{class:$.badgesRow,children:[X(Li,{tone:e.mode===`random`?`flow`:`learn`,children:e.nom}),X(`span`,{class:$.orderBadgeWrap,title:Fi(e.playOrders),children:X(Li,{tone:`learn`,class:`font-normal opacity-90`,children:Fi(e.playOrders)})}),X(Li,{tone:`flow`,class:`font-normal opacity-90`,children:Ni(e.playQtype)}),e.playInfinite?X(Li,{tone:`learn`,class:`font-normal opacity-90`,children:`Session infinie`}):null]})]})}function jo({data:e,index:t,total:n,q:r,pickedId:i,revealed:a,anecdote:o,correct:s,draftVerifier:c,nextBusy:l,fetchingMore:u,onPick:d,onOpenCreateLinkedQuestionModal:f,onOpenEditQuestionModal:p,onCopyCurrentQuestionJson:m,onDraftVerifier:h,onNext:g,onEndInfiniteSession:_}){return X(`div`,{class:$.bodyLayout,children:[X(`aside`,{class:$.aside,"aria-label":`Actions sur la question`,children:[X(Z,{variant:`outline`,class:$.actionButton,type:`button`,title:`Ajouter une question liée à celle-ci (relation parent → enfant)`,onClick:()=>f(r),children:[X(Hr,{class:`h-4 w-4 shrink-0`,"aria-hidden":!0}),X(`span`,{class:$.actionButtonText,children:`Ajouter`})]}),X(Z,{variant:`outline`,class:$.actionButton,type:`button`,title:`Modifier la question affichée`,onClick:()=>p(r),children:[X(Vr,{class:`h-4 w-4 shrink-0`,"aria-hidden":!0}),X(`span`,{class:$.actionButtonText,children:`Modifier`})]}),X(Z,{variant:`outline`,class:$.actionButton,type:`button`,title:`Copier un JSON (question, commentaire, réponses)`,onClick:()=>void m(r),children:[X(jr,{class:`h-4 w-4 shrink-0`,"aria-hidden":!0}),X(`span`,{class:$.actionButtonText,children:`Copier`})]}),X(`label`,{"aria-label":`Fake-checker, ${c?`oui`:`non`}. Cliquer pour basculer.`,class:or($.verifierToggleBase,c?$.verifierToggleOn:$.verifierToggleOff,l&&`pointer-events-none opacity-50`),children:[X(`input`,{type:`checkbox`,class:`sr-only`,checked:c,disabled:l,onChange:e=>h(e.target.checked)}),X(`span`,{class:`pointer-events-none select-none`,children:[`verifier : `,c?`Oui`:`Non`]})]})]}),X(ur,{class:$.card,children:[X(`p`,{class:$.questionMeta,children:[`Question `,t+1,` / `,n]}),X(`h2`,{class:$.questionTitle,children:r.question}),X(`div`,{class:$.answers,children:r.reponses.map(e=>X(Do,{label:e.reponse,reponseId:e.id,pickedId:i,revealed:a,isCorrectAnswer:e.bonne_reponse,disabled:i!=null,onPick:()=>d(e.id)},e.id))}),a?X(`div`,{class:$.revealBox,children:[X(`p`,{class:$.revealText,children:o?X(`span`,{class:`block`,children:o}):s?X(b,{children:[`Bien vu — `,X(`span`,{class:`font-semibold text-flow`,children:`c’est la bonne réponse`}),`.`]}):X(b,{children:`Ce n’était pas la bonne proposition.`})}),X(`div`,{class:$.revealActions,children:[X(Z,{variant:`flow`,class:$.nextButton,disabled:l||u,onClick:g,children:l?`Enregistrement…`:u?`Chargement…`:t>=n-1?e.playInfinite?`Nouvelles questions`:`Voir le résultat`:`Suivant`}),e.playInfinite&&t>=n-1?X(Z,{variant:`outline`,class:$.nextButton,disabled:l||u,onClick:_,children:`Terminer la session`}):null]})]}):null]})]})}function Mo({playInfinite:e,progressValue:t,total:n}){return e?null:X(wo,{value:t,max:n,class:`mb-6`})}function No(e){let t=xo(e);if(t.kind===`loading`)return X(Oo,{});if(t.kind===`error`)return X(ko,{loadError:t.loadError});let{session:n,progress:r,questionCard:i,editModal:a}=t;return X(`div`,{class:$.pageShell,children:[X(ti,{}),X(`main`,{class:$.contentMain,children:[X(Ao,{data:n.data,backTarget:n.backTarget}),n.actionMessage?X(`p`,{class:$.actionMessage,children:n.actionMessage}):null,X(Mo,{playInfinite:r.playInfinite,progressValue:r.progressValue,total:r.total}),X(jo,{...i})]}),X(ri,{}),X(Xa,{settings:a.settings,actions:a.actions,status:a.status,data:a.data,drafts:a.drafts})]})}function Po(e){let{sessionId:t}=e,{userId:n}=Pt(),[r,i]=V(null),[a,o]=V(!0),[s,c]=V(!1);return H(()=>{if(t==null||t===``){c(!0),o(!1);return}let e=!1;return(async()=>{o(!0),c(!1);try{let r=await Dt(t,n);if(e)return;r==null?c(!0):i(r)}catch{e||c(!0)}finally{e||o(!1)}})(),()=>{e=!0}},[t,n]),{status:{loading:a,notFound:s||r==null},navigation:{toDashboard:()=>U(`/dashboard`)},session:r}}var Fo={root:`flex min-h-dvh flex-col`,centeredMain:`fl-page-enter mx-auto flex max-w-lg flex-1 flex-col items-center justify-center gap-4 px-4 py-12`,contentMain:`fl-page-enter mx-auto w-full max-w-2xl flex-1 px-4 py-6 md:py-8`};function Io(e){let{status:t,navigation:n,session:r}=Po(e);return t.loading?X(`div`,{class:Fo.root,children:[X(ti,{}),X(`main`,{class:Fo.centeredMain,children:X(`p`,{class:`text-sm text-base-content/60`,children:`Chargement...`})}),X(ri,{})]}):t.notFound||r==null?X(`div`,{class:Fo.root,children:[X(ti,{}),X(`main`,{class:Fo.centeredMain,children:[X(`p`,{class:`text-base text-base-content/70`,children:`Session introuvable.`}),X(Z,{variant:`flow`,onClick:n.toDashboard,children:`Retour au dashboard`})]}),X(ri,{})]}):X(`div`,{class:Fo.root,children:[X(ti,{}),X(`main`,{class:Fo.contentMain,children:[X(Z,{variant:`ghost`,class:`btn-sm mb-6 gap-1 px-3`,onClick:n.toDashboard,children:[X(Or,{class:`h-4 w-4`,"aria-hidden":!0}),`Dashboard`]}),X(`h1`,{class:`mb-2 text-2xl font-bold tracking-tight text-base-content`,children:`Detail de session`}),X(`p`,{class:`mb-6 text-sm text-base-content/60`,children:r.date}),X(ur,{class:`mb-6`,children:[X(`div`,{class:`flex flex-wrap items-center gap-2`,children:[X(Li,{tone:`flow`,children:[`id `,r.id]}),X(`span`,{class:`font-semibold text-base-content`,children:r.collectionName})]}),X(`p`,{class:`mt-4 text-3xl font-semibold text-flow`,children:r.scoreLabel}),X(`p`,{class:`mt-2 text-sm text-base-content/55`,children:[`Reponses enregistrees ce jour pour cette collection (agregat `,X(`code`,{class:`text-xs`,children:`user_kpi`}),`).`]})]}),X(ur,{children:[X(`p`,{class:`mb-3 text-sm font-medium text-base-content`,children:`Questions concernees`}),X(`ul`,{class:`space-y-2 text-sm text-base-content/75`,children:r.questionsPreview.map(e=>X(`li`,{class:`rounded-lg bg-base-200/40 px-3 py-2`,children:[X(`span`,{class:`text-xs text-base-content/45`,children:[`#`,e.id]}),` `,e.question]},e.id))})]})]}),X(ri,{})]})}var Lo={wrapper:`relative overflow-hidden`,glow:`pointer-events-none absolute -right-6 -top-6 h-24 w-24 rounded-full opacity-20 blur-2xl`,content:`relative flex items-start gap-3`};function Ro({title:e,value:t,hint:n,icon:r,accent:i=`flow`}){return X(ur,{padding:`sm`,class:Lo.wrapper,children:[X(`div`,{class:or(Lo.glow,i===`flow`?`bg-flow`:`bg-learn`)}),X(`div`,{class:Lo.content,children:[r?X(`div`,{class:i===`flow`?`text-flow`:`text-learn`,children:r}):null,X(`div`,{class:`min-w-0 flex-1 space-y-1`,children:[X(`p`,{class:`text-xs font-medium uppercase tracking-wide text-base-content/50`,children:e}),X(`p`,{class:or(`text-2xl font-semibold tracking-tight`,i===`flow`?`text-flow`:`text-learn`),children:t}),n?X(`p`,{class:`text-xs text-base-content/55`,children:n}):null]})]})]})}function zo(e){return e.length===0?0:e.reduce((e,t)=>e+t,0)/e.length}function Bo(e){let t=e.trim().replace(`,`,`.`);if(!t)return null;let n=Number.parseFloat(t);return Number.isFinite(n)&&n>=0?n:null}function Vo(e){return String(e).padStart(2,`0`)}function Ho(e){return`${e.getFullYear()}-${Vo(e.getMonth()+1)}-${Vo(e.getDate())}`}function Uo(e,t=new Date){let n=new Date(t.getFullYear(),t.getMonth(),t.getDate()),r=[];for(let e=6;e>=0;--e){let t=new Date(n);t.setDate(t.getDate()-e);let i=Ho(t),a=t.toLocaleDateString(`fr-FR`,{weekday:`short`}),o=a.length>0?a.charAt(0).toUpperCase()+a.slice(1):i;r.push({key:i,label:o,count:0,h:0})}for(let t of e){let e=new Date(t.create_at);if(Number.isNaN(e.getTime()))continue;let n=Ho(e),i=r.find(e=>e.key===n);i&&(i.count+=1)}let i=Math.max(...r.map(e=>e.count),0);for(let e of r)e.h=i===0?0:Math.round(e.count/i*100);return r}function Wo(e,t){let n=e.length,r=e.filter(e=>e.correct).length,i=n===0?0:Math.round(r/n*100),a=e.map(e=>Bo(e.duree_session)).filter(e=>e!=null),o=zo(a),s=new Set(e.map(e=>e.question_id)).size,c=t.reduce((e,t)=>e+t.good,0),l=t.reduce((e,t)=>e+t.total,0),u=t.length===0?`${s} question${s===1?``:`s`} distincte${s===1?``:`s`} · aucune session dans l historique`:`${s} question${s===1?``:`s`} distincte${s===1?``:`s`} · sessions : ${c}/${l} bonnes reponses`;return{total:n,good:r,ratio:i,avgSecLabel:a.length===0?`—`:`${o.toFixed(1)} s`,uniqueQuestions:s,sessionsHint:u}}var Go={root:`flex min-h-dvh flex-col`};function Ko(){let{userId:e}=Pt(),[t,n]=V([]),[r,i]=V([]),[a,o]=V(!0),[s,c]=V(null),l=Ce(async()=>{o(!0),c(null);try{let[t,r]=await Promise.all([wt(e),Et(e)]);n(t),i(r)}catch{c(`fetch`)}finally{o(!1)}},[e]);H(()=>{l()},[l]);let u=Se(()=>Wo(t,r),[t,r]),d=Se(()=>Uo(t),[t]);return X(`div`,{class:Go.root,children:[X(ti,{}),X(ai,{children:[X(`h1`,{class:`mb-2 text-2xl font-bold tracking-tight text-base-content sm:text-3xl`,children:`Dashboard`}),X(`p`,{class:`mb-6 text-sm text-base-content/60`,children:[`Donnees live : `,X(`code`,{class:`text-xs`,children:`/stats/kpis`}),` et `,X(`code`,{class:`text-xs`,children:`/stats/sessions`}),` pour l utilisateur connecte (id `,e,`).`]}),a?X(`p`,{class:`text-sm text-base-content/60`,children:`Chargement...`}):s?X(`div`,{class:`rounded-[var(--radius-box)] border border-base-content/10 bg-base-200/40 px-4 py-8 text-center text-sm text-base-content/65`,children:[X(`p`,{class:`mb-3`,children:`Impossible de charger les statistiques.`}),X(Z,{variant:`flow`,class:`btn-sm`,onClick:()=>void l(),children:`Reessayer`})]}):X(b,{children:[X(`div`,{class:`mb-6 grid gap-3 sm:grid-cols-3`,children:[X(Ro,{title:`Ratio correct`,value:`${u.ratio}%`,hint:`${u.good} / ${u.total} reponses`,accent:`flow`,icon:X(Kr,{class:`h-5 w-5`,"aria-hidden":!0})}),X(Ro,{title:`Temps moyen`,value:u.avgSecLabel,hint:`Moyenne des durees enregistrees (champ duree_session)`,accent:`learn`,icon:X(Mr,{class:`h-5 w-5`,"aria-hidden":!0})}),X(Ro,{title:`Volume & couverture`,value:`${u.total}`,hint:u.sessionsHint,accent:`flow`,icon:X(kr,{class:`h-5 w-5`,"aria-hidden":!0})})]}),X(ur,{class:`mb-6 transition duration-300`,children:[X(`p`,{class:`mb-4 text-sm font-medium text-base-content`,children:`Reponses enregistrees (7 derniers jours)`}),X(`p`,{class:`mb-3 text-xs text-base-content/50`,children:[`Agregat des lignes `,X(`code`,{class:`text-[10px]`,children:`user_kpi`}),` par date locale (fuseau du navigateur).`]}),X(`div`,{class:`flex h-36 items-end justify-between gap-2`,children:d.map(e=>{let t=d.some(e=>e.count>0)?Math.max(e.h,e.count>0?14:10):8;return X(`div`,{class:`flex flex-1 flex-col items-center gap-2`,children:[X(`div`,{class:`w-full max-w-10 rounded-t-full bg-gradient-to-t from-flow to-learn/80 transition-all duration-300 ease-out hover:opacity-100 ${e.count>0?`opacity-90`:`opacity-35`}`,style:{height:`${t}%`},title:`${e.count} reponse(s) · ${e.key}`}),X(`span`,{class:`text-[10px] font-medium text-base-content/50`,children:e.label})]},e.key)})})]}),X(ur,{class:`transition duration-300`,children:[X(`p`,{class:`mb-4 text-sm font-medium text-base-content`,children:`Historique des sessions`}),r.length===0?X(`p`,{class:`text-sm text-base-content/55`,children:`Aucune session agregee pour cet utilisateur.`}):X(`ul`,{class:`space-y-3`,children:r.map(e=>X(`li`,{children:X(`button`,{type:`button`,class:`flex w-full flex-col gap-2 rounded-full border border-base-content/10 bg-base-200/30 px-4 py-4 text-left transition-all duration-300 ease-out hover:border-flow/30 hover:bg-base-200/55 sm:flex-row sm:items-center sm:justify-between`,onClick:()=>U(`/dashboard/session/${encodeURIComponent(e.id)}`),children:[X(`div`,{children:[X(`p`,{class:`font-medium text-base-content`,children:e.collectionName}),X(`p`,{class:`text-xs text-base-content/50`,children:e.date})]}),X(`div`,{class:`flex items-center gap-2`,children:[X(Li,{tone:`learn`,children:e.scoreLabel}),X(`span`,{class:`text-xs text-flow`,children:`Details →`})]})]})},e.id))})]})]})]}),X(ri,{})]})}function qo(){return typeof window>`u`?`/`:`${window.location.pathname||`/`}${window.location.search||``}`}function Jo(){let[e,t]=V(qo);return X(gr,{children:X(_r.Provider,{value:e,children:X(Qe,{onChange:()=>{t(qo())},children:[X($e,{path:`/`,component:ta}),X($e,{path:`/collections`,component:Ji}),X($e,{path:`/database`,component:$i}),X($e,{path:`/questions/:collectionId`,component:so}),X($e,{path:`/questions`,component:so}),X($e,{path:`/dashboard/session/:sessionId`,component:Io}),X($e,{path:`/dashboard`,component:Ko}),X($e,{path:`/play/:collectionId`,component:No}),X($e,{path:`/results`,component:vo})]})})})}oe(X(Jo,{}),document.getElementById(`app`));
